% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R, R/compounds-set.R
\name{assignMobilities_comp}
\alias{assignMobilities_comp}
\alias{assignMobilities,compounds-method}
\alias{assignMobilities,compoundsSet-method}
\title{Assign IMS data to a \code{compounds} object.}
\usage{
\S4method{assignMobilities}{compounds}(
  obj,
  fGroups,
  IMS = TRUE,
  from = NULL,
  matchFromBy = "InChIKey1",
  overwrite = FALSE,
  adduct = NULL,
  CCSParams = NULL,
  prefCalcChemProps = TRUE,
  neutralChemProps = FALSE,
  virtualenv = "patRoon-C3SDB"
)

\S4method{assignMobilities}{compoundsSet}(obj, fGroups, IMS = TRUE, from = NULL, ...)
}
\arguments{
\item{obj}{The \code{\link{compounds}} object for which IMS assignments should be performed.}

\item{fGroups}{The \code{\link{featureGroups}} object for which the compound candidates (\code{obj}) were calculated.}

\item{IMS}{\IMSWF Specifies which feature groups are considered for IMS assignments in IMS workflows. The following options
  are valid: \itemize{

  \item \code{"both"}: Selects IMS and non-IMS features.

  \item \code{"maybe"}: Selects non-IMS features and IMS features without assigned IMS parent.

  \item \code{FALSE}: Selects only non-IMS features.

  \item \code{TRUE}: Selects only IMS features.

  }}

\item{from, matchFromBy, overwrite, CCSParams, prefCalcChemProps, neutralChemProps, virtualenv}{Passed to the
\link[=assignMobilities_susp]{method for suspects}}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
  Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}. If the \code{featureGroups} object has
  adduct annotations then these are used if \code{adducts=NULL}.

  \setsWF The \code{adduct} argument is not supported for sets workflows, since the
  adduct annotations will then always be used.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}
}
\description{
Assigns ion mobility and \acronym{CCS} values to the candidates in a \code{\link{compounds}} object.
}
\details{
The \code{assignMobilities} method for \code{\link{compounds}} is used to (1) add predicted or known IMS data to
annotation candidates and (2) convert (previously added) mobility <--> \acronym{CCS} values. Internally, the
\link[=assignMobilities_susp]{assignMobilities method for suspects} is used to perform these operations, please see
its documentation for more details.

If both adduct specific and non-adduct specific mobility and \acronym{CCS} data is available (and not \code{NA}),
then the non-adduct specific data is assigned to the compound candidate. Otherwise, data corresponding to the
\code{adduct} argument or the adduct assigned to the feature group is taken. The
\code{\link[=filter,compounds-method]{filter}} method can be used to filter out candidates with deviating IMS data.
}
\note{
\command{SIRIUS} does currently not report \acronym{InChIKey} values, hence, \code{matchFromBy="InChIKey"} is
  not supported in this case.

  If compound annotation is performed with \code{\link{generateCompoundsMetFrag}} with \code{database="pubchemlite"}
  then \acronym{CCS} data is already added, provided the local database has it. If you want to overwrite this data,
  set \code{overwrite=TRUE}.
}
\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} the calculations are performed and stored
  independently for each set, as adducts, charges and \emph{m/z} values typically differ.

  The \code{from} argument may be a \code{list} that specifies the \code{from} values for each set. This is primarily
  intended when tables are used to set \code{from}.
}

\seealso{
The \link[=assignMobilities_susp]{assignMobilities method for suspects}.
}
