% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups.R, R/feature_groups-set.R,
%   R/feature_groups-screening.R, R/feature_groups-screening-set.R
\name{assignMobilities_feat}
\alias{assignMobilities_feat}
\alias{assignMobilities,featureGroups-method}
\alias{assignMobilities,featureGroupsSet-method}
\alias{assignMobilities,featureGroupsScreening-method}
\alias{assignMobilities,featureGroupsScreeningSet-method}
\title{Assign Ion Mobility and \acronym{CCS} values to features}
\usage{
\S4method{assignMobilities}{featureGroups}(
  obj,
  mobPeakParams = NULL,
  chromPeakParams = NULL,
  EIMParams = getDefEIMParams(),
  EICParams = getDefEICParams(),
  peakRTWindow = defaultLim("retention", "narrow"),
  fallbackEIC = TRUE,
  calcArea = "integrate",
  mobWindow = defaultLim("mobility", "medium"),
  scoreWeights = c(mobility = 1, intensity = 1),
  CCSParams = NULL,
  parallel = "maybe"
)

\S4method{assignMobilities}{featureGroupsSet}(
  obj,
  mobPeakParams = NULL,
  chromPeakParams = NULL,
  EIMParams = getDefEIMParams(),
  EICParams = getDefEICParams(),
  peakRTWindow = defaultLim("retention", "narrow"),
  fallbackEIC = TRUE,
  calcArea = "integrate",
  mobWindow = defaultLim("mobility", "medium"),
  scoreWeights = c(mobility = 1, intensity = 1),
  CCSParams = NULL,
  parallel = "maybe"
)

\S4method{assignMobilities}{featureGroupsScreening}(
  obj,
  mobPeakParams = NULL,
  chromPeakParams = NULL,
  EIMParams = getDefEIMParams(),
  EICParams = getDefEICParams(),
  peakRTWindow = defaultLim("retention", "narrow"),
  fallbackEIC = TRUE,
  calcArea = "integrate",
  mobWindow = defaultLim("mobility", "medium"),
  scoreWeights = c(mobility = 1, intensity = 1),
  CCSParams = NULL,
  parallel = "maybe",
  fromSuspects = FALSE,
  IMSMatchParams = NULL
)

\S4method{assignMobilities}{featureGroupsScreeningSet}(
  obj,
  mobPeakParams = NULL,
  chromPeakParams = NULL,
  EIMParams = getDefEIMParams(),
  EICParams = getDefEICParams(),
  peakRTWindow = defaultLim("retention", "narrow"),
  fallbackEIC = TRUE,
  calcArea = "integrate",
  mobWindow = defaultLim("mobility", "medium"),
  scoreWeights = c(mobility = 1, intensity = 1),
  CCSParams = NULL,
  parallel = "maybe",
  fromSuspects = FALSE,
  IMSMatchParams = NULL
)
}
\arguments{
\item{obj}{A \code{\link{featureGroups}} (derived) object to which IMS data should be assigned.}

\item{mobPeakParams}{A \code{list} with peak detection parameters for mobility peaks, generated with
\code{\link{getDefPeakParams}}. Set to \code{NULL} to skip detection of mobility peaks.}

\item{chromPeakParams}{A \code{list} with peak detection parameters for the re-integration of mobility features,
generated with \code{\link{getDefPeakParams}}. Set to \code{NULL} to skip re-integration by peak detection.}

\item{EIMParams, EICParams}{Parameters to be used for the generation of EIMs (for mobility peak detection) and EICs
(for mobility feature re-integration), generated by \code{\link{getDefEIMParams}} and
\code{\link{getDefEICParams}}.}

\item{peakRTWindow}{The retention time tolerance (in seconds) for detected peaks to be used for re-integration (Step
3, see \verb{Mobility assignment}). This should be kept low to ensure no closely eluting features from other
compounds are used. Note that peaks with retention times outside the RT window of the IMS parent are always
ignored.}

\item{fallbackEIC}{Set to \code{TRUE} to allow updating intensities and areas from raw EICs in Step 3 (see
\verb{Mobility assignment}).}

\item{calcArea}{controls how the area is calculated when updating from raw EIC data (see \code{fallbackEIC}):
\code{"integrate"} (performs area integration) or \code{"sum"} (sums up all intensity data points). Note: these
methods may be different than what is used by the feature detection of the IMS parent or the peak detection
algorithm used for re-integration, hence, comparing areas should be done with care.}

\item{mobWindow}{The mobility tolerance window.}

\item{scoreWeights}{A \code{numeric} with the weights used by \link[=groupFeaturesGreedy]{greedy grouping}.
\strong{NOTE}: only the \code{mobility} and \code{intensity} weights are used. See the documentation of
\code{\link{groupFeaturesGreedy}} for more details on the scoring and grouping.}

\item{CCSParams}{A \code{list} with parameters for mobility <--> \acronym{CCS} conversion. See
\code{\link{getCCSParams}} for details and to make such parameter lists. Set to \code{NULL} to skip conversions.}

\item{parallel}{If set to \code{TRUE} then code is executed in parallel through the \CRANpkg{future} package. Please
see the parallelization section in the handbook for more details.

Alternatively, set \code{parallel="maybe"} to disable parallelization if this normally will not}

\item{fromSuspects}{If \code{TRUE} then mobilities values are directly copied from suspect list data (if available).
See the \verb{Post mobility assignment} section.}

\item{IMSMatchParams}{\IMSWF A \code{list} with parameters to be used for matching IMS data. See
\code{\link{getIMSMatchParams}} for details and how to make such a parameter list.}
}
\description{
Various approaches to assign mobilities to features and perform \acronym{CCS} conversions.
}
\details{
The \code{assignMobilities} method function for features is used (1) to assign Ion Mobility values and (2) calculate
\acronym{CCS} values from these mobilities.

In \pkg{patRoon}, two approaches are supported to assign mobilities to features: \emph{direct} and \emph{post}
assignment (\acronym{DMA} and \acronym{PMA}). With the \acronym{DMA} the mobility values are directly assigned during
feature detection. This is currently only supported by the \link[=findFeaturesPiek]{piek} and
\link[=groupFeaturesGreedy]{greedy} algorithms or by \link[=importFeaturesTable]{importing feature data}. With
\acronym{PMA}, the mobility values are assigned after feature detection and grouping (and possibly other steps such
as filtering). Thus, the \acronym{PMA} approach is supported by all available feature detection algorithms in
\pkg{patRoon}. The \acronym{PMA} approach is further described below. Only the \acronym{CCS} conversion functionality
of \code{assignMobilities} should be used in \acronym{PMA} mobility assignment workflows.

The assignment of \acronym{CCS} values is controlled by the \code{CCSParams} argument (see above).

In suspect screening workflows \code{assignMobilities} also assigns reference mobility and \acronym{CCS}
  values to suspect hits, and can filter hits if \code{IMSMatchParams} is set. This is similarly performed as
  \code{\link{screenSuspects}}, please see its documentation for more details.
}
\section{Post mobility assignment}{
 The \emph{post} assignment of mobilities occurs in the following steps:
  \enumerate{

  \item Extracted ion mobilograms (EIMs) are generated for all features and subjected to automatic peak detection to
  obtain \emph{mobility peaks}.

  \item The detected mobility peaks in an EIM are then used to form \emph{mobility features}. These features inherit
  their LC-MS properties (RT, \emph{m/z}, etc) from the corresponding \emph{IMS parent}, \emph{i.e.} the feature for
  which the EIM was created. The mobility peak centroids and ranges are used to assign IMS data to the mobility
  features. Multiple mobility peaks within the same EIM result in multiple mobility features, and each are linked to
  the same IMS parent. The linkage is especially useful to keep a relation between \emph{e.g.} protomers.

  \item LC-MS properties such as the area, intensity and RTs are (optionally) updated by re-integration of detected
  peaks from mobility filtered extracted ion chromatograms. If peak detection is disabled or fails, then the
  intensity and areas can be estimated directly from raw EIC data (\code{fallbackEIC} argument).

  \item Any mobility features that could not be re-integrated (either by peak detection or EIC fallback) are removed.

  \item The feature grouping is updated: the mobility features with close mobilities (defined by \code{mobWindow})
  within a feature group are split-off into new feature groups and linked to the original \emph{IMS parent} feature
  group. This is performed by the \link[=groupFeaturesGreedy]{greedy grouping algorithm}. LC-MS properties and most
  other data such as feature group qualities and scores (\code{\link{calculatePeakQualities}}), adduct annotations
  (\emph{e.g.} \code{\link{selectIons}}), predicted concentrations and toxicities (\code{\link{calculateConcs}} and
  \code{\link{calculateTox}}) and internal standards for intensity normalization (\code{\link{normInts}}) are copied
  from the IMS parents to the mobility feature groups.

  }

  Note that re-running \code{assignMobilities} will first remove any existing mobility features.

 \subsection{Suspect screening workflows}{In suspect screening workflows the
  \code{fromSuspects} arguments can be set to alternatively perform mobility assignment directly from the suspect
  list data (replacing Steps 1-2). The feature mobility is simply assigned from the suspect data and the mobility
  range is derived from the \code{mobWindow} argument. Relationships with IMS parents (Step 2) are similarly formed.
  An advantage of this approach is that no mobility peak detection is needed, which may useful for low intensity
  features where this could be difficult. Setting \code{fromSuspects=TRUE} is primarily intended for workflows where
  (1) the mobility of a suspect is accurately known upfront or (2) IMS data should only be used as a rough filtering
  step for feature data. In the latter case accurate feature mobility assignment is not of interest and the suspect
  IMS data is typically not accurately known (\emph{e.g.} predicted), hence, for these workflows the tolerance
  specified by \code{mobWindow} should be increased.

  With \code{fromSuspects=TRUE} no mobility peak detection is performed, hence, the actual presence of the feature is
  only verified in Step 3. For this reason, falling back to EIC data (\code{fallbackEIC} argument) is never performed
  for mobility features from suspects, and \code{chromPeakParams} must always be defined to allow chromatographic
  peak detection.

  If both \code{fromSuspects} and \code{mobPeakParams} are set, regular mobility assignment (Steps 1-2) is performed
  for features without suspect hit. If multiple suspects were assigned to a feature group then suspect data is
  \emph{never} used to form mobility features.}
}

\section{Mobility features}{
 The features (and feature groups) with IMS properties are referred to \emph{mobility
  features} (and \emph{mobility feature groups}). These are referred to \emph{orphans} if their link to the original
  IMS parent is removed (in \emph{post} workflows, see previous section) or non-existent (\emph{direct} workflows).
  The formation of orphans in \emph{post} workflows typically occurs by removal of IMS parents by subsetting or
  filtering operations.

  Most data-processing functionality, such as subsetting, plotting, filtering, etc., allows to selectively operate
  either on the mobility features, their IMS parents, both or either the parent or orphans (controlled by the
  \code{IMS} argument to the corresponding functions).
}

\section{Use of raw HRMS data}{
 The \link[=msdata]{raw data interface} of \pkg{patRoon} is used by \code{assignMobilities} to
  process HRMS (or IMS-HRMS) data. Please see \link[=msdata]{its documentation} for more information on the supported
  formats and available configuration options.
}

\section{Sets workflows}{
 Mobility assignment in a \link[=sets-workflow]{sets workflow} is equivalent to non-sets
  workflows. However, currently there is no (known) way to relate mobilities across MS polarities (+/- mode). Thus,
  mobility features will always be formed for each set separately. However, \emph{IMS parents} will still be grouped
  across polarities (like a non-IMS workflow) and their links to mobility features can therefore be used to relate
  mobility features across MS polarities.
}

