% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{assignMobilities_susp}
\alias{assignMobilities_susp}
\alias{assignMobilities,data.table-method}
\alias{assignMobilities,data.frame-method}
\title{Assign IMS data to suspects}
\usage{
\S4method{assignMobilities}{data.table}(
  obj,
  from = NULL,
  matchFromBy = "InChIKey1",
  overwrite = FALSE,
  adducts = c("[M+H]+", "[M-H]-", NA),
  predictAdductOnly = TRUE,
  CCSParams = NULL,
  prepareChemProps = TRUE,
  prefCalcChemProps = TRUE,
  neutralChemProps = FALSE,
  virtualenv = "patRoon-C3SDB"
)

\S4method{assignMobilities}{data.frame}(obj, ...)
}
\arguments{
\item{obj}{The suspect list to which the mobility and/or CCS data should be added. Should be a \code{data.frame} or
\code{data.table}.}

\item{from}{Specifies from where IMS data is added to the suspect list. This can be the following: \itemize{

  \item \code{"pubchemlite"}: \acronym{CCS} data is matched from predicted values of the
  \href{https://pubchemlite.lcsb.uni.lu/}{PubChemLite} database. \strong{Note}: this requires a local copy of the
  \href{https://zenodo.org/records/15311000}{CCS amended PubChemLite database} (see the Handbook for more details),
  which is automatically installed by \pkg{patRoonExt}.

  \item \code{"c3sdb"}: Uses the \href{https://github.com/dylanhross/c3sdb}{C3SDB} \command{Python} package to
  predict \acronym{CCS} values. This requires a local installation of \command{C3SDB}, \emph{e.g.} performed with
  \code{\link{installC3SDB}}.

  \item A \code{data.table} or \code{data.frame} to which IMS data is matched. Should contain the column defined by
  \code{matchFromBy} and columns storing (non-)adduct specific mobility/\acronym{CCS} columns (see Details).

  \item \code{NULL}: No IMS data is added to the suspect list.

  }

  Any \code{NA} values in \code{from} are ignored.}

\item{matchFromBy}{Which column should be used to match the IMS data from \code{from} and suspects. Valid options are
  \code{"InChIKey"}, \code{"InChIKey1"} (first block InChIKey), \code{"InChI"}, \code{"SMILES"}, \code{"name"}.
  However, this also depends on which columns are available in either of the data sources. \code{InChIKey1} values
  are automatically calculated from \code{InChIKey}s, if possible.

  Matching by \code{InChiKey1} is recommended by default to allow tolerance between different datasources. Note that
  most compound annotation algorithms also match by \code{InChIKey1} and current IMS resolving power is generally
  insufficient to distinguish the different stereoisomers/tautomers specified by the full \code{InChIKey}.}

\item{overwrite}{Set to \code{TRUE} to overwrite any existing suspect IMS data with data from \code{from}.}

\item{adducts}{A \code{character} with the adduct(s) to consider for assigning mobility data to suspects and mobility
  <--> \acronym{CCS} conversions. This may be limited by what is available in the data source specified by
  \code{from} (see \code{\link{C3SDBAdducts}} for \code{from="c3sdb"}). Inclusion of \code{NA} in \code{adducts}
  allows the use of non-adduct specific values (see Details).

  The value for \code{adducts} is automatically expanded by the adducts specified in the \code{adduct} column of the
  suspect list. Hence, \code{adducts} can be empty (\code{character()}) if no calculations for other adducts are
  desired.}

\item{predictAdductOnly}{If \code{from="c3sdb"} and \code{predictAdductOnly=TRUE} then only predictions are performed
for the adduct specified in the \code{adduct} column in the suspect list (if present).}

\item{CCSParams}{A \code{list} with parameters for mobility <--> \acronym{CCS} conversion. See
\code{\link{getCCSParams}} for details and to make such parameter lists. Set to \code{NULL} to skip conversions.}

\item{prepareChemProps}{Set to \code{TRUE} to perform chemical property calculation and validation on the suspect
list (described below).}

\item{prefCalcChemProps}{If \code{TRUE} then calculated chemical properties such as the formula and
\acronym{InChIKey} are preferred over what is already present in the suspect data (if \code{prepareChemProps=TRUE}) and \code{from} data (if a table). For efficiency reasons it is
recommended to set this to \code{TRUE}. See the \verb{Validating and calculating chemical properties} section for
more details.}

\item{neutralChemProps}{If \code{TRUE} then the neutral form of the molecule is considered to calculate
\acronym{SMILES}, formulae etc. Enabling this may improve feature matching when considering common adducts
(\emph{e.g.} \code{[M+H]+}, \code{[M-H]-}). See the \verb{Validating and calculating chemical properties} section
for more details.}

\item{virtualenv}{The virtual \command{Python} environment in which \command{C3SDB} is installed. This is passed to
\code{\link[reticulate:use_virtualenv]{reticulate::use_virtualenv}}. Set to \code{NULL} to skip this and not setup
the environment.}

\item{\dots}{Arguments passed to \code{data.table} method.}
}
\description{
Adds calculated mobility and/or \acronym{CCS} data to a suspect list.
}
\details{
The \code{assignMobilities} method for suspect lists is used to (1) add IMS data to suspects from predictions or
library data and (2) convert (previously added) mobility <--> \acronym{CCS} values. These steps are controlled by the
\code{from} and \code{CCSParams} arguments, respectively.

Mobility and \acronym{CCS} values assigned in the suspect list are either adduct specific or not. Adduct specific
values are preferred, as the 'correct' value can be automatically selected during suspect screening based on the
adduct assigned to the feature (or passed as the \code{adduct} argument to \code{\link{screenSuspects}}). The
non-adduct specific values are typically used when the corresponding adduct for the mobility/\acronym{CCS} value is
unknown (or not of interest). These values get precedence over adduct specific values. The adduct specific values are
stored in \code{mobility_<adduct>} and \code{CCS_<adduct>} columns, where \code{<adduct>} is the adduct name
(\emph{e.g.} \code{[M+H]+}, \code{[M-H]-}). The \code{mobility} and \acronym{CCS} columns store any non-adduct
specific values. The \code{adducts} argument ultimately defines the use of adduct and non-adduct specific values.

The mobility <--> \acronym{CCS} conversions occur both ways, \emph{i.e.} missing \acronym{CCS} values will be
converted from mobility values and \emph{vice versa}. If adduct specific values are converted then the charge value
used for these calculations is taken from the corresponding adduct. For non-adduct specific values the charge is
taken from the adduct specified in suspect list if present, or from the default charge specified in \code{CCSParams}
otherwise.
}
\section{Validating and calculating chemical properties}{
 Chemical properties such as \acronym{SMILES},
  \acronym{InChIKey} and formulae in the suspect data (if \code{prepareChemProps=TRUE}) and \code{from} data (if a table) are automatically validated and calculated if missing/invalid.

  The internal validation/calculation process performs the following steps: \itemize{

  \item Validation of \acronym{SMILES}, \acronym{InChI}, \acronym{InChIKey} and formula data (if present). Invalid
  entries will be set to \code{NA}.

  \item If \code{neutralChemProps=TRUE} then chemical data (\acronym{SMILES}, formulae etc.) is neutralized by
  (de-)protonation (using the \command{--neutralized} option of \command{OpenBabel}). An additional column
  \code{molNeutralized} is added to mark those molecules that were neutralized. Note that neutralization requires
  either \acronym{SMILES} or \acronym{InChI} data to be available.

  \item The \acronym{SMILES} and \acronym{InChI} data are used to calculate missing or invalid \acronym{SMILES},
  \acronym{InChI}, \acronym{InChIKey} and formula data. If \code{prefCalcChemProps=TRUE} then existing
  \acronym{InChIKey} and formula data is overwritten by calculated values whenever possible.

  \item The chemical formulae which were \emph{not} calculated are verified and normalized. This process may be time
  consuming, and is potentially largely avoided by setting \code{prefCalcChemProps=TRUE}.

  \item Neutral masses are calculated for missing values (\code{prefCalcChemProps=FALSE}) or whenever possible
  (\code{prefCalcChemProps=TRUE}).

  }

  Note that calculation of formulae for molecules that are isotopically labelled is currently only supported for
  deuterium (2H) elements.

  This functionality relies heavily on \href{https://github.com/openbabel/openbabel}{OpenBabel}, please make sure it
  is installed.
}

\references{
\insertRef{OBoyle2011}{patRoon}

\insertRef{Schymanski2021}{patRoon} \cr\cr \insertRef{Elapavalore2025}{patRoon} \cr\cr
  \insertRef{Ross2020}{patRoon}
}
