% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/utils-bruker.R
\name{bruker-utils}
\alias{bruker-utils}
\alias{showDataAnalysis}
\alias{setDAMethod}
\alias{revertDAAnalyses}
\alias{recalibrarateDAFiles}
\alias{getDACalibrationError}
\alias{exportDAFiles}
\alias{addDAEIC}
\alias{addAllDAEICs}
\title{Bruker DataAnalysis utilities}
\usage{
showDataAnalysis()

setDAMethod(anaInfo, method, close = TRUE)

revertDAAnalyses(anaInfo, close = TRUE, save = close)

recalibrarateDAFiles(anaInfo, close = TRUE, save = close)

getDACalibrationError(anaInfo)

exportDAFiles(anaInfo, format = "mzML", exportLine = TRUE,
  outPath = anaInfo$path, overWrite = FALSE)

addDAEIC(analysis, path, mz, mzWindow = 0.005, ctype = "EIC",
  mtype = "MS", polarity = "both", bgsubtr = FALSE, fragpath = "",
  name = NULL, hideDA = TRUE, close = FALSE, save = close)

addAllDAEICs(fGroups, mzWindow = 0.005, ctype = "EIC",
  bgsubtr = FALSE, name = TRUE, onlyPresent = TRUE, hideDA = TRUE,
  close = FALSE, save = close)
}
\arguments{
\item{anaInfo}{\link[=analysis-information]{Analysis info table}}

\item{method}{The full path of the DataAnalysis method.}

\item{close, save}{If \code{TRUE} then Bruker files are closed and saved after
processing with DataAnalysis, respectively. Setting \code{close=TRUE}
prevents that many analyses might be opened simultaneously in DataAnalysis,
which otherwise may use excessive memory or become slow. By default
\code{save} is \code{TRUE} when \code{close} is \code{TRUE}, which is
likely what you want as otherwise any processed data is lost.}

\item{format}{The output format of exported files. Should be either
\code{"mzXML"}, \code{"mzML"} or \code{"mzData"}.}

\item{exportLine}{Export line spectra (\code{TRUE}) or profile spectra
(\code{FALSE}). Usually line spectra are preferred, since profile spectra
use signficantly more disk space and increase required memory during
processing.}

\item{outPath}{Character vector of output paths for exported analyses. Will
be recycled if necessary.}

\item{overWrite}{If \code{TRUE} existing files will be overwritten.}

\item{analysis}{Analysis name (without file extension).}

\item{path}{path of the analysis.}

\item{mz}{\emph{m/z} (Da) value used for the chromatographic trace (if
applicable).}

\item{mzWindow}{\emph{m/z} window (in Da) used for the chromatographic trace
(if applicable).}

\item{ctype}{Type of the chromatographic trace. Valid options are:
\code{"EIC"} (extracted ion chromatogram), \code{"TIC"} (total ion
chromatogram, only for \code{addDAEIC}) and \code{"BPC"} (Base Peak
Chromatogram).}

\item{mtype}{MS filter for chromatographic trace. Valid values are:
\code{"all"}, \code{"MS"}, \code{"MSMS"}, \code{"allMSMS"} and
\code{"BBCID"}.}

\item{polarity}{Polarity filter for chromatographic trace. Valid values:
\code{"both"}, \code{"positive"} and \code{"negative"}.}

\item{bgsubtr}{If \code{TRUE} then background subtraction ('Spectral'
algorithm) will be performed.}

\item{fragpath}{Precursor \emph{m/z} used for MS/MS traces (\code{""} for
none).}

\item{name}{For \code{addDAEIC}: the name for the chromatographic trace. For
\code{addAllEICs}: \code{TRUE} to automatically set EIC names. Set to
\code{NULL} for none.}

\item{hideDA}{Hides DataAnalysis while adding the chromatographic trace
(faster).}

\item{fGroups}{The \code{\link{featureGroups}} object for which EICs should be made.}

\item{onlyPresent}{If \code{TRUE} then EICs are only generated for analyses
where the feature was detected.}
}
\value{
\code{getDACalibrationError} returns a \code{data.frame} with a
  column of all analyses (named \code{analysis}) and their mass error (named
  \code{error}).
}
\description{
Miscellaneous utility functions which interface with Bruker DataAnalysis
}
\details{
These functions communicate directly with Bruker DataAnalysis to provide
various functionality, such as calibrating and exporting data and adding
chromatographic traces. For this the \pkg{RDCOMClient} package is
required to be installed.

\code{showDataAnalysis} makes a hidden DataAnalysis window visible
  again. Most functions using DataAnalysis will hide the window during
  processing for efficiency reasons. If the window remains hidden
  (\emph{e.g.} because there was an error) this function can be used to make
  it visible again. This function can also be used to start DataAnalysis if
  it is not running yet.

\code{setDAMethod} Sets a given DataAnalysis method (\file{.m} file)
  to a set of analyses. \strong{NOTE}: as a workaround for a bug in
  DataAnalysis, this function will save(!), close and re-open any analyses
  that are already open prior to setting the new method. The \code{close}
  argument only controls whether the file should be closed after setting the
  method (files are always saved).

\code{revertDAAnalyses} Reverts a given set of analyses to their
  unprocessed raw state.

\code{recalibrarateDAFiles} Performs automatic mass recalibration of
  a given set of analyses. The current method settings for each analyses will
  be used.

\code{getDACalibrationError} is used to obtain the standard
  deviation of the current mass calibration (in ppm).

\code{exportDAFiles} will export a set of analyses either in
  \file{.mzXML} or \file{.mzML} formats.

\code{addDAEIC} adds an Extracted Ion Chromatogram (EIC) or other
  chromatographic trace to a given analysis which can be used directly with
  DataAnalysis.

\code{addAllDAEICs} adds Extracted Ion Chromatograms (EICs) for all
  features within a \code{\link{featureGroups}} object.
}
\seealso{
\code{\link{analysis-information}}
}
