% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{C3SDBAdducts}
\alias{C3SDBAdducts}
\title{Returns the adducts supported by C3SDB}
\usage{
C3SDBAdducts()
}
\description{
Returns the adducts supported by the \href{https://github.com/dylanhross/c3sdb}{C3SDB} \command{Python} package.
}
