% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{caching}
\alias{caching}
\alias{makeHash}
\alias{makeFileHash}
\alias{loadCacheData}
\alias{saveCacheData}
\alias{clearCache}
\title{Utilities to caching of workflow data.}
\usage{
makeHash(..., checkDT = TRUE)

makeFileHash(...)

loadCacheData(category, hashes, dbArg = NULL)

saveCacheData(category, data, hash, dbArg = NULL)

clearCache(what = NULL, file = NULL, vacuum = TRUE)
}
\arguments{
\item{\dots}{Arguments/objects to be used for hashing.}

\item{checkDT}{\code{logical}, set to \code{TRUE} with (a \code{list} with) \code{data.table}s to ensure reproducible
hashing. Otherwise can be set to \code{FALSE} to improve performance.}

\item{category}{The category of the object to be cached.}

\item{hashes}{A \code{character} with one more hashes (\emph{e.g.} obtained with \code{makeHash}) of the objects to
be loaded.}

\item{dbArg}{Alternative connection to database. Default is \code{NULL} and uses the cache options as defined by
\option{patRoon.cache.fileName}. Mainly used internally to improve performance.}

\item{data}{The object to be cached.}

\item{hash}{The hash string of the object to be cached (\emph{e.g.} obtained with \code{makeHash}).}

\item{what}{This argument describes what should be done. When \code{what = NULL} this function will list which tables
are present along with an indication of their size (database rows). If \code{what = "all"} then the complete file
will be removed. Otherwise, \code{what} should be a character string (a regular expression) that is used to match
the table names that should be removed.}

\item{file}{The cache file. If \code{NULL} then the value of the \code{patRoon.cache.fileName} option is used.}

\item{vacuum}{If \code{TRUE} then the \code{VACUUM} operation will be run on the cache database to reduce the file
size. Setting this to \code{FALSE} might be handy to avoid long processing times on large cache databases.}
}
\description{
Several utiulity functions related to cache workflow data. The most important function is \code{clearCache}; other
functions are primarily for internal use.
}
\details{
\code{makeHash} Make a hash string of given arguments.

\code{makeFileHash} Generates a hash from the contents of one or more files.

\code{loadCacheData} Loads cached data from a database.

\code{saveCacheData} caches data in a database.

\code{clearCache} will either remove one or more tables within the cache \code{sqlite} database or simply
  wipe the whole cache file. Removing tables will \code{VACUUM} the database (unless \code{vacuum=FALSE}), which may
  take some time for large cache files.
}
