% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{CCS-Conversion}
\alias{CCS-Conversion}
\alias{convertMobilityToCCS}
\alias{convertCCSToMobility}
\title{Conversion between mobility and CCS}
\usage{
convertMobilityToCCS(mobility, mz, CCSParams, charge = NULL)

convertCCSToMobility(ccs, mz, CCSParams, charge = NULL)
}
\arguments{
\item{mobility, ccs}{A \code{numeric} vector with mobility or \acronym{CCS} values that should be converted.}

\item{mz}{A \code{numeric} vector with the \emph{m/z} values that map to the input mobility or \acronym{CCS} values.}

\item{CCSParams}{A \code{list} with parameters for mobility <--> \acronym{CCS} conversion. See
\code{\link{getCCSParams}} for details and to make such parameter lists. Set to \code{NULL} to skip conversions.}

\item{charge}{A \code{numeric} vector with the ion charges that map to the input mobility or \acronym{CCS} data. Will
be recycled if necessary. If \code{NULL} then the charge configured in \code{CCSParams} is used.}
}
\description{
Utility functions to convert between mobility and \acronym{CCS} data.
}
