% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\docType{class}
\name{clusterInfo-class}
\alias{clusterInfo-class}
\alias{clusterInfo}
\alias{length,clusterInfo-method}
\alias{show,clusterInfo-method}
\alias{clusterProperties,clusterInfo-method}
\alias{clusterProperties}
\title{Hierarchical clustering information}
\usage{
\S4method{length}{clusterInfo}(x)

\S4method{show}{clusterInfo}(object)

\S4method{clusterProperties}{clusterInfo}(cInfo)
}
\arguments{
\item{x, object, cInfo}{A \code{clusterInfo} object.}
}
\value{
\code{clusterProperties} returns a \code{list}
}
\description{
Objects from this class hold information required for
\link[=h-cluster]{hierarchical clustering}.
}
\details{
Objects from this class are returned by \code{\link{makeHCluster}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: \code{length} returns the total number of clusters.

\item \code{show}: \code{show} prints general information about this
object.

\item \code{clusterProperties}: returns a list with data and properties relevant to
the clustering that was performed (see \verb{Slots} section for a
description).
}}

\section{Slots}{

\describe{
\item{\code{clusterm}}{Numeric matrix with normalized feature group intensities that
was used for clustering.}

\item{\code{distm}}{Distance matrix that was used for clustering (obtained with
\code{\link{daisy}}).}

\item{\code{clust}}{Object returned by \code{\link{hclust}}.}

\item{\code{metric}}{Distance metric that was used (\emph{e.g.} euclidean).}

\item{\code{method}}{Clustering method that was used.}

\item{\code{average}}{If \code{TRUE} then feature groups from the same replicate
groups were averaged prior to clustering.}
}}

