% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-intclust.R
\docType{class}
\name{componentsIntClust-class}
\alias{componentsIntClust-class}
\alias{componentsIntClust}
\alias{plotHeatMap,componentsIntClust-method}
\alias{plotHeatMap}
\alias{plotInt,componentsIntClust-method}
\title{Components based on clustered intensity profiles.}
\usage{
\S4method{plotHeatMap}{componentsIntClust}(
  obj,
  interactive = FALSE,
  col = NULL,
  margins = c(6, 2),
  cexCol = 1,
  ...
)

\S4method{plotInt}{componentsIntClust}(obj, index, pch = 20, type = "b", lty = 3, col = NULL, ...)
}
\arguments{
\item{interactive}{If \code{TRUE} an interactive heatmap will be drawn (with
\code{\link{heatmaply}}).}

\item{col}{The colour used for plotting. Set to \code{NULL} for automatic
colours.}

\item{margins, cexCol}{Passed to \code{\link{heatmap.2}}}

\item{\dots}{Further options passed to \code{\link{heatmap.2}} /
\code{\link{heatmaply}} (\code{plotHeatMap}), \code{\link{plot.dendrogram}}
(\code{plot}) or \code{\link[graphics]{plot}} (\code{plotInt}).}

\item{index}{Numeric component/cluster index.}

\item{lty}{Passed to \code{\link{lines}}.}

\item{x, obj}{A \code{componentsIntClust} object.}
}
\value{
\code{plotHeatMap} returns the same as \code{\link{heatmap.2}} or
  \code{\link{heatmaply}}.
}
\description{
This class is derived from \code{\link{components}} and is used to store
hierarchical clustering information from intensity profiles of feature
groups.
}
\details{
Objects from this class are generated by
\code{\link{generateComponentsIntClust}}
}
\section{Methods (by generic)}{
\itemize{
\item \code{plotHeatMap}: draws a heatmap using the
\code{\link{heatmap.2}} or \code{\link{heatmaply}} function.

\item \code{plotInt}: makes a plot for all (normalized) intensity
profiles of the feature groups within a given cluster.
}}

\section{Slots}{

\describe{
\item{\code{clusterm}}{Numeric matrix with normalized feature group intensities that
was used for clustering.}

\item{\code{distm}}{Distance matrix that was used for clustering (obtained with
\code{\link{daisy}}).}

\item{\code{clust}}{Object returned by \code{\link{hclust}}.}

\item{\code{cutClusters}}{A \code{list} with assigned clusters (same format as what
\code{\link{cutree}} returns).}

\item{\code{gInfo}}{The \code{\link{groupInfo}} of the feature groups object that was
used.}

\item{\code{properties}}{A list containing general properties and parameters used for
clustering.}
}}

\note{
The intensity values for components (used by \code{plotSpectrum}) are set
  to a dummy value (1) as no single intensity value exists for this kind of
  components.
}
\references{
\insertRef{Scholle2018}{patRoon}
}
\seealso{
\code{\link{components}} and \link{component-generation}
}
