% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-nontarget.R,
%   R/components-nontarget-set.R
\docType{class}
\name{componentsNT-class}
\alias{componentsNT-class}
\alias{componentsNT}
\alias{plotGraph,componentsNT-method}
\alias{componentsNTSet-class}
\alias{componentsNTSet}
\alias{plotGraph,componentsNTSet-method}
\alias{componentsNTUnset-class}
\alias{componentsNTUnset}
\alias{unset,componentsNTSet-method}
\title{Components class for homologous series.}
\usage{
\S4method{plotGraph}{componentsNT}(obj, onlyLinked)

\S4method{plotGraph}{componentsNTSet}(obj, onlyLinked, set)

\S4method{unset}{componentsNTSet}(obj, set)
}
\arguments{
\item{obj}{The \code{componentsNT} object to plot.}

\item{onlyLinked}{If \code{TRUE} then only components with links are plotted.}

\item{set}{\setsWF The name of the set.}
}
\value{
\code{plotGraph} returns the result of \code{\link{visNetwork}}.
}
\description{
This class is derived from \code{\link{components}} and is used to store
results from unsupervised homolog detection with the \pkg{\link{nontarget}}
package.
}
\details{
Objects from this class are generated by
\code{\link{generateComponentsNontarget}}
}
\section{Methods (by generic)}{
\itemize{
\item \code{plotGraph}: Plots an interactive network graph for linked
homologous series (\emph{i.e.} series with (partial) overlap which could
not be merged). The resulting graph can be browsed interactively and allows
quick inspection of series which may be related. The graph is constructed
with the \pkg{\link{igraph}} package and rendered with
\pkg{\link{visNetwork}}.
}}

\section{Slots}{

\describe{
\item{\code{homol}}{A \code{list} with \code{homol} objects for each replicate group
as returned by \code{\link{homol.search}}}
}}

\section{Sets workflows}{
 \setsWFClass{componentsNTSet}{componentsNT}

  \setsWFNewMethodsSO{componentsNTUnset}{Only the components in the specified set are kept. Furthermore, the
  component names are restored to non-set specific names (see \code{\link{generateComponents}} for more details).}

  \setsWFChangedMethods{

  \item \code{plotGraph} Currently can only create graph networks from one set (specified by the \code{set}
  argument).

  }

  Note that the \code{componentsNTSet} class does not have a \code{homol} slot. Instead, the \code{\link{setObjects}}
  method can be used to access this data for a specific set.
}

\references{
\addCitations{nontarget}{1} \cr\cr \addCitations{enviPat}{1}

\addCitations{igraph}{1} \cr \cr \addCitations{visNetwork}{1}
}
\seealso{
\code{\link{components}} and \link{component-generation}
}
