% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-specclust.R
\docType{class}
\name{componentsSpecClust-class}
\alias{componentsSpecClust-class}
\alias{componentsSpecClust}
\title{Components based on MS/MS similarity.}
\description{
This class is derived from \code{\link{componentsClust}} and is used to store components from feature groups that
were clustered based on their MS/MS similarities.
}
\details{
Objects from this class are generated by \code{\link{generateComponentsSpecClust}}
}
\note{
When the object is altered (\emph{e.g.} by filtering or subsetting it), methods that need the original
  clustered data such as plotting methods do not work anymore and stop with an error.
}
\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{componentsClust}}}   \itemize{     \item{\strong{\code{\link{componentsSpecClust}}}}   } }
}

\seealso{
\code{\link{componentsClust}} for other relevant methods and \code{\link{generateComponents}}
}
