% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/compounds.R, R/compounds-metfrag.R,
%   R/compounds-sirius.R
\docType{methods}
\name{compound-generation}
\alias{compound-generation}
\alias{generateCompounds,featureGroups-method}
\alias{generateCompounds}
\alias{generateCompoundsMetfrag}
\alias{generateCompoundsSirius}
\title{Automatic compound identification}
\usage{
\S4method{generateCompounds}{featureGroups}(fGroups, MSPeakLists, algorithm,
  ...)

generateCompoundsMetfrag(fGroups, MSPeakLists, method = "CL",
  logPath = file.path("log", "metfrag"), timeout = 300,
  timeoutRetries = 2, errorRetries = 2, topMost = 100, dbRelMzDev = 5,
  fragRelMzDev = 5, fragAbsMzDev = 0.002, isPositive, adduct,
  database = "PubChem", chemSpiderToken = "",
  scoreTypes = c("FragmenterScore", "OfflineMetFusionScore"),
  scoreWeights = 1, preProcessingFilters = c("UnconnectedCompoundFilter",
  "IsotopeFilter"), postProcessingFilters = c("InChIKeyFilter"),
  maxCandidatesToStop = 2500, addTrivialNames = TRUE, identifiers = NULL,
  extraOpts = NULL, maxProcAmount = getOption("patRoon.maxProcAmount"))

generateCompoundsSirius(fGroups, MSPeakLists, maxMzDev = 5,
  adduct = "[M+H]+", elements = "CHNOP", profile = "qtof",
  formulaDatabase = NULL, fingerIDDatabase = "pubchem", noise = NULL,
  errorRetries = 2, topMost = 100, logPath = file.path("log", "sirius"),
  maxProcAmount = getOption("patRoon.maxProcAmount"))
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which compounds should
be identified. This should be the same or a subset of the object that was
used to create the specified \code{MSPeakLists}. In the case of a subset
only the remaining feature groups in the subset are considered.}

\item{MSPeakLists}{A \code{\link{MSPeakLists}} object that was generated for
the supplied \code{fGroups}.}

\item{algorithm}{A character string describing the algorithm that should be
used: \code{"metfrag"}, \code{"sirius"}}

\item{...}{Any parameters to be passed to the selected compound generation
algorithm.}

\item{method}{Which method should be used for MetFrag execution: \code{"CL"}
for \command{MetFragCL} and \code{"R"} for \command{MetFragR}. The former
might be faster.}

\item{logPath}{Destination directory for log files with output from executed
commands. Will be created if non-existant. Set to \code{NULL} to disable
logging.}

\item{timeout}{Maximum time (in seconds) before a metFrag query for a feature
group is stopped. Also see \code{timeoutRetries} argument.}

\item{timeoutRetries}{Maximum number of retries after reaching a timeout
before completely skipping the metFrag query for a feature group. Also see
\code{timeout} argument.}

\item{errorRetries}{Maximum number of retries after an error occurred. This
may be useful to handle e.g. connection errors.}

\item{topMost}{Only keep this number of candidates (per feature group) with
highest score. Set to \code{NULL} to always keep all candidates, however,
please note that this may result in significant usage of CPU/RAM resources
for large numbers of candidates.}

\item{dbRelMzDev}{Relative mass deviation (in ppm) for database search. Sets
the \option{DatabaseSearchRelativeMassDeviation} option.}

\item{fragRelMzDev}{Relative mass deviation (in ppm) for fragment matching.
Sets the \option{FragmentPeakMatchRelativeMassDeviation} option.}

\item{fragAbsMzDev}{Absolute mass deviation (in Da) for fragment matching.
Sets the \option{FragmentPeakMatchAbsoluteMassDeviation} option.}

\item{isPositive}{Set to \code{TRUE} for data measured with positive
ionization. Sets the \option{IsPositiveIonMode} option.}

\item{adduct}{The assumed MS adduct. For \code{generateCompoundsMetfrag}: a
numeric value (\emph{e.g.} \samp{1} for [M+H]+, \samp{-1} for [M-H]- and
\samp{0} for [M]+/-, see the
\href{http://c-ruttkies.github.io/MetFrag/projects/metfragcl/}{MetFrag CL}
homepage for more). Please note that \code{isPositive} also needs to be
set. For \code{generateCompoundsSirius}, a character string such as
\code{"[M+H]+"}, \code{"[M-H]-"}, \code{"[M+Na]+"} and \code{"[M]+"}. Sets
the \option{PrecursorIonMode} and \option{--ion} options, respectively.}

\item{database}{Compound database to use. Valid values are: \code{"PubChem"},
\code{"ExtendedPubChem"}, \code{"ChemSpider"} and \code{"KEGG"}. The
\code{"ExtendedPubChem"} is a PubChem database which includes number of
patents and references information, which an be used for further scoring
(see \code{scoreTypes} parameter.) In order to use the \code{ChemSpider}
database the \code{chemSpiderToken} should be set. Sets the
\option{MetFragDatabaseType} option.}

\item{chemSpiderToken}{A character string with the
\href{http://www.chemspider.com/AboutServices.aspx}{ChemSpider security
token} that should be set when the ChemSpider database is used. Sets the
\option{ChemSpiderToken} option.}

\item{scoreTypes}{A character vector defining the scoring methods
(\emph{e.g.} \code{"FragmenterScore"}, \code{"OfflineMetFusionScore"},
\code{"RetentionTimeScore"}). Some methods require further options to be
set. Additional scoring methods become available when the
\option{"ExtendedPubChem"} (\emph{i.e.} \option{"PubChemNumberPatents"} and
\option{"PubChemNumberPubMedReferences"}) or \option{"ChemSpider"}
(\emph{i.e.} \option{"ChemSpiderReferenceCount"},
\option{"ChemSpiderNumberExternalReferences"},
\option{"ChemSpiderRSCCount"}, \option{"ChemSpiderNumberPubMedReferences"}
and \option{"ChemSpiderDataSourceCount"}) database type is used. For all
scoring types and more information refer to the \verb{Candidate Scores}
section on the
\href{http://c-ruttkies.github.io/MetFrag/projects/metfragr/}{MetFragR
homepage}. Sets the \option{MetFragScoreTypes} option.}

\item{scoreWeights}{Numeric vector containing weights of the used scoring
types. Order is the same as set in \code{scoreTypes}. Values are recycled
if necessary. Sets the \option{MetFragScoreWeights} option.}

\item{preProcessingFilters, postProcessingFilters}{A character vector defining
pre/post filters applied before/after fragmentation and scoring
(\emph{e.g.} \code{"UnconnectedCompoundFilter"}, \code{"IsotopeFilter"},
\code{"ElementExclusionFilter"}). Some methods require further options to
be set. For all filters and more information refer to the \verb{Candidate
Filters} section on the
\href{http://c-ruttkies.github.io/MetFrag/projects/metfragr/}{MetFragR
homepage}. Sets the \option{MetFragPreProcessingCandidateFilter} and
\code{MetFragPostProcessingCandidateFilter} options.}

\item{maxCandidatesToStop}{If more than this number of candidate structures
are found then processing will be aborted and no results this feature group
will be reported. Low values increase the chance of missing data, whereas
too high values will use too much computer resources and signficantly
slowdown the process. Sets the \option{MaxCandidateLimitToStop} option.}

\item{addTrivialNames}{If \code{TRUE} and the PubChem database is used then
trivial names will be added after compound search.}

\item{identifiers}{A \code{list} containing for each feature group a
character vector with database identifiers that should be used to find
candidates for a feature group (the list should be named by feature group
names). If \code{NULL} all relevant candidates will be retrieved from the
specified database. An example usage scenario is to obtain the list of
candidate identifiers from a \code{\link{compounds}} object obtained with
\code{\link{generateCompoundsSirius}} using the \code{\link{identifiers}}
method. This way, only those candidates will be searched by MetFrag that
were generated by SIRIUS+CSI:FingerID. Sets the
\option{PrecursorCompoundIDs} option.}

\item{extraOpts}{A named \code{list} containing further settings to be passed
to \code{\link[metfRag]{run.metfrag}}. See the
\href{http://c-ruttkies.github.io/MetFrag/projects/metfragr/}{MetFragR} and
\href{http://c-ruttkies.github.io/MetFrag/projects/metfragcl/}{MetFrag CL}
homepages for all available options.}

\item{maxProcAmount}{Maximum number of processes to run for parallelization.
Usually a number close to the amount of physical cores yields most
efficient results.}

\item{maxMzDev}{Maximum deviation between the measured and candidate formula
\emph{m/z} values (in ppm). Sets the \option{--ppm-max} commandline
option.}

\item{elements}{Elements to be considered for formulae calculation. This will
heavily affects the number of candidates! Always try to work with a minimal
set by excluding elements you don't expect. The minimum/maximum number
of elements can also be specified, for example: a value of
\code{"C[5]H[10-15]O"} will only consider formulae with up to five carbon
atoms, between ten and fifteen hydrogen atoms and any amount of oxygen
atoms. Sets the \option{--elements} commandline option.}

\item{profile}{Name of the configuration profile, for example:
\option{"qtof"}, \option{"orbitrap"}, \option{"fticr"}. Sets the
\option{--profile} commandline option.}

\item{formulaDatabase}{If not \code{NULL}, use a database for retrieval of formula
candidates. Possible values are: \option{"pubchem"}, \option{"bio"},
\option{"kegg"}, \option{"hmdb"}. Sets the \option{--database} commandline
option.}

\item{fingerIDDatabase}{Database specifically used for
\command{CSI:FingerID}. If \code{NULL}, the value of the
\code{formulaDatabase} parameter will be used or \code{"pubchem"} when that
is also \code{NULL}. Sets the \option{--fingerid-db} option.}

\item{noise}{Median intensity of the noise (\code{NULL} ignores this
parameter). Sets the \option{--noise} commandline option.}
}
\value{
A \code{\link{compounds}} object containing all tentatively
  identified compounds.
}
\description{
Functionality to automatically identify chemical compounds from feature
groups.
}
\details{
Several algorithms are provided to automatically identify compounds for given
feature groups. To this end, each measured masses for all feature groups are
searched within online database(s) (\emph{e.g.}
\href{https://pubchem.ncbi.nlm.nih.gov/}{PubChem}) to retrieve a list of
potential candidate chemical compounds. Depending on the algorithm and its
parameters, further scoring of candidates is then performed using, for
instance, matching of measured and theoretical isotopic patterns, presence
within other data sources such as patent databases and similarity of measured
and in-silico predicted MS/MS fragments. Note that this process is often
quite time consuming, especially for large feature group sets. Therefore,
this is often one of the last steps within the workflow and not performed
before feature groups have been prioritized.

\code{generateCompounds} is a generic function that will generate compounds
  using one of the supported algorithms. The actual functionality is provided
  by algorithm specific functions such as \code{generateCompoundsMetfrag} and
  \code{generateCompoundsSirius}. While these functions may be called directly,
  \code{generateCompounds} provides a generic interface and is therefore usually
  preferred.

\code{generateCompoundsMetfrag} uses the \pkg{metfRag} package for
  compound identification (see \url{http://c-ruttkies.github.io/MetFrag/}).
  Several online compound databases such as
  \href{https://pubchem.ncbi.nlm.nih.gov/}{PubChem} and
  \href{http://www.chemspider.com/}{ChemSpider} may be chosen for retrieval
  of candidate structures. In addition, many options exist to score and
  filter resulting data, and it is highly suggested to optimize these to
  improve results. While MS/MS data is not mandatory, it will usually greatly
  improve candidate scoring. The \command{MetFrag} options \code{PeakList},
  \code{IonizedPrecursorMass} and \code{ExperimentalRetentionTimeValue} (in
  minutes) fields are automatically set from feature data.

\code{generateCompoundsSirius} uses
  \href{https://bio.informatik.uni-jena.de/software/sirius/}{SIRIUS} in
  combination with \href{https://www.csi-fingerid.uni-jena.de/}{CSI:FingerID}
  for compound identification. Similar to
  \code{\link{generateFormulasSirius}}, candidate formulae are generated with
  SIRIUS. These results are then feed to CSI:FingerID to acquire candidate
  structures. Candidates are then scored on similarity of measured and
  in-silico predicted MS/MS fragments and if the PubChem database is used, on
  the number of references in patents. This method requires the availability
  of MS/MS data, and feature groups without it will be ignored.
}
\note{
At this point only mass spectrometric data from the most intense
  feature is used for compound identification.
}
\references{
\insertRef{Ruttkies2016}{patRoon}

\insertRef{Duhrkop2015}{patRoon} \cr\cr
  \insertRef{Duhrkop2015-2}{patRoon} \cr\cr
  \insertRef{Bcker2008}{patRoon}
}
\seealso{
\code{\link{compounds-class}}
}
