% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds-cluster.R
\docType{class}
\name{compoundsCluster-class}
\alias{compoundsCluster-class}
\alias{compoundsCluster}
\alias{clusters,compoundsCluster-method}
\alias{cutClusters,compoundsCluster-method}
\alias{clusterProperties,compoundsCluster-method}
\alias{groupNames,compoundsCluster-method}
\alias{length,compoundsCluster-method}
\alias{lengths,compoundsCluster-method}
\alias{show,compoundsCluster-method}
\alias{[,compoundsCluster,ANY,missing,missing-method}
\alias{treeCut,compoundsCluster-method}
\alias{treeCutDynamic,compoundsCluster-method}
\alias{plot,compoundsCluster,ANY-method}
\alias{getMCS,compoundsCluster-method}
\alias{plotStructure,compoundsCluster-method}
\title{Compounds cluster class}
\usage{
\S4method{clusters}{compoundsCluster}(obj)

\S4method{cutClusters}{compoundsCluster}(obj)

\S4method{clusterProperties}{compoundsCluster}(obj)

\S4method{groupNames}{compoundsCluster}(obj)

\S4method{length}{compoundsCluster}(x)

\S4method{lengths}{compoundsCluster}(x, use.names = TRUE)

\S4method{show}{compoundsCluster}(object)

\S4method{[}{compoundsCluster,ANY,missing,missing}(x, i, j, ...,
  drop = TRUE)

\S4method{treeCut}{compoundsCluster}(obj, k = NULL, h = NULL,
  groupName)

\S4method{treeCutDynamic}{compoundsCluster}(obj, maxTreeHeight, deepSplit,
  minModuleSize, groupName)

\S4method{plot}{compoundsCluster,ANY}(x, groupName, pal = "Paired",
  colourBranches = lengths(x)[groupName] < 50,
  showLegend = lengths(x)[groupName] < 20, ...)

\S4method{getMCS}{compoundsCluster}(obj, groupName, cluster)

\S4method{plotStructure}{compoundsCluster}(obj, groupName, cluster,
  width = 500, height = 500, withTitle = TRUE)
}
\arguments{
\item{obj, x, object}{A \code{compoundsCluster} object.}

\item{use.names}{A logical value specifying whether the returned vector
should be named with the feature group names.}

\item{i}{Either a numeric, character or logical \code{vector} that is used to select feature groups by their index, name and logical selection, respectively (for the order/names see \code{groupNames()}). If missing all feature groups are selected.}

\item{j}{ignored.}

\item{\dots}{Any arguments directly given to \code{\link{plot.dendrogram}}.}

\item{drop}{ignored.}

\item{k, h}{Desired number of clusters or tree height to be used for cutting
the dendrogram, respecitively. One or the other must be specified.
Analogous to \code{\link{cutree}}.}

\item{groupName}{A character specifying the feature group name.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}

\item{pal}{Colour palette to be used from \pkg{\link{RColorBrewer}}.}

\item{colourBranches}{Whether branches from cut clusters (and their labels)
should be coloured. Might be slow with large numbers of clusters, hence,
the default is only \code{TRUE} when this is not the case.}

\item{showLegend}{If \code{TRUE} and \code{colourBranches} is also
\code{TRUE} then a legend will be shown which outlines cluster numbers and
their colours. By default \code{TRUE} for small amount of clusters to avoid
overflowing the plot.}

\item{cluster}{A numeric value specifying the cluster.}

\item{width, height}{The dimensions (in pixels) of the raster image that
should be plotted.}

\item{withTitle}{A logical value specifying whether a title should be added.}
}
\value{
\code{cutTree} and \code{cutTreeDynamic} return the modified
  \code{compoundsCluster} object.

\code{getMCS} returns an \CRANpkg{rcdk} molecule object
  (\code{IAtomContainer}).
}
\description{
Objects from this class are used to store hierarchical clustering data of
candidate structures within \code{\link{compounds}} objects.
}
\details{
Objects from this type are returned by the \code{compounds} method for
\code{\link[=makeHCluster,compounds-method]{makeHCluster}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{clusters}: Accessor method to the \code{clusters} slot.
Returns a list that contains for each feature group an object as returned
by \code{\link{hclust}}.

\item \code{cutClusters}: Accessor method to the \code{cutClusters} slot.
Returns a list that contains for each feature group a vector with cluster
membership for each candidate (format as \code{\link{cutree}}).

\item \code{clusterProperties}: Returns a list with properties on how the
clustering was performed.

\item \code{groupNames}: returns a \code{character} vector with the names of the
feature groups from which data is present in this object.

\item \code{length}: Returns the total number of clusters.

\item \code{lengths}: Returns a \code{vector} with the number of
clusters per feature group.

\item \code{show}: Show summary information for this object.

\item \code{[}: Operator to subset on feature groups .

\item \code{treeCut}: Manually (re-)cut a dendrogram that was
generated for a feature group.

\item \code{treeCutDynamic}: Automatically (re-)cut a dendrogram that was
generated for a feature group using the \code{\link{cutreeDynamicTree}}
function from \pkg{\link{dynamicTreeCut}}.

\item \code{plot}: Plot the dendrogram for clustered compounds of a
feature group. Clusters are highlighted using \CRANpkg{dendextend}.

\item \code{getMCS}: Calculates the maximum common substructure (MCS)
for all candidate structures within a specified cluster. This method uses
the \code{\link{get.mcs}} function from \CRANpkg{rcdk}.

\item \code{plotStructure}: Plots the maximum common substructure (MCS) for
all candidate structures within a specified cluster.
}}

\section{Slots}{

\describe{
\item{\code{clusters}}{A \code{list} with \code{\link{hclust}} objects for each
feature group.}

\item{\code{SMILES}}{A \code{list} containing a vector with \code{SMILES} for all
candidate structures per feature group.}

\item{\code{cutClusters}}{A \code{list} with assigned clusters for all candidates per
feature group (same format as what \code{\link{cutree}} returns).}

\item{\code{properties}}{A list containing general properties and parameters used for
clustering.}
}}

