% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds-sirius.R
\docType{class}
\name{compoundsSIRIUS-class}
\alias{compoundsSIRIUS-class}
\alias{compoundsSIRIUS}
\title{Compounds class for SIRIUS results.}
\description{
This class is derived from \code{\link{compounds}} and contains additional specific \command{SIRIUS} data.
}
\details{
Objects from this class are generated by \code{\link{generateCompoundsSIRIUS}}
}
\section{Slots}{

\describe{
\item{\code{fingerprints}}{A \code{list} with for each feature group result a \code{data.table} containing fingerprints
obtained with \command{CSI:FingerID}.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{compounds}}}   \itemize{     \item{\strong{\code{\link{compoundsSIRIUS}}}}   } }
}

\references{
\insertRef{Dhrkop2019}{patRoon} \cr\cr \insertRef{Duhrkop2015}{patRoon} \cr\cr
  \insertRef{Duhrkop2015-2}{patRoon} \cr\cr \insertRef{Bcker2008}{patRoon}
}
\seealso{
\code{\link{compounds}} and \code{\link{generateCompoundsSIRIUS}}
}
