% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{COMStop}
\alias{COMStop}
\title{Internal fix for \pkg{RDCOMClient}, ignore.}
\usage{
COMStop(...)
}
\arguments{
\item{...}{ignore}
}
\description{
Internal fix for \pkg{RDCOMClient}, ignore.
}
\keyword{internal}
