% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertMSFiles}
\alias{convertMSFiles}
\title{Conversion of analyses between several open formats}
\usage{
convertMSFiles(anaInfo, formatFrom, formatTo, outPath = anaInfo$path,
  overWrite = FALSE)
}
\arguments{
\item{anaInfo}{A table with \link[=analysis-information]{analysis
information} that contains the analyses to be converted.}

\item{formatFrom, formatTo}{A \code{character} that specifies the source and
destination format, respectively. Valid options are: \code{"mzXML"},
\code{"mzML"} and \code{"mzData"}.}

\item{outPath}{A character vector specifying directories that should be used for the output. Usually
this is the same as the source (otherwise the
\link[=analysis-information]{analysis information} should be changed for
further processing). Will be re-cycled if necessary.}

\item{overWrite}{Should existing destination file be overwritten
(\code{TRUE}) or not (\code{FALSE})?}
}
\description{
This function uses the \command{FileConverter} command from
\href{http://www.openms.de/}{OpenMS} to convert analyses files from one open
format to another. The supported file formats are (both directions):
\file{.mzXML}, \file{.mzML} and \file{.mzData}.
}
\references{
\insertRef{Rst2016}{patRoon}
}
