% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{createCOMReference}
\alias{createCOMReference}
\title{Internal fix for \pkg{RDCOMClient}, ignore.}
\usage{
createCOMReference(ref, className)
}
\arguments{
\item{ref, className}{ignore}
}
\description{
Internal fix for \pkg{RDCOMClient}, ignore.
}
\keyword{internal}
