% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-openms.R
\name{defaultOpenMSAdducts}
\alias{defaultOpenMSAdducts}
\title{Default adducts for OpenMS componentization}
\usage{
defaultOpenMSAdducts(ionization)
}
\arguments{
\item{ionization}{The ionization polarity: either \code{"positive"} or \code{"negative"}.}
}
\description{
Returns the default adducts and their probabilities when the OpenMS algorithm is used for componentization.
}
\details{
See the \code{potentialAdducts} argument of \code{\link{generateComponentsOpenMS}} for more details.
}
