% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/feature_groups-filter.R
\name{feature-filtering}
\alias{feature-filtering}
\alias{filter,featureGroups-method}
\alias{filter,featureGroupsSet-method}
\alias{replicateSubtract,featureGroups-method}
\alias{replicateSubtract}
\title{Filtering of grouped features}
\usage{
replicateSubtract(fGroups, replicates, threshold = 0)

\S4method{filter}{featureGroups}(
  obj,
  absMinIntensity = NULL,
  relMinIntensity = NULL,
  preAbsMinIntensity = NULL,
  preRelMinIntensity = NULL,
  absMinMaxIntensity = NULL,
  relMinMaxIntensity = NULL,
  absMinAnalyses = NULL,
  relMinAnalyses = NULL,
  absMinReplicates = NULL,
  relMinReplicates = NULL,
  absMinFeatures = NULL,
  relMinFeatures = NULL,
  absMinReplicateAbundance = NULL,
  relMinReplicateAbundance = NULL,
  absMinConc = NULL,
  relMinConc = NULL,
  absMaxTox = NULL,
  relMaxTox = NULL,
  absMinConcTox = NULL,
  relMinConcTox = NULL,
  maxReplicateIntRSD = NULL,
  maxReplicateIntRSDPres = NULL,
  blankThreshold = NULL,
  retentionRange = NULL,
  mzRange = NULL,
  mzDefectRange = NULL,
  chromWidthRange = NULL,
  featQualityRange = NULL,
  groupQualityRange = NULL,
  replicates = NULL,
  IMS = NULL,
  withIMSPrecursor = FALSE,
  IMSRangeParams = NULL,
  results = NULL,
  removeBlanks = FALSE,
  removeISTDs = FALSE,
  checkFeaturesSession = NULL,
  predAggrParams = getDefPredAggrParams(),
  removeNA = FALSE,
  negate = FALSE,
  applyIMS = "both"
)

\S4method{filter}{featureGroupsSet}(
  obj,
  ...,
  negate = FALSE,
  applyIMS = "both",
  sets = NULL,
  absMinSets = NULL,
  relMinSets = NULL
)

\S4method{replicateSubtract}{featureGroups}(fGroups, replicates, threshold = 0)
}
\arguments{
\item{fGroups, obj}{\code{\link{featureGroups}} object to which the filter is applied.}

\item{replicates}{A character vector of replicates that should be kept (\code{filter}) or subtracted from
(\code{replicateSubtract}).}

\item{threshold}{Minimum relative threshold (compared to mean intensity of
replicate being subtracted) for a feature group to be \emph{not}
removed. When \samp{0} a feature group is always removed when present in
the given replicates.}

\item{absMinIntensity, relMinIntensity}{Minimum absolute/relative intensity for features to be kept. The relative
intensity is determined from the feature with highest intensity (of
all features from all groups). Set to \samp{0} or \code{NULL} to skip this step.}

\item{preAbsMinIntensity, preRelMinIntensity}{As \code{absMinIntensity}/\code{relMinIntensity}, but applied
\emph{before} any other filters. This is typically used to speed-up subsequent filter steps. However, care must be
taken that a sufficiently low value is chosen that is not expected to affect subsequent filtering steps. See below
why this may be important.}

\item{absMinMaxIntensity, relMinMaxIntensity}{Feature groups are only kept if at least one feature in the group has an
intensity above this absolute/relative threshold.}

\item{absMinAnalyses, relMinAnalyses}{Feature groups are only kept when they contain data for at least this (absolute
or relative) amount of analyses. Set to \code{NULL} to ignore.}

\item{absMinReplicates, relMinReplicates}{Feature groups are only kept when they contain data for at least this
(absolute or relative) amount of replicates. Set to \code{NULL} to ignore.}

\item{absMinFeatures, relMinFeatures}{Analyses are only kept when they contain at least this (absolute or relative)
amount of features. Set to \code{NULL} to ignore.}

\item{absMinReplicateAbundance, relMinReplicateAbundance}{Minimum absolute/relative abundance that a grouped feature
should be present within a replicate. If this minimum is not met all features within the replicate are removed. Set
to \code{NULL} to skip this step.}

\item{absMinConc, relMinConc}{The minimum absolute/relative predicted concentration (calculated by
\code{\link{calculateConcs}}) assigned to a feature. The toxicities are first aggregated prior to filtering, as
controlled by the \code{predAggrParams} argument. Also see the \code{removeNA} argument.}

\item{absMaxTox, relMaxTox}{The maximum absolute/relative predicted toxicity (LC50) (calculated by
\code{\link{calculateTox}}) assigned to a feature group. The concentrations are first aggregated prior to
filtering, as controlled by the \code{predAggrParams} argument. Also see the \code{removeNA} argument.}

\item{absMinConcTox, relMinConcTox}{Like \code{absMinConc}/\code{relMinConc}, but instead considers the ratio between
feature concentrations and the toxicity of the feature group. For instance, \code{absMinConcTox=0.1} means that the
calculated concentration of a feature should be at least \samp{10\%} of its toxicity.}

\item{maxReplicateIntRSD, maxReplicateIntRSDPres}{Maximum relative standard deviation (RSD) of intensity values for
features within a replicate. If the RSD is above this value all features within the replicate are removed. The
\code{maxReplicateIntRSD} includes absent features in the RSD calculation (\emph{i.e.} zero intensities), whereas
these are ignored with \code{maxReplicateIntRSDPres}. Set to \code{NULL} to ignore.}

\item{blankThreshold}{Feature groups that are also present in blank analyses (see
\link[=analysis-information]{analysis info}) are filtered out unless their relative intensity is above this
threshold. For instance, a value of \samp{5} means that only features with an intensity five times higher than that
of the blank are kept. The relative intensity values between blanks and non-blanks are determined from the mean of
all non-zero blank intensities. Set to \code{NULL} to skip this step.}

\item{retentionRange, mzRange, mzDefectRange, chromWidthRange}{Range of retention time (in seconds), \emph{m/z}, mass
defect (defined as the decimal part of \emph{m/z} values) or chromatographic peak width (in seconds), respectively.
Features outside this range will be removed. Should be a numeric vector with length of two containing the min/max
values. The maximum can be \code{Inf} to specify no maximum range. Set to \code{NULL} to skip this step.}

\item{featQualityRange}{Used to filter features by their peak qualities/scores
(see \code{\link{calculatePeakQualities}}). Should be a named \code{list} with min/max ranges for each
quality/score to be filtered (the \code{\link{getFeatureQualityNames}} function can be used to obtain valid names).
Example: \code{featQualityRange=list(ModalityScore=c(0.3, Inf),
SymmetryScore=c(0.5, Inf))}. Set to \code{NULL} to ignore.}

\item{groupQualityRange}{Like \code{featQualityRange}, but filters on group specific or averaged qualities/scores.}

\item{IMS}{\IMSWF Specifies which feature groups are considered to be kept in IMS workflows. The following options
  are valid: \itemize{

  \item \code{"both"}: Selects IMS and non-IMS features.

  \item \code{"maybe"}: Selects non-IMS features and IMS features without assigned IMS precursor.

  \item \code{FALSE}: Selects only non-IMS features.

  \item \code{TRUE}: Selects only IMS features.

  }

  Set to \code{NULL} to ignore.}

\item{withIMSPrecursor}{\IMSWF only keep IMS feature groups with IMS precursors, \emph{i.e.} remove all orphans.
Unaffected by \code{negate=TRUE}.}

\item{IMSRangeParams}{\IMSWF A \code{list} with parameters to be used for filtering IMS range data. See
\code{\link{getIMSRangeParams}} for details and how to make such a parameter list.}

\item{results}{Only keep feature groups that have results in the object specified by \code{results}. Valid classes
are \code{\link{featureAnnotations}} (\emph{e.g.} formula/compound annotations) and \code{\link{components}}. Can
also be a \code{list} with multiple objects: in this case a feature group is kept if it has a result in \emph{any}
of the objects. Set to \code{NULL} to ignore.}

\item{removeBlanks}{Set to \code{TRUE} to remove all analyses that belong to replicates that are specified as a blank
in the \link{analysis-information}. This is useful to simplify the analyses in the specified
\code{\link{featureGroups}} object after blank subtraction. When both \code{blankThreshold} and this argument are
set, blank subtraction is performed prior to removing any analyses.}

\item{removeISTDs}{If \code{TRUE} then all feature groups marked as internal standard (IS) are removed. This requires
IS assignments done by \code{\link{normInts}}, see its documentation for more details.}

\item{checkFeaturesSession}{If set then features and/or feature groups are removed that were selected for removal
(see \link{check-GUI}). The session files are typically generated with the \code{\link{checkFeatures}} and
\code{\link{predictCheckFeaturesSession}} functions. The value of \code{checkFeaturesSession} should either by a
path to the session file or \code{TRUE}, in which case the default session file name is used. If \code{negate=TRUE}
then all non-selected features/feature groups are removed instead.}

\item{predAggrParams}{Parameters to aggregate calculated concentrations/toxicities (obtained with
\code{\link{calculateConcs}}/\code{\link{calculateTox}}) prior to filtering data. See \link[=pred-aggr-params]{prediction aggregation
parameters} for more information.}

\item{removeNA}{Set to \code{TRUE} to remove \code{NA} values. Currently only applicable to the concentration and
toxicity filters.}

\item{negate}{If set to \code{TRUE} then filtering operations are performed in opposite manner.}

\item{applyIMS}{\IMSWF whether the filters are only applied to IMS precursors (\code{applyIMS=FALSE}), only to IMS
features (\code{applyIMS=TRUE}) or to both (\code{applyIMS="both"}). Other feature groups will always be kept. The
\code{negate} option does not affect \code{applyIMS}.}

\item{\dots}{\setsPassedArgs1{featureGroups}}

\item{sets}{\setsWF A \code{character} with name(s) of the sets to keep (or remove if \code{negate=TRUE}).}

\item{absMinSets, relMinSets}{\setsWF Feature groups are only kept when they contain data for at least this (absolute
or relative) amount of sets. Set to \code{NULL} to ignore.}
}
\value{
A filtered \code{\link{featureGroups}} object. Feature groups that are filtered away have their intensity set
  to zero. In case a feature group is not present in any of the analyses anymore it will be removed completely.
}
\description{
Basic rule based filtering of feature groups.
}
\details{
\code{filter} performs common rule based filtering of feature groups such as blank subtraction, minimum
  intensity and minimum replicate abundance. Removing of features occurs by zeroing their intensity values.
  Furthermore, feature groups that are left completely empty (\emph{i.e.} all intensities are zero) will be
  automatically removed.

\code{replicateSubtract} removes feature groups present in a
  given set of replicates (unless intensities are above a given
  threshold). The replicates that are subtracted will be removed.
}
\section{Sets workflows}{
 \setsWFChangedMethods{

  \item \code{filter} has specific arguments to filter by (feature presence in) sets. See the argument descriptions.
  
  \item \strong{Important}: the \code{mzRange}, \code{mzDefectRange} and \code{IMSRangeParams} filters use neutral
  feature masses, whereas non-sets workflows use \emph{m/z} values. Hence, adjust accordingly to avoid (slightly)
  different results!

  }
}

\section{Filter order}{
 When multiple arguments are specified to \code{filter}, multiple filters are applied in
  sequence. Since some of these filters may affect each other, choosing their order correctly may be important for
  effective data filtering. For instance, when an intensity filter removes features from blank analyses, a subsequent
  blank filter may not adequately perform blank subtraction. Similarly, when intensity and blank filters are executed
  after the replicate abundance filter it may be necessary to ensure minimum replicate abundance again as the
  intensity and blank filters may have removed some features within a replicate.

  With this in mind, filters (if specified) occur in the following order:

  \enumerate{

  \item Features/feature groups selected for removal by the session specified by \code{checkFeaturesSession}.

  \item Pre-Intensity filters (\emph{i.e.} \code{preAbsMinIntensity} and \code{preRelMinIntensity}).

  \item Chromatography and mass filters (\emph{i.e} \code{retentionRange}, \code{mzRange}, \code{mzDefectRange},
  \code{chromWidthRange}, \code{featQualityRange} and \code{groupQualityRange}).

  \item Replicate abundance filters (\emph{i.e.} \code{absMinReplicateAbundance}, \code{relMinReplicateAbundance} and
  \code{maxReplicateIntRSD}/\code{maxReplicateIntRSDPres}).

  \item Blank filter (\emph{i.e.} blankThreshold).

  \item Intensity filters (\emph{i.e.} \code{absMinIntensity} and \code{relMinIntensity}).

  \item Replicate abundance filters (2nd time, only if previous filters affected results).

  \item Minimum-maximum intensity filters (\emph{i.e.} \code{absMinMaxIntensity} and \code{relMinMaxIntensity}).

  \item General abundance filters (\emph{i.e.} \code{absMinAnalyses}, \code{relMinAnalyses}, \code{absMinReplicates},
  \code{relMinReplicates}, \code{absMinFeatures}, \code{relMinFeatures}), \code{absMinConc}, \code{relMinConc},
  \code{absMaxTox} and \code{relMaxTox}.

  \item Replicate filter (\emph{i.e.} \code{replicates}), results filter (\emph{i.e.} \code{results}) and blank
  analyses / internal standard removal (\emph{i.e.} \code{removeBlanks=TRUE} / \code{removeISTDs=TRUE}).

  }

  If another filtering order is desired then \code{filter} should be called multiple times with only one filter
  argument at a time.
}

\seealso{
\code{\link{featureGroups-class}} and \code{\link{groupFeatures}}
}
