% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/feature_groups-plot.R
\name{feature-plotting}
\alias{feature-plotting}
\alias{plot,featureGroups,missing-method}
\alias{plotInt,featureGroups-method}
\alias{plotChord,featureGroups-method}
\alias{plotChroms,featureGroups-method}
\alias{plotChroms3D,featureGroups-method}
\alias{plotChroms3D,featureGroupsSet-method}
\alias{plotMobilograms,featureGroups-method}
\alias{plotMobilograms}
\alias{plotVenn,featureGroups-method}
\alias{plotUpSet,featureGroups-method}
\alias{plotVolcano,featureGroups-method}
\alias{plotGraph,featureGroups-method}
\alias{plotGraph,featureGroupsSet-method}
\alias{plotTICs,featureGroups-method}
\alias{plotBPCs,featureGroups-method}
\title{Plotting of grouped features}
\usage{
plotMobilograms(obj, ...)

\S4method{plot}{featureGroups,missing}(
  x,
  groupBy = NULL,
  onlyUnique = FALSE,
  retMin = FALSE,
  showLegend = TRUE,
  IMS = "maybe",
  col = NULL,
  pch = NULL,
  ...
)

\S4method{plotInt}{featureGroups}(
  obj,
  average = FALSE,
  averageFunc = mean,
  areas = FALSE,
  normalized = FALSE,
  xBy = NULL,
  xNames = TRUE,
  groupBy = "fGroups",
  regression = FALSE,
  showLegend = FALSE,
  IMS = "maybe",
  pch = 20,
  type = if (regression) "p" else "b",
  lty = 3,
  xlim = NULL,
  ylim = NULL,
  col = NULL,
  plotArgs = NULL,
  linesArgs = NULL
)

\S4method{plotChord}{featureGroups}(
  obj,
  addSelfLinks = FALSE,
  addRetMzPlots = TRUE,
  aggregate = FALSE,
  groupBy = NULL,
  addIntraOuterGroupLinks = FALSE,
  IMS = "maybe",
  ...
)

\S4method{plotChroms}{featureGroups}(
  obj,
  analysis = analyses(obj),
  groupName = names(obj),
  retMin = FALSE,
  showPeakArea = FALSE,
  showFGroupRect = TRUE,
  title = NULL,
  groupBy = NULL,
  showLegend = TRUE,
  annotate = c("none", "ret", "mz", "mob"),
  intMax = "eic",
  EICParams = getDefEICParams(),
  showProgress = FALSE,
  IMS = "maybe",
  xlim = NULL,
  ylim = NULL,
  EICs = NULL,
  ...
)

\S4method{plotChroms3D}{featureGroups}(
  obj,
  analysis = analyses(obj),
  groupName = names(obj),
  dim3 = "mz",
  retMin = FALSE,
  showLimits = TRUE,
  rtWindow = defaultLim("retention", "medium"),
  mzWindow = defaultLim("mz", "medium"),
  mobWindow = defaultLim("mobility", "medium"),
  gridSize = 50,
  title = NULL,
  ...
)

\S4method{plotChroms3D}{featureGroupsSet}(obj, analysis = analyses(obj), ...)

\S4method{plotMobilograms}{featureGroups}(
  obj,
  analysis = analyses(obj),
  groupName = names(obj),
  showPeakArea = FALSE,
  showFGroupRect = TRUE,
  title = NULL,
  groupBy = NULL,
  showLegend = TRUE,
  annotate = c("none", "ret", "mz", "mob"),
  EIMParams = getDefEIMParams(),
  showProgress = FALSE,
  xlim = NULL,
  ylim = NULL,
  EIMs = NULL,
  ...
)

\S4method{plotVenn}{featureGroups}(obj, which = NULL, aggregate = TRUE, IMS = "maybe", ...)

\S4method{plotUpSet}{featureGroups}(
  obj,
  which = NULL,
  aggregate = TRUE,
  IMS = "maybe",
  nsets = NULL,
  nintersects = NA,
  ...
)

\S4method{plotVolcano}{featureGroups}(
  obj,
  FCParams,
  showLegend = TRUE,
  averageFunc = mean,
  normalized = FALSE,
  IMS = "maybe",
  col = NULL,
  pch = 19,
  ...
)

\S4method{plotGraph}{featureGroups}(obj, onlyPresent = TRUE, width = NULL, height = NULL)

\S4method{plotGraph}{featureGroupsSet}(obj, onlyPresent = TRUE, set, ...)

\S4method{plotTICs}{featureGroups}(
  obj,
  retentionRange = NULL,
  MSLevel = 1,
  retMin = FALSE,
  title = NULL,
  groupBy = NULL,
  showLegend = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)

\S4method{plotBPCs}{featureGroups}(
  obj,
  retentionRange = NULL,
  MSLevel = 1,
  retMin = FALSE,
  title = NULL,
  groupBy = NULL,
  showLegend = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{obj, x}{\code{featureGroups} object to be used for plotting.}

\item{\dots}{passed to \code{\link[base]{plot}} (\code{plot}, \code{plotChroms}, \code{plotMobilograms},
\code{plotTICs} and \code{plotBPCs}), \code{\link{filled.contour}} (\code{plotChroms3D}), \CRANpkg{VennDiagram}
plotting functions (\code{plotVenn}), \code{\link{chordDiagram}} (\code{plotChord}), \code{\link[UpSetR]{upset}}
(\code{plotUpSet}).}

\item{groupBy}{Specifies how results are grouped in the plot. Should be a name of a column in the
  \link[=analysis-information]{analysis information} table which is used to make analysis groups (\emph{e.g.}
  \code{"replicate"}). Set to \code{NULL} for no grouping. The \code{groupBy} argument can also be set to
  \code{"fGroups"} to group by feature groups (except \code{plotChord}).

  For \code{plotInt}: see the \verb{Data aggregation in intensity plots} section for more details.

  For \code{plotChord}: the grouping is used to generate 'outer groups'.}

\item{onlyUnique}{If \code{TRUE} and \code{groupBy} is set to a column name of the
\link[=analysis-information]{analysis information} then only feature groups that are unique to a replicate are
plotted.}

\item{retMin}{Plot retention time in minutes (instead of seconds).}

\item{showLegend}{Plot a legend if \code{TRUE}.}

\item{IMS}{\IMSWF Specifies which feature groups are considered for plotting in IMS workflows. The following options
  are valid: \itemize{

  \item \code{"both"}: Selects IMS and non-IMS features.

  \item \code{"maybe"}: Selects non-IMS features and IMS features without assigned IMS precursor.

  \item \code{FALSE}: Selects only non-IMS features.

  \item \code{TRUE}: Selects only IMS features.

  }

  For \code{plotChroms}: if the \code{groupName} argument is set to a subset of the feature groups, then the \code{IMS} argument is forced to \code{"both"} to ensure that the selected feature groups are plotted.}

\item{col}{Colour(s) used. If \code{col=NULL} then colours are automatically generated.}

\item{pch, type, lty}{Common plotting parameters passed to \emph{e.g.} \code{\link[graphics]{plot}}.

  For \code{plot}: if \code{pch=NULL} then values are automatically assigned.

  for \code{plotInt}: the \code{type} argument defaults to single points (\code{"p"}) when \code{regression=TRUE},
  and lines+points (\code{"b"}) otherwise. For non-aggregated data, it probably makes more sense to set \emph{e.g.}
  \code{type="p"} to avoid lines being drawn to points with equal x-values.}

\item{average}{Controls plot data averaging: see the \verb{Data aggregation in intensity plots} section.}

\item{averageFunc, normalized}{Used for intensity data treatment, see the documentation for the
\code{\link[=as.data.table,featureGroups-method]{as.data.table method}}.}

\item{areas}{Set to \code{TRUE} to use feature areas instead of peak intensities for plotting.}

\item{xBy}{Controls x-value grouping in the plot: see the \verb{Data aggregation in intensity plots} section.}

\item{xNames}{Plot names (\code{xNames=TRUE}) or a number sequence (\code{xNames=FALSE}) on the x axis.}

\item{regression}{If \code{TRUE} then a regression line is plotted for each group, using the x values set by
\code{xBy} (see \verb{Data aggregation in intensity plots}). if \code{showLegend=TRUE} then the R-squared value is
show in the legend for each group (unless there are multiple feature groups and \code{groupBy != "fGroups"}).}

\item{xlim, ylim}{Sets the plot size limits used by
\code{\link[graphics]{plot}}. Set to \code{NULL} for automatic plot sizing.}

\item{plotArgs, linesArgs}{A \code{list} with further arguments passed to \code{\link[base]{plot}} and
\code{\link[graphics]{lines}}, respectively.}

\item{addSelfLinks}{If \code{TRUE} then 'self-links' are added which
represent non-shared data.}

\item{addRetMzPlots}{Set to \code{TRUE} to enable \emph{m/z} \emph{vs}
retention time scatter plots.}

\item{aggregate}{Specifies how data should be aggregated prior to comparison. Set to \code{FALSE} to compare
analyses, \code{TRUE} to compare replicates or to the name of a column in the \link[=analysis-information]{analysis
information} to compare by a custom grouping of analyses.}

\item{addIntraOuterGroupLinks}{If \code{TRUE} then links will be added within
outer groups.}

\item{analysis, groupName}{\code{character} vector with the analyses/group names to be considered for plotting.

  For \code{plotChroms} and \code{plotMobilograms}: Compared to subsetting the \code{featureGroups} object
  (\code{obj}) upfront this is slightly faster. Furthermore, if \code{onlyPresent=FALSE} in \code{EICParams} or
  \code{EIMParams}, this allows plotting chromatograms for feature groups where none of the specified analyses
  contain the feature (which is impossible otherwise since subsetting leads to removal of 'empty' feature groups).
  For IMS workflows: if \code{IMS!="both"} then the \code{analysis} and \code{groupName} arguments are adjusted for
  the remaining data after IMS selection.

  For \code{plotChroms3D}: a single analysis and group name should be specified (unless \code{obj} contains only a
  single feature, \emph{e.g.} created as \code{fGroups[1, 1]}).}

\item{showPeakArea}{Set to \code{TRUE} to display integrated peak ranges by filling (shading) their areas.}

\item{showFGroupRect}{Set to \code{TRUE} to mark the full integration/intensity range of all features within a
feature group by drawing a rectangle around it.}

\item{title}{Character string used for title of the plot. If \code{NULL} a title will be automatically generated.}

\item{annotate}{Set to \code{"ret"}, \code{"mz"} and/or \code{"mob"} to annotate peaks with the retention time,
\emph{m/z} and/or ion mobility (if available) feature group values, respectively.}

\item{intMax}{Method used to determine the maximum intensity plot limit. Should be \code{"eic"} (from EIC data) or
\code{"feature"} (from feature data). Ignored if the \code{ylim} parameter is specified.}

\item{EICParams}{A named \code{list} with parameters used for extracted ion chromatogram (EIC) creation. See the
\link[=EIXParams]{EIC parameters} documentation for more details.}

\item{showProgress}{if set to \code{TRUE} then a text progressbar will be displayed when all EICs are being plot. Set
to \code{"none"} to disable any annotation.}

\item{EICs, EIMs}{Internal parameter for now and should be kept at \code{NULL} (default).}

\item{dim3}{The third dimension to plot besides retention time and intensity. Can be either \code{"mz"} or
\code{"mobility"}.}

\item{showLimits}{If \code{TRUE}, a rectangle is drawn to indicate the feature limits.}

\item{rtWindow, mzWindow, mobWindow}{Numeric values specifying the window size around the feature for retention time,
m/z, and ion mobility respectively. Values \code{>0} will effectively zoom out.}

\item{gridSize}{The size of the grid for interpolation.}

\item{EIMParams}{A named \code{list} with parameters used for extracted ion mobilogram (EIM) creation. See the
\link[=EIXParams]{EIM parameters} documentation for more details.}

\item{which}{A character vector with the selection to compare (\emph{e.g.} replicates, as set by the \code{aggregate}
argument). Set to \code{NULL} to select everything.}

\item{nsets, nintersects}{See \code{\link[UpSetR]{upset}}. If \code{nsets == NULL} then it will be set to the number
of compared items.}

\item{FCParams}{A parameter list to calculate Fold change data. See \code{getFCParams} for more details.}

\item{onlyPresent}{Only plot feature groups of internal standards that are still present in the \code{featureGroups}
input object (which may be otherwise be removed by \emph{e.g.} subsetting or
\code{\link[=filter,featureGroups-method]{filter}}).}

\item{width, height}{Passed to \code{\link{visNetwork}}.}

\item{set}{\setsWF The set for which data must be plotted.}

\item{retentionRange}{Range of retention time (in seconds) to collect TIC traces. Should be a numeric vector with
length of two containing the min/max values. Set to NULL to ignore.}

\item{MSLevel}{Integer vector with the ms levels (i.e., 1 for MS1 and 2 for MS2) to obtain TIC traces.}
}
\value{
\code{plotChroms3D} returns the interpolated grid used for plotting (generated with
  \link[MBA:mba.surf]{mba.surf}).

\code{plotVenn} (invisibly) returns a list with the following fields: \itemize{
\item \code{gList} the \code{gList} object that was returned by
  the utilized \CRANpkg{VennDiagram} plotting function.
\item \code{areas} The total area for each plotted group.
\item \code{intersectionCounts} The number of intersections between groups.
}

The order for the \code{areas} and \code{intersectionCounts} fields is the same as the parameter order
from the used plotting function (see \emph{e.g.} \code{\link{draw.pairwise.venn}} and
\code{\link{draw.triple.venn}}).

\code{plotGraph} returns the result of \code{\link{visNetwork}}.
}
\description{
Various plotting functions for feature group data.
}
\details{
\code{plot} Generates an \emph{m/z} \emph{vs} retention time plot for all featue groups. Optionally
  highlights unique/overlapping presence amongst replicates.

\code{plotInt} Generates a line plot with feature intensities.

\code{plotChord} Generates a chord diagram which can be used to
  visualize shared presence of feature groups between analyses or replicate
  groups. In addition, analyses/replicates sharing similar properties
  (\emph{e.g.} location, age, type) may be grouped to enhance visualization
  between these 'outer groups'.

\code{plotChroms} Plots extracted ion chromatograms (EICs) of feature groups.

\code{plotChroms3D} generates a 3D plot of chromatographic data for a single feature group in a single
  analysis. The plot shows retention time on the x-axis, m/z or mobility on the y-axis, and intensity as color in a
  contour plot. The plot is made with the \code{\link{filled.contour}} function.

\code{plotMobilograms} Plots extracted ion mobilograms (EIMs) of feature groups.

\code{plotVenn} plots a Venn diagram (using \CRANpkg{VennDiagram}) outlining unique and shared feature
  groups between up to five replicates.

\code{plotUpSet} plots an UpSet diagram (using the \code{\link[UpSetR]{upset}} function) outlining unique
  and shared feature groups between given replicates.

\code{plotVolcano} Plots Fold change data in a 'Volcano plot'.

\code{plotGraph} generates an interactive network plot which is used to explore internal standard (IS)
  assignments to each feature group. This requires the availability of IS assignments, see the documentation for
  \code{\link{normInts}} for details. The graph is rendered with \pkg{\link{visNetwork}}.

\code{plotTICs} Plots the total ion chromatogram/s (TICs) of the analyses.

\code{plotTICs} Plots the base peak chromatogram/s (BPCs) of the analyses.
}
\section{Use of raw HRMS data}{
 The \link[=msdata]{raw data interface} of \pkg{patRoon} is used by \code{plotChroms}, \code{plotMobilograms}, \code{plotTICs} and \code{plotBPCs} to
  process HRMS (or IMS-HRMS) data. Please see \link[=msdata]{its documentation} for more information on the supported
  formats and available configuration options.
}

\section{Sets workflows}{
 \setsWFChangedMethods{

  \item \code{plotGraph} only plots data per set, and requires the \code{set} argument to be set.

  }

  In sets workflows the \link[=analysis-information]{analysis information} contains an additional \code{"set"}
  column, which can be used for arguments that involve grouping of analyses. For instance, if \code{groupBy="set"}
  then plotting data is grouped per set.
}

\section{Data aggregation in intensity plots}{
 the \code{average}, \code{xBy} and \code{groupBy} arguments control how
  data is aggregated in intensity plots: \itemize{

  \item \code{average}: controls the averaging of feature intensities prior to plotting.

  \item \code{xBy}: can map the x value of individual points to analysis metadata. For example, exposure time or
  sample location. Non-numeric values are allowed (unless \code{regression=TRUE}).

  \item \code{groupBy}: controls the grouping of points in the plot. Equal groups are plotted in sequence so they can
  be connected with lines and are coloured equally. Examples include experiment type or feature groups.

  }

  The following values are valid: \itemize{

  \item \code{FALSE} (\code{average}) or \code{NULL} (\code{xBy} and \code{groupBy}): aggregation is disabled.

  \item \code{TRUE} (only \code{average}): results are averaged for each replicate

  \item a name of a column in the \link[=analysis-information]{analysis information}: results are aggregated for
  analyses with the same table column value.

  \item \code{"fGroups"} (only \code{groupBy}): plots are grouped by feature groups.

  }
}

\references{
\addCitations{circlize}

\insertRef{Conway2017}{patRoon} \cr\cr \insertRef{Lex2014}{patRoon}
}
\seealso{
\code{\link{featureGroups-class}}, \code{\link{groupFeatures}}
}
\author{
Rick Helmus <\email{r.helmus@uva.nl}> and Ricardo Cunha <\email{cunha@iuta.de}> (\code{plotTICs} and
  \code{plotBPCs} functions)
}
