% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-ADT.R
\name{feature-table}
\alias{feature-table}
\alias{as.data.table,featureGroups-method}
\alias{as.data.table,featureGroupsScreening-method}
\alias{as.data.table,featureGroupsScreeningSet-method}
\title{Feature group to table conversion}
\usage{
\S4method{as.data.table}{featureGroups}(
  x,
  average = FALSE,
  areas = FALSE,
  features = FALSE,
  qualities = FALSE,
  regression = FALSE,
  regressionBy = NULL,
  averageFunc = mean,
  normalized = FALSE,
  FCParams = NULL,
  concAggrParams = getDefPredAggrParams(),
  toxAggrParams = getDefPredAggrParams(),
  normConcToTox = FALSE,
  anaInfoCols = NULL,
  IMS = "both"
)

\S4method{as.data.table}{featureGroupsScreening}(x, ..., collapseSuspects = ",", onlyHits = FALSE)

\S4method{as.data.table}{featureGroupsScreeningSet}(x, ..., collapseSuspects = ",", onlyHits = FALSE)
}
\arguments{
\item{x}{The \code{featureGroups} like object to be exported as table.}

\item{average}{Controls the averaging of feature intensities. Averaging also influences the calculation of regression
  parameters. Set to \code{FALSE} to disable averaging, \code{TRUE} to average per replicate or to the name of a
  column in the \link[=analysis-information]{analysis information} to compare by a custom grouping of analyses.

  If \code{features=TRUE} all numerical properties are averaged and non-numericals are collapsed. Each row represents
  the feature aggregated data and the groups used for aggregation (\emph{e.g.} replicates) are stored in the
  \code{average_group} column. It is also possible to average these data per feature group, by setting
  \code{average="fGroups"}.}

\item{areas}{If set to \code{TRUE} then areas are output instead of peak intensities. Ignored if
\code{features=TRUE}, as areas of features are always reported.}

\item{features}{If \code{TRUE} then feature specific data will be added. Also see the \code{average} argument.}

\item{qualities}{Adds feature (group) qualities (\code{qualities="quality"}), scores (\code{qualities="score"}) or
both (\code{qualities="both"}), if this data is available (\emph{i.e.} from \code{calculatePeakQualities}). If
\code{qualities=FALSE} then nothing is reported.}

\item{regression, regressionBy}{Used for regression calculations. See the \verb{Regression calculation} section below.
Set to \code{NULL} to ignore.}

\item{averageFunc}{Function used for averaging. Only used when data is averaged or \code{FCParams != NULL}.}

\item{normalized}{If \code{TRUE} then normalized intensities are used (see the \verb{Feature intensity normalization}
section). If no normalization data is available (\emph{e.g.} because \code{normInts} was not used) then an
automatic group normalization is performed.}

\item{FCParams}{A parameter list to calculate fold change data. See \code{getFCParams} for more details. Set to
\code{NULL} to not perform FC calculations. \strong{NOTE}: the intensity data is always averaged to calculate fold
changes (using \code{averageFunc}) by replicates, irrespective of the \code{average} argument.}

\item{concAggrParams, toxAggrParams}{Parameters to aggregate calculated concentrations/toxicities (obtained with
\code{\link{calculateConcs}}/\code{\link{calculateTox}}). See \link[=pred-aggr-params]{prediction aggregation
parameters} for more information. Set to \code{NULL} to omit this data.}

\item{normConcToTox}{Set to \code{TRUE} to normalize concentrations to toxicities. Only relevant if this data is
present (see \code{\link{calculateConcs}}/\code{\link{calculateTox}}).}

\item{anaInfoCols}{A \code{character} with any additional columns from the \link[=analysis-information]{analysis
information} table. Only supported if \code{features=TRUE}. If averaging is performed then the data in the
specified columns should be numeric. Set to \code{NULL} to ignore.}

\item{IMS}{\IMSWF Specifies which feature groups are considered to be returned in IMS workflows. The following options
  are valid: \itemize{

  \item \code{"both"}: Selects IMS and non-IMS features.

  \item \code{"maybe"}: Selects non-IMS features and IMS features without assigned IMS parent.

  \item \code{FALSE}: Selects only non-IMS features.

  \item \code{TRUE}: Selects only IMS features.

  }}

\item{\dots}{Passed to the parent \code{as.data.table} method.}

\item{collapseSuspects}{If a \code{character} then any suspects that were matched to the same feature group are
collapsed to a single row and suspect names are separated by the value of \code{collapseSuspects}. If \code{NULL}
then no collapsing occurs, and each suspect match is reported on a single row. See the \verb{Suspect collapsing}
section below for additional details.}

\item{onlyHits}{If \code{TRUE} then only feature groups with suspect hits are reported.}
}
\description{
Convert feature group data to a \code{\link{data.table}} (or \code{data.frame}).
}
\details{
The \code{as.data.table} generic function converts most feature group data to a highly customizable
\code{\link{data.table}}. If a classical \code{data.frame} is preferred, the \code{as.data.frame} generic function
can be used instead and accepts the exact same arguments. The methods defined for \link[=suspect-screening]{suspect
screening workflows} will merge the information from \code{screenInfo}, such as
suspect names and other properties and annotation data.
}
\section{Regression calculation}{
 The \code{regression} argument controls the calculation of regression parameters
  from a regression model calculated with feature intensities (or areas if \code{areas=TRUE}). Here, simple linear
  regression is used, \emph{i.e.} \samp{y=ax+b} with \samp{a} the slope and \samp{b} the intercept. The value for
  \code{regression} should be the name of a column in the \link[=analysis-information]{analysis information} table
  with numerical data to be used for x-values. Alternatively, if \code{regression=TRUE} then the \code{"conc"} column
  is used. Any \code{NA} x-values are ignored, and no regression will be calculated if less than two (non-NA)
  x-values are available. The output table will contain properties such as the slope and correlation coefficient
  (R-squared). Furthermore, if \code{features=TRUE} then x-values will be calculated from the model and stored in the
  \code{x_reg} column.

  The \code{regressionBy} argument can be used to construct separate regression models for different groups of
  analysis. It should be set to the name of a column in the \link[=analysis-information]{analysis information} table
  which defines the grouping between samples. If \code{features=TRUE} then the grouping is stored in the
  \code{regression_group} column of the output table.

  Please see the handbook for examples on how to use the regression functionality.
}

\section{{Suspect collapsing}}{
 The \code{as.data.table} method for \code{featureGroupsScreening} supports an
  additional format where each suspect hit is reported on a separate row (enabled by setting
  \code{collapseSuspects=NULL}). In this format the suspect
  properties from the \code{screenInfo} method are merged with each suspect row. Alternatively, if \emph{suspect
  collapsing} is enabled (the default) then the regular \code{as.data.table} format is used, and amended with the
  names and estimated ID levels (if available) of the suspects matched to a feature group (each separated by the
  value of the \code{collapseSuspects} argument).

  Suspect collapsing also influences the reporting of predicted \link[=pred-quant]{feature concentrations} and
  \link[=pred-tox]{toxicities}. In the case that (1) suspects are \emph{not} collapsed in the output table and (2)
  predictions are available for a specific suspect hit (\emph{i.e.} if \code{\link{predictRespFactors}} or
  \code{\link{predictTox}} was called on the feature groups object), then only the suspect specific data is reported
  and no aggregation is performed. Hence, this allows you to obtain specific concentration/toxicity values for each
  suspect/feature group pair.
}

\section{IMS workflows}{
 If the \code{IMS} argument is set to \code{"both"} or \code{"maybe"} then
  \code{"mobility_collapsed"} and \code{"CCS_collapsed"} columns will be added that summarize all
  mobility/\acronym{CCS} values of the mobility features (or feature groups) assigned to this IMS parent. These
  numbers are currently rounded to \samp{3} decimals.
}

\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} normalization of feature intensities occur per
  set.

  In sets workflows the \link[=analysis-information]{analysis information} contains an additional \code{"set"}
  column, which can be used for arguments that involve grouping of analyses. For instance, if
  \code{regressionBy="set"} then regression models will be calculated for each set.
}

