% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_annotations.R
\docType{class}
\name{featureAnnotations-class}
\alias{featureAnnotations-class}
\alias{featureAnnotations}
\alias{delete,featureAnnotations-method}
\title{Base feature annotations class}
\usage{
\S4method{delete}{featureAnnotations}(obj, i = NULL, j = NULL, ...)
}
\arguments{
\item{obj, x, object}{\code{featureAnnotations} object to be accessed}

\item{i, j}{For \code{[}/\code{[[}: A numeric or character value which is used to select feature groups by
their index or name, respectively (for the order/names see \code{groupNames()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all feature groups are selected.\cr\cr For \code{[[}: should be a scalar value.\cr\cr For \code{delete}: The data to remove from. \code{i} are the
feature groups as numeric index, logical or character, \code{j} the candidates as numeric indices (rows). If either is
\code{NULL} then data for all is removed. \code{j} may also be a function: it will be called for each 
feature group, with the annotation table (a \code{data.table}) as first argument, the feature group name as second argument, and any other arguments passed as
\code{\dots} to \code{delete}. The return value of this function specifies the candidate indices (rows) to be removed (specified as an \code{integer} or \code{logical} vector).}

\item{\dots}{Ignored.}

\item{drop, j}{ignored.}

\item{name}{The feature group name (partially matched).}
}
\value{
\code{delete} returns the object for which the specified data was removed.
}
\description{
Holds information for all feature (group) annotations.
}
\details{
This class provides a way to store intensity, retention times, \emph{m/z} and
other data for all features in a set of analyses. The class is \code{virtual}
and derived objects are created by 'feature finders' such as
\code{findFeaturesOpenMS}, \code{findFeaturesXCMS} and
\code{findFeaturesBruker}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{delete}: Completely deletes specified annotations.
}}

\section{Slots}{

\describe{
\item{\code{features}}{List of features per analysis file. Use the
\code{featureTable} method for access.}

\item{\code{analysisInfo}}{Analysis group information. Use the \code{analysisInfo} method
for access.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{workflowStep}}}   \itemize{     \item{\strong{\code{\link{features}}}}     \itemize{       \item{\code{\link{featuresSet}}}       \item{\code{\link{featuresUnset}}}       \item{\code{\link{featuresFromFeatGroups}}}       \item{\code{\link{featuresConsensus}}}       \item{\code{\link{featuresBruker}}}       \item{\code{\link{featuresEnviPick}}}       \item{\code{\link{featuresKPIC2}}}       \item{\code{\link{featuresOpenMS}}}       \item{\code{\link{featuresSAFD}}}       \item{\code{\link{featuresSIRIUS}}}       \item{\code{\link{featuresBrukerTASQ}}}       \item{\code{\link{featuresXCMS}}}       \item{\code{\link{featuresXCMS3}}}     }   } }
}

\seealso{
\code{\link{feature-finding}}
}
