% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/feature_groups-screening.R,
%   R/feature_groups-screening-set.R
\docType{class}
\name{featureGroupsScreening-class}
\alias{featureGroupsScreening-class}
\alias{featureGroupsScreening}
\alias{screenInfo,featureGroupsScreening-method}
\alias{screenInfo}
\alias{show,featureGroupsScreening-method}
\alias{[,featureGroupsScreening,ANY,ANY,missing-method}
\alias{delete,featureGroupsScreening-method}
\alias{filter,featureGroupsScreening-method}
\alias{featureGroupsScreeningSet-class}
\alias{featureGroupsScreeningSet}
\alias{screenInfo,featureGroupsScreeningSet-method}
\alias{show,featureGroupsScreeningSet-method}
\alias{[,featureGroupsScreeningSet,ANY,ANY,missing-method}
\alias{delete,featureGroupsScreeningSet-method}
\alias{filter,featureGroupsScreeningSet-method}
\alias{featureGroupsSetScreeningUnset-class}
\alias{featureGroupsSetScreeningUnset}
\alias{unset,featureGroupsScreeningSet-method}
\title{Class for suspect screened feature groups.}
\usage{
screenInfo(obj)

\S4method{screenInfo}{featureGroupsScreening}(obj)

\S4method{show}{featureGroupsScreening}(object)

\S4method{[}{featureGroupsScreening,ANY,ANY,missing}(x, i, j, ..., suspects = NULL, reorder = FALSE, drop = TRUE)

\S4method{delete}{featureGroupsScreening}(obj, i = NULL, j = NULL, k = NULL, ...)

\S4method{filter}{featureGroupsScreening}(
  obj,
  ...,
  onlyHits = NULL,
  IMSMatchParams = NULL,
  selectHitsBy = NULL,
  selectBestFGroups = FALSE,
  maxLevel = NULL,
  maxFormRank = NULL,
  maxCompRank = NULL,
  minAnnSimForm = NULL,
  minAnnSimComp = NULL,
  minAnnSimBoth = NULL,
  absMinFragMatches = NULL,
  relMinFragMatches = NULL,
  minRF = NULL,
  maxLC50 = NULL,
  negate = FALSE,
  applyIMS = "both"
)

\S4method{screenInfo}{featureGroupsScreeningSet}(obj)

\S4method{show}{featureGroupsScreeningSet}(object)

\S4method{[}{featureGroupsScreeningSet,ANY,ANY,missing}(x, i, j, ..., suspects = NULL, sets = NULL, reorder = FALSE, drop = TRUE)

\S4method{delete}{featureGroupsScreeningSet}(obj, i = NULL, j = NULL, k = NULL, ...)

\S4method{filter}{featureGroupsScreeningSet}(
  obj,
  ...,
  onlyHits = NULL,
  IMSMatchParams = NULL,
  selectHitsBy = NULL,
  selectBestFGroups = FALSE,
  maxLevel = NULL,
  maxFormRank = NULL,
  maxCompRank = NULL,
  minAnnSimForm = NULL,
  minAnnSimComp = NULL,
  minAnnSimBoth = NULL,
  absMinFragMatches = NULL,
  relMinFragMatches = NULL,
  minRF = NULL,
  maxLC50 = NULL,
  negate = FALSE,
  applyIMS = "both"
)

\S4method{unset}{featureGroupsScreeningSet}(obj, set)
}
\arguments{
\item{obj, object, x}{The \code{featureGroupsScreening} object.}

\item{i, j, reorder}{See \code{\link{featureGroups}}.}

\item{\dots}{Further arguments passed to the base method.}

\item{suspects}{An optional \code{character} vector with suspect names. If
specified, only \code{featureGroups} will be kept that are assigned to
these suspects.}

\item{drop}{Ignored.}

\item{k}{The \code{k} argument is used to delete screening results (instead of features) and should be: \itemize{

  \item a \code{character} vector with suspect names that should be removed

  \item a \code{function} that is called with the screening info table and should return a \code{logical} vector for
  each suspect row to be removed

  \item \code{NA} to remove all screening results, which is especially useful when paired with the \code{j} argument,
  \emph{i.e.} to remove all screening results for a particular set of feature groups.

  \item \code{NULL} to not touch screening results and only perform deletion as the \code{\link{featureGroups}}
  method.

  }

  Setting both \code{i} and \code{k} is currently not supported.}

\item{onlyHits}{If \itemize{

  \item \code{negate=FALSE} and \code{onlyHits=TRUE} then all feature groups without suspect hits will be removed.
  Otherwise nothing will be done.

  \item \code{negate=TRUE} then \code{onlyHits=TRUE} will select feature groups without suspect hits,
  \code{onlyHits=FALSE} will only retain feature groups with suspect matches and this filter is ignored if
  \code{onlyHits=NULL}.

  }}

\item{IMSMatchParams}{\IMSWF A \code{list} with parameters to be used for matching IMS data. See
\code{\link{getIMSMatchParams}} for details and how to make such a parameter list.}

\item{selectHitsBy}{Should be \code{"intensity"} or \code{"level"}. For cases where the same suspect is matched to
multiple feature groups, only the suspect to the feature group with highest mean intensity
(\code{selectHitsBy="intensity"}) or best identification level (\code{selectHitsBy="level"}) is kept. In case of
ties only the first hit is kept. Set to \code{NULL} to ignore this filter. If \code{negate=TRUE} then only those
hits with lowest mean intensity/poorest identification level are kept.}

\item{selectBestFGroups}{If \code{TRUE} then for any cases where a single feature group is matched to several
suspects only the suspect assigned to the feature group with best identification score is kept. In case of ties
only the first is kept.}

\item{maxLevel, maxFormRank, maxCompRank, minAnnSimForm, minAnnSimComp, minAnnSimBoth}{Filter suspects by maximum
identification level (\emph{e.g.} \code{"3a"}), formula/compound rank or with minimum formula/compound/combined
annotation similarity. Set to \code{NULL} to ignore.}

\item{absMinFragMatches, relMinFragMatches}{Only retain suspects with this minimum number MS/MS matches with the
fragments specified in the suspect list (\emph{i.e.} \code{fragments_mz}/\code{fragments_formula}).
\code{relMinFragMatches} sets the minimum that is relative (\samp{0-1}) to the maximum number of MS/MS fragments
specified in the \code{fragments_*} columns of the suspect list. Set to \code{NULL} to ignore.}

\item{minRF}{Filter suspect hits by the given minimum predicted response factor (as calculated by
\code{\link[=predictRespFactors]{predictRespFactors}}). Set to \code{NULL} to ignore.}

\item{maxLC50}{Filter suspect hits by the given maximum toxicity (LC50) (as calculated by
\code{\link[=predictTox]{predictTox}}). Set to \code{NULL} to ignore.}

\item{negate}{If set to \code{TRUE} then filtering operations are performed in opposite manner.}

\item{applyIMS}{\IMSWF whether the filters are only applied to IMS precursors (\code{applyIMS=FALSE}), only to IMS
features (\code{applyIMS=TRUE}) or to both (\code{applyIMS="both"}). Other feature groups will always be kept. The
\code{negate} option does not affect \code{applyIMS}.}

\item{sets}{\setsWF A \code{character} with name(s) of the sets to keep (or remove if \code{negate=TRUE}).}

\item{set}{\setsWF The name of the set.}
}
\value{
\code{delete} returns the object for which the specified data was removed.

\code{filter} returns a filtered \code{featureGroupsScreening} object.
}
\description{
This class derives from \code{\link{featureGroups}} and adds suspect screening information.
}
\section{Methods (by generic)}{
\itemize{
\item \code{screenInfo(featureGroupsScreening)}: Returns a table with screening information
(see \code{screenInfo} slot).

\item \code{show(featureGroupsScreening)}: Shows summary information for this object.

\item \code{x[i}: Subset on analyses, feature groups and/or
suspects.

\item \code{delete(featureGroupsScreening)}: Completely deletes specified feature groups or screening results.

\item \code{filter(featureGroupsScreening)}: Performs rule based filtering. This method builds on the comprehensive filter
functionality from the base \code{\link{filter,featureGroups-method}}. It adds several filters to select
\emph{e.g.} the best ranked suspects or those with a minimum estimated identification level. \strong{NOTE}: most
filters \emph{only} affect suspect hits, not feature groups. Set \code{onlyHits=TRUE} to subsequently remove any
feature groups that lost any suspect matches due to these filter steps.

}}
\section{Slots}{

\describe{
\item{\code{screenInfo}}{A (\code{\link{data.table}}) with results from suspect screening. This table will be amended with
ID confidence data when \code{\link{estimateIDConfidence}} is run.}

\item{\code{MS2QuantMeta}}{Metadata from \pkg{MS2Quant} filled in by \code{predictRespFactors}.

\setsWF A named \code{list} with the metadata stored for each set.}
}}

\note{
\code{filter} removes suspect hits with \code{NA} values when any of the filters related to minimum or maximum
  values are applied (unless \code{negate=TRUE}).
}
\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{featureGroups}}}   \itemize{     \item{\strong{\code{\link{featureGroupsScreening}}}}     \itemize{       \item{\code{\link{featureGroupsSetScreeningUnset}}}     }   } }
}

\section{Sets workflows}{
 \setsWFClass{featureGroupsScreeningSet}{featureGroupsScreening}

  \setsWFNewMethodsSO{featureGroupsScreeningUnset}{Only the screening results present in the specified set are kept.}

  \setsWFChangedMethods{

  \item \code{estimateIDConfidence} See the \verb{Sets workflows} section in the documentation for
  \code{\link{estimateIDConfidence}}.

  \item \code{filter} All filters related to estimated identification levels and formula/compound rankings  are
  applied to the overall set data (see above). All others are applied to set specific data: in this case candidates
  are only removed if none of the set data confirms to the filter.

  }

  This class derives also from \code{\link{featureGroupsSet}}. Please see its documentation for more relevant details
  with sets workflows.

  Note that the \code{formRank} and \code{compRank} columns are \emph{not} updated when the data is subset.
}

\seealso{
\code{\link{featureGroups}}
}
