% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\docType{class}
\name{features-class}
\alias{features-class}
\alias{features}
\alias{length,features-method}
\alias{show,features-method}
\alias{featureTable,features-method}
\alias{analysisInfo,features-method}
\alias{filter,features-method}
\title{Base features class}
\usage{
\S4method{length}{features}(x)

\S4method{show}{features}(object)

\S4method{featureTable}{features}(obj)

\S4method{analysisInfo}{features}(obj)

\S4method{filter}{features}(obj, intensityThreshold = NULL)
}
\arguments{
\item{obj, x, object}{\code{features} object to be accessed}
}
\value{
\code{featureTable}: A \code{list} containing a
  \code{\link{data.table}} for each analysis with feature data

\code{analysisInfo}: A \code{data.frame} containing a column with
  analysis name (\code{analysis}), its path (\code{path}), and other columns
  such as replicate group name (\code{group}) and blank reference
  (\code{ref}).
}
\description{
Holds information for all features present within a set of analysis.
}
\details{
This class provides a way to store intensity, retention times, \emph{m/z} and
other data for all features in a set of analyses. The class is \code{virtual}
and derived objects are created by 'feature finders' such as
\code{findFeaturesOpenMS}, \code{findFeaturesXCMS} and
\code{findFeaturesBruker}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: Obtain total number of features.

\item \code{show}: Shows summary information for this object.

\item \code{featureTable}: Get table with feature information

\item \code{analysisInfo}: Get analysis information

\item \code{filter}: Performs common rule based filtering on of features,
such as intensity thresholds.
}}

\section{Slots}{

\describe{
\item{\code{features}}{List of features per analysis file. Use the
\code{featureTable} method for access.}

\item{\code{analysisInfo}}{Analysis group information. Use the \code{analysisInfo} method
for access.}
}}

\seealso{
\code{\link{feature-finding}}
}
