% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-bruker.R
\name{findFeaturesBruker}
\alias{findFeaturesBruker}
\title{Find features using Bruker DataAnalysis}
\usage{
findFeaturesBruker(
  analysisInfo,
  doFMF = "auto",
  startRange = 0,
  endRange = 0,
  save = TRUE,
  close = save,
  verbose = TRUE
)
}
\arguments{
\item{analysisInfo}{A \code{data.frame} with \link[=analysis-information]{Analysis information}.}

\item{doFMF}{Run the 'Find Molecular Features' algorithm before loading compounds. Valid options are: \code{"auto"}
(run FMF automatically if current results indicate it is necessary) and \code{"force"} (run FMF \emph{always}, even
if cached results exist). Note that checks done if \code{doFMF="auto"} are fairly simplistic, hence set
\code{doFMF="force"} if feature data needs to be updated.}

\item{startRange, endRange}{Start/End retention range (seconds) from which to collect features. A 0 (zero) for
\code{endRange} marks the end of the analysis.}

\item{close, save}{If \code{TRUE} then Bruker files are closed and saved after
processing with DataAnalysis, respectively. Setting \code{close=TRUE}
prevents that many analyses might be opened simultaneously in DataAnalysis,
which otherwise may use excessive memory or become slow. By default
\code{save} is \code{TRUE} when \code{close} is \code{TRUE}, which is
likely what you want as otherwise any processed data is lost.}

\item{verbose}{If set to \code{FALSE} then no text output is shown.}
}
\value{
An object of a class which is derived from \code{\link{features}}.
}
\description{
Uses the 'Find Molecular Features' (FMF) algorithm of Bruker DataAnalysis vendor software to find features.
}
\details{
This function uses Bruker to automatically find features. This function is called when calling \code{findFeatures} with
  \code{algorithm="bruker"}.

The resulting 'compounds' are transferred from DataAnalysis and stored as features.

  This algorithm only works with Bruker data files (\code{.d} extension) and requires Bruker DataAnalysis
  and the \pkg{RDCOMClient} package to be installed. Furthermore, DataAnalysis combines multiple related masses in a
  feature (\emph{e.g.} isotopes, adducts) but does not report the actual (monoisotopic) mass of the feature.
  Therefore, it is simply assumed that the feature mass equals that of the highest intensity mass peak.
}
\note{
If any errors related to \command{DCOM} appear it might be necessary to
  terminate DataAnalysis (note that DataAnalysis might still be running as a
  background process). The \command{ProcessCleaner} application installed
  with DataAnalayis can be used for this.
}
\seealso{
\code{\link{findFeatures}} for more details and other algorithms.
}
