% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-kpic2.R
\name{findFeaturesKPIC2}
\alias{findFeaturesKPIC2}
\title{Find features using KPIC2}
\usage{
findFeaturesKPIC2(
  analysisInfo,
  kmeans = TRUE,
  level = 1000,
  ...,
  parallel = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{analysisInfo}{A \code{data.frame} with \link[=analysis-information]{Analysis information}.}

\item{kmeans}{If \code{TRUE} then \code{\link[KPIC]{getPIC.kmeans}} is used to obtain PICs, otherwise it is
\code{\link[KPIC]{getPIC}}.}

\item{level}{Passed to \code{\link[KPIC]{getPIC}} or \code{\link[KPIC]{getPIC.kmeans}}}

\item{\dots}{Further parameters passed to \code{\link[KPIC]{getPIC}}/\code{\link[KPIC]{getPIC.kmeans}}}

\item{parallel}{If set to \code{TRUE} then code is executed in parallel through the \CRANpkg{futures} package. Please
see the parallelization section in the handbook for more details.}

\item{verbose}{If set to \code{FALSE} then no text output is shown.}
}
\value{
An object of a class which is derived from \code{\link{features}}.
}
\description{
Uses the \href{https://github.com/hcji/KPIC2}{KPIC2} \R package to extract features.
}
\details{
This function uses KPIC2 to automatically find features. This function is called when calling \code{findFeatures} with
  \code{algorithm="kpic2"}.

The MS files should be in the \code{mzML} or \code{mzXML} format.

The input MS data files need to be centroided. The \code{\link{convertMSFiles}} function can be used to
  centroid data.
}
\references{
\insertRef{Ji2017}{patRoon}
}
\seealso{
\code{\link{findFeatures}} for more details and other algorithms.
}
