% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-qalgorithms.R
\name{findFeaturesQAlgorithms}
\alias{findFeaturesQAlgorithms}
\title{Find features using qAlgorithms}
\usage{
findFeaturesQAlgorithms(analysisInfo, ppm = 5, verbose = FALSE)
}
\arguments{
\item{analysisInfo}{A \code{data.frame} with \link[=analysis-information]{Analysis information}.}

\item{ppm}{The mass accuracy in parts per million (ppm) for the qPeaks when using centroid data. Default is 5 ppm.}

\item{verbose}{If set to \code{FALSE} then no text output is shown.}
}
\value{
An object of a class which is derived from \code{\link{features}}.
}
\description{
uses the \href{https://github.com/odea-project/qAlgorithms}{qAlgorithms} software to find features in LC-MS data.
}
\details{
This function uses qalgorithms to automatically find features. This function is called when calling \code{findFeatures} with
  \code{algorithm="qalgorithms"}.

This functionality has been tested with qAlgorithms version >= 0.1.1. Please make sure it is installed and
  configured, e.g. by configuring the path of the binaries with the \code{patRoon.path.qAlgorithms} option.

  The file format of analyses must be \file{mzML}. Both profile and centroid data are supported. Yet, the profile 
  data is recommended for better results.
}
\seealso{
\code{\link{findFeatures}} for more details and other algorithms.
}
