% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-safd.R
\name{findFeaturesSAFD}
\alias{findFeaturesSAFD}
\title{Find features using SAFD}
\usage{
findFeaturesSAFD(
  analysisInfo,
  prefCentroid = FALSE,
  mzRange = c(0, 400),
  maxNumbIter = 1000,
  maxTPeakW = 300,
  resolution = 30000,
  minMSW = 0.02,
  RThreshold = 0.75,
  minInt = 2000,
  sigIncThreshold = 5,
  S2N = 2,
  minPeakWS = 3,
  centroidMethod = "RFM",
  centroidDM = 0.005,
  verbose = TRUE
)
}
\arguments{
\item{analysisInfo}{A \code{data.frame} (or \code{data.table}) with \link[=analysis-information]{Analysis
information}.}

\item{prefCentroid}{Set to \code{TRUE} to prefer centroided data over other MS data specified in \code{analysisInfo}.

  \strong{NOTE}: if \code{prefCentroid=FALSE} but the package option \option{patRoon.MS.preferIMS=TRUE} (see
  \code{\link{msdata}}), then centroided data will still be preferred over IMS data.}

\item{mzRange}{The \emph{m/z} window to be imported.}

\item{maxNumbIter, maxTPeakW, resolution, minMSW, RThreshold, minInt, sigIncThreshold, S2N, minPeakWS}{Parameters directly
passed to the \code{safd_s3D} function.}

\item{centroidMethod, centroidDM}{Passed to the \code{safd_s3d_cent} function (\code{method} and \code{mdm} arguments,
respectively).}

\item{verbose}{If set to \code{FALSE} then no text output is shown.}
}
\value{
An object of a class which is derived from \code{\link{features}}.
}
\description{
Uses \href{https://bitbucket.org/SSamanipour/safd.jl/src/master/}{SAFD} to obtain features. This functionality is
still experimental. Please see the details below.
}
\details{
This function uses SAFD to automatically find features. This function is called when calling \code{findFeatures} with
  \code{algorithm="safd"}.

The support for SAFD is still experimental, and its interface might change in the future.

  In order to use SAFD, please make sure that its \command{Julia} packages are installed and you have verified that
  everything works, \emph{e.g.} by running the test data with \command{SAFD}.

  As of \pkg{patRoon} \samp{3.0}, \code{findFeaturesSAFD} uses the \link{msdata} interface instead of the
  \pkg{MS_Import.jl} \command{Julia} package to read HRMS data. This means that \pkg{MS_Import.jl} does not need to
  be installed, and all file formats supported by \link{msdata} are also supported for \command{SAFD} feature
  detection. This includes IMS-HRMS data, however, in that case IMS resolved spectra are summed and the IMS dimension
  is removed to make the data compatible for \command{SAFD}.

  The \command{SAFD} algorithm was primarily developed to detect features in profile \emph{m/z} data, but centroided
  data is also supported. To use profile data, ensure that the paths are correctly set up in the
  \link[=analysis-information]{analysisInfo}. Furthermore, when using profile data you probably also need to specify
  centroided data in the \code{analysisInfo}, as \emph{e.g.} \code{\link{generateMSPeakLists}} currently does not
  support profile data. If IMS-HRMS data is used it is treated as profile data, as this data is typically not or
  partially centroided (\code{generateMSPeakLists} supports IMS-HRMS data directly).
}
\section{Parallelization}{
 \code{findFeaturesSAFD} uses multiprocessing to parallelize
  computations. Please see the parallelization section in the handbook for
  more details and \link[=patRoon-package]{patRoon options} for configuration
  options.

 Note that for caching purposes, the analyses files must always exist on the local host
  computer, even if it is not participating in computations.
}

\references{
\insertRef{Samanipour2019}{patRoon}
}
\seealso{
\code{\link{findFeatures}} for more details and other algorithms.
}
