% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-xcms.R
\name{findFeaturesXCMS}
\alias{findFeaturesXCMS}
\title{Find features using XCMS (old interface)}
\usage{
findFeaturesXCMS(analysisInfo, method = "centWave", ..., verbose = TRUE)
}
\arguments{
\item{analysisInfo}{A \code{data.frame} with \link[=analysis-information]{Analysis information}.}

\item{method}{The method setting used by XCMS peak finding, see \code{\link[xcms:findPeaks-methods]{xcms::findPeaks}}}

\item{\dots}{Further parameters passed to \code{\link[xcms]{xcmsSet}}.}

\item{verbose}{If set to \code{FALSE} then no text output is shown.}
}
\value{
An object of a class which is derived from \code{\link{features}}.
}
\description{
Uses the legacy \code{\link[xcms]{xcmsSet}} function from the \pkg{xcms} package to find features.
}
\details{
This function uses XCMS to automatically find features. This function is called when calling \code{findFeatures} with
  \code{algorithm="xcms"}.

This function uses the legacy interface of \pkg{xcms}. It is recommended to use
  \code{\link{findFeaturesXCMS3}} instead.

  The file format of analyses must be \code{mzML} or \code{mzXML}.

The input MS data files need to be centroided. The \code{\link{convertMSFiles}} function can be used to
  centroid data.
}
\references{
\addCitations{xcms}{1} \cr\cr \addCitations{xcms}{2} \cr\cr \addCitations{xcms}{3}
}
\seealso{
\code{\link{findFeatures}} for more details and other algorithms.

\code{\link{findFeaturesXCMS3}}
}
