% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/formulas.R, R/formula-bruker.R,
%   R/formula-genform.R, R/formulas-sirius.R
\docType{methods}
\name{formula-generation}
\alias{formula-generation}
\alias{generateFormulas,featureGroups-method}
\alias{generateFormulas}
\alias{generateFormulasDA}
\alias{generateFormulasGenForm}
\alias{generateFormulasSirius}
\title{Automatic chemical formula generation}
\usage{
\S4method{generateFormulas}{featureGroups}(fGroups, algorithm, ...)

generateFormulasDA(fGroups, precursorMzSearchWindow = 0.002,
  MSMode = "both")

generateFormulasGenForm(fGroups, MSPeakLists, maxMzDev = 5, adduct = "M+H",
  elements = "CHNOP", hetero = TRUE, MSMode = "both", extraOpts = NULL,
  maxProcAmount = getOption("patRoon.maxProcAmount"), maxCmdsPerProc = 25)

generateFormulasSirius(fGroups, MSPeakLists, maxMzDev = 5,
  adduct = "[M+H]+", elements = "CHNOP", profile = "qtof",
  database = NULL, noise = NULL, logPath = file.path("log", "sirius"),
  maxProcAmount = getOption("patRoon.maxProcAmount"))
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which formulae should
be generated. This should be the same or a subset of the object that was
used to create the specified \code{MSPeakLists} (only relevant for
algorithms using \code{MSPeakLists}). In the case of a subset only the
remaining feature groups in the subset are considered.}

\item{algorithm}{A character string describing the algorithm that should be
used: \code{"bruker"}, \code{"genform"}, \code{"sirius"}}

\item{...}{Any parameters to be passed to the selected formula generation
algorithm.}

\item{precursorMzSearchWindow}{Maximum \emph{m/z} deviation (Da) to find back feature data
of precursor/parent ions from MS/MS spectra (this data is not readily
available from \command{SmartFormula3D} results).}

\item{MSMode}{Whether formulae should be generated only from MS data
(\code{"ms"}), MS/MS data (\code{"msms"}) or both (\code{"both"}). Using
the latter option, unique formulae from MS data will still be reported if
not predicted from MS/MS data. Formulae calculated from MS data that were
also generated from MS/MS data will be removed.}

\item{MSPeakLists}{An \code{\link{MSPeakLists}} object that was generated for
the supplied \code{fGroups}.}

\item{maxMzDev}{Maximum deviation between the measured and candidate formula
\emph{m/z} values (in ppm). Sets the \option{ppm} and
\option{--ppm-max} commandline options for \command{GenForm} and
\command{SIRIUS}, respectively.}

\item{adduct}{Sets the adduct form of measured ions. For
\code{generateFormulasGenForm} the format is without brackets and charges
(based on
\href{http://fiehnlab.ucdavis.edu/staff/kind/Metabolomics/MS-Adduct-Calculator}{this
 website}), examples: \code{"M+H"}, \code{"M-H"}, \code{"M+Na"},
\code{"M-e"}. For \code{generateFormulasSirius} the format is with brackets
and charge, for example: \code{"[M+H]+"}, \code{"[M-H]-"},
\code{"[M+Na]+"}, \code{"[M]+"}. Sets the \option{ion} and \option{--ion}
commandline options for \command{GenForm} and \command{SIRIUS},
respectively.}

\item{elements}{Elements to be considered for formulae calculation. This will
heavily affects the number of candidates! Always try to work with a minimal
set by excluding elements you don't expect. For
\code{generateFormulasSirius}, the minimum/maximum number
of elements can also be specified, for example: a value of
\code{"C[5]H[10-15]O"} will only consider formulae with up to five carbon
atoms, between ten and fifteen hydrogen atoms and any amount of oxygen
atoms. Sets the \option{el} and \option{--elements}
commandline options for \command{GenForm} and \command{SIRIUS},
respectively.}

\item{hetero}{Only consider formulae with at least one hetero atom. Sets the
\option{het} commandline option.}

\item{extraOpts}{An optional character vector with any other commandline
options that will be passed to \command{GenForm}. See the \verb{GenForm
options} section for all available commandline options.}

\item{maxProcAmount}{Maximum number of processes to run for parallelization.
Usually a number close to the amount of physical cores yields most
efficient results.}

\item{maxCmdsPerProc}{Maximum number of commands that should be combined for
each executed process. Combining commands with short runtimes (such as
\command{GenForm}) can significantly increase parallel performance.}

\item{profile}{Name of the configuration profile, for example:
\option{"qtof"}, \option{"orbitrap"}, \option{"fticr"}. Sets the
\option{--profile} commandline option.}

\item{database}{If not \code{NULL}, use a database for retrieval of formula
candidates. Possible values are: \option{"pubchem"}, \option{"bio"},
\option{"kegg"}, \option{"hmdb"}. Sets the \option{--database} commandline
option.}

\item{noise}{Median intensity of the noise (\code{NULL} ignores this
parameter). Sets the \option{--noise} commandline option.}

\item{logPath}{Destination directory for log files with output from executed
commands. Will be created if non-existant. Set to \code{NULL} to disable
logging.}
}
\value{
A \code{\link{formulas}} object containing all generated formulae.
}
\description{
Functionality to automatically calculate chemical formulae from feature
groups and generate a consensus over all analyses.
}
\details{
Several algorithms are provided to automatically generate formulae for given
feature groups. All tools use the accurate mass of a feature to
back-calculate candidate formulae. Depending on the algorithm and data
availability, other data such as isotopic pattern and MS/MS fragments may be
used to further improve formula assignment.

\code{generateFormulas} is a generic function that will generate formulae
  using one of the supported algorithms. The actual functionality is provided
  by algorithm specific functions such as \code{generateFormulasDA} and
  \code{generateFormulasGenForm}. While these functions may be called directly,
  \code{generateFormulas} provides a generic interface and is therefore usually
  preferred.

\code{generateFormulasDA} uses Bruker DataAnalysis to generate
  chemical formulae. This method supports scoring based on overlap between
  measured and theoretical isotopic patterns (both MS and MS/MS data) and the
  presence of 'fitting' MS/MS fragments. The method will iterate through all
  features (or "Compounds" in DataAnalysis terms) and call
  \command{SmartFormula} (and \command{SmartFormula3D} if MS/MS data is
  available) to generate all formulae. Parameters affecting formula
  calculation have to be set in advance within the DataAnalysis method for
  each analysis (\emph{e.g.} by \code{\link{setDAMethod}}). This method
  requires that features were obtained with
  \code{\link{findFeaturesBruker}}. Unlike other algorithms, there is no
  prior need to generate \code{\link[=MSPeakLists]{MS peak lists}}.

\code{generateFormulasGenForm} uses
  \href{https://sourceforge.net/projects/genform/}{GenForm} to generate
  chemical formulae. When MS/MS data is available it will be used to score
  candidate formulae by presence of 'fitting' fragments.

\code{generateFormulasSirius} uses
  \href{https://bio.informatik.uni-jena.de/software/sirius/}{SIRIUS} to
  generate chemical formulae. Similarity of measured and theoretical isotopic
  patterns will be used for scoring candidates. Note that \command{SIRIUS}
  requires availability of MS/MS data.
}
\note{
\code{generateFormulasGenForm} always sets the \option{exist} and
  \option{oei} \command{GenForm} commandline options.
}
\section{GenForm options}{
 Below is a list of options (generated by running
  \command{GenForm} without commandline options) which can be set by the
  \code{extraOpts} parameter.
  \Sexpr[results=verbatim,echo=FALSE,stage=build]{cat(patRoon:::readAllFile(system.file("misc",
  "genform.txt", package = "patRoon")))}
}

\references{
\insertRef{Meringer2011}{patRoon}

\insertRef{Duhrkop2015}{patRoon} \cr\cr
  \insertRef{Bcker2008}{patRoon}
}
\seealso{
\code{\link{formulas-class}} and
  \code{\link{formulaConsensus-class}}. The
  \href{https://www.researchgate.net/publication/307964728_MOLGEN-MSMS_Software_User_Manual}{GenForm
  manual} (formerly known as MOLGEN-MSMS).
}
