% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\docType{class}
\name{formulaConsensus-class}
\alias{formulaConsensus-class}
\alias{formulaConsensus}
\alias{formulaTable,formulaConsensus-method}
\alias{algorithm,formulaConsensus-method}
\alias{length,formulaConsensus-method}
\alias{show,formulaConsensus-method}
\alias{filterBy,formulaConsensus-method}
\alias{plotSpec,formulaConsensus-method}
\title{Formula consensus class}
\usage{
\S4method{formulaTable}{formulaConsensus}(obj)

\S4method{algorithm}{formulaConsensus}(obj)

\S4method{length}{formulaConsensus}(x)

\S4method{show}{formulaConsensus}(object)

\S4method{filterBy}{formulaConsensus}(obj, fGroups, negate)

\S4method{plotSpec}{formulaConsensus}(obj, precursor, groupName,
  MSPeakLists, useGGPlot2 = FALSE)
}
\arguments{
\item{obj, x, object}{The \code{formulaConsensus} object.}

\item{fGroups}{A \code{\link{featureGroups}} object which should be filtered.}

\item{negate}{If set to \code{TRUE} then the filter is applied in opposite
manner.}

\item{precursor}{The formula of the precursor (in ionic form, \emph{i.e.} as
detected by the MS).}

\item{groupName}{The name of the feature group to which the candidate
belongs.}

\item{MSPeakLists}{The \code{\link{MSPeakLists}} object that was used to
generate the candidate}

\item{useGGPlot2}{If \code{TRUE} then \pkg{\link{ggplot2}} is used for
plotting, otherwise base plot used. For \code{plotSpec}, \code{ggplot2}
allows nicely repelled text for annotation. However, base plot is generally
faster.}
}
\value{
\code{formulaTable} returns a \code{\link{data.table}} containing all
  generated formulae, scoring information, MS/MS fragments, etc.

\code{filterBy} returns a filtered \code{\link{featureGroups}}
  object.

\code{plotSpec} will return a \code{\link[=ggplot2]{ggplot object}}
  if \code{useGGPlot2} is \code{TRUE}.
}
\description{
Contains all formulae for given feature groups after a consensus was made.
}
\details{
Objects of \code{formulaConsensus} are generated by the
\code{\link{consensus}} method.
}
\section{Methods (by generic)}{
\itemize{
\item \code{formulaTable}: Accessor method to obtain generated formulae.

\item \code{algorithm}: Accessor method for the algorithm (a character
string) used to generate formulae.

\item \code{length}: Obtain total number of formulae entries.

\item \code{show}: Show summary information for this object.

\item \code{filterBy}: Only keep feature groups in a given
\code{\link{featureGroups}} object that are also present in the specified
\code{formulaConsensus} object.

\item \code{plotSpec}: Plots an annotated spectrum for a given
candidate formula of a feature group.
}}

\section{Slots}{

\describe{
\item{\code{formulas}}{A \code{\link{data.table}} containing an overview of all
consensus formulae. Use the \code{formulaTable} method for access.}

\item{\code{algorithm}}{The algorithm that was used for generation of formulae. Use
the \code{algorithm} method for access.}
}}

