% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-formulas.R
\name{formulaScorings}
\alias{formulaScorings}
\title{Scorings terms for formula candidates}
\usage{
formulaScorings()
}
\description{
Returns a \code{data.frame} with information on which scoring terms are used and what their algorithm specific name
is.
}
\seealso{
generateFormulas
}
