% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-sirius.R
\docType{class}
\name{formulasSIRIUS-class}
\alias{formulasSIRIUS-class}
\alias{formulasSIRIUS}
\title{Formulas class for SIRIUS results.}
\description{
This class is derived from \code{\link{formulas}} and contains additional specific \command{SIRIUS} data.
}
\details{
Objects from this class are generated by \code{\link{generateFormulasSIRIUS}}
}
\section{Slots}{

\describe{
\item{\code{fingerprints}}{A \code{list} with for each feature group result a \code{data.table} containing fingerprints
obtained with \command{CSI:FingerID}. Will be empty unless the \code{getFingerprints} argument to
\code{\link{generateFormulasSIRIUS}} was set to \code{TRUE}.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{formulas}}}   \itemize{     \item{\strong{\code{\link{formulasSIRIUS}}}}   } }
}

\references{
\insertRef{Dhrkop2019}{patRoon} \cr\cr \insertRef{Duhrkop2015}{patRoon} \cr\cr
  \insertRef{Duhrkop2015-2}{patRoon} \cr\cr \insertRef{Bcker2008}{patRoon}
}
\seealso{
\code{\link{formulas}} and \code{\link{generateFormulasSIRIUS}}
}
