% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-camera.R
\name{generateComponentsCAMERA}
\alias{generateComponentsCAMERA}
\alias{generateComponentsCAMERA,featureGroups-method}
\alias{generateComponentsCAMERA,featureGroupsSet-method}
\title{Componentization of adducts, isotopes etc. with CAMERA}
\usage{
\S4method{generateComponentsCAMERA}{featureGroups}(
  fGroups,
  ionization = NULL,
  onlyIsotopes = FALSE,
  minSize = 2,
  relMinReplicates = 0.5,
  extraOpts = NULL
)

\S4method{generateComponentsCAMERA}{featureGroupsSet}(fGroups, ionization = NULL, ...)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which components should be generated.}

\item{ionization}{Which ionization polarity was used to generate the data: should be \code{"positive"}
  or \code{"negative"}. If the \code{featureGroups} object has adduct annotations, and \code{ionization=NULL}, the
  ionization will be detected automatically.

  \setsWF This parameter is not supported for sets workflows, as the ionization will always be detected
  automatically.}

\item{onlyIsotopes}{Logical value. If \code{TRUE} only isotopes are considered when generating components (faster).
Corresponds to \code{quick} argument of \code{\link[CAMERA:annotate-methods]{CAMERA::annotate}}.}

\item{minSize}{The minimum size of a component. Smaller components than this size will be removed. See note below.}

\item{relMinReplicates}{Feature groups within a component are only kept when they contain data for at least this
(relative) amount of replicate analyses. For instance, \samp{0.5} means that at least half of the replicates should
contain data for a particular feature group in a component. In this calculation replicates that are fully absent
within a component are not taken in to account. See note below.}

\item{extraOpts}{Named character vector with extra arguments directly passed to
\code{\link[CAMERA:annotate-methods]{CAMERA::annotate}}. Set to \code{NULL} to ignore.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}
}
\value{
A \code{\link{components}} (derived) object containing all generated components.
}
\description{
Interfaces with \href{https://bioconductor.org/packages/release/bioc/html/CAMERA.html}{CAMERA} to generate components
from known adducts, isotopes and in-source fragments.
}
\details{
This function uses CAMERA to generate components. This function is called when calling \code{generateComponents} with
  \code{algorithm="camera"}.

The specified \code{featureGroups} object is automatically converted to an \code{\link{xcmsSet}} object
  using \code{\link{getXCMSSet}}.
}
\note{
The default value for \code{minSize} and \code{relMinReplicates} results in
  extra filtering, hence, the final results may be different than what the algorithm normally would return.
}
\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} the componentization is first performed for each
  set independently. The resulting components are then all combined in a \code{\link{componentsSet}} object. Note that
  the components themselves are never merged. The components are renamed to include the set name from which they were
  generated (\emph{e.g.} \code{"CMP1"} becomes \code{"CMP1-positive"}).
}

\references{
\addCitations{CAMERA}{1}
}
\seealso{
\code{\link{generateComponents}} for more details and other algorithms.
}
