% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/components-intclust.R
\name{generateComponentsIntClust}
\alias{generateComponentsIntClust}
\alias{generateComponentsIntClust,featureGroups-method}
\title{Generate components based on intensity profiles}
\usage{
generateComponentsIntClust(fGroups, ...)

\S4method{generateComponentsIntClust}{featureGroups}(
  fGroups,
  method = "complete",
  metric = "euclidean",
  normalized = TRUE,
  average = TRUE,
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1,
  IMS = "maybe"
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which components should be generated.}

\item{...}{Any parameters to be passed to the selected component generation algorithm.}

\item{method}{Clustering method that should be applied (passed to
\code{\link[fastcluster:hclust]{fastcluster::hclust}}).}

\item{metric}{Distance metric used to calculate the distance matrix (passed to \code{\link{daisy}}).}

\item{normalized, average}{Passed to \code{\link[=as.data.table,featureGroups-method]{as.data.table}} to perform
normalization and averaging of data.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}

\item{IMS}{\IMSWF Specifies which feature groups are considered for componentization in IMS workflows. The following options
  are valid: \itemize{

  \item \code{"both"}: Selects IMS and non-IMS features.

  \item \code{"maybe"}: Selects non-IMS features and IMS features without assigned IMS parent.

  \item \code{FALSE}: Selects only non-IMS features.

  \item \code{TRUE}: Selects only IMS features.

  }}
}
\value{
The components are stored in objects derived from \code{\link{componentsIntClust}}.
}
\description{
Generates components based on intensity profiles of feature groups.
}
\details{
This function uses hierarchical clustering of intensity profiles to generate components. This function is called when calling \code{generateComponents} with
  \code{algorithm="intclust"}.

Hierarchical clustering is performed on normalized (and optionally replicate averaged) intensity data and
  the resulting dendrogram is automatically cut with \code{\link{cutreeDynamicTree}}. The distance matrix is
  calculated with \code{\link{daisy}} and clustering is performed with
  \code{\link[fastcluster:hclust]{fastcluster::hclust}}. The clustering of the resulting components can be further
  visualized and modified using the methods defined for \code{\link{componentsIntClust}}.
}
\section{IMS workflows}{
 In IMS workflows with post mobility assignment (see
  \code{\link[=assignMobilities_feat]{assignMobilities}}) it may be necessary to call \code{\link{expandForIMS}} when
  componentization was performed \emph{prior} to mobility assignments, see its documentation for more details.

  If mobilities were already assigned prior to componentization, then the \code{IMS} argument selects which feature
  groups are subjected to componentization. Data for mobility feature groups that were not considered (\emph{i.e.}
  when \code{IMS} is \code{FALSE} or \code{"maybe"}), will be expanded similarly as is done by
  \code{\link{expandForIMS}}.
}

\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} normalization of feature intensities occur per
  set.
}

\references{
\addCitations{fastcluster}

\insertRef{Scholle2018}{patRoon}
}
\seealso{
\code{\link{generateComponents}} for more details and other algorithms.
}
