% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-nontarget.R
\name{generateComponentsNontarget}
\alias{generateComponentsNontarget}
\alias{generateComponentsNontarget,featureGroups-method}
\alias{generateComponentsNontarget,featureGroupsSet-method}
\title{Componentization of homologous series with nontarget}
\usage{
\S4method{generateComponentsNontarget}{featureGroups}(
  fGroups,
  ionization = NULL,
  rtRange = c(-120, 120),
  mzRange = c(5, 120),
  elements = c("C", "H", "O"),
  rtDev = 30,
  absMzDev = 0.002,
  absMzDevLink = absMzDev * 2,
  traceHack = all(R.Version()[c("major", "minor")] >= c(3, 4)),
  ...
)

\S4method{generateComponentsNontarget}{featureGroupsSet}(fGroups, ionization = NULL, ...)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which components should be generated.}

\item{ionization}{Which ionization polarity was used to generate the data: should be \code{"positive"}
  or \code{"negative"}. If the \code{featureGroups} object has adduct annotations, and \code{ionization=NULL}, the
  ionization will be detected automatically.

  \setsWF This parameter is not supported for sets workflows, as the ionization will always be detected
  automatically.}

\item{rtRange}{A numeric vector containing the minimum and maximum retention time (in seconds) between homologues.
Series are always considered from low to high \emph{m/z}, thus, a negative minimum retention time allows detection
of homologous series with increasing \emph{m/z} and decreasing retention times. These values set the \code{minrt}
and \code{maxrt} arguments of \code{\link[nontarget]{homol.search}}.}

\item{mzRange}{A numeric vector specifying the minimum and maximum \emph{m/z} increment of a homologous series. Sets
the \code{minmz} and \code{maxmz} arguments of \code{\link[nontarget]{homol.search}}.}

\item{elements}{A character vector with elements to be considered for detection of repeating units. Sets the
\code{elements} argument of \code{\link[nontarget]{homol.search}} function.}

\item{rtDev}{Maximum retention time deviation. Sets the \code{rttol} to \code{\link[nontarget]{homol.search}}.}

\item{absMzDev}{Maximum absolute \emph{m/z} deviation. Sets the \code{mztol} argument to \code{\link[nontarget]{homol.search}}}

\item{absMzDevLink}{Maximum absolute \emph{m/z} deviation when linking series. This should usually be a bit higher
than \code{absMzDev} to ensure proper linkage.}

\item{traceHack}{Currently \code{\link[nontarget]{homol.search}} does not work with \R \samp{>3.3.3}. This flag, which is
enabled by default on these R versions, implements a (messy) workaround
(\href{https://github.com/blosloos/nontarget/issues/6}{more details here}).}

\item{\dots}{Any further arguments passed to \code{\link[nontarget]{homol.search}}.\cr\cr \setsWF Further arguments passed to the non-sets workflow method.}
}
\value{
The generated comnponents are returned as an object from the \code{\link{componentsNT}} class.
}
\description{
Uses \href{https://cran.r-project.org/web/packages/nontarget/index.html}{the nontarget R package} to generate
components by unsupervised detection of homologous series.
}
\details{
This function uses nontarget to generate components. This function is called when calling \code{generateComponents} with
  \code{algorithm="nontarget"}.

In the first step the \code{\link[nontarget]{homol.search}} function is used to detect all homologous series
  within each replicate group (analyses within each replicate group are averaged prior to detection). Then,
  homologous series across replicate groups are merged in case of full overlap or when merging of partial overlapping
  series causes no conflicts.
}
\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} the componentization is first performed for each
  set independently. The resulting components are then all combined in a \code{\link{componentsNTSet}} object. Note that
  the components themselves are never merged. The components are renamed to include the set name from which they were
  generated (\emph{e.g.} \code{"CMP1"} becomes \code{"CMP1-positive"}).

 The output class supports additional methods such as \code{plotGraph}.
}

\references{
\addCitations{nontarget}{1} \cr\cr \addCitations{enviPat}{1}
}
\seealso{
\code{\link{generateComponents}} for more details and other algorithms.
}
