% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/components-ramclustr.R
\name{generateComponentsRAMClustR}
\alias{generateComponentsRAMClustR}
\alias{generateComponentsRAMClustR,featureGroups-method}
\alias{generateComponentsRAMClustR,featureGroupsSet-method}
\title{Componentization of adducts, isotopes etc. with RAMClustR}
\usage{
generateComponentsRAMClustR(fGroups, ...)

\S4method{generateComponentsRAMClustR}{featureGroups}(
  fGroups,
  ionization = NULL,
  st = NULL,
  sr = NULL,
  maxt = 12,
  hmax = 0.3,
  normalize = "TIC",
  absMzDev = 0.002,
  relMzDev = 5,
  minSize = 2,
  relMinReplicates = 0.5,
  RCExperimentVals = list(design = list(platform = "LC-MS"), instrument = list(ionization
    = ionization, MSlevs = 1)),
  extraOptsRC = NULL,
  extraOptsFM = NULL
)

\S4method{generateComponentsRAMClustR}{featureGroupsSet}(fGroups, ionization = NULL, ...)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which components should be generated.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{ionization}{Which ionization polarity was used to generate the data: should be \code{"positive"}
  or \code{"negative"}. If the \code{featureGroups} object has adduct annotations, and \code{ionization=NULL}, the
  ionization will be detected automatically.

  \setsWF This parameter is not supported for sets workflows, as the ionization will always be detected
  automatically.}

\item{st, sr, maxt, hmax, normalize}{Arguments to tune the behaviour of feature group clustering. See their documentation
from \code{\link[RAMClustR]{ramclustR}}. When \code{st} is \code{NULL} it will be automatically calculated as the
half of the median for all chromatographic peak widths.}

\item{absMzDev}{Maximum absolute \emph{m/z} deviation. Sets the \code{mzabs.error} argument to \code{\link[RAMClustR]{do.findmain}}}

\item{relMzDev}{Maximum relative mass deviation (\acronym{ppm}). Sets the \code{ppm.error} argument to
\code{\link[RAMClustR]{do.findmain}}.}

\item{minSize}{The minimum size of a component. Smaller components than this size will be removed. See note below. Sets the \code{minModuleSize} argument to \code{\link[RAMClustR]{ramclustR}}.}

\item{relMinReplicates}{Feature groups within a component are only kept when they contain data for at least this
(relative) amount of replicate analyses. For instance, \samp{0.5} means that at least half of the replicates should
contain data for a particular feature group in a component. In this calculation replicates that are fully absent
within a component are not taken in to account. See note below.}

\item{RCExperimentVals}{A named \code{list} containing two more \code{list}s: \code{design} and \code{instrument}.
These are used to construct the \code{ExpDes} argument passed to \code{\link[RAMClustR]{ramclustR}}.}

\item{extraOptsRC, extraOptsFM}{Named \code{list} with further arguments to be passed to
\code{\link[RAMClustR]{ramclustR}} and \code{\link[RAMClustR]{do.findmain}}. Set to \code{NULL} to ignore.}
}
\value{
A \code{\link{components}} (derived) object containing all generated components.
}
\description{
Uses \href{https://github.com/cbroeckl/RAMClustR}{RAMClustR} to generate components from feature groups which follow
similar chromatographic retention profiles and annotate their relationships (\emph{e.g.} adducts and isotopes).
}
\details{
This function uses RAMClustR to generate components. This function is called when calling \code{generateComponents} with
  \code{algorithm="ramclustr"}.

This method uses the \code{\link[RAMClustR]{ramclustR}} functions for generating the components, whereas
  \code{\link[RAMClustR]{do.findmain}} is used for annotation.
}
\note{
The default value for  \code{relMinReplicates} results in
  extra filtering, hence, the final results may be different than what the algorithm normally would return.
}
\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} the componentization is first performed for each
  set independently. The resulting components are then all combined in a \code{\link{componentsSet}} object. Note that
  the components themselves are never merged. The components are renamed to include the set name from which they were
  generated (\emph{e.g.} \code{"CMP1"} becomes \code{"CMP1-positive"}).
}

\references{
\insertRef{Broeckling2013}{patRoon} \cr\cr \insertRef{Broeckling2014}{patRoon}
}
\seealso{
\code{\link{generateComponents}} for more details and other algorithms.
}
