% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/components-specclust.R
\name{generateComponentsSpecClust}
\alias{generateComponentsSpecClust}
\alias{generateComponentsSpecClust,featureGroups-method}
\title{Generate components based on MS/MS similarity}
\usage{
generateComponentsSpecClust(fGroups, ...)

\S4method{generateComponentsSpecClust}{featureGroups}(
  fGroups,
  MSPeakLists,
  method = "complete",
  specSimParams = getDefSpecSimParams(),
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1,
  IMS = "maybe"
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which components should be generated.}

\item{...}{Any parameters to be passed to the selected component generation algorithm.}

\item{MSPeakLists}{The \code{\link{MSPeakLists}} object for the given feature groups that should be used for MS
spectral similarity calculations.}

\item{method}{Clustering method that should be applied (passed to
\code{\link[fastcluster:hclust]{fastcluster::hclust}}).}

\item{specSimParams}{A named \code{list} with parameters that influence the calculation of MS spectra similarities.
See the \link[=specSimParams]{spectral similarity parameters} documentation for more details.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}

\item{IMS}{\IMSWF Specifies which feature groups are considered for componentization in IMS workflows. The following options
  are valid: \itemize{

  \item \code{"both"}: Selects IMS and non-IMS features.

  \item \code{"maybe"}: Selects non-IMS features and IMS features without assigned IMS parent.

  \item \code{FALSE}: Selects only non-IMS features.

  \item \code{TRUE}: Selects only IMS features.

  }}
}
\value{
The components are stored in objects derived from \code{\link{componentsSpecClust}}.
}
\description{
Generates components based on MS/MS similarity between feature groups.
}
\details{
This function uses hierarchical clustering of MS/MS spectra to generate components. This function is called when calling \code{generateComponents} with
  \code{algorithm="specclust"}.

The similarities are converted to a distance matrix and used as input for hierarchical clustering, and the
  resulting dendrogram is automatically cut with \code{\link{cutreeDynamicTree}}. The clustering is performed with
  \code{\link[fastcluster:hclust]{fastcluster::hclust}}.
}
\section{IMS workflows}{
 In IMS workflows with post mobility assignment (see
  \code{\link[=assignMobilities_feat]{assignMobilities}}) it may be necessary to call \code{\link{expandForIMS}} when
  componentization was performed \emph{prior} to mobility assignments, see its documentation for more details.

  If mobilities were already assigned prior to componentization, then the \code{IMS} argument selects which feature
  groups are subjected to componentization. Data for mobility feature groups that were not considered (\emph{i.e.}
  when \code{IMS} is \code{FALSE} or \code{"maybe"}), will be expanded similarly as is done by
  \code{\link{expandForIMS}}.
}

\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} the spectral similarities for each set are
  combined as is described for the \code{\link[=spectrumSimilarity,MSPeakListsSet-method]{spectrumSimilarity}} method
  for sets workflows.
}

\references{
\addCitations{fastcluster}
}
\seealso{
\code{\link{generateComponents}} for more details and other algorithms.
}
\author{
Rick Helmus <\email{r.helmus@uva.nl}> and Bas van de Velde (major contributions to spectral binning and
  similarity calculation).
}
