% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/components-tps.R
\name{generateComponentsTPs}
\alias{generateComponentsTPs}
\alias{generateComponentsTPs,featureGroups-method}
\alias{generateComponentsTPs,featureGroupsSet-method}
\title{Generate components of transformation products}
\usage{
generateComponentsTPs(fGroups, ...)

\S4method{generateComponentsTPs}{featureGroups}(
  fGroups,
  fGroupsTPs = fGroups,
  TPs = NULL,
  MSPeakLists = NULL,
  formulas = NULL,
  compounds = NULL,
  ignoreParents = FALSE,
  minRTDiff = 20,
  specSimParams = getDefSpecSimParams(),
  IMS = "maybe"
)

\S4method{generateComponentsTPs}{featureGroupsSet}(
  fGroups,
  fGroupsTPs = fGroups,
  TPs = NULL,
  MSPeakLists = NULL,
  formulas = NULL,
  compounds = NULL,
  ignoreParents = FALSE,
  minRTDiff = 20,
  specSimParams = getDefSpecSimParams(),
  IMS = "maybe"
)
}
\arguments{
\item{fGroups}{The input \code{\link{featureGroups}} for componentization. See \code{fGroupsTPs}.}

\item{\dots}{Further arguments specified to the methods.}

\item{fGroupsTPs}{A \code{\link{featureGroups}} object containing the feature groups that are expected to be
transformation products. If a distinction between parents and TPs is not yet known, \code{fGroupsTPs} should equal
the \code{fGroups} argument. Otherwise, \code{fGroups} should only contain the parent feature groups, and both
\code{fGroups} and \code{fGroupsTPs} \emph{must} be a subset of the same \code{\link{featureGroups}} object.}

\item{TPs}{A \code{\link{transformationProducts}} object. Set to \code{NULL} to perform linking without this data.}

\item{MSPeakLists, formulas, compounds}{A \code{\link{MSPeakLists}}/\code{\link{formulas}}/\code{\link{compounds}}
object to calculate MS/MS or annotation similarities between parents and TPs. If \code{NULL} then this data is not
calculated. For more details see the \verb{Linking parents and transformation products} section below.}

\item{ignoreParents}{If \code{TRUE} then feature groups present in both \code{fGroups} and \code{fGroupsTPs} are not
considered as TPs.}

\item{minRTDiff}{Minimum retention time (in seconds) difference between the parent and a TP to calculate the
\link[=retDir]{retention order direction}.}

\item{specSimParams}{A named \code{list} with parameters that influence the calculation of MS spectra similarities.
See the \link[=specSimParams]{spectral similarity parameters} documentation for more details.}

\item{IMS}{\IMSWF Specifies which feature groups are considered for componentization in IMS workflows. The following options
  are valid: \itemize{

  \item \code{"both"}: Selects IMS and non-IMS features.

  \item \code{"maybe"}: Selects non-IMS features and IMS features without assigned IMS precursor.

  \item \code{FALSE}: Selects only non-IMS features.

  \item \code{TRUE}: Selects only IMS features.

  }}
}
\value{
The components are stored in objects derived from \code{\link{componentsTPs}}.
}
\description{
Generates components by linking feature groups of transformation products and their parents.
}
\details{
This function uses transformation product screening to generate components. This function is called when calling \code{generateComponents} with
  \code{algorithm="tp"}.

This method typically employs data from \link[=generateTPs]{generated transformation products} to find
  parents and their TPs. However, this data is not necessary, and components can also be made based on MS/MS
  similarity and/or other annotation similarities between the parent and its TPs. For more details see the
  \verb{Linking parents and transformation products} section below.
}
\note{
The \code{shift} parameter of \code{specSimParams} is ignored by \code{generateComponentsTPs}, since it always
  calculates similarities with all supported options.
}
\section{Linking parents and transformation products}{
 Each component consists of feature groups that are considered
  to be transformation products for one parent (the parent that 'belongs' to the component can be retrieved with the
  \code{\link{componentInfo}} method). The parent feature groups are taken from the \code{fGroups} parameter, while
  the feature groups for TPs are taken from \code{fGroupsTPs}. If a feature group occurs in both variables, it may
  therefore be considered as both a parent or TP.

  If transformation product data is given, \emph{i.e.} the \code{TPs} argument is set, then a suspect screening of
  the parents and/or TPs may need to be performed in advance to facilitate linkage. This depends on the algorithm
  that was used to generate the TPs: \itemize{

    \item the parents need to be screened for all algorithms except \code{logic} (\code{\link{generateTPsLogic}})

    \item the TPs need to be screened for all algorithms except \code{ann_form} and \code{ann_comp}
    (\code{\link{generateTPsAnnForm}} and \code{\link{generateTPsAnnComp}}).
  }

  See \code{\link{screenSuspects}} to perform the screening and \code{\link{convertToSuspects}} to create the suspect
  list. To include parents make sure to set \code{includeParents=TRUE} when calling \code{convertToSuspects} or first
  screen for the parents and then amend the screening object with TP screening results by setting \code{amend=TRUE}
  to \code{screenSuspects}. If the the suspect screening yields multiple TP hits, all will be reported. Similarly, if
  the suspect screening contains multiple hits for a parent, a component is made for each of the parent hits.

  In case no transformation product data is provided (\code{TPs=NULL}), the componentization algorithm simply assumes
  that each feature group from \code{fGroupsTPs} is a potential TP for every parent feature group in \code{fGroups}.
  For this reason, it is highly recommended to specify which feature groups are parents/TPs (see the
  \code{fGroupsTPs} argument description above) and \emph{crucial} that the data is post-processed, for instance by
  only retaining TPs that have high annotation similarity with their parents (see the
  \code{\link[=filter,componentsTPs-method]{filter}} method for \code{\link{componentsTPs}}).

  A typical way to distinguish which feature groups are parents or TPs from two different (groups of) samples is by
  calculating Fold Changes (see the \code{\link[=as.data.table,featureGroups-method]{as.data.table}} method for
  feature groups and \code{\link{plotVolcano}}). Of course, other statistical techniques from \R are also suitable.
}

\section{Result columns}{
 During componentization, several characteristics are calculated which may be useful for
  post-processing. These can be obtained with \emph{e.g.}
  \code{\link[=as.data.table,componentsTPs-method]{as.data.table}} or \code{\link{componentTable}}. The properties
  are either reported for each feature group in a component, or for each candidate of a feature group in a component
  (only if \code{TPs} was set).

  The following properties may be reported for each feature group: \itemize{

    \item \code{specSimilarity}: the MS/MS spectral similarity between the feature groups of the TP and its parent
    (\samp{0-1}).

    \item \code{specSimilarityPrec},\code{specSimilarityBoth}: as \code{specSimilarity}, but calculated with binned
    data using the \code{"precursor"} and \code{"both"} method, respectively (see \link[=specSimParams]{MS spectral
    similarity parameters} for more details).

    \item \code{totalFragmentMatches} The total number of MS/MS fragment annotations that overlap between \emph{all}
    feature annotation candidates for the TP feature group and the feature annotations specifically for the parent
    (based on the assigned fragment formula). If both the \code{formulas} and \code{compounds} arguments are
    specified then the annotation data is pooled prior to calculation. Each unique match is only counted once.

    \item \code{totalNeutralLossMatches} As \code{totalFragmentMatches}, but counting overlapping neutral loss
    formulae.

    \item \code{retDir},\code{TP_retDir} The \link[=retDir]{retention order direction} derived from the feature groups
    (\code{retDir}) or the (expected) value from TP data (\code{TP_retDir}).

    \item \code{retDiff},\code{mzDiff}, The retention time and \emph{m/z} difference between the parent and TP.

  }

  The candidate specific properties are stored inside the \code{candidates} column in component tables, and can be
  obtained with \code{as.data.table} by setting \code{candidates=TRUE}. The following properties may be present: \itemize{

    \item \code{fragmentMatches},\code{neutralLossMatches} As \code{totalFragmentMatches} and
    \code{totalNeutralLossMatches}, but only considering the feature annotations specifically for this candidate.

    \item \code{formulaDiff} The formula difference between the parent and TP (if formula data is available).

    \item \code{TPScore},\code{annSim},\code{fitFormula},\code{fitCompound},\code{simSusps}: TP scoring properties,
    see \code{\link{generateTPsAnnForm}} and \code{\link{generateTPsAnnComp}}.

  }
}

\section{IMS workflows}{
 In IMS workflows with post mobility assignment (see
  \code{\link[=assignMobilities_feat]{assignMobilities}}) it may be necessary to call \code{\link{expandForIMS}} when
  componentization was performed \emph{prior} to mobility assignments, see its documentation for more details.

  If mobilities were already assigned prior to componentization, then the \code{IMS} argument selects which feature
  groups are subjected to componentization. Data for IMS feature groups that were not considered (\emph{i.e.}
  when \code{IMS} is \code{FALSE} or \code{"maybe"}), will be expanded similarly as is done by
  \code{\link{expandForIMS}}.

 \strong{NOTE}: IMS expansion by \code{\link{expandForIMS}} only expands results for TP
  candidates, \emph{i.e.} no new components from parents assigned to IMS feature groups will be added.
}

\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} the component tables are amended with extra
  information such as overall/specific set spectrum similarities. As sets data is mixed, transformation products are
  able to be linked with a parent, even if they were not measured in the same set.
}

\seealso{
\code{\link{generateComponents}} for more details and other algorithms.
}
