% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds-library.R
\name{generateCompoundsLibrary}
\alias{generateCompoundsLibrary}
\alias{generateCompoundsLibrary,featureGroups-method}
\alias{generateCompoundsLibrary,featureGroupsSet-method}
\title{Compound annotation with an MS library}
\usage{
\S4method{generateCompoundsLibrary}{featureGroups}(
  fGroups,
  MSPeakLists,
  MSLibrary,
  minSim = 0.75,
  minAnnSim = minSim,
  absMzDev = 0.002,
  adduct = NULL,
  checkIons = "adduct",
  spectrumType = "MS2",
  specSimParams = getDefSpecSimParams(),
  specSimParamsLib = getDefSpecSimParams()
)

\S4method{generateCompoundsLibrary}{featureGroupsSet}(
  fGroups,
  MSPeakLists,
  MSLibrary,
  minSim = 0.75,
  minAnnSim = minSim,
  absMzDev = 0.002,
  adduct = NULL,
  ...,
  setThreshold = 0,
  setThresholdAnn = 0
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object which should be annotated. This should be the same or a subset of
the object that was used to create the specified \code{MSPeakLists}. In the case of a subset only the remaining
feature groups in the subset are considered.}

\item{MSPeakLists}{A \code{\link{MSPeakLists}} object that was generated for the supplied \code{fGroups}.}

\item{MSLibrary}{The \code{\link{MSLibrary}} object that should be used to find candidates.}

\item{minSim}{The minimum spectral similarity for candidate records.}

\item{minAnnSim}{The minimum spectral similarity of a record for it to be used to find annotations (see the
\verb{Details} section).}

\item{absMzDev}{The maximum absolute \emph{m/z} deviation between the feature group and library record \emph{m/z}
values for candidate selection.}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
  Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}. If the \code{featureGroups} object has
  adduct annotations then these are used if \code{adducts=NULL}.

  \setsWF The \code{adduct} argument is not supported for sets workflows, since the
  adduct annotations will then always be used.}

\item{checkIons}{A \code{character} that excludes library records with different adduct (\code{checkIons="adduct"})
or MS ionization polarity (\code{checkIons="polarity"}). If \code{checkIons="none"} then these filters are not
applied.}

\item{spectrumType}{A \code{character} vector which limits library records to the given spectrum types
(\code{Spectrum_type} field, \emph{e.g.} \code{"MS2"}). Set to \code{NULL} to allow all spectrum types.}

\item{specSimParams}{A named \code{list} with parameters that influence the calculation of MS spectra similarities.
See the \link[=specSimParams]{spectral similarity parameters} documentation for more details.}

\item{specSimParamsLib}{Like \code{specSimParams}, but these parameters \emph{only} influence pre-treatment of
library spectra (only the \code{removePrecursor}, \code{relMinIntensity} and \code{minPeaks} parameters are used).}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{setThreshold}{\setsWF Minimum abundance for a candidate among all sets (\samp{0-1}). For instance, a value of
\samp{1} means that the candidate needs to be present in all the set data.}

\item{setThresholdAnn}{\setsWF As \code{setThreshold}, but only taking into account the set data that contain
annotations for the feature group of the candidate.}
}
\description{
Uses a MS library loaded by \code{\link{loadMSLibrary}} for compound annotation.
}
\details{
This function uses MS library spectra to generate compound candidates. This function is called when calling \code{generateCompounds} with
  \code{algorithm="library"}.

This method matches measured MS/MS data (peak lists) with those from an MS library to find candidate
  structures. Hence, only feature groups with MS/MS peak list data are annotated.

  The library is searched for candidates with the following criteria: \enumerate{

  \item Only records with ion \emph{m/z} (\code{PrecursorMZ}), \acronym{SMILES}, \acronym{InChI}, \acronym{InChIKey}
  and \code{formula} data are considered.

  \item Depending on the value of the \code{checkIons} argument, records with different adduct
  (\code{Precursor_type}) or polarity (\code{Ion_mode}) may be ignored.

  \item The \emph{m/z} values of the candidate and feature group should match (tolerance set by \code{absMzDev}
  argument).

  \item The spectral similarity should not be lower than the value defined for the \code{minSim} argument.

  \item If multiple candidates with the same first-block \acronym{InChIKey} are found then only the candidate with
  the best spectral match is kept.

  }

  If the library contains annotations these will be added to the matched MS/MS peaks. However, since the candidate
  selected from criterion #5 above may not contain all the annotation data available from the MS library, annotations
  from other records are also considered (controlled by the \code{minAnnSim} argument). If this leads to different
  annotations for the same mass peak then only the most abundant annotation is kept.
}
\seealso{
\code{\link{generateCompounds}} for more details and other algorithms.

\code{\link{loadMSLibrary}} to obtain MS library data and the methods for \code{\link{MSLibrary}} to treat
  the data before using it for annotation.
}
