% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds-metfrag.R
\name{generateCompoundsMetFrag}
\alias{generateCompoundsMetFrag}
\alias{generateCompoundsMetFrag,featureGroups-method}
\alias{generateCompoundsMetFrag,featureGroupsSet-method}
\title{Compound annotation with MetFrag}
\usage{
\S4method{generateCompoundsMetFrag}{featureGroups}(
  fGroups,
  MSPeakLists,
  method = "CL",
  timeout = 300,
  timeoutRetries = 2,
  errorRetries = 2,
  topMost = 100,
  dbRelMzDev = 5,
  fragRelMzDev = 5,
  fragAbsMzDev = 0.002,
  adduct = NULL,
  database = "pubchem",
  extendedPubChem = "auto",
  chemSpiderToken = "",
  scoreTypes = compoundScorings("metfrag", database, onlyDefault = TRUE)$name,
  scoreWeights = 1,
  preProcessingFilters = c("UnconnectedCompoundFilter", "IsotopeFilter"),
  postProcessingFilters = c("InChIKeyFilter"),
  maxCandidatesToStop = 2500,
  identifiers = NULL,
  extraOpts = NULL
)

\S4method{generateCompoundsMetFrag}{featureGroupsSet}(
  fGroups,
  MSPeakLists,
  method = "CL",
  timeout = 300,
  timeoutRetries = 2,
  errorRetries = 2,
  topMost = 100,
  dbRelMzDev = 5,
  fragRelMzDev = 5,
  fragAbsMzDev = 0.002,
  adduct = NULL,
  ...,
  setThreshold = 0,
  setThresholdAnn = 0,
  setAvgSpecificScores = FALSE
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object which should be annotated. This should be the same or a subset of
the object that was used to create the specified \code{MSPeakLists}. In the case of a subset only the remaining
feature groups in the subset are considered.}

\item{MSPeakLists}{A \code{\link{MSPeakLists}} object that was generated for the supplied \code{fGroups}.}

\item{method}{Which method should be used for MetFrag execution: \code{"CL"} for \command{MetFragCL} and \code{"R"}
for \command{MetFragR}. The former is usually much faster and recommended.}

\item{timeout}{Maximum time (in seconds) before a metFrag query for a feature group is stopped. Also see
\code{timeoutRetries} argument.}

\item{timeoutRetries}{Maximum number of retries after reaching a timeout before completely skipping the metFrag query
for a feature group. Also see \code{timeout} argument.}

\item{errorRetries}{Maximum number of retries after an error occurred. This may be useful to handle e.g. connection
errors.}

\item{topMost}{Only keep this number of candidates (per feature group) with highest score. Set to \code{NULL} to
always keep all candidates, however, please note that this may result in significant usage of CPU/RAM resources for
large numbers of candidates.}

\item{dbRelMzDev}{Relative mass deviation (in ppm) for database search. Sets the
\option{DatabaseSearchRelativeMassDeviation} option.}

\item{fragRelMzDev}{Relative mass deviation (in ppm) for fragment matching. Sets the
\option{FragmentPeakMatchRelativeMassDeviation} option.}

\item{fragAbsMzDev}{Absolute mass deviation (in Da) for fragment matching. Sets the
\option{FragmentPeakMatchAbsoluteMassDeviation} option.}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
  Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}. If the \code{featureGroups} object has
  adduct annotations then these are used if \code{adducts=NULL}.

  \setsWF The \code{adduct} argument is not supported for sets workflows, since the
  adduct annotations will then always be used.}

\item{database}{Compound database to use. Valid values are: \code{"pubchem"}, \code{"chemspider"},
\code{"for-ident"}, \code{"comptox"}, \code{"pubchemlite"}, \code{"kegg"}, \code{"sdf"}, \code{"psv"} and
\code{"csv"}. See section below for more information. Sets the \code{MetFragDatabaseType} option.}

\item{extendedPubChem}{If \code{database="pubchem"}: whether to use the \emph{extended} database that includes
information for compound scoring (\emph{i.e.} number of patents/PubMed references). Note that downloading
candidates from this database might take extra time. Valid values are: \code{FALSE} (never use it), \code{TRUE}
(always use it) or \code{"auto"} (default, use if specified scorings demand it).}

\item{chemSpiderToken}{A character string with the \href{http://www.chemspider.com/AboutServices.aspx}{ChemSpider
security token} that should be set when the ChemSpider database is used. Sets the \option{ChemSpiderToken} option.}

\item{scoreTypes}{A character vector defining the scoring types. See the \verb{Scorings} section below for more
information. Note that both generic and \command{MetFrag} specific names are accepted (\emph{i.e.} \code{name} and
\code{metfrag} columns returned by \code{\link{compoundScorings}}). When a local database is used, the name should
match what is given there (\code{e.g} column names when \code{database=csv}). Note that MetFrag may still report
other scoring data, however, these are not used for ranking. Sets the \option{MetFragScoreTypes} option.}

\item{scoreWeights}{Numeric vector containing weights of the used scoring types. Order is the same as set in
\code{scoreTypes}. Values are recycled if necessary. Sets the \option{MetFragScoreWeights} option.}

\item{preProcessingFilters, postProcessingFilters}{A character vector defining pre/post filters applied before/after
fragmentation and scoring (\emph{e.g.} \code{"UnconnectedCompoundFilter"}, \code{"IsotopeFilter"},
\code{"ElementExclusionFilter"}). Some methods require further options to be set. For all filters and more
information refer to the \verb{Candidate Filters} section on the
\href{http://ipb-halle.github.io/MetFrag/projects/metfragr/}{MetFragR homepage}. Sets the
\option{MetFragPreProcessingCandidateFilter} and \code{MetFragPostProcessingCandidateFilter} options.}

\item{maxCandidatesToStop}{If more than this number of candidate structures are found then processing will be aborted
and no results this feature group will be reported. Low values increase the chance of missing data, whereas too
high values will use too much computer resources and signficantly slowdown the process. Sets the
\option{MaxCandidateLimitToStop} option.}

\item{identifiers}{A \code{list} containing for each feature group a character vector with database identifiers that
should be used to find candidates for a feature group (the list should be named by feature group names). If
\code{NULL} all relevant candidates will be retrieved from the specified database. An example usage scenario is to
obtain the list of candidate identifiers from a \code{\link{compounds}} object obtained with
\code{\link{generateCompoundsSIRIUS}} using the \code{\link{identifiers}} method. This way, only those candidates
will be searched by MetFrag that were generated by SIRIUS+CSI:FingerID. Sets the \option{PrecursorCompoundIDs}
option.}

\item{extraOpts}{A named \code{list} containing further settings \command{MetFrag}. See the
\href{http://ipb-halle.github.io/MetFrag/projects/metfragr/}{MetFragR} and
\href{http://ipb-halle.github.io/MetFrag/projects/metfragcl/}{MetFrag CL} homepages for all available options. Set
to \code{NULL} to ignore.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{setThreshold}{\setsWF Minimum abundance for a candidate among all sets (\samp{0-1}). For instance, a value of
\samp{1} means that the candidate needs to be present in all the set data.}

\item{setThresholdAnn}{\setsWF As \code{setThreshold}, but only taking into account the set data that contain
annotations for the feature group of the candidate.}

\item{setAvgSpecificScores}{\setsWF If \code{TRUE} then set specific scorings (\emph{e.g.} MS/MS match) are also
averaged.}
}
\value{
\code{generateCompoundsMetFrag} returns a \code{\link{compoundsMF}} object.
}
\description{
Uses the \pkg{metfRag} package or \command{MetFrag CL} for compound identification (see
\url{http://ipb-halle.github.io/MetFrag/}).
}
\details{
This function uses MetFrag to generate compound candidates. This function is called when calling \code{generateCompounds} with
  \code{algorithm="metfrag"}.

Several online compound databases such as \href{https://pubchem.ncbi.nlm.nih.gov/}{PubChem} and
  \href{http://www.chemspider.com/}{ChemSpider} may be chosen for retrieval of candidate structures. This method
  requires the availability of MS/MS data, and feature groups without it will be ignored. Many options exist to score
  and filter resulting data, and it is highly suggested to optimize these to improve results. The \command{MetFrag}
  options \code{PeakList}, \code{IonizedPrecursorMass} and \code{ExperimentalRetentionTimeValue} (in minutes) fields
  are automatically set from feature data.
}
\section{Scorings}{
 \command{MetFrag} supports \emph{many} different scorings to rank candidates. The
  \code{\link{compoundScorings}} function can be used to get an overview: (some columns are omitted)

 \tabular{lll}{
  \strong{name}       \tab \strong{metfrag}                                \tab \strong{database}\cr
  score                \tab Score                                            \tab                   \cr
  fragScore            \tab FragmenterScore                                  \tab                   \cr
  metFusionScore       \tab OfflineMetFusionScore                            \tab                   \cr
  individualMoNAScore  \tab OfflineIndividualMoNAScore                       \tab                   \cr
  numberPatents        \tab PubChemNumberPatents                             \tab pubchem           \cr
  numberPatents        \tab Patent_Count                                     \tab pubchemlite       \cr
  pubMedReferences     \tab PubChemNumberPubMedReferences                    \tab pubchem           \cr
  pubMedReferences     \tab ChemSpiderNumberPubMedReferences                 \tab chemspider        \cr
  pubMedReferences     \tab NUMBER_OF_PUBMED_ARTICLES                        \tab comptox           \cr
  pubMedReferences     \tab PubMed_Count                                     \tab pubchemlite       \cr
  extReferenceCount    \tab ChemSpiderNumberExternalReferences               \tab chemspider        \cr
  dataSourceCount      \tab ChemSpiderDataSourceCount                        \tab chemspider        \cr
  referenceCount       \tab ChemSpiderReferenceCount                         \tab chemspider        \cr
  RSCCount             \tab ChemSpiderRSCCount                               \tab chemspider        \cr
  smartsInclusionScore \tab SmartsSubstructureInclusionScore                 \tab                   \cr
  smartsExclusionScore \tab SmartsSubstructureExclusionScore                 \tab                   \cr
  suspectListScore     \tab SuspectListScore                                 \tab                   \cr
  retentionTimeScore   \tab RetentionTimeScore                               \tab                   \cr
  CPDATCount           \tab CPDAT_COUNT                                      \tab comptox           \cr
  TOXCASTActive        \tab TOXCAST_PERCENT_ACTIVE                           \tab comptox           \cr
  dataSources          \tab DATA_SOURCES                                     \tab comptox           \cr
  pubChemDataSources   \tab PUBCHEM_DATA_SOURCES                             \tab comptox           \cr
  EXPOCASTPredExpo     \tab EXPOCAST_MEDIAN_EXPOSURE_PREDICTION_MG/KG-BW/DAY \tab comptox           \cr
  ECOTOX               \tab ECOTOX                                           \tab comptox           \cr
  NORMANSUSDAT         \tab NORMANSUSDAT                                     \tab comptox           \cr
  MASSBANKEU           \tab MASSBANKEU                                       \tab comptox           \cr
  TOX21SL              \tab TOX21SL                                          \tab comptox           \cr
  TOXCAST              \tab TOXCAST                                          \tab comptox           \cr
  KEMIMARKET           \tab KEMIMARKET                                       \tab comptox           \cr
  MZCLOUD              \tab MZCLOUD                                          \tab comptox           \cr
  pubMedNeuro          \tab PubMedNeuro                                      \tab comptox           \cr
  CIGARETTES           \tab CIGARETTES                                       \tab comptox           \cr
  INDOORCT16           \tab INDOORCT16                                       \tab comptox           \cr
  SRM2585DUST          \tab SRM2585DUST                                      \tab comptox           \cr
  SLTCHEMDB            \tab SLTCHEMDB                                        \tab comptox           \cr
  THSMOKE              \tab THSMOKE                                          \tab comptox           \cr
  ITNANTIBIOTIC        \tab ITNANTIBIOTIC                                    \tab comptox           \cr
  STOFFIDENT           \tab STOFFIDENT                                       \tab comptox           \cr
  KEMIMARKET_EXPO      \tab KEMIMARKET_EXPO                                  \tab comptox           \cr
  KEMIMARKET_HAZ       \tab KEMIMARKET_HAZ                                   \tab comptox           \cr
  REACH2017            \tab REACH2017                                        \tab comptox           \cr
  KEMIWW_WDUIndex      \tab KEMIWW_WDUIndex                                  \tab comptox           \cr
  KEMIWW_StpSE         \tab KEMIWW_StpSE                                     \tab comptox           \cr
  KEMIWW_SEHitsOverDL  \tab KEMIWW_SEHitsOverDL                              \tab comptox           \cr
  ZINC15PHARMA         \tab ZINC15PHARMA                                     \tab comptox           \cr
  PFASMASTER           \tab PFASMASTER                                       \tab comptox           \cr
  peakFingerprintScore \tab AutomatedPeakFingerprintAnnotationScore          \tab                   \cr
  lossFingerprintScore \tab AutomatedLossFingerprintAnnotationScore          \tab                   \cr
  agroChemInfo         \tab AgroChemInfo                                     \tab pubchemlite       \cr
  bioPathway           \tab BioPathway                                       \tab pubchemlite       \cr
  drugMedicInfo        \tab DrugMedicInfo                                    \tab pubchemlite       \cr
  foodRelated          \tab FoodRelated                                      \tab pubchemlite       \cr
  pharmacoInfo         \tab PharmacoInfo                                     \tab pubchemlite       \cr
  safetyInfo           \tab SafetyInfo                                       \tab pubchemlite       \cr
  toxicityInfo         \tab ToxicityInfo                                     \tab pubchemlite       \cr
  knownUse             \tab KnownUse                                         \tab pubchemlite       \cr
  disorderDisease      \tab DisorderDisease                                  \tab pubchemlite       \cr
  identification       \tab Identification                                   \tab pubchemlite       \cr
  annoTypeCount        \tab FPSum                                            \tab pubchemlite       \cr
  annoTypeCount        \tab AnnoTypeCount                                    \tab pubchemlite       \cr
  annotHitCount        \tab AnnotHitCount                                    \tab pubchemlite       
}

 In addition, the \code{\link{compoundScorings}} function is also useful to programmatically
  generate a set of scorings to be used for ranking with \command{MetFrag}. For instance, the following can be given
  to the \code{scoreTypes} argument to use all default scorings for PubChem: \code{compoundScorings("metfrag",
  "pubchem", onlyDefault=TRUE)$name}.

  For all \command{MetFrag} scoring types refer to the \verb{Candidate Scores} section on the
  \href{http://ipb-halle.github.io/MetFrag/projects/metfragr/}{MetFragR homepage}.
}

\section{Usage of MetFrag databases}{
 When \code{database="chemspider"} setting the \code{chemSpiderToken} argument is
  mandatory.

  When a local database is set (\emph{i.e.} \code{sdf}, \code{psv}, \code{csv}, \code{comptox}, \code{pubchemlite})
  the file location of the database should be set in the \code{LocalDatabasePath} value via the \code{extraOpts}
  argument or using the \code{patRoon.path.MetFragCompTox}/\code{patRoon.path.MetFragPubChemLite} option (only when
  \code{database="comptox"} or \code{database="pubchemlite"}).

  Examples: \verb{options(patRoon.path.MetFragCompTox = "C:/CompTox_17March2019_SelectMetaData.csv")}

  \verb{extraOpts = list(LocalDatabasePath = "C:/myDB.csv")}.

  For \code{database="comptox"} the files can be obtained from
  \href{ftp://newftp.epa.gov/COMPTOX/Sustainable_Chemistry_Data/Chemistry_Dashboard/MetFrag_metadata_files}{here}.
  Furthermore, the files with additions for \href{https://zenodo.org/record/3364464#.XnjM-XLvKUk}{smoking} and
  \href{https://zenodo.org/record/3472781#.XnjMAHLvKUk}{wastewater} metadata are also supported. For
  \code{database="pubchemlite"} the \file{.csv} database can be downloaded from
  \href{https://zenodo.org/record/6503754}{here} or from \href{https://zenodo.org/record/6385954}{here} for the OECD
  PFAS database. Note that only recent \command{MetFrag} versions (>= \samp{2.4.5}) support these libraries.
}

\section{Parallelization}{
 generateCompoundsMetFrag uses multiprocessing to parallelize
  computations. Please see the parallelization section in the handbook for
  more details and \link[=patRoon-package]{patRoon options} for configuration
  options.

 When local database files are used with \code{generateCompoundsMetFrag} (\emph{e.g.} when
  \code{database} is set to \code{"pubchemlite"}, \code{"csv"} etc.) and \option{patRoon.MP.method="future"}, then
  the database file must be present on all the nodes. When \code{pubchemlite} or \code{comptox} is used, the location
  for these databases can be configured on the host with the respective package options
  (\option{patRoon.path.MetFragPubChemLite} and \option{patRoon.path.MetFragCompTox}). Note that these files must
  \emph{also} be present on the local host computer, even if it is not participating in computations.
}

\references{
\insertRef{Ruttkies2016}{patRoon}
}
\seealso{
\code{\link{generateCompounds}} for more details and other algorithms.
}
