% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-bruker.R
\name{generateFormulasDA}
\alias{generateFormulasDA}
\alias{generateFormulasDA,featureGroups-method}
\alias{generateFormulasDA,featureGroupsSet-method}
\title{Generate formula with Bruker DataAnalysis}
\usage{
\S4method{generateFormulasDA}{featureGroups}(
  fGroups,
  MSPeakLists,
  precursorMzSearchWindow = 0.002,
  MSMode = "both",
  adduct = NULL,
  featThreshold = 0,
  featThresholdAnn = 0.75,
  absAlignMzDev = 0.002,
  save = TRUE,
  close = save
)

\S4method{generateFormulasDA}{featureGroupsSet}(
  fGroups,
  MSPeakLists,
  precursorMzSearchWindow = 0.002,
  MSMode = "both",
  adduct = NULL,
  ...,
  setThreshold = 0,
  setThresholdAnn = 0
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which formulae should be generated. This should be the same or
a subset of the object that was used to create the specified \code{MSPeakLists}. In the case of a subset only the
remaining feature groups in the subset are considered.}

\item{MSPeakLists}{An \code{\link{MSPeakLists}} object that was generated for the supplied \code{fGroups}.}

\item{precursorMzSearchWindow}{Search window for \emph{m/z} values (+/- the feature \emph{m/z}) used to find back
feature data of precursor/parent ions from MS/MS spectra (this data is not readily available from
\command{SmartFormula3D} results).}

\item{MSMode}{Whether formulae should be generated only from MS data (\code{"ms"}), MS/MS data (\code{"msms"}) or
both (\code{"both"}). Selecting "both" will calculate formulae from MS data and MS/MS data and combines the results
(duplicated formulae are removed). This is useful when poor MS/MS data would exclude proper candidates.}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
  Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}. If the \code{featureGroups} object has
  adduct annotations then these are used if \code{adducts=NULL}.

  \setsWF The \code{adduct} argument is not supported for sets workflows, since the
  adduct annotations will then always be used.}

\item{featThreshold}{If \code{calculateFeatures=TRUE}: minimum presence (\samp{0-1}) of a formula in all features
before it is considered as a candidate for a feature group. For instance, \code{featThreshold=0.75} dictates that a
formula should be present in at least 75\% of the features inside a feature group.}

\item{featThresholdAnn}{As \code{featThreshold}, but only considers features with annotations. For instance,
\code{featThresholdAnn=0.75} dictates that a formula should be present in at least 75\% of the features with
annotations inside a feature group.}

\item{absAlignMzDev}{When the group formula annotation consensus is made from feature annotations, the \emph{m/z}
values of annotated MS/MS fragments may slightly deviate from those of the corresponding group MS/MS peak list. The
\code{absAlignMzDev} argument specifies the maximum \emph{m/z} window used to re-align the mass peaks.}

\item{close, save}{If \code{TRUE} then Bruker files are closed and saved after
processing with DataAnalysis, respectively. Setting \code{close=TRUE}
prevents that many analyses might be opened simultaneously in DataAnalysis,
which otherwise may use excessive memory or become slow. By default
\code{save} is \code{TRUE} when \code{close} is \code{TRUE}, which is
likely what you want as otherwise any processed data is lost.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{setThreshold}{\setsWF Minimum abundance for a candidate among all sets (\samp{0-1}). For instance, a value of
\samp{1} means that the candidate needs to be present in all the set data.}

\item{setThresholdAnn}{\setsWF As \code{setThreshold}, but only taking into account the set data that contain
annotations for the feature group of the candidate.}
}
\value{
A \code{\link{formulas}} object containing all generated formulae.
}
\description{
Uses Bruker DataAnalysis to generate chemical formulae.
}
\details{
This function uses bruker to generate formula candidates. This function is called when calling \code{generateFormulas} with
  \code{algorithm="bruker"}.

This method supports scoring based on overlap between measured and theoretical isotopic patterns (both MS
  and MS/MS data) and the presence of 'fitting' MS/MS fragments. The method will iterate through all features (or
  "Compounds" in DataAnalysis terms) and call \command{SmartFormula} (and \command{SmartFormula3D} if MS/MS data is
  available) to generate all formulae. Parameters affecting formula calculation have to be set in advance within the
  DataAnalysis method for each analysis (\emph{e.g.} by \code{\link{setDAMethod}}).

  This method requires that features were obtained with \code{\link{findFeaturesBruker}}. It is recommended, but not
  mandatory, that the \code{\link{MSPeakLists}} are also generated by DataAnalysis.

  Calculation of formulae with DataAnalysis always occurs with the 'feature approach' (see \verb{Candidate
  assignment} in \code{\link{generateFormulas}}).
}
\note{
If any errors related to \command{DCOM} appear it might be necessary to
  terminate DataAnalysis (note that DataAnalysis might still be running as a
  background process). The \command{ProcessCleaner} application installed
  with DataAnalayis can be used for this.
}
\seealso{
\code{\link{generateFormulas}} for more details and other algorithms.
}
