% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-genform.R
\name{generateFormulasGenForm}
\alias{generateFormulasGenForm}
\alias{generateFormulasGenForm,featureGroups-method}
\alias{generateFormulasGenForm,featureGroupsSet-method}
\title{Generate formula with GenForm}
\usage{
\S4method{generateFormulasGenForm}{featureGroups}(
  fGroups,
  MSPeakLists,
  relMzDev = 5,
  adduct = NULL,
  elements = "CHNOP",
  hetero = TRUE,
  oc = FALSE,
  thrMS = NULL,
  thrMSMS = NULL,
  thrComb = NULL,
  maxCandidates = Inf,
  extraOpts = NULL,
  calculateFeatures = TRUE,
  featThreshold = 0,
  featThresholdAnn = 0.75,
  absAlignMzDev = 0.002,
  MSMode = "both",
  isolatePrec = TRUE,
  timeout = 120,
  topMost = 50,
  batchSize = 8
)

\S4method{generateFormulasGenForm}{featureGroupsSet}(
  fGroups,
  MSPeakLists,
  relMzDev = 5,
  adduct = NULL,
  ...,
  setThreshold = 0,
  setThresholdAnn = 0
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which formulae should be generated. This should be the same or
a subset of the object that was used to create the specified \code{MSPeakLists}. In the case of a subset only the
remaining feature groups in the subset are considered.}

\item{MSPeakLists}{An \code{\link{MSPeakLists}} object that was generated for the supplied \code{fGroups}.}

\item{relMzDev}{Maximum relative deviation between the measured and candidate formula \emph{m/z} values (in ppm).
Sets the \option{ppm} command line option.}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
  Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}. If the \code{featureGroups} object has
  adduct annotations then these are used if \code{adducts=NULL}.

  \setsWF The \code{adduct} argument is not supported for sets workflows, since the
  adduct annotations will then always be used.}

\item{elements}{Elements to be considered for formulae calculation. This will heavily affects the number of
candidates! Always try to work with a minimal set by excluding elements you don't expect. Sets the \option{el}
command line option.}

\item{hetero}{Only consider formulae with at least one hetero atom. Sets the \option{het} commandline option.}

\item{oc}{Only consider organic formulae (\emph{i.e.} with at least one carbon atom). Sets the \option{oc}
commandline option.}

\item{thrMS, thrMSMS, thrComb}{Sets the thresholds for the \command{GenForm} MS score (\code{isoScore}), MS/MS score
(\code{MSMSScore}) and combined score (\code{combMatch}). Sets the \option{thms}/\option{thmsms}/\option{thcomb}
command line options, respectively. Set to \code{NULL} for no threshold.}

\item{maxCandidates}{If this number of candidates are found then \command{GenForm} aborts any further formula
calculations. The number of candidates is determined \emph{after} any formula filters, hence, the properties and
'quality' of the candidates is influenced by options such as \code{oc} and \code{thrMS} arguments. Note that this
is different than \code{topMost}, which selects the candidates after \command{GenForm} finished. Sets the
\option{max} command line option. Set to \samp{0} or \code{Inf} for no maximum.}

\item{extraOpts}{An optional character vector with any other command line options that will be passed to
\command{GenForm}. See the \verb{GenForm options} section for all available command line options.}

\item{calculateFeatures}{If \code{TRUE} fomulae are first calculated for all features
prior to feature group assignment (see \verb{Candidate assignment} in \code{\link{generateFormulas}}).}

\item{featThreshold}{If \code{calculateFeatures=TRUE}: minimum presence (\samp{0-1}) of a formula in all features
before it is considered as a candidate for a feature group. For instance, \code{featThreshold=0.75} dictates that a
formula should be present in at least 75\% of the features inside a feature group.}

\item{featThresholdAnn}{As \code{featThreshold}, but only considers features with annotations. For instance,
\code{featThresholdAnn=0.75} dictates that a formula should be present in at least 75\% of the features with
annotations inside a feature group. @param topMost Only keep this number of candidates
(per feature group) with highest score.}

\item{absAlignMzDev}{When the group formula annotation consensus is made from feature annotations, the \emph{m/z}
values of annotated MS/MS fragments may slightly deviate from those of the corresponding group MS/MS peak list. The
\code{absAlignMzDev} argument specifies the maximum \emph{m/z} window used to re-align the mass peaks.}

\item{MSMode}{Whether formulae should be generated only from MS data (\code{"ms"}), MS/MS data (\code{"msms"}) or
both (\code{"both"}). Selecting \code{"both"} will fall back to formula calculation with only MS data in case no
MS/MS data is available.}

\item{isolatePrec}{Settings used for isolation of precursor mass peaks and their isotopes. This isolation is highly
important for accurate isotope scoring of candidates, as non-relevant mass peaks will dramatically decrease the
score. The value of \code{isolatePrec} should either be a \code{list} with parameters (see the
\code{\link[=filter,MSPeakLists-method]{filter method}} for \code{MSPeakLists} for more details), \code{TRUE} for
default parameters or \code{FALSE} for no isolation (\emph{e.g.} when you already performed isolation with the
\code{filter} method). The \code{z} parameter (charge) is automatically deduced from the adduct used for annotation
(unless \code{isolatePrec=FALSE}), hence any custom \code{z} setting is ignored.}

\item{timeout}{Maximum time (in seconds) that a \command{GenForm} command is allowed to execute. If this time is
exceeded a warning is emitted and the command is terminated. See the notes section for more information on the need
of timeouts.}

\item{topMost}{Only keep this number of candidates (per feature group) with highest
score.}

\item{batchSize}{Maximum number of \command{GenForm} commands that should be run sequentially in each parallel
process. Combining commands with short runtimes (such as \command{GenForm}) can significantly increase parallel
performance. For more information see \code{\link{executeMultiProcess}}. Note that this is ignored if
\option{patRoon.MP.method="future"}.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{setThreshold}{\setsWF Minimum abundance for a candidate among all sets (\samp{0-1}). For instance, a value of
\samp{1} means that the candidate needs to be present in all the set data.}

\item{setThresholdAnn}{\setsWF As \code{setThreshold}, but only taking into account the set data that contain
annotations for the feature group of the candidate.}
}
\value{
A \code{\link{formulas}} object containing all generated formulae.
}
\description{
Uses \href{https://sourceforge.net/projects/genform/}{GenForm} to generate chemical formula candidates.
}
\details{
This function uses genform to generate formula candidates. This function is called when calling \code{generateFormulas} with
  \code{algorithm="genform"}.

When MS/MS data is available it will be used to score candidate formulae by presence of 'fitting' fragments.
}
\note{
This function always sets the \option{exist} and \option{oei} \command{GenForm} command line options.

  Formula calculation with \command{GenForm} may produce an excessive number of candidates for high \emph{m/z} values
  (\emph{e.g.} above 600) and/or many elemental combinations (set by \code{elements}). In this scenario formula
  calculation may need a very long time. Timeouts are used to avoid excessive computational times by terminating long
  running commands (set by the \code{timeout} argument).
}
\section{GenForm options}{
 Below is a list of options (generated by running \command{GenForm} without commandline
  options) which can be set by the \code{extraOpts} parameter.

 \preformatted{Formula calculation from MS and MS/MS data as described in
Meringer et al (2011) MATCH Commun Math Comput Chem 65: 259-290
Usage: GenForm ms=<filename> [msms=<filename>] [out=<filename>]
        [exist[=mv]] [m=<number>] [ion=-e|+e|-H|+H|+Na] [cha=<number>]
        [ppm=<number>] [msmv=ndp|nsse|nsae] [acc=<number>] [rej=<number>]
        [thms=<number>] [thmsms=<number>] [thcomb=<number>]
        [sort[=ppm|msmv|msmsmv|combmv]] [el=<elements> [oc]] [ff=<fuzzy formula>]
        [vsp[=<even|odd>]] [vsm2mv[=<value>]] [vsm2ap2[=<value>]] [hcf] [kfer[=ex]]
        [wm[=lin|sqrt|log]] [wi[=lin|sqrt|log]] [exp=<number>] [oei]
        [dbeexc=<number>] [ivsm2mv=<number>] [vsm2ap2=<number>]
        [oms[=<filename>]] [omsms[=<filename>]] [oclean[=<filename>]]
        [analyze [loss] [intens]] [dbe] [cm] [pc] [sc] [max]
Explanation:
        ms      : filename of MS data (*.txt)
        msms    : filename of MS/MS data (*.txt)
        out     : output generated formulas
        exist   : allow only molecular formulas for that at least one
                  structural formula exists;overrides vsp, vsm2mv, vsm2ap2;
                  argument mv enables multiple valencies for P and S
        m       : experimental molecular mass (default: mass of MS basepeak)
        ion     : type of ion measured (default: M+H)
        ppm     : accuracy of measurement in parts per million (default: 5)
        msmv    : MS match value based on normalized dot product, normalized
                  sum of squared or absolute errors (default: nsae)
        acc     : allowed deviation for full acceptance of MS/MS peak in ppm
                  (default: 2)
        rej     : allowed deviation for total rejection of MS/MS peak in ppm
                  (default: 4)
        thms    : threshold for the MS match value
        thmsms  : threshold for the MS/MS match value
        thcomb  : threshold for the combined match value
        sort    : sort generated formulas according to mass deviation in ppm,
                  MS match value, MS/MS match value or combined match value
        el      : used chemical elements (default: CHBrClFINOPSSi)
        oc      : only organic compounds, i.e. with at least one C atom
        ff      : overwrites el and oc and uses fuzzy formula for limits of
                  element multiplicities
        het     : formulas must have at least one hetero atom
        vsp     : valency sum parity (even for graphical formulas)
        vsm2mv  : lower bound for valency sum - 2 * maximum valency
                  (>=0 for graphical formulas)
        vsm2ap2 : lower bound for valency sum - 2 * number of atoms + 2
                  (>=0 for graphical connected formulas)
        hcf     : apply Heuerding-Clerc filter
        kfer    : apply Kind-Fiehn element ratio (extended) ranges
        wm      : m/z weighting for MS/MS match value
        wi      : intensity weighting for MS/MS match value
        exp     : exponent used, when wi is set to log
        oei     : allow odd electron ions for explaining MS/MS peaks
        dbeexc  : excess of double bond equivalent for ions
        ivsm2mv : lower bound for valency sum - 2 * maximum valency
                  for fragment ions
        ivsm2ap2: lower bound for valency sum - 2 * number of atoms + 2
                  for fragment ions
        oms     : write scaled MS peaks to output
        omsms   : write weighted MS/MS peaks to output
        oclean  : write explained MS/MS peaks to output
        analyze : write explanations for MS/MS peaks to output
        loss    : for analyzing MS/MS peaks write losses instead of fragments
        intens  : write intensities of MS/MS peaks to output
        dbe     : write double bond equivalents to output
        cm      : write calculated ion masses to output
        pc      : output match values in percent
        sc      : strip calculated isotope distributions
        noref   : hide the reference information
        max     : maximum number of final candidates (0 is no limit)
}
}

\section{Parallelization}{
 generateFormulasGenForm uses multiprocessing to parallelize
  computations. Please see the parallelization section in the handbook for
  more details and \link[=patRoon-package]{patRoon options} for configuration
  options.

 When \code{futures} are used for parallel processing (\code{patRoon.MP.method="future"}),
  calculations with \command{GenForm} are done with batch mode disabled (see \code{batchSize} argument), which
  generally limit overall performance.
}

\references{
\insertRef{Meringer2011}{patRoon}
}
\seealso{
\code{\link{generateFormulas}} for more details and other algorithms.
}
