% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mspeaklists-bruker.R
\name{generateMSPeakListsDA}
\alias{generateMSPeakListsDA}
\alias{generateMSPeakListsDA,featureGroups-method}
\alias{generateMSPeakListsDA,featureGroupsSet-method}
\title{Generate peak lists with Bruker DataAnalysis}
\usage{
generateMSPeakListsDA(fGroups, ...)

\S4method{generateMSPeakListsDA}{featureGroups}(
  fGroups,
  bgsubtr = TRUE,
  maxMSRtWindow = 5,
  minMSIntensity = 500,
  minMSMSIntensity = 500,
  clear = TRUE,
  close = TRUE,
  save = close,
  MSMSType = "MSMS",
  avgFGroupParams = getDefAvgPListParams()
)

\S4method{generateMSPeakListsDA}{featureGroupsSet}(fGroups, ...)
}
\arguments{
\item{fGroups}{The \code{\link{featureGroups}} object from which MS peak lists should be extracted.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{bgsubtr}{If \code{TRUE} background will be subtracted using the 'spectral' algorithm.}

\item{maxMSRtWindow}{Maximum chromatographic peak window used for spectrum averaging (in seconds, +/- retention
time). If \code{NULL} all spectra from a feature will be taken into account. Lower to decrease processing time.}

\item{minMSIntensity, minMSMSIntensity}{Minimum intensity for peak lists obtained with DataAnalysis. Highly
recommended to set \samp{>0} as DA tends to report many very low intensity peaks.}

\item{clear}{Remove any existing chromatogram traces/mass spectra prior to making new ones.}

\item{close, save}{If \code{TRUE} then Bruker files are closed and saved after
processing with DataAnalysis, respectively. Setting \code{close=TRUE}
prevents that many analyses might be opened simultaneously in DataAnalysis,
which otherwise may use excessive memory or become slow. By default
\code{save} is \code{TRUE} when \code{close} is \code{TRUE}, which is
likely what you want as otherwise any processed data is lost.}

\item{MSMSType}{The type of MS/MS experiment performed: \code{"MSMS"} for MRM/AutoMSMS or \code{"BBCID"} for
broadband CID.}

\item{avgFGroupParams}{A \code{list} with parameters used for averaging of peak lists for feature groups. See
\code{\link{getDefAvgPListParams}} for more details.}
}
\value{
A \code{\link{MSPeakLists}} object.
}
\description{
Uses Bruker DataAnalysis to read the data needed to generate MS peak lists.
}
\details{
This function uses Bruker DataAnalysis to generate MS peak lists. This function is called when calling \code{generateMSPeakLists} with
  \code{algorithm="bruker"}.

The MS data should be in the Bruker data format (\file{.d}). This function leverages DataAnalysis
  functionality to support averaging of spectra, background subtraction and identification of isotopes. In order to
  obtain mass spectra TICs will be added in DataAnalysis of the MS and relevant MS/MS signals.
}
\note{
The \option{Component} column should be active (Method-->Parameters-->Layouts-->Mass List Layout) in order to
  add isotopologue information.

If any errors related to \command{DCOM} appear it might be necessary to
  terminate DataAnalysis (note that DataAnalysis might still be running as a
  background process). The \command{ProcessCleaner} application installed
  with DataAnalayis can be used for this.
}
\seealso{
\code{\link{generateMSPeakLists}} for more details and other algorithms.
}
