% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/deprecated-mspeaklists-bruker.R
\name{generateMSPeakListsDAFMF}
\alias{generateMSPeakListsDAFMF}
\alias{generateMSPeakListsDAFMF,featureGroups-method}
\alias{generateMSPeakListsDAFMF,featureGroupsSet-method}
\title{Generate peak lists with Bruker DataAnalysis from bruker features (deprecated)}
\usage{
generateMSPeakListsDAFMF(fGroups, ...)

\S4method{generateMSPeakListsDAFMF}{featureGroups}(
  fGroups,
  minMSIntensity = 500,
  minMSMSIntensity = 500,
  close = TRUE,
  save = close,
  avgFGroupParams = getDefAvgPListParams()
)

\S4method{generateMSPeakListsDAFMF}{featureGroupsSet}(fGroups, ...)
}
\arguments{
\item{fGroups}{The \code{\link{featureGroups}} object for which MS peak lists should be generated.}

\item{...}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{minMSIntensity, minMSMSIntensity}{Minimum intensity for peak lists obtained with DataAnalysis. Highly
recommended to set \samp{>0} as DA tends to report many very low intensity peaks.}

\item{close, save}{If \code{TRUE} then Bruker files are closed and saved after
processing with DataAnalysis, respectively. Setting \code{close=TRUE}
prevents that many analyses might be opened simultaneously in DataAnalysis,
which otherwise may use excessive memory or become slow. By default
\code{save} is \code{TRUE} when \code{close} is \code{TRUE}, which is
likely what you want as otherwise any processed data is lost.}

\item{avgFGroupParams}{A \code{list} with parameters used for averaging of peak lists for feature groups. See
\code{\link{getDefAvgPListParams}} for more details.}
}
\value{
An \code{\link{MSPeakLists}} object.
}
\description{
Uses 'compounds' that were generated by the Find Molecular Features (FMF) algorithm of Bruker DataAnalysis to extract
MS peak lists. This function is now deprecated, please use \code{\link{generateMSPeakLists}} instead.
}
\details{
This function is similar to \code{\link{generateMSPeakListsDA}}, but uses 'compounds' that were generated by
  the Find Molecular Features (FMF) algorithm to extract MS peak lists. This is generally much faster , however, it
  only works when features were obtained with the \code{\link{findFeaturesBruker}} function. Since all MS spectra are
  generated in advance by Bruker DataAnalysis, only few parameters exist to customize its operation.
}
\note{
If any errors related to \command{DCOM} appear it might be necessary to
  terminate DataAnalysis (note that DataAnalysis might still be running as a
  background process). The \command{ProcessCleaner} application installed
  with DataAnalayis can be used for this.
}
\keyword{internal}
