% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mspeaklists-mzr.R
\name{generateMSPeakListsMzR}
\alias{generateMSPeakListsMzR}
\alias{generateMSPeakListsMzR,featureGroups-method}
\alias{generateMSPeakListsMzR,featureGroupsSet-method}
\title{Generate peak lists with mzR}
\usage{
generateMSPeakListsMzR(fGroups, ...)

\S4method{generateMSPeakListsMzR}{featureGroups}(
  fGroups,
  maxMSRtWindow = 5,
  precursorMzWindow = 4,
  topMost = NULL,
  avgFeatParams = getDefAvgPListParams(),
  avgFGroupParams = getDefAvgPListParams()
)

\S4method{generateMSPeakListsMzR}{featureGroupsSet}(fGroups, ...)
}
\arguments{
\item{fGroups}{The \code{\link{featureGroups}} object from which MS peak lists should be extracted.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{maxMSRtWindow}{Maximum chromatographic peak window used for spectrum averaging (in seconds, +/- retention
time). If \code{NULL} all spectra from a feature will be taken into account. Lower to decrease processing time.}

\item{precursorMzWindow}{The \emph{m/z} window (in Da) to find MS/MS spectra of a precursor. This is typically used
for Data-Dependent like MS/MS data and should correspond to the isolation \emph{m/z} window (\emph{i.e.} +/- the
precursor \emph{m/z}) that was used to collect the data. For Data-Independent MS/MS experiments, where precursor
ions are not isolated prior to fragmentation (\emph{e.g.} bbCID, MSe, all-ion, ...) the value should be
\code{NULL}.}

\item{topMost}{Only extract MS peak lists from a maximum of \code{topMost} analyses with highest intensity. If
\code{NULL} all analyses will be used.}

\item{avgFeatParams}{Parameters used for averaging MS peak lists of individual features. Analogous to
\code{avgFGroupParams}.}

\item{avgFGroupParams}{A \code{list} with parameters used for averaging of peak lists for feature groups. See
\code{\link{getDefAvgPListParams}} for more details.}
}
\value{
A \code{\link{MSPeakLists}} object.
}
\description{
Uses the \pkg{mzR} package to read the MS data needed for MS peak lists.
}
\details{
This function uses mzR to generate MS peak lists. This function is called when calling \code{generateMSPeakLists} with
  \code{algorithm="mzr"}.

The MS data files should be either in \file{.mzXML} or \file{.mzML} format.

The input MS data files need to be centroided. The \code{\link{convertMSFiles}} function can be used to
  centroid data.
}
\references{
\addCitations{mzR}{1} \cr\cr

  \addCitations{mzR}{2} \cr\cr

  \addCitations{mzR}{3} \cr\cr

  \addCitations{mzR}{4} \cr\cr

  \addCitations{mzR}{5} \cr\cr
}
\seealso{
\code{\link{generateMSPeakLists}} for more details and other algorithms.
}
