% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP.R
\name{generateTPs}
\alias{generateTPs}
\title{Generation of transformation products (TPs)}
\usage{
generateTPs(algorithm, ...)
}
\arguments{
\item{algorithm}{A character string describing the algorithm that should be
used: \code{"biotransformer"}, \code{"logic"}, \code{"library"}, \code{"cts"}}

\item{\dots}{Any parameters to be passed to the selected TP generation algorithm.}
}
\value{
A \code{\link{transformationProducts}} (derived) object containing all generated TPs.
}
\description{
Functionality to automatically obtain transformation products for a given set of parent compounds.
}
\details{
\code{generateTPs} is a generic function that will generate transformation products by one of the supported algorithms. The actual
  functionality is provided by algorithm specific functions such as \code{generateTPsBioTransformer} and \code{generateTPsLogic}. While these
  functions may be called directly, \code{generateTPs} provides a generic interface and is therefore usually preferred.
}
\seealso{
The \code{\link{transformationProducts}} output class and its methods and the algorithm specific functions:
  \code{\link{generateTPsBioTransformer}}, \code{\link{generateTPsLogic}}, \code{\link{generateTPsLibrary}}, \code{\link{generateTPsCTS}}

The derived class \code{\link{transformationProductsStructure}} for more specific methods to post-process TP
  data.
}
