% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-ann_comp.R
\name{generateTPsAnnComp}
\alias{generateTPsAnnComp}
\title{Obtain transformation products (TPs) from compound annotation candidates}
\usage{
generateTPsAnnComp(
  parents,
  compounds,
  TPsRef = NULL,
  fGroupsComps = NULL,
  minRTDiff = 20,
  minFitFormula = 0.94,
  minFitCompound = 0,
  minSimSusp = 0,
  minFitCompOrSimSusp = c(0.54, 0.65),
  extraOptsFMCSR = NULL,
  skipInvalid = TRUE,
  prefCalcChemProps = TRUE,
  neutralChemProps = FALSE,
  neutralizeTPs = TRUE,
  TPStructParams = getDefTPStructParams(),
  parallel = TRUE
)
}
\arguments{
\item{parents}{The parents for which transformation products should be obtained. This can be \itemize{

  \item a suspect list (see \link[=suspect-screening]{suspect screening} for more information)

  \item the output of \code{\link{screenSuspects}} in which case the suspects hits are used as parents

  

  

  }

  The parents need to have \acronym{SMILES} or \acronym{InChI} information available.}

\item{compounds}{The \code{\link{compounds}} object containing the compound candidates.}

\item{TPsRef}{A \code{\link{transformationProductsStructure}} object containing suspect TP candidates obtained for
the same \code{parents} from another TP generation algorithm (\emph{e.g.} \code{\link{generateTPsBioTransformer}}).
This is used for the calculation of \code{simSusps} (see Details). Set to \code{NULL} to skip its calculation.}

\item{fGroupsComps}{The \code{\link{featureGroups}} object for which the \code{compounds} were generated. This is
used to obtain retention times for the calculation for \link[=retDir]{retention order directions}. Set to
\code{NULL} to skip its calculation.}

\item{minRTDiff}{Minimum retention time (in seconds) difference between the parent and a TP to calculate the
\link[=retDir]{retention order direction}. Candidates with unexpected retention orders are filtered out.}

\item{minFitFormula, minFitCompound, minSimSusp}{Thresholds to filter out unlikely candidates. For \code{fitFormula}:
see \code{\link{generateTPsAnnForm}}, for the others see the Details section.}

\item{minFitCompOrSimSusp}{A two-sized numeric vector specifying the thresholds for \code{fitCompound} \emph{or}
\code{simSusp}, respectively.}

\item{extraOptsFMCSR}{A \code{list} with additional options passed to the \code{\link[fmcsR:fmcs]{fmcsR::fmcs}}
function. The following defaults are set: \code{au=1, bu=4, matching.mode="aromatic"}.}

\item{skipInvalid}{If set to \code{TRUE} then the parents will be skipped (with a warning) for which insufficient
information (\emph{e.g.} SMILES) is available.}

\item{prefCalcChemProps}{If \code{TRUE} then calculated chemical properties such as the formula and
\acronym{InChIKey} are preferred over what is already present in the parent suspect list. For efficiency reasons it is
recommended to set this to \code{TRUE}. See the \verb{Validating and calculating chemical properties} section for
more details.}

\item{neutralChemProps}{If \code{TRUE} then the neutral form of the molecule is considered to calculate
\acronym{SMILES}, formulae etc. Enabling this may improve feature matching when considering common adducts
(\emph{e.g.} \code{[M+H]+}, \code{[M-H]-}). See the \verb{Validating and calculating chemical properties} section
for more details.}

\item{neutralizeTPs}{If \code{TRUE} then all resulting TP structure information is
neutralized. This argument has a similar meaning as \code{neutralChemProps}. This is defaulted to \code{TRUE} for
prediction algorithms, as these may output charged molecules. \strong{NOTE:} if neutralization results in
duplicate TPs, \emph{i.e.} when the neutral form of the TP was also generated by the algorithm, then the
neutralized TP \emph{will be removed}.}

\item{TPStructParams}{Parameters that influence the calculation of structural properties. See
\code{\link{getDefTPStructParams}}.}

\item{parallel}{If set to \code{TRUE} then code is executed in parallel through the \CRANpkg{future} package. Please
see the parallelization section in the handbook for more details.}
}
\value{
\code{generateTPsAnnComp} returns an object of the class \code{\link{transformationProductsAnnComp}}. Please
  see its documentation for \emph{e.g.} filtering steps that can be performed on this object.
}
\description{
Transforms and prioritizes \link[=compounds]{compound annotation candidates} to obtain TPs.
}
\details{
This function uses compound annotations to obtain transformation products. This function is called when calling \code{generateTPs} with
  \code{algorithm="ann_comp"}.

The \code{generateTPsAnnComp} function implements the unknown TP screening from compound candidates approach
  as described in \insertCite{Helmus2025}{patRoon}. This algorithm does not rely on any known or predicted TPs and is
  therefore suitable for 'full non-target' workflows. \emph{All} compound candidates are considered as potential TPs
  and are ranked by the \code{TP score}:

  \deqn{TP score = max(fitCompound,simSusp) + annSim}

With: \itemize{

  \item \code{annSim}: the \link[=id-conf]{annotation similarity}

  \item \code{fitCompound}: the structural fit of the compound candidate into the parent (or vice versa, maximum is
  taken). Calculated as the \code{"Overlap coefficient"} with \code{\link[fmcsR:fmcs]{fmcsR::fmcs}}. The molecular
  data is prepared with \CRANpkg{rcdk} and \pkg{ChemmineR}.

  \item \code{simSusp}: the maximum structural similarity with TP suspect candidates for this parent, \emph{i.e.}
  obtained from other algorithms of \code{\link{generateTPs}}). The calculation is configured by the
  \link[=getDefTPStructParams]{TPStructParams}.

}

  To speed up the calculation process, several thresholds are applied to rule out unlikely candidates. These
  thresholds are defaulted to those derived in \insertCite{Helmus2025}{patRoon}. Nevertheless, calculations can take
  a very long time (multiple hours), especially when processing large numbers of candidates from \emph{e.g.} PubChem.

  Unlike most other TP generation algorithms, no additional suspect screening step is required.
}
\note{
Setting \code{parallel=TRUE} can speed up calculations considerably on multi-core systems. but will also add to
  RAM usage. Furthermore, parallelization is only favorable for long calculations due to the overhead of setting up
  multiple \R processes. Note that the parallel workers must be on the same system, \emph{i.e.} this will not work on
  \emph{e.g.} clusters.

  It is possible that candidates are equal to their parent. To remove these the \code{removeParentIsomers}
  \code{\link[=filter,transformationProductsStructure-method]{filter}} can be used afterwards.
}
\section{Validating and calculating chemical properties}{
 Chemical properties such as \acronym{SMILES},
  \acronym{InChIKey} and formulae in the parent suspect list are automatically validated and calculated if missing/invalid.

  The internal validation/calculation process performs the following steps: \itemize{

  \item Validation of \acronym{SMILES}, \acronym{InChI}, \acronym{InChIKey} and formula data (if present). Invalid
  entries will be set to \code{NA}.

  \item If \code{neutralChemProps=TRUE} then chemical data (\acronym{SMILES}, formulae etc.) is neutralized by
  (de-)protonation (using the \command{--neutralized} option of \command{OpenBabel}). An additional column
  \code{molNeutralized} is added to mark those molecules that were neutralized. Note that neutralization requires
  either \acronym{SMILES} or \acronym{InChI} data to be available.

  \item The \acronym{SMILES} and \acronym{InChI} data are used to calculate missing or invalid \acronym{SMILES},
  \acronym{InChI}, \acronym{InChIKey} and formula data. If \code{prefCalcChemProps=TRUE} then existing
  \acronym{InChIKey} and formula data is overwritten by calculated values whenever possible.

  \item The chemical formulae which were \emph{not} calculated are verified and normalized. This process may be time
  consuming, and is potentially largely avoided by setting \code{prefCalcChemProps=TRUE}.

  \item Neutral masses are calculated for missing values (\code{prefCalcChemProps=FALSE}) or whenever possible
  (\code{prefCalcChemProps=TRUE}).

  }

  Note that calculation of formulae for molecules that are isotopically labelled is currently only supported for
  deuterium (2H) elements.

  This functionality relies heavily on \href{https://github.com/openbabel/openbabel}{OpenBabel}, please make sure it
  is installed.
}

\references{
\insertRef{OBoyle2011}{patRoon}

\insertAllCited{} \cr\cr \insertRef{Wang2013}{patRoon} \cr\cr \addCitations{rcdk} \cr\cr \insertRef{Cao2008}{patRoon}
}
\seealso{
\code{\link{generateTPs}} for more details and other algorithms.
}
