% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-ann_form.R
\name{generateTPsAnnForm}
\alias{generateTPsAnnForm}
\title{Obtain transformation products (TPs) from formula annotation candidates}
\usage{
generateTPsAnnForm(
  parents,
  formulas,
  minFitFormula = 0.94,
  skipInvalid = TRUE,
  prefCalcChemProps = TRUE,
  neutralChemProps = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{parents}{The parents for which transformation products should be obtained. This can be \itemize{

  \item a suspect list (see \link[=suspect-screening]{suspect screening} for more information)

  \item the output of \code{\link{screenSuspects}} in which case the suspects hits are used as parents

  

  

  }

  The parents need to have formula information available.}

\item{formulas}{The \code{\link{formulas}} object containing the formula candidates.}

\item{minFitFormula}{Minimum \code{fitFormula} (see Details sections) to filter out unlikely candidates.}

\item{skipInvalid}{If set to \code{TRUE} then the parents will be skipped (with a warning) for which insufficient
information (\emph{e.g.} SMILES) is available.}

\item{prefCalcChemProps}{If \code{TRUE} then calculated chemical properties such as the formula and
\acronym{InChIKey} are preferred over what is already present in the parent suspect list. For efficiency reasons it is
recommended to set this to \code{TRUE}. See the \verb{Validating and calculating chemical properties} section for
more details.}

\item{neutralChemProps}{If \code{TRUE} then the neutral form of the molecule is considered to calculate
\acronym{SMILES}, formulae etc. Enabling this may improve feature matching when considering common adducts
(\emph{e.g.} \code{[M+H]+}, \code{[M-H]-}). See the \verb{Validating and calculating chemical properties} section
for more details.}

\item{parallel}{If set to \code{TRUE} then code is executed in parallel through the \CRANpkg{future} package. Please
see the parallelization section in the handbook for more details.}
}
\value{
\code{generateTPsAnnForm} returns an object of the class \code{\link{transformationProductsAnnForm}}. Please
  see its documentation for \emph{e.g.} filtering steps that can be performed on this object.
}
\description{
Transforms and prioritizes \link[=formulas]{formula annotation candidates} to obtain TPs.
}
\details{
This function uses formula annotations to obtain transformation products. This function is called when calling \code{generateTPs} with
  \code{algorithm="ann_form"}.

The \code{generateTPsAnnForm} function implements the unknown TP screening from formula candidates approach
  as described in \insertCite{Helmus2025}{patRoon}. This algorithm does not rely on any known or predicted TPs and is
  therefore suitable for 'full non-target' workflows. \emph{All} formula candidates are considered as potential TPs
  and are ranked by the \code{TP score}: \deqn{TP score = fitFormula + annSim}

With: \itemize{

  \item \code{annSim}: the \link[=id-conf]{annotation similarity}

  \item \code{fitFormula}: the common element count divided by the total element count for the formulae of the
  parent/TP or TP/parent (maximum is taken)

}

  To speed up the calculation process, a threshold for \code{fitFormula} is applied to rule out unlikely candidates.
  The default was derived in \insertCite{Helmus2025}{patRoon}.

  Unlike most other TP generation algorithms, no additional suspect screening step is required.
}
\note{
Setting \code{parallel=TRUE} may speed up calculations, but is only favorable for long calculations due to the
  overhead of setting up multiple \R processes.
}
\section{Validating and calculating chemical properties}{
 Chemical properties such as \acronym{SMILES},
  \acronym{InChIKey} and formulae in the parent suspect list are automatically validated and calculated if missing/invalid.

  The internal validation/calculation process performs the following steps: \itemize{

  \item Validation of \acronym{SMILES}, \acronym{InChI}, \acronym{InChIKey} and formula data (if present). Invalid
  entries will be set to \code{NA}.

  \item If \code{neutralChemProps=TRUE} then chemical data (\acronym{SMILES}, formulae etc.) is neutralized by
  (de-)protonation (using the \command{--neutralized} option of \command{OpenBabel}). An additional column
  \code{molNeutralized} is added to mark those molecules that were neutralized. Note that neutralization requires
  either \acronym{SMILES} or \acronym{InChI} data to be available.

  \item The \acronym{SMILES} and \acronym{InChI} data are used to calculate missing or invalid \acronym{SMILES},
  \acronym{InChI}, \acronym{InChIKey} and formula data. If \code{prefCalcChemProps=TRUE} then existing
  \acronym{InChIKey} and formula data is overwritten by calculated values whenever possible.

  \item The chemical formulae which were \emph{not} calculated are verified and normalized. This process may be time
  consuming, and is potentially largely avoided by setting \code{prefCalcChemProps=TRUE}.

  \item Neutral masses are calculated for missing values (\code{prefCalcChemProps=FALSE}) or whenever possible
  (\code{prefCalcChemProps=TRUE}).

  }

  Note that calculation of formulae for molecules that are isotopically labelled is currently only supported for
  deuterium (2H) elements.

  This functionality relies heavily on \href{https://github.com/openbabel/openbabel}{OpenBabel}, please make sure it
  is installed.
}

\references{
\insertRef{OBoyle2011}{patRoon}

\insertAllCited{}
}
\seealso{
\code{\link{generateTPs}} for more details and other algorithms.
}
