% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-CTS.R
\name{generateTPsCTS}
\alias{generateTPsCTS}
\title{Obtain transformation products (TPs) with Chemical Transformation Simulator (CTS)}
\usage{
generateTPsCTS(
  parents,
  transLibrary,
  generations = 1,
  errorRetries = 3,
  skipInvalid = TRUE,
  prefCalcChemProps = TRUE,
  neutralChemProps = FALSE,
  neutralizeTPs = TRUE,
  TPStructParams = getDefTPStructParams(),
  parallel = TRUE
)
}
\arguments{
\item{parents}{The parents for which transformation products should be obtained. This can be \itemize{

  \item a suspect list (see \link[=suspect-screening]{suspect screening} for more information)

  \item the output of \code{\link{screenSuspects}} in which case the suspects hits are used as parents

  \item a \code{\link{compounds}} object in which case all candidates are used
  parents

  

  }

  The parents need to have \acronym{SMILES} or \acronym{InChI} information available.}

\item{transLibrary}{A \code{character} specifying which transformation library should be used. Currently supported
are: \code{"hydrolysis"}, \code{"abiotic_reduction"}, \code{"photolysis_unranked"}, \code{"photolysis_ranked"},
\code{"mammalian_metabolism"}, \code{"combined_abioticreduction_hydrolysis"},
\code{"combined_photolysis_abiotic_hydrolysis"}, \code{"pfas_environmental"}, \code{"pfas_metabolism"}.}

\item{generations}{An \code{integer} that specifies the number of transformation generations to predict.}

\item{errorRetries}{The maximum number of connection retries. Sets the \code{times} argument to the
\code{\link[httr:RETRY]{http::RETRY}} function.}

\item{skipInvalid}{If set to \code{TRUE} then the parents will be skipped (with a warning) for which insufficient
information (\emph{e.g.} SMILES) is available.}

\item{prefCalcChemProps}{If \code{TRUE} then calculated chemical properties such as the formula and
\acronym{InChIKey} are preferred over what is already present in the parent suspect list. For efficiency reasons it is
recommended to set this to \code{TRUE}. See the \verb{Validating and calculating chemical properties} section for
more details.}

\item{neutralChemProps}{If \code{TRUE} then the neutral form of the molecule is considered to calculate
\acronym{SMILES}, formulae etc. Enabling this may improve feature matching when considering common adducts
(\emph{e.g.} \code{[M+H]+}, \code{[M-H]-}). See the \verb{Validating and calculating chemical properties} section
for more details.}

\item{neutralizeTPs}{If \code{TRUE} then all resulting TP structure information is
neutralized. This argument has a similar meaning as \code{neutralChemProps}. This is defaulted to \code{TRUE} for
prediction algorithms, as these may output charged molecules. \strong{NOTE:} if neutralization results in
duplicate TPs, \emph{i.e.} when the neutral form of the TP was also generated by the algorithm, then the
neutralized TP \emph{will be removed}.}

\item{TPStructParams}{Parameters that influence the calculation of structural properties. See
\code{\link{getDefTPStructParams}}.}

\item{parallel}{If set to \code{TRUE} then code is executed in parallel through the \CRANpkg{future} package. Please
see the parallelization section in the handbook for more details.}
}
\value{
The TPs are stored in an object derived from the \code{\link{transformationProductsStructure}} class.
}
\description{
Uses \href{https://qed.epa.gov/cts/}{Chemical Transformation Simulator (CTS)} to predict TPs.
}
\details{
This function uses CTS to obtain transformation products. This function is called when calling \code{generateTPs} with
  \code{algorithm="cts"}.

This function uses the \CRANpkg{httr} package to access the Web API of CTS for automatic TP prediction.
  Hence, an Internet connection is mandatory. Please take care to not 'abuse' the CTS servers, \emph{e.g.} by running
  very large batch calculations in parallel, as this may result in rejected connections.
}
\note{
When the \code{parents} argument is a \code{\link{compounds}} object, the
candidate library \code{identifier} is used in case the candidate has no defined \code{compoundName}.
}
\section{Validating and calculating chemical properties}{
 Chemical properties such as \acronym{SMILES},
  \acronym{InChIKey} and formulae in the parent suspect list are automatically validated and calculated if missing/invalid.

  The internal validation/calculation process performs the following steps: \itemize{

  \item Validation of \acronym{SMILES}, \acronym{InChI}, \acronym{InChIKey} and formula data (if present). Invalid
  entries will be set to \code{NA}.

  \item If \code{neutralChemProps=TRUE} then chemical data (\acronym{SMILES}, formulae etc.) is neutralized by
  (de-)protonation (using the \command{--neutralized} option of \command{OpenBabel}). An additional column
  \code{molNeutralized} is added to mark those molecules that were neutralized. Note that neutralization requires
  either \acronym{SMILES} or \acronym{InChI} data to be available.

  \item The \acronym{SMILES} and \acronym{InChI} data are used to calculate missing or invalid \acronym{SMILES},
  \acronym{InChI}, \acronym{InChIKey} and formula data. If \code{prefCalcChemProps=TRUE} then existing
  \acronym{InChIKey} and formula data is overwritten by calculated values whenever possible.

  \item The chemical formulae which were \emph{not} calculated are verified and normalized. This process may be time
  consuming, and is potentially largely avoided by setting \code{prefCalcChemProps=TRUE}.

  \item Neutral masses are calculated for missing values (\code{prefCalcChemProps=FALSE}) or whenever possible
  (\code{prefCalcChemProps=TRUE}).

  }

  Note that calculation of formulae for molecules that are isotopically labelled is currently only supported for
  deuterium (2H) elements.

  This functionality relies heavily on \href{https://github.com/openbabel/openbabel}{OpenBabel}, please make sure it
  is installed.
}

\references{
\insertRef{OBoyle2011}{patRoon}

\insertRef{Wolfe2016}{patRoon} \cr\cr \insertRef{TebesStevens2017}{patRoon} \cr\cr
  \insertRef{Yuan2020}{patRoon} \cr\cr \insertRef{Yuan2021}{patRoon} \cr\cr \insertRef{OBoyle2011}{patRoon}
}
\seealso{
\code{\link{generateTPs}} for more details and other algorithms.

The website: \url{https://qed.epa.gov/cts/} and the
  \href{https://www.epa.gov/chemical-research/users-guide-chemical-transformation-simulator-cts}{CTS User guide}.
}
