% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-library.R
\name{generateTPsLibrary}
\alias{generateTPsLibrary}
\title{Obtain transformation products (TPs) from a library}
\usage{
generateTPsLibrary(
  parents = NULL,
  TPLibrary = NULL,
  generations = 1,
  skipInvalid = TRUE,
  prefCalcChemProps = TRUE,
  neutralChemProps = FALSE,
  neutralizeTPs = FALSE,
  matchParentsBy = "InChIKey",
  matchGenerationsBy = "InChIKey",
  TPStructParams = getDefTPStructParams()
)
}
\arguments{
\item{parents}{The parents for which transformation products should be obtained. This can be \itemize{

  \item a suspect list (see \link[=suspect-screening]{suspect screening} for more information)

  \item the output of \code{\link{screenSuspects}} in which case the suspects hits are used as parents

  \item a \code{\link{compounds}} object in which case all candidates are used
  parents

  \item \code{NULL} in which case all parents from the library are used.

  }

  The parents need to have \acronym{SMILES} or \acronym{InChI} information available.}

\item{TPLibrary}{If \code{NULL}, a default \href{https://doi.org/10.5281/zenodo.5644560}{PubChem} based library is
used. Otherwise, \code{TPLibrary} should be a \code{data.frame}. See the details below.}

\item{generations}{An \code{integer} that specifies the number of transformation generations. TPs for subsequent
iterations obtained by repeating the library search where the TPs from the previous generation are considered
parents.}

\item{skipInvalid}{If set to \code{TRUE} then the parents will be skipped (with a warning) for which insufficient
information (\emph{e.g.} SMILES) is available.}

\item{prefCalcChemProps}{If \code{TRUE} then calculated chemical properties such as the formula and
\acronym{InChIKey} are preferred over what is already present in the parent suspect list. For efficiency reasons it is
recommended to set this to \code{TRUE}. See the \verb{Validating and calculating chemical properties} section for
more details.}

\item{neutralChemProps}{If \code{TRUE} then the neutral form of the molecule is considered to calculate
\acronym{SMILES}, formulae etc. Enabling this may improve feature matching when considering common adducts
(\emph{e.g.} \code{[M+H]+}, \code{[M-H]-}). See the \verb{Validating and calculating chemical properties} section
for more details.}

\item{neutralizeTPs}{If \code{TRUE} then all resulting TP structure information is
neutralized. This argument has a similar meaning as \code{neutralChemProps}. This is defaulted to \code{TRUE} for
prediction algorithms, as these may output charged molecules. \strong{NOTE:} if neutralization results in
duplicate TPs, \emph{i.e.} when the neutral form of the TP was also generated by the algorithm, then the
neutralized TP \emph{will be removed}.}

\item{matchParentsBy}{A \code{character} that specifies how the input parents are matched with the data from the TP
library. Valid options are: \code{"InChIKey"}, \code{"InChIKey1"}, \code{"InChI"}, \code{"SMILES"},
\code{"formula"}, \code{"name"}. If the parent from the TP library is matched with multiple input parents then only
the first is considered.}

\item{matchGenerationsBy}{Similar to \code{matchParentsBy}, but specifies how parents/TPs are matched when
\code{generations>1}.}

\item{TPStructParams}{Parameters that influence the calculation of structural properties. See
\code{\link{getDefTPStructParams}}.}
}
\value{
The TPs are stored in an object derived from the \code{\link{transformationProductsStructure}} class.
}
\description{
Automatically obtains transformation products from a library.
}
\details{
This function uses a library to obtain transformation products. This function is called when calling \code{generateTPs} with
  \code{algorithm="library"}.

By default, a library is used that is based on data from
  \href{https://doi.org/10.5281/zenodo.5644560}{PubChem}. However, it also possible to use your own library.
}
\note{
When the \code{parents} argument is a \code{\link{compounds}} object, the
candidate library \code{identifier} is used in case the candidate has no defined \code{compoundName}.
}
\section{TP libraries}{
 The \code{TPLibrary} argument is used to specify a custom TP library. This should be a
  \code{data.frame} where each row specifies a TP for a parent, with the following columns: \itemize{

  \item \code{parent_name} and \code{TP_name}: The name of the parent/TP.

  \item \code{parent_SMILES} and \code{TP_SMILES} The SMILES of the parent/TP structure.

  \item \code{retDir} The expected \link[=retDir]{retention order direction}. (\strong{optional})

  For \code{generateTPsLibrary}: If not specified or \code{forceCalcRetDir=TRUE} from
  \code{\link[=getDefTPStructParams]{TPStructParams}}, then the \code{log P} values below may be used to calculate
  retention order directions.
  
  \item \code{parent_LogP} and \code{TP_LogP} The \code{log P} values for the parent/TP. (\strong{optional})

  \item \code{logPDiff} The difference between parent and TP \code{Log P} values. Ignored if \emph{both}
  \code{parent_LogP} and \code{TP_LogP} are specified. (\strong{optional})

  }

  Other columns are allowed, and will be included in the final object. Multiple TPs for a single parent are specified
  by repeating the value within \code{parent_} columns.
}

\section{Validating and calculating chemical properties}{
 Chemical properties such as \acronym{SMILES},
  \acronym{InChIKey} and formulae in the parent suspect list are automatically validated and calculated if missing/invalid.

  The internal validation/calculation process performs the following steps: \itemize{

  \item Validation of \acronym{SMILES}, \acronym{InChI}, \acronym{InChIKey} and formula data (if present). Invalid
  entries will be set to \code{NA}.

  \item If \code{neutralChemProps=TRUE} then chemical data (\acronym{SMILES}, formulae etc.) is neutralized by
  (de-)protonation (using the \command{--neutralized} option of \command{OpenBabel}). An additional column
  \code{molNeutralized} is added to mark those molecules that were neutralized. Note that neutralization requires
  either \acronym{SMILES} or \acronym{InChI} data to be available.

  \item The \acronym{SMILES} and \acronym{InChI} data are used to calculate missing or invalid \acronym{SMILES},
  \acronym{InChI}, \acronym{InChIKey} and formula data. If \code{prefCalcChemProps=TRUE} then existing
  \acronym{InChIKey} and formula data is overwritten by calculated values whenever possible.

  \item The chemical formulae which were \emph{not} calculated are verified and normalized. This process may be time
  consuming, and is potentially largely avoided by setting \code{prefCalcChemProps=TRUE}.

  \item Neutral masses are calculated for missing values (\code{prefCalcChemProps=FALSE}) or whenever possible
  (\code{prefCalcChemProps=TRUE}).

  }

  Note that calculation of formulae for molecules that are isotopically labelled is currently only supported for
  deuterium (2H) elements.

  This functionality relies heavily on \href{https://github.com/openbabel/openbabel}{OpenBabel}, please make sure it
  is installed.
}

\references{
\insertRef{OBoyle2011}{patRoon}
}
\seealso{
\code{\link{generateTPs}} for more details and other algorithms.
}
