% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/TP-logic.R
\name{generateTPsLogic}
\alias{generateTPsLogic}
\alias{generateTPsLogic,featureGroups-method}
\alias{generateTPsLogic,featureGroupsSet-method}
\title{Obtain transformation products (TPs) with metabolic logic}
\usage{
generateTPsLogic(fGroups, minMass = 40, ...)

\S4method{generateTPsLogic}{featureGroups}(fGroups, minMass = 40, adduct = NULL, transformations = NULL)

\S4method{generateTPsLogic}{featureGroupsSet}(fGroups, minMass = 40, transformations = NULL)
}
\arguments{
\item{fGroups}{A \code{\link{featureGroups}} object for which TPs should be calculated.}

\item{minMass}{A \code{numeric} that specifies the minimum mass of calculated TPs. If below this mass it will be
removed.}

\item{\dots}{Further arguments specified to the methods.}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
  Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}. If the \code{featureGroups} object has
  adduct annotations then these are used if \code{adducts=NULL}.

  \setsWF The \code{adduct} argument is not supported for sets workflows, since the
  adduct annotations will then always be used.}

\item{transformations}{A \code{data.frame} with transformation reactions to be used for calculating the TPs (see
details below). If \code{NULL}, a default table from Schollee \emph{et al.} is used (see references).}
}
\value{
A \code{\link{transformationProducts}} (derived) object containing all generated TPs.
}
\description{
Automatically calculate potential transformation products with \emph{metabolic logic}.
}
\details{
This function uses metabolic logic to obtain transformation products. This function is called when calling \code{generateTPs} with
  \code{algorithm="logic"}.

With this algorithm TPs are predicted from common (environmental) chemical reactions, such as hydroxylation,
  demethylation etc. The generated TPs result from calculating the mass differences between a parent feature after it
  underwent the reaction. While this only results in little information on chemical properties of the TP, an
  advantage of this method is that it does not rely on structural information of the parent, which may be unknown in
  a full non-target analysis.
}
\section{Transformation reactions}{
 The \code{transformations} argument specifies custom rules to calculate
  transformation products. This should be a \code{data.frame} with the following columns: \itemize{

  \item \code{transformation} The name of the chemical transformation

  \item \code{add} The elements that are added by this reaction (\emph{e.g.} \code{"O"}).

  \item \code{sub} The elements that are removed by this reaction (\emph{e.g.} \code{"H2O"}).

  \item \code{retDir} The expected retention time direction relative to the parent (assuming a reversed phase like LC
  separation). Valid values are: \samp{-1} (elutes before the parent), \samp{1} (elutes after the parent) or \samp{0}
  (no significant change or unknown).

  }
}

\section{Source}{
 The algorithms using transformation reactions are directly based on the work done by Schollee
  \emph{et al.} (see references).
}

\references{
\insertRef{Scholle2015}{patRoon}
}
\seealso{
\code{\link{generateTPs}} for more details and other algorithms.
}
