% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-library_formula.R
\name{genFormulaTPLibrary}
\alias{genFormulaTPLibrary}
\title{Automatically generate a transformation product library with formula data.}
\usage{
genFormulaTPLibrary(
  parents,
  transformations = NULL,
  minMass = 40,
  generations = 1,
  skipInvalid = TRUE,
  prefCalcChemProps = TRUE,
  neutralChemProps = FALSE
)
}
\arguments{
\item{parents}{The parents to which the given transformation rules should be used to generate the TP library. Should
be either a suspect list (see \link[=suspect-screening]{suspect screening} for more information) or the resulting
output of \code{\link{screenSuspects}}.}

\item{transformations}{A \code{data.frame} with transformation reactions to be used for calculating the TPs (see
details below). If \code{NULL}, a default table from Schollee \emph{et al.} is used (see references).}

\item{minMass}{The minimum mass for a TP to be kept.}

\item{generations}{An \code{integer} that specifies the number of transformation generations that should be
calculated. If \code{generations>1} then TPs are calculated by applying the transformation rules to the TPs
generated in the previous generation.}

\item{skipInvalid}{Set to \code{TRUE} to skip parents without formula information. Otherwise an error is thrown.}

\item{prefCalcChemProps}{If \code{TRUE} then calculated chemical properties such as the formula and
\acronym{InChIKey} are preferred over what is already present in the parent suspect list. For efficiency reasons it is
recommended to set this to \code{TRUE}. See the \verb{Validating and calculating chemical properties} section for
more details.}

\item{neutralChemProps}{If \code{TRUE} then the neutral form of the molecule is considered to calculate
\acronym{SMILES}, formulae etc. Enabling this may improve feature matching when considering common adducts
(\emph{e.g.} \code{[M+H]+}, \code{[M-H]-}). See the \verb{Validating and calculating chemical properties} section
for more details.}
}
\value{
A \code{data.table} that is suitable for the \code{TPLibrary} argument to
  \code{\link{generateTPsLibraryFormula}}.
}
\description{
Functionality to automatically generate a TP library with formula data from a set of transformation rules, which can
be used with \code{\link{generateTPsLibraryFormula}}. TP calculation will be skipped if the transformation involves
subtraction of elements not present in the parent.
}
\section{Transformation reactions}{
 The \code{transformations} argument specifies custom rules to calculate
  transformation products. This should be a \code{data.frame} with the following columns: \itemize{

  \item \code{transformation} The name of the chemical transformation

  \item \code{add} The elements that are added by this reaction (\emph{e.g.} \code{"O"}).

  \item \code{sub} The elements that are removed by this reaction (\emph{e.g.} \code{"H2O"}).

  \item \code{retDir} The expected retention time direction relative to the parent (assuming a reversed phase like LC
  separation). Valid values are: \samp{-1} (elutes before the parent), \samp{1} (elutes after the parent) or \samp{0}
  (no significant change or unknown).

  }
}

\section{Source}{
 The algorithms using transformation reactions are directly based on the work done by Schollee
  \emph{et al.} (see references).
}

\section{Validating and calculating chemical properties}{
 Chemical properties such as \acronym{SMILES},
  \acronym{InChIKey} and formula in the parent suspect list are automatically validated and calculated if missing/invalid.

  The internal validation/calculation process performs the following steps: \itemize{

  \item Validation of \acronym{SMILES}, \acronym{InChI}, \acronym{InChIKey} and formula data (if present). Invalid
  entries will be set to \code{NA}.

  \item If \code{neutralChemProps=TRUE} then chemical data (\acronym{SMILES}, formulae etc.) is neutralized by
  (de-)protonation (using the \command{--neutralized} option of \command{OpenBabel}). An additional column
  \code{molNeutralized} is added to mark those molecules that were neutralized. Note that neutralization requires
  either \acronym{SMILES} or \acronym{InChI} data to be available.

  \item The \acronym{SMILES} and \acronym{InChI} data are used to calculate missing or invalid \acronym{SMILES},
  \acronym{InChI}, \acronym{InChIKey} and formula data. If \code{prefCalcChemProps=TRUE} then existing
  \acronym{InChIKey} and formula data is overwritten by calculated values whenever possible.

  \item The chemical formulae which were \emph{not} calculated are verified and normalized. This process may be time
  consuming, and is potentially largely avoided by setting \code{prefCalcChemProps=TRUE}.

  \item Neutral masses are calculated for missing values (\code{prefCalcChemProps=FALSE}) or whenever possible
  (\code{prefCalcChemProps=TRUE}).

  }

  Note that calculation of formulae for molecules that are isotopically labelled is currently only supported for
  deuterium (2H) elements.

  This functionality relies heavily on \href{https://github.com/openbabel/openbabel}{OpenBabel}, please make sure it
  is installed.
}

\references{
\insertRef{Scholle2015}{patRoon}

\insertRef{OBoyle2011}{patRoon}
}
\seealso{
\code{\link{generateTPsLibraryFormula}} and \code{\link{generateTPsLogic}}
}
