% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{getCCSParams}
\alias{getCCSParams}
\title{Parameters for CCS calculation}
\source{
The calculation formulas was derived from \insertCite{Haler2017}{patRoon}, \insertCite{George2024}{patRoon}
  and the implementation used by \href{https://systemsomicslab.github.io/compms/msdial/main.html}{MS-DIAL}
  \insertCite{Tsugawa2020}{patRoon}. (\code{MobilityToCrossSection} method from the \code{IonMobilityUtility} class).
}
\usage{
getCCSParams(method, ..., calibrant = NULL)
}
\arguments{
\item{method, calibrant}{Sets the CCS calculation method and the calibrant data (only required if
\code{method="agilent"}). See details.}

\item{\dots}{optional named arguments that override defaults.}
}
\description{
Configuration and description of parameters used for CCS calculation.
}
\details{
The following parameters exist to configure the CCS calculation: \itemize{

  \item \code{method} The CCS calculation method. Should be \code{"bruker"}, \code{"mason-schamp_k"},
  \code{"mason-schamp_1/k"} or \code{"agilent"}. See details below.

  \item \code{defaultCharge} The default charge of the ions. This is used when no charge information is available.

  \item \code{temperature},\code{massGas} The temperature (Kelvin) and exact mass of the drift gas. See calculation
  details below.

  \item \code{MasonSchampConstant} The Mason-Schamp constant. See calculation details below.

  \item \code{calibrant} If \code{method="agilent"}: the calibrant data to be used for CCS calculation. This should
  either be \itemize{

    \item A path to an Agilent \file{.d} file.

    \item A path to an \file{OverrideImsCal.xml} file (found in \file{sample.d/AcqData}).

    \item A named \code{list} with the elements \code{massGas}, \code{TFix} and \code{beta}.

   }

}

The CCS calculation depends on the \code{method} parameter: \itemize{

  \item \code{bruker}: uses the Bruker \command{TDF-SDK} for calculations. See \link{msdata} for configuration
  options. Only applicable to TIMS data.

  \item \code{mason-schamp_k}: uses the Mason-Schamp equation:
    \deqn{CCS = C \cdot \frac{charge}{\sqrt{u \cdot T}} \cdot \frac{1}{mobility}}

    With \itemize{

      \item \emph{C} the Mason-Schamp constant, can be changed by setting the \code{MasonSchampConstant} parameter.
      See \insertCite{George2024}{patRoon} for details.

      \item \emph{u} the reduced mass of the drift gas and the ion:
      \deqn{u = \frac{m_{gas} \cdot m_{ion}}{m_{gas} + m_{ion}}}
      The mass of the drift gas is defined by the \code{massGas} parameter.

      \item \emph{T} the temperature (Kelvin) as defined by the \code{temperature} parameter.

    }

  \item \code{mason-schamp_1/k}: as \code{mason-schamp_k} but assuming an inversed mobility (\eqn{\frac{1}{k}}).
  This is meant for TIMS data. Compared to \code{method="bruker"}, this doesn't rely on the \command{TDF-SDK} but may
  produce results with very minor differences \insertCite{George2024}{patRoon}.

  \item \code{agilent}: uses Agilent calibration data with the following equation:

    \deqn{CCS = (mobility - t_{fix}) \cdot \frac{charge}{\beta} \cdot \frac{1}{\sqrt{\frac{m_{ion}}{m_{ion} + m_{gas}}}}}

    With \eqn{t_{fix}} and \eqn{\beta} the \code{TFix} and \code{beta} values from the calibration data. The
    \code{massGas} parameter sets the \eqn{m_{gas}} value.

}

The \code{getCCSParams} function generates such parameter list with defaults.
}
\references{
\insertAllCited{}
}
