% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{getFCParams}
\alias{getFCParams}
\title{Fold change calculation}
\usage{
getFCParams(rGroups, ...)
}
\arguments{
\item{rGroups}{A \code{character} vector with the names of the two replicate groups to be compared.}

\item{\dots}{Optional named arguments that override defaults.}
}
\description{
Fold change calculation
}
\details{
Fold change calculation can be used to easily identify significant changes between replicate groups. The
  calculation process is configured through a paramater list, which can be constructed with the \code{getFCParams}
  function. The parameter list has the following entries: \itemize{

  \item \code{rGroups} the name of the two replicate groups to compare (taken from the \code{rGroups} argument to
  \code{getFCParams}).

  \item \code{thresholdFC}: the threshold log FC for a feature group to be classified as increasing/decreasing.

  \item \code{thresholdPV}: the threshold log P for a feature group to be significantly different.

  \item \code{zeroMethod},\code{zeroValue}: how to handle zero values when calculating the FC: \code{add} adds an
  offset to zero values, \code{"fixed"} sets zero values to a fixed number and \code{"omit"} removes zero data. The
  number that is added/set by the former two options is defined by \code{zeroValue}.

  \item \code{PVTestFunc}: a function that is used to calculate P values (usually using \code{\link{t.test}}).

  \item \code{PVAdjFunc}: a function that is used to adjust P values (usually using \code{\link{p.adjust}})

  }
}
\seealso{
\code{\link{featureGroups-class}} and \link{feature-plotting}
}
\author{
The code to calculate and plot Fold change data was created by Bas van de Velde.
}
