% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{getIMSMatchParams}
\alias{getIMSMatchParams}
\title{Parameters for IMS matching}
\usage{
getIMSMatchParams(param, ...)
}
\arguments{
\item{param}{Should be \code{"mobility"} or \code{"CCS"} to match by mobility or \acronym{CCS}, respectively.}

\item{\dots}{optional named arguments that override defaults.}
}
\description{
Parameters that define how mobility or \acronym{CCS} values between \emph{e.g.} features and suspects should be
matched.
}
\details{
The following parameters should be defined: \itemize{

  \item \code{param} Should be \code{"mobility"} or \code{"CCS"} to match by mobility or \acronym{CCS}, respectively.

  \item \code{window},\code{relative} The \code{window} parameter sets the tolerance window size used for matching.
  If \code{relative=TRUE} then the tolerance is relative (\samp{0-1}). The defaults for \code{window} are
  (see \link{limits}): \itemize{

    \item \code{defaultLim("mobility", "medium")} (\code{param="mobility"} and \code{relative=FALSE})

    \item \code{defaultLim("mobility", "medium_rel")} (\code{param="mobility"} and \code{relative=TRUE})

    \item \code{defaultLim("CCS", "medium")} (\code{param="CCS"} and \code{relative=FALSE})

    \item \code{defaultLim("CCS", "medium_rel")} (\code{param="CCS"} and \code{relative=TRUE})

  }

  \item \code{minMatches} The minimum number of mobility/\acronym{CCS} matches for a suspect hit. If the number of
  available reference mobility/\acronym{CCS} values in the suspect list is less than \code{minMatches}, then that
  number is used as threshold. Set to \code{0} to disable.

}

These parameters are passed as a named \code{list} as the \code{IMSMatchParams} argument to functions.

The \code{getIMSMatchParams} function generates such parameter list with defaults.
}
\note{
If negation is enabled with suspect filtering and \code{minMatches>0}, then the \code{window} match filter is
  \emph{not} negated. Negating both would lead to unexpected results, \emph{i.e.} suspects outside \code{window} are
  kept and increase the number of matched suspects as seen by the \code{minMatches} filter.
}
