% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{getIMSRangeParams}
\alias{getIMSRangeParams}
\title{Parameters to specify a IMS data range}
\usage{
getIMSRangeParams(param, lower, upper, mzRelative = FALSE)
}
\arguments{
\item{param, lower, upper, mzRelative}{Arguments to specify the IMS range parameters, see Details.}
}
\description{
Parameters that define a range of mobility or \acronym{CCS} values.
}
\details{
The following parameters are used to define an IMS range: \itemize{

  \item \code{param} Should be \code{"mobility"} or \code{"CCS"} to specify a mobility or \acronym{CCS} range,
  respectively.
  
  \item \code{lower},\code{upper} The lower and upper range.
  
  \item \code{mzRelative} Set to \code{TRUE} to specify an IMS range that is normalized by \emph{m/z}.

}

These parameters are passed as a named \code{list} as the \code{IMSRangeParams} argument to functions.

The \code{getIMSRangeParams} function generates such parameter list with defaults.
}
