% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{getMSFileFormats}
\alias{getMSFileFormats}
\title{Get supported MS file formats}
\usage{
getMSFileFormats(fileType = NULL)
}
\arguments{
\item{fileType}{The type of file for which formats should be returned (see \code{\link{getMSFileTypes}}). If
\code{NULL} then all file formats are returned.}
}
\description{
Returns the supported file formats of MS files in \pkg{patRoon}.
}
