% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{getSIRIUSConfig}
\alias{getSIRIUSConfig}
\title{Obtain a SIRIUS job configuration}
\usage{
getSIRIUSConfig(
  config = NULL,
  import = NULL,
  login = "check",
  alwaysLogin = FALSE,
  SIRIUSAPI = NULL
)
}
\arguments{
\item{config}{A configuration specification. Use \code{NULL} to obtain the server default,
a \code{character} string to fetch a named configuration, \code{NA} to list available
configuration names, or a \code{list} / \code{RSirius::JobSubmission} object to use directly.}

\item{import}{Path to a JSON file to import the configuration from, \emph{e.g.} exported from the \command{SIRIUS}
GUI. Cannot be used together with \code{config}.}

\item{login, alwaysLogin}{Specifies if and how account logging of SIRIUS should be handled:

  \code{login=FALSE}: no automatic login is performed and the active login status is not checked.

  \code{login="check"}: aborts if no active login is present.

  \code{login="interactive"}: interactively ask for login (using \CRANpkg{getPass}).

  \code{login=c(username="...", password="...")}: perform the login with the given details. For security reasons,
  please do not enter the details directly, but use e.g. environment variables or store/retrieve them with the
  \CRANpkg{keyring} package.
  
  if \code{alwaysLogin=TRUE} then a login is always performed, otherwise only if SIRIUS reports no active login.

  See the \href{https://v6.docs.sirius-ms.io/account-and-license/}{SIRIUS website} and \pkg{patRoon} handbook for
  more information.
  
  \strong{NOTE}: By loggin in you will accept the terms of the Service and Privacy Policy of the SIRIUS Webservice.}

\item{SIRIUSAPI}{An \code{rsirius_api} object for connecting to the \command{SIRIUS} API. If \code{NULL}, a new
connection will be started automatically.}
}
\value{
A \code{RSirius::JobSubmission} object with the job configuration, or a \code{character}
  vector of configuration names when \code{config = NA}.
}
\description{
This function obtains a \link[RSirius]{JobSubmission} configuration object for use with
\code{\link{generateFormulasSIRIUS}} and \code{\link{generateCompoundsSIRIUS}}.
}
\details{
The function can return the server default configuration, fetch a named configuration, import a configuration from a
JSON file, or use a provided configuration object/list. If needed, the SIRIUS API is started automatically and a
login can be performed. There are \emph{many} configuration options available in \command{SIRIUS}. It is probably
easiest to explore and configure them through the \command{SIRIUS} GUI, save or export the configuration profile and
then load it via the \code{config} or \code{import} arguments.
}
