% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{getSIRIUSToken}
\alias{getSIRIUSToken}
\title{Obtains a SIRIUS refresh token}
\usage{
getSIRIUSToken(user)
}
\arguments{
\item{user}{A \code{character} string with the user name/email.}
}
\value{
A \code{character} string with the (secret) refresh token or \code{NULL} if the input was cancelled.
}
\description{
This function is used to obtain a \command{SIRIUS} refresh token with your login details, which allows
\code{\link{generateCompoundsSIRIUS}} to automatically log in to use \emph{e.g.} \command{CSI:FingerID}.
}
\details{
\command{SIRIUS} version \samp{5} requires the user to be logged in when using web services such as
\command{CSI:FingerID}. To allow secure log in by external software tools such as \pkg{patRoon}, a \emph{referesh
token} is needed. This function uses your user name (email) and asks for a password (using \code{\link{getPass}}) to
obtain such a token more easily.

More details for \emph{e.g.} account creation can be found on
\url{https://boecker-lab.github.io/docs.sirius.github.io/account-and-license/}.
}
\references{
\insertRef{Dhrkop2019}{patRoon}
}
