% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils-xcms.R
\docType{methods}
\name{getXcmsSet}
\alias{getXcmsSet}
\alias{getXcmsSet,features-method}
\alias{getXcmsSet,featuresOpenMS-method}
\alias{getXcmsSet,featuresXCMS-method}
\alias{getXcmsSet,featureGroups-method}
\alias{getXcmsSet,featureGroupsXCMS-method}
\title{Conversion to xcmsSet objects}
\usage{
getXcmsSet(obj, ...)

\S4method{getXcmsSet}{features}(obj, exportedData)

\S4method{getXcmsSet}{featuresOpenMS}(obj, exportedData)

\S4method{getXcmsSet}{featuresXCMS}(obj, exportedData)

\S4method{getXcmsSet}{featureGroups}(obj, exportedData)

\S4method{getXcmsSet}{featureGroupsXCMS}(obj, exportedData)
}
\arguments{
\item{obj}{The object that should be converted.}

\item{exportedData, \dots}{Set to \code{TRUE} if analyses were exported as
\code{mzXML} or \code{mzML} files (ignored for \code{featuresOpenMS} and
\code{featuresXCMS} methods).}
}
\description{
Converts a \code{\link{features}} or \code{\link{featureGroups}} object to an
\code{\link{xcmsSet}} object.
}
