% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/feature_groups-greedy.R
\name{groupFeaturesGreedy}
\alias{groupFeaturesGreedy}
\alias{groupFeaturesGreedy,features-method}
\title{Group features using greedy algorithm}
\usage{
groupFeaturesGreedy(feat, ...)

\S4method{groupFeaturesGreedy}{features}(
  feat,
  rtalign = FALSE,
  rtWindow = defaultLim("retention", "medium"),
  mzWindow = defaultLim("mz", "medium"),
  mobWindow = defaultLim("mobility", "medium"),
  scoreWeights = c(retention = 1, mz = 1, mobility = 1, intensity = 1),
  verbose = TRUE
)
}
\arguments{
\item{feat}{The \code{\link{features}} object with the features to be grouped.}

\item{...}{Further parameters passed to the selected grouping algorithm.}

\item{rtalign}{Not yet supported. Provided for consistency with other grouping methods.}

\item{rtWindow, mzWindow, mobWindow}{Numeric tolerances for retention time (seconds), \emph{m/z}, and mobility,
respectively. The scoring terms are normalized to these values. Defaults to \code{defaultLim("retention",
"medium")}, \code{defaultLim("mz", "medium")}, and \code{defaultLim("mobility", "medium")}, respectively (see \link{limits}).}

\item{scoreWeights}{Numeric vector specifying the scoring weights. Should contain the following named elements:
\code{"retention"}, \code{"mz"}, \code{"mobility"}, and \code{"intensity"}. Missing elements are defaulted to one.}

\item{verbose}{if \code{FALSE} then no text output will be shown.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.
}
\description{
Group features using a greedy algorithm that maximizes group scores based on retention time, m/z, mobility, and
intensity similarities.
}
\details{
This function uses greedy to group features. This function is called when calling \code{groupFeatures} with
  \code{algorithm="greedy"}.

The \code{greedy} algorithm is a simple feature grouping algorithm that can work with both HRMS and IMS-HRMS
  data. The algorithm groups features by iteratively building the best possible groups. Features are processed in
  order of decreasing intensity. For each feature, candidate groups are formed from all other (ungrouped) features
  within the specified retention time, \emph{m/z} and mobility windows. Each candidate group only contains a maximum
  of one feature per analysis. The candidates are then scored and the group with the lowest overall variations in
  retention time, \emph{m/z}, mobility and replicate intensity is then selected. This process is repeated until all
  features have been assigned to a group. The weights for each of the scoring terms can be configured.
}
\note{
Any links between IMS precursors and IMS features are removed. This can occur \emph{e.g.} when \code{greedy}
is used to generate a \link[=featureGroups-compare]{feature consensus} from a \link[=assignMobilities_feat]{post
mobility assignment} workflow.
}
\seealso{
\code{\link{groupFeatures}} for more details and other algorithms.
}
