% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/feature_groups-kpic2.R
\name{groupFeaturesKPIC2}
\alias{groupFeaturesKPIC2}
\alias{groupFeaturesKPIC2,features-method}
\alias{groupFeaturesKPIC2,featuresSet-method}
\title{Group features using KPIC2}
\usage{
groupFeaturesKPIC2(feat, ...)

\S4method{groupFeaturesKPIC2}{features}(
  feat,
  rtalign = TRUE,
  loadRawData = TRUE,
  groupArgs = list(tolerance = c(0.005, 12)),
  alignArgs = list(),
  verbose = TRUE
)

\S4method{groupFeaturesKPIC2}{featuresSet}(
  feat,
  groupArgs = list(tolerance = c(0.005, 12)),
  verbose = TRUE
)
}
\arguments{
\item{feat}{The \code{\link{features}} object with the features to be grouped.}

\item{...}{Further parameters passed to the selected grouping algorithm.}

\item{rtalign}{Set to \code{TRUE} to enable retention time alignment.}

\item{loadRawData}{Set to \code{TRUE} if analyses are available as \code{mzXML} or \code{mzML} files. Otherwise MS
data is not loaded, and some dummy data (\emph{e.g.} file paths) is used in the returned object.}

\item{groupArgs, alignArgs}{Named \code{character} vector that may contain extra parameters to be used by
\code{\link[KPIC:PICset.group]{KPIC::PICset.group}} and \code{\link[KPIC:PICset.align]{KPIC::PICset.align}},
respectively.}

\item{verbose}{if \code{FALSE} then no text output will be shown.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.

The \code{featuresSet} method (for \link[=sets-workflow]{sets workflows}) returns a
  \code{\link{featureGroupsSet}} object.
}
\description{
Uses the the \href{https://github.com/hcji/KPIC2}{KPIC2} \R package for grouping of features.
}
\details{
This function uses KPIC2 to group features. This function is called when calling \code{groupFeatures} with
  \code{algorithm="kpic2"}.

Grouping of features and alignment of their retention times are performed with the
  \code{\link[KPIC:PICset.group]{KPIC::PICset.group}} and \code{\link[KPIC:PICset.align]{KPIC::PICset.align}}
  functions, respectively.
}
\section{Sets workflows}{
 \code{loadRawData} and arguments related to retention time alignment are currently not
  supported for \link[=sets-workflow]{sets workflows}.
}

\references{
\insertRef{Ji2017}{patRoon}
}
\seealso{
\code{\link{groupFeatures}} for more details and other algorithms.
}
