% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/feature_groups-openms.R
\name{groupFeaturesOpenMS}
\alias{groupFeaturesOpenMS}
\alias{groupFeaturesOpenMS,features-method}
\title{Group features using OpenMS}
\usage{
groupFeaturesOpenMS(feat, ...)

\S4method{groupFeaturesOpenMS}{features}(
  feat,
  rtalign = TRUE,
  QT = FALSE,
  maxAlignRT = 30,
  maxAlignMZ = 0.005,
  maxGroupRT = 12,
  maxGroupMZ = 0.005,
  extraOptsRT = NULL,
  extraOptsGroup = NULL,
  verbose = TRUE
)
}
\arguments{
\item{feat}{The \code{\link{features}} object with the features to be grouped.}

\item{...}{Further parameters passed to the selected grouping algorithm.}

\item{rtalign}{Set to \code{TRUE} to enable retention time alignment.}

\item{QT}{If enabled, use \command{FeatureLinkerUnlabeledQT} instead of \command{FeatureLinkerUnlabeled} for feature
grouping.}

\item{maxAlignRT, maxAlignMZ}{Used for retention alignment. Maximum retention time or m/z difference (seconds/Dalton)
for feature pairing. Sets \code{-algorithm:pairfinder:distance_RT:max_difference} and
\code{-algorithm:pairfinder:distance_MZ:max_difference} otpions, respectively.}

\item{maxGroupRT, maxGroupMZ}{as \code{maxAlignRT} and \code{maxAlignMZ}, but for grouping of features. Sets
\code{-algorithm:distance_RT:max_difference} and \code{-algorithm:distance_MZ:max_difference} options,
respectively.}

\item{extraOptsRT, extraOptsGroup}{Named \code{list} containing extra options that will be passed to
\command{MapAlignerPoseClustering} or \command{FeatureLinkerUnlabeledQT/FeatureLinkerUnlabeled}, respectively. Any
options specified here will override any of the above. Example:
\code{extraOptsGroup=list("-algorithm:distance_RT:max_difference"=12)} (corresponds to setting
\code{maxGroupRT=12}). Set to \code{NULL} to ignore.}

\item{verbose}{if \code{FALSE} then no text output will be shown.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.

The \code{featuresSet} method (for \link[=sets-workflow]{sets workflows}) returns a
  \code{\link{featureGroupsSet}} object.
}
\description{
Group and align features with OpenMS tools
}
\details{
This function uses OpenMS to group features. This function is called when calling \code{groupFeatures} with
  \code{algorithm="openms"}.

Retention times may be aligned by the
  \href{https://abibuilder.cs.uni-tuebingen.de/archive/openms/Documentation/nightly/html/TOPP_MapAlignerPoseClustering.html}{MapAlignerPoseClustering}
   TOPP tool. Grouping is achieved by either the
  \href{https://abibuilder.cs.uni-tuebingen.de/archive/openms/Documentation/nightly/html/TOPP_FeatureLinkerUnlabeled.html}{FeatureLinkerUnlabeled}
   or
  \href{https://abibuilder.cs.uni-tuebingen.de/archive/openms/Documentation/nightly/html/TOPP_FeatureLinkerUnlabeledQT.html}{FeatureLinkerUnlabeledQT}
   TOPP tools.
}
\references{
\insertRef{Rst2016}{patRoon} \cr\cr
  \href{https://pugixml.org/}{pugixml} (via \href{http://www.rcpp.org/}{Rcpp}) is used to process OpenMS XML output. \cr\cr
  \addCitations{Rcpp}{1} \cr\cr
  \addCitations{Rcpp}{2} \cr\cr
  \addCitations{Rcpp}{3}
}
\seealso{
\code{\link{groupFeatures}} for more details and other algorithms.
}
