% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/feature_groups-xcms.R
\name{groupFeaturesXCMS}
\alias{groupFeaturesXCMS}
\alias{groupFeaturesXCMS,features-method}
\alias{groupFeaturesXCMS,featuresSet-method}
\title{Group features using XCMS (old interface)}
\usage{
groupFeaturesXCMS(feat, ...)

\S4method{groupFeaturesXCMS}{features}(
  feat,
  rtalign = TRUE,
  loadRawData = TRUE,
  groupArgs = list(mzwid = 0.015),
  retcorArgs = list(method = "obiwarp"),
  verbose = TRUE
)

\S4method{groupFeaturesXCMS}{featuresSet}(feat, groupArgs = list(mzwid = 0.015), verbose = TRUE)
}
\arguments{
\item{feat}{The \code{\link{features}} object with the features to be grouped.}

\item{...}{Further parameters passed to the selected grouping algorithm.}

\item{rtalign}{Set to \code{TRUE} to enable retention time alignment.}

\item{loadRawData}{Set to \code{TRUE} if analyses are available as \code{mzXML} or \code{mzML} files. Otherwise MS
data is not loaded, and some dummy data (\emph{e.g.} file paths) is used in the returned object.}

\item{groupArgs}{named \code{character vector} that may contain extra grouping parameters to be used by
\code{\link[xcms:group-methods]{xcms::group}}}

\item{retcorArgs}{named \code{character vector} that may contain extra parameters to be used by
\code{\link[xcms:retcor-methods]{xcms::retcor}}.}

\item{verbose}{if \code{FALSE} then no text output will be shown.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.

The \code{featuresSet} method (for \link[=sets-workflow]{sets workflows}) returns a
  \code{\link{featureGroupsSet}} object.
}
\description{
Group and align features with the legacy \code{\link[xcms]{xcmsSet}} function from the \pkg{xcms} package.
}
\details{
This function uses XCMS to group features. This function is called when calling \code{groupFeatures} with
  \code{algorithm="xcms"}.

Grouping of features and
  alignment of their retention times are performed with the \code{\link[xcms:group-methods]{xcms::group}} and
  \code{\link[xcms:retcor-methods]{xcms::retcor}} functions, respectively. Both functions have an extensive list of
  parameters to modify their behavior and may therefore be used to potentially optimize results.
}
\section{Sets workflows}{
 \code{loadRawData} and arguments related to retention time alignment are currently not
  supported for \link[=sets-workflow]{sets workflows}.
}

\references{
\addCitations{xcms}{1} \cr\cr \addCitations{xcms}{2} \cr\cr \addCitations{xcms}{3}
}
\seealso{
\code{\link{groupFeatures}} for more details and other algorithms.
}
