% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/feature_groups-xcms3.R
\name{groupFeaturesXCMS3}
\alias{groupFeaturesXCMS3}
\alias{groupFeaturesXCMS3,features-method}
\alias{groupFeaturesXCMS3,featuresSet-method}
\title{Group features using XCMS (new interface)}
\usage{
groupFeaturesXCMS3(feat, ...)

\S4method{groupFeaturesXCMS3}{features}(
  feat,
  rtalign = TRUE,
  loadRawData = TRUE,
  groupParam = xcms::PeakDensityParam(sampleGroups = analysisInfo(feat)$group),
  preGroupParam = groupParam,
  retAlignParam = xcms::ObiwarpParam(),
  verbose = TRUE
)

\S4method{groupFeaturesXCMS3}{featuresSet}(
  feat,
  groupParam = xcms::PeakDensityParam(sampleGroups = analysisInfo(feat)$group),
  verbose = TRUE
)
}
\arguments{
\item{feat}{The \code{\link{features}} object with the features to be grouped.}

\item{...}{Further parameters passed to the selected grouping algorithm.}

\item{rtalign}{Set to \code{TRUE} to enable retention time alignment.}

\item{loadRawData}{Set to \code{TRUE} if analyses are available as \code{mzXML} or \code{mzML} files. Otherwise MS
data is not loaded, and some dummy data (\emph{e.g.} file paths) is used in the returned object.}

\item{groupParam, retAlignParam}{parameter object that is directly passed to
\code{\link[xcms:groupChromPeaks]{xcms::groupChromPeaks}} and \code{\link[xcms:adjustRtime]{xcms::adjustRtime}},
respectively.}

\item{preGroupParam}{grouping parameters applied when features are grouped \emph{prior} to alignment (only with peak
groups alignment).}

\item{verbose}{if \code{FALSE} then no text output will be shown.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.

The \code{featuresSet} method (for \link[=sets-workflow]{sets workflows}) returns a
  \code{\link{featureGroupsSet}} object.
}
\description{
Uses the new \code{xcms3} interface from the \pkg{xcms} package to find features.
}
\details{
This function uses XCMS3 to group features. This function is called when calling \code{groupFeatures} with
  \code{algorithm="xcms3"}.

Grouping of features and alignment of their retention times are performed with the
  \code{\link[xcms:groupChromPeaks]{xcms::groupChromPeaks}} and \code{\link[xcms:adjustRtime]{xcms::adjustRtime}}
  functions, respectively. Both of these functions support an extensive amount of parameters that modify their
  behavior and may therefore require optimization.
}
\section{Sets workflows}{
 \code{loadRawData} and arguments related to retention time alignment are currently not
  supported for \link[=sets-workflow]{sets workflows}.
}

\references{
\addCitations{xcms}{1} \cr\cr \addCitations{xcms}{2} \cr\cr \addCitations{xcms}{3}
}
\seealso{
\code{\link{groupFeatures}} for more details and other algorithms.
}
