% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/check_components.R,
%   R/check_features.R, R/project-tool.R
\name{GUI-utils}
\alias{GUI-utils}
\alias{checkComponents,components-method}
\alias{checkComponents}
\alias{checkFeatures,featureGroups-method}
\alias{checkFeatures}
\alias{newProject}
\title{Interactive GUI utilities}
\usage{
\S4method{checkComponents}{components}(
  components,
  fGroups,
  session = "checked-components.yml",
  rtWindow = 30,
  clearSession = FALSE
)

\S4method{checkFeatures}{featureGroups}(
  fGroups,
  session = "checked-features.yml",
  rtWindow = 30,
  clearSession = FALSE
)

newProject(destPath = NULL)
}
\arguments{
\item{fGroups}{A \code{\link{featureGroups}} object.}

\item{destPath}{Set destination path value to this value (useful for
debugging). Set to \code{NULL} for a default value.}

\item{mzWindow}{Default \emph{m/z} window to be used for creating extracted
ion chromatograms (EICs).}

\item{enabledFGroups}{A logical vector that states for each feature group
whether it should be kept (\code{TRUE}) or not (\code{FALSE}). The order is
the same as the \code{fGroups} parameter. If \code{NULL} then all feature
groups are considered to be kept.}
}
\value{
\code{checkFeatures} returns a logical vector for all feature
  groups that were selected to be kept (\code{TRUE}) or not (\code{FALSE}).
  This result can be passed to the \code{enabledFGroups} parameter for
  subsequent calls to \code{checkFeatures} in order to restore the
  keep/not keep state from a previous call. To actually remove unwanted
  feature groups the object should be subset by the subsetting
  (\code{\link{[}}) operator to which the return value should be passed as
  the second parameter.
}
\description{
Interactive utilities using \pkg{\link{shiny}} to provide a graphical user
interface (GUI).
}
\details{
\code{checkFeatures} is used to review chromatographic
  information for feature groups. This is especially useful to get a visual
  impression of the quality of detected features. In addition, this function
  may be used to remove unwanted (\emph{e.g.} outlier) features. Better
  performance is often obtained when an external browser is used to use this
  Shiny application. Furthermore, when a large \code{featureGroups} object is
  used it is recommended to limit the number of analyses/feature groups by
  subsetting the object.

The \code{newProject} function is used to quickly generate a
  processing R script. This tool allows the user to quickly select the
  targeted analyses, workflow steps and configuring some of their common
  parameters. This function requires to be run within a
  \href{https://www.rstudio.com/}{RStudio} session. The resulting script is
  either added to the current open file or to a new file. The
  \link[=analysis-information]{analysis information} will be written to a
  \file{.csv} file so that it can easily be modified afterwards.
}
