% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\docType{methods}
\name{h-cluster}
\alias{h-cluster}
\alias{makeHCluster,featureGroups-method}
\alias{makeHCluster}
\alias{hClusterFilter,featureGroups,clusterInfo-method}
\alias{hClusterFilter}
\alias{drawHeatMap,clusterInfo-method}
\alias{drawHeatMap}
\alias{getSilhouetteInfo,clusterInfo-method}
\alias{getSilhouetteInfo}
\alias{plotInt,clusterInfo-method}
\alias{plot,clusterInfo,ANY-method}
\alias{plot,silhouetteInfo,ANY-method}
\title{Hierarchical clustering of feature groups}
\usage{
\S4method{makeHCluster}{featureGroups}(obj, method, metric = "euclidean",
  normFunc = max, average = TRUE)

\S4method{hClusterFilter}{featureGroups,clusterInfo}(fGroups, cInfo, k, c)

\S4method{drawHeatMap}{clusterInfo}(obj, col = colorRampPalette(c("black",
  "yellow"))(100), interactive = FALSE, ...)

\S4method{getSilhouetteInfo}{clusterInfo}(cInfo, ranges = seq(10, 100, 10))

\S4method{plotInt}{clusterInfo}(obj, k, c, ...)

\S4method{plot}{clusterInfo,ANY}(x, k = NULL, ...)

\S4method{plot}{silhouetteInfo,ANY}(x, y, ...)
}
\arguments{
\item{method}{Clustering method that should be applied (passed to
\code{\link{hclust}}).}

\item{metric}{Distance metric used to calculate the distance matrix (passed
to \code{\link{daisy}}).}

\item{normFunc}{Function that should be used for normalization of data.
Intensitity values of a feature group will be divided by the result of this
function when it is called with all intensity values of that feature group.
For example, when \code{\link{max}} is used normalized intensities will be
between zero and one.}

\item{average}{If \code{TRUE} then all intensitity data will be averaged for
each replicate group.}

\item{fGroups}{A \code{\link{featureGroups}} object.}

\item{cInfo}{A \code{\link{clusterInfo}} object with relevant clustering
information.}

\item{k}{The desired number of branches from a given cluster object.}

\item{c}{The numeric identifier of the desired branch from a cut cluster.}

\item{col}{Colours used for drawing the heatmap. See \code{\link{heatmap}} /
\code{\link{d3heatmap}}.}

\item{interactive}{If \code{TRUE} an interactive heatmap will be drawn (with
\code{\link{d3heatmap}}).}

\item{\dots}{Further options passed to \code{\link{heatmap}} /
\code{\link{d3heatmap}} (\code{drawHeatMap}), \code{\link{barplot}}
(\code{plot}, \code{\link{silhouetteInfo}} signature) or
\code{\link[graphics]{plot}} (\code{plot}, \code{\link{clusterInfo}}
signature and \code{plotInt}).}

\item{ranges}{An integer vector containing a sequence of desired amounts of
branches to which silhouette information should be calculated.}

\item{x, obj}{A \code{\link{clusterInfo}}, \code{\link{silhouetteInfo}}
or \code{\link{featureGroups}} object.}

\item{y}{Ignored.}
}
\value{
\code{makeHCluster} Returns an object of the
  \code{\link{clusterInfo}} class.

\code{drawHeatMap} returns the same as \code{\link{heatmap}} or
  \code{\link{d3heatmap}}.

\code{getSilhouetteInfo} returns a \code{\link{silhouetteInfo}}
  object.

\code{plot} (\code{\link{silhouetteInfo}} signature) returns the result of
  \code{\link{barplot}}.
}
\description{
Utilities to perform and process hierarchical clustering with feature groups.
}
\details{
These functions provide an interface to common hierarchical clustering tools
such as \code{\link{hclust}} that can be used for clustering feature groups.

\code{makeHCluster} generates hierarchical clustering information
  for feature groups. Intensity data will optionally be averaged and then
  normalized. A distance matrix is calculated with \code{\link{daisy}} and
  clustering is performed with \code{\link{hclust}}.

\code{hClusterFilter} isolates feature groups within a
  \code{\link{featureGroups}} object which are present within a given branch
  of a cluster object.

\code{drawHeatMap} draws a heatmap using the \code{\link{heatmap}}
  or \code{\link{d3heatmap}} function.

\code{getSilhouetteInfo} is used to obtain silhouette information of a
  \code{\link{clusterInfo}} object which is 'cut' by a given sequence of
  desired number of branches.

\code{plotInt} makes a plot for all (normalized) intensity profiles
  of the feature groups within a given cluster.

\code{plot} (\code{\link{clusterInfo}} signature) generates a dendrogram
  from a given cluster object and optionally highlights resulting branches
  when the cluster is cut.

\code{plot} (\code{\link{silhouetteInfo}} signature) plots the various mean
  silhouette widths stored inside a \code{\link{silhouetteInfo}} object. The
  resulting barplot can be used to assess the optimal number of clusters.
}
