% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups.R
\name{importFeatureGroups}
\alias{importFeatureGroups}
\title{Import feature groups from files}
\usage{
importFeatureGroups(input, type, ...)
}
\arguments{
\item{input}{The input object or path that should be imported. See the algorithm specific functions for more details.}

\item{type}{What type of data should be imported: \code{"xcms"}, \code{"xcms3"}, \code{"kpic2"}, \code{"sirius"},
\code{"table"}, \code{"brukerpa"} (Bruker ProfileAnalysis), \code{"brukertasq"} (Bruker TASQ) or \code{"envimass"}.}

\item{\dots}{Further arguments passed to the selected import algorithm function.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.
}
\description{
Generic function to import feature groups produced by other software from files.
}
\details{
\code{importFeatureGroups} is a generic function that will import feature groups from files by one of the supported algorithms. The actual
  functionality is provided by algorithm specific functions such as \code{importFeatureGroupsXCMS3} and \code{importFeatureGroupsTable}. While these
  functions may be called directly, \code{importFeatureGroups} provides a generic interface and is therefore usually preferred.
}
\seealso{
The \code{\link{featureGroups}} output class and its methods and the algorithm specific functions:
  \code{\link{importFeatureGroupsXCMS}}, \code{\link{importFeatureGroupsXCMS3}}, \code{\link{importFeatureGroupsKPIC2}}, \code{\link{importFeatureGroupsSIRIUS}}, \code{\link{importFeatureGroupsTable}}, \code{\link{importFeatureGroupsBrukerPA}}, \code{\link{importFeatureGroupsBrukerTASQ}}, \code{\link{importFeatureGroupsEnviMass}}

\code{\link{groupFeatures}} to group features.
}
