% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-bruker.R
\name{importFeatureGroupsBrukerPA}
\alias{importFeatureGroupsBrukerPA}
\title{Imports feature groups from Bruker ProfileAnalysis}
\usage{
importFeatureGroupsBrukerPA(
  path,
  feat,
  rtWindow = 12,
  mzWindow = 0.005,
  intWindow = 5,
  warn = TRUE
)
}
\arguments{
\item{path}{The file path to a exported 'bucket table' \file{.txt} file from PA.}

\item{feat}{The \code{\link{features}} object obtained with \code{\link{findFeaturesBruker}}.}

\item{rtWindow, mzWindow, intWindow}{Search window values for retention time (seconds), \emph{m/z} (Da) and intensity
used to find back features within feature groups from PA (+/- the retention/mass/intensity value of a feature).}

\item{warn}{Warn about missing or duplicate features when relating them back from grouped features.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.

The \code{featuresSet} method (for \link[=sets-workflow]{sets workflows}) returns a
  \code{\link{featureGroupsSet}} object.
}
\description{
Imports a 'bucket table' produced by Bruker ProfileAnalysis (PA)
}
\details{
This function imports data from Bruker ProfileAnalysis. This function is called when calling \code{importFeatureGroups} with
  \code{type="brukerpa"}.

The 'bucket table' should be exported as \file{.txt} file. Please note that this function only supports
  features generated by \code{\link{findFeaturesBruker}} and it is \strong{crucial} that DataAnalysis files remain
  unchanged when features are collected and the bucket table is generated. Furthermore, please note that PA does not
  retain information about originating features for generated buckets. For this reason, this function tries to find
  back the original features and care must be taken to correctly specify search parameters (\code{rtWindow},
  \code{mzWindow}, \code{intWindow}).
}
\seealso{
\code{\link{importFeatureGroups}} for more details and other algorithms.
}
