% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-tasq.R
\name{importFeatureGroupsBrukerTASQ}
\alias{importFeatureGroupsBrukerTASQ}
\title{Imports feature groups from Bruker TASQ}
\usage{
importFeatureGroupsBrukerTASQ(path, analysisInfo, clusterRTWindow = 12)
}
\arguments{
\item{path}{The file path to an Excel export of the Global results table from TASQ, converted to \file{.csv} format.}

\item{analysisInfo}{A \code{data.frame} with \link[=analysis-information]{Analysis information}.}

\item{clusterRTWindow}{This retention time window (in seconds) is used to group hits across analyses together. See
also the details section.}
}
\value{
A new \code{featureGroups} object containing converted screening results from Bruker TASQ.
}
\description{
Imports screening results from Bruker TASQ as feature groups.
}
\details{
This function imports data from Bruker TASQ. This function is called when calling \code{importFeatureGroups} with
  \code{type="brukertasq"}.

The feature groups across analyses are formed based on the name of suspects and their closeness in retention
  time. The latter is necessary because TASQ does not necessarily perform checks on retention times and may therefore
  assign a suspect to peaks with different retention times across analyses (or within a single analysis). Hence,
  suspects with equal names are hierarchically clustered on their retention times (using \pkg{\link{fastcluster}}) to
  form the feature groups. The cut-off value for this is specified by the \code{clusterRTWindow} argument. The input
  for this function is obtained by generating an Excel export of the 'global' results and subsequently converting the
  file to \file{.csv} format.
}
\note{
This function uses estimated min/max values for retention times and dummy min/max \emph{m/z} values for
  conversion to features, since this information is not (readily) available. Hence, when plotting, for instance,
  extracted ion chromatograms (with \code{\link{plotChroms}}) the integrated chromatographic peak range shown is
  incorrect.

  This function may use suspect names to base file names used for reporting, logging etc. Therefore, it is important
  that these are file-compatible names.
}
\references{
\addCitations{fastcluster}{1}
}
\seealso{
\code{\link{importFeatureGroups}} for more details and other algorithms.
}
