% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-envimass.R
\name{importFeatureGroupsEnviMass}
\alias{importFeatureGroupsEnviMass}
\title{Imports feature groups from enviMass}
\usage{
importFeatureGroupsEnviMass(path, feat, positive)
}
\arguments{
\item{path}{The path of the enviMass project.}

\item{feat}{The \code{\link{features}} object obtained with \code{\link{importFeaturesEnviMass}}.}

\item{positive}{Whether data from positive (\code{TRUE}) or negative (\code{FALSE}) should be loaded.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.

The \code{featuresSet} method (for \link[=sets-workflow]{sets workflows}) returns a
  \code{\link{featureGroupsSet}} object.
}
\description{
Imports a 'profiles' produced by \pkg{enviMass}.
}
\details{
This function imports data from enviMass. This function is called when calling \code{importFeatureGroups} with
  \code{type="envimass"}.

This function \emph{only} imports 'raw' profiles, \emph{not} any results from further componentization steps
  performed in \pkg{enviMass}. Furthermore, this functionality has only been tested with older versions of
  \pkg{enviMass}. Finally, please note that this function only supports features imported by
  \code{\link{importFeaturesEnviMass}} (obviously, the same project should be used for both importing functions).
}
\seealso{
\code{\link{importFeatureGroups}} for more details and other algorithms.
}
