% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-kpic2.R
\name{importFeatureGroupsKPIC2}
\alias{importFeatureGroupsKPIC2}
\title{Imports feature groups from KPIC2}
\usage{
importFeatureGroupsKPIC2(input, analysisInfo)
}
\arguments{
\item{input}{A grouped \code{PIC set} object (\emph{e.g.} as returned by
\code{\link[KPIC:PICset.group]{KPIC::PICset.group}}).}

\item{analysisInfo}{A \code{data.frame} (or \code{data.table}) with \link[=analysis-information]{Analysis
information}.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.
}
\description{
Imports feature groups from KPIC2
}
\details{
This function imports data from KPIC2. This function is called when calling \code{importFeatureGroups} with
  \code{type="kpic2"}.
}
\references{
\insertRef{Ji2017}{patRoon}
}
\seealso{
\code{\link{importFeatureGroups}} for more details and other algorithms.

\code{\link{groupFeatures}}
}
