% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-table.R
\name{importFeatureGroupsTable}
\alias{importFeatureGroupsTable}
\title{Import feature groups from a table}
\usage{
importFeatureGroupsTable(
  input,
  analysisInfo,
  addCols = NULL,
  groupAlgo,
  groupArgs = NULL
)
}
\arguments{
\item{input}{The input to be imported: either a \code{data.frame}, \code{data.table} or file path to a file that can
be imported with \code{\link{fread}}.}

\item{analysisInfo}{A \code{data.frame} (or \code{data.table}) with \link[=analysis-information]{Analysis
information}.}

\item{addCols}{Passed to \code{\link{importFeaturesTable}}.}

\item{groupAlgo, groupArgs}{\setsWF The grouping algorithm and a \code{list} with parameters that is used to re-group
features when adduct assignments are changed, \emph{e.g.} by \code{\link{selectIons}}. Hence, these are often
unused and dummy values can be set here, such as \code{groupAlgo="openms", groupArgs = list()}.}
}
\value{
An object derived from the class \code{\link{featureGroupsSet}} (if the imported data is from a
  \link[=sets-workflow]{sets workflow}) or \code{\link{featureGroups}} object otherwise.
}
\description{
This function imports grouped features from a table, which can be either a \code{data frame} or a path to a file with
tabular data (\emph{e.g.} \file{.csv}).
}
\details{
This function imports data from a table. This function is called when calling \code{importFeatureGroups} with
  \code{type="table"}.

This function can be used to import feature group data from a table generated by the
  \code{\link[=as.data.table,featureGroups-method]{as.data.table}} (and \code{as.data.frame}) function, or the output
  from any other feature detection and grouping software. The column format mostly follows the format used by the
  \code{as.data.table} function with the \code{features} argument set to \code{TRUE}.

  The function first imports the feature data from the table using \code{\link{importFeaturesTable}}. Please see its
  documentation for an overview of the columns that are expected in the input table.

  In addition to the columns for feature data, the following columns are expected in the input table: \itemize{

    \item \code{group}: a string naming the feature group in which the feature is part of. Values in this column
    should be unique per analysis. (\strong{mandatory})

    \item \code{group_ret},\code{group_mz}: the retention time (in seconds) and m/z assigned to the feature group.
    Will be calculated from mean values of feature data if missing. (\strong{optional})

  }

  For data from an \link[=assignMobilities_feat]{IMS workflow}, the following additional columns are relevant: \itemize{

    \item \code{group_mobility},\code{group_CCS}: the mobility and \acronym{CCS} assigned to the feature group. Will
    be calculated from mean values of feature data if missing. (\strong{optional})

    \item \code{ims_parent_group}: a string naming the IMS parent group of the feature group (\code{NA} if none or
    not a mobility feature). (\strong{optional})

  }

  For data from a \link[=sets-workflow]{sets workflow}, the following additional columns are relevant: \itemize{

    \item \code{group_ion_mz-<set>},\code{group_neutralMass},\code{group_adduct-<set>}: the ion m/z, neutral mass and
    adduct assigned to the feature group. Columns should be present for each set (\emph{e.g.}
    \code{ion_mz-positive}). Missing columns will be calculated from feature data. \strong{NOTE} If feature columns
    are missing, then the group columns will be used to create them.

  }
}
\note{
This function does not yet allow importing more advanced feature group properties, such as normalized
  intensities and predicted concentrations.
}
\seealso{
\code{\link{importFeatureGroups}} for more details and other algorithms.

\code{\link{importFeaturesTable}} for importing features from a table.
  \code{\link[=as.data.table,featureGroups-method]{as.data.table}} for converting feature groups data to a
  \code{data.table} format. \code{\link{groupFeatures}} to generate feature groups.
}
