% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-xcms.R
\name{importFeatureGroupsXCMS}
\alias{importFeatureGroupsXCMS}
\title{Imports feature groups from XCMS (old interface)}
\usage{
importFeatureGroupsXCMS(input, analysisInfo)
}
\arguments{
\item{input}{An \code{\link{xcmsSet}} object.}

\item{analysisInfo}{A \code{data.frame} (or \code{data.table}) with \link[=analysis-information]{Analysis
information}.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.
}
\description{
Imports feature groups from XCMS (old interface)
}
\details{
This function imports data from XCMS. This function is called when calling \code{importFeatureGroups} with
  \code{type="xcms"}.
}
\references{
\addCitations{xcms}
}
\seealso{
\code{\link{importFeatureGroups}} for more details and other algorithms.

\code{\link{groupFeaturesXCMS}}
}
