% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-xcms3.R
\name{importFeatureGroupsXCMS3}
\alias{importFeatureGroupsXCMS3}
\title{Imports feature groups from XCMS (new interface)}
\usage{
importFeatureGroupsXCMS3(input, analysisInfo)
}
\arguments{
\item{input}{An \code{\link{XCMSnExp}} object.}

\item{analysisInfo}{A \code{data.frame} (or \code{data.table}) with \link[=analysis-information]{Analysis
information}.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.
}
\description{
Imports feature groups from XCMS (new interface)
}
\details{
This function imports data from XCMS3. This function is called when calling \code{importFeatureGroups} with
  \code{type="xcms3"}.
}
\references{
\addCitations{xcms}
}
\seealso{
\code{\link{importFeatureGroups}} for more details and other algorithms.

\code{\link{groupFeaturesXCMS3}}
}
