% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-kpic2.R
\name{importFeaturesKPIC2}
\alias{importFeaturesKPIC2}
\title{Imports features from KPIC2}
\usage{
importFeaturesKPIC2(picsList, analysisInfo)
}
\arguments{
\item{picsList}{A \code{list} with a \code{pics} objects obtained with \code{\link[KPIC]{getPIC}} or
\code{\link[KPIC]{getPIC.kmeans}} for each analysis.}

\item{analysisInfo}{A \code{data.frame} with \link[=analysis-information]{Analysis information}.}
}
\value{
An object of a class which is derived from \code{\link{features}}.
}
\description{
Imports feature data generated by the \pkg{KPIC2} package.
}
\details{
This function imports data from KPIC2. This function is called when calling \code{importFeatures} with
  \code{type="kpic2"}.
}
\references{
\insertRef{Ji2017}{patRoon}
}
\seealso{
\code{\link{importFeatures}} for more details and other algorithms.
}
