% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-table.R
\name{importFeaturesTable}
\alias{importFeaturesTable}
\title{Import features from a table}
\usage{
importFeaturesTable(input, analysisInfo, addCols = NULL)
}
\arguments{
\item{input}{The input to be imported: either a \code{data.frame}, \code{data.table} or file path to a file that can
be imported with \code{\link{fread}}.}

\item{analysisInfo}{A \code{data.frame} (or \code{data.table}) with \link[=analysis-information]{Analysis
information}.}

\item{addCols}{A \code{character} vector with additional columns that should be included in the imported features
object. These columns are not used by \pkg{patRoon} and should not interfere with internally used columns (these
will be ignored with a warning if they do).}
}
\value{
An object derived from the class \code{\link{featuresSet}} (if the imported features are from a
  \link[=sets-workflow]{sets workflow}) or \code{\link{features}} object otherwise.
}
\description{
This function imports features from a table, which can be either a \code{data frame} or a path to a file with tabular
data (\emph{e.g.} \file{.csv}).
}
\details{
This function imports data from a table. This function is called when calling \code{importFeatures} with
  \code{type="table"}.

This function can be used to import features from a table generated by the
  \code{\link[=as.data.table,features-method]{as.data.table}} (and \code{as.data.frame}) function, or the output from
  any other feature detection software. The column format mostly follows the format used by the \code{as.data.table}
  function.

  The following columns must be present: \itemize{

    \item \code{analysis}: the analysis name (corresponding the the \link[=analysis-information]{analysis information}).

    \item \code{ret}: the retention time of the feature (in seconds).

    \item \code{mz}: the m/z value of the feature.

    \item \code{intensity}: the peak intensity of the feature.

  }

  The following columns are optional, but usually recommended: \itemize{

    \item \code{ID}: the feature ID. Should be unique across the features from the same analysis. If not present, a
    sequential ID is automatically generated.

    \item \code{area}: the peak area of the feature. If not present, it is calculated as \code{2.5*intensity}.

    \item \code{retmin} and \code{retmax}: the minimum and maximum retention time range of the feature. If not present,
    they are derived by subtraction or addition of \code{ret} by \code{defaultLim("retention", "narrow")} (see
    \link{limits}).

    \item \code{mzmin} and \code{mzmax}: the minimum and maximum m/z range of the feature. If not present, they are
    derived by subtraction or addition of \code{mz} by \code{defaultLim("mz", "narrow")} (see \link{limits}).

  }

  If a \code{mobility} column is present, it is assumed that the features are from an \link[=assignMobilities_feat]{IMS
  workflow}. In this case \itemize{

    \item \code{mobility}: specifies the mobility of the feature.

    \item \code{mobmin} and \code{mobmax}: the minimum and maximum mobility range of the feature. If not present,
    they are derived by subtraction or addition of \code{mobility} by \code{defaultLim("mobility", "narrow")} (see
    \link{limits}). (\strong{optional})

    \item \code{mob_area} and \code{mob_intensity}: the peak area and intensity of the peak in the mobilogram for the
    feature. (\strong{optional})

    \item \code{ims_precursor_ID}: the ID of the IMS precursor in a \link[=assignMobilities_feat]{post mobility
    assignment} workflow. (\strong{optional})

    \item \code{mob_assign_method}: a string that names the method used to assign the mobility to the feature.
    (\strong{optional})

  }

  If a \code{set} column is present, it is assumed that the features are from a \link[=sets-workflow]{sets workflow}.
  In this case \itemize{

    \item \code{set}: specifies the set name in which the feature is present.

    \item \code{mz} and \code{ion_mz}: specify the \emph{neutral mass} and \emph{ionized m/z} of the feature,
    respectively.

    \item \code{adduct}: specifies the \link{adduct} of the feature (generic textual format), \emph{e.g.}
    \code{"[M+H]+"}.

  }
}
\note{
The \code{"set"} column in the \link[=analysis-information]{analysis information} is not used when importing
  data (this column is present when obtaining the analysis information from a sets object with
  \code{\link{analysisInfo}}).
}
\seealso{
\code{\link{importFeatures}} for more details and other algorithms.

\code{\link{importFeatureGroupsTable}} to import feature group data from a table. 
  \code{\link[=as.data.table,features-method]{as.data.table}} for converting features to a \code{data.table} format.
  \code{\link{findFeatures}} to generate feature data.
}
