% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-xcms.R
\name{importFeaturesXCMS}
\alias{importFeaturesXCMS}
\title{Imports features from XCMS (old interface)}
\usage{
importFeaturesXCMS(input, analysisInfo)
}
\arguments{
\item{input}{An \code{\link{xcmsSet}} object.}

\item{analysisInfo}{A \code{data.frame} (or \code{data.table}) with \link[=analysis-information]{Analysis
information}.}
}
\value{
An object of a class which is derived from \code{\link{features}}.
}
\description{
Imports feature data generated with the legacy \code{\link[xcms]{xcmsSet}} function from the \pkg{xcms} package.
}
\details{
This function imports data from XCMS. This function is called when calling \code{importFeatures} with
  \code{type="xcms"}.
}
\references{
\addCitations{xcms}
}
\seealso{
\code{\link{importFeatures}} for more details and other algorithms.

\code{\link{importFeaturesXCMS3}}
}
