% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{installC3SDB}
\alias{installC3SDB}
\title{Automatically installs C3SDB}
\usage{
installC3SDB(envname = "patRoon-C3SDB", clearEnv = FALSE, ...)
}
\arguments{
\item{envname}{The name of the virtual \command{Python} environment to install \command{C3SDB} into. Passed to
\link[reticulate:py_install]{reticulate::py_install}. Set to \code{NULL} to not set any virtual \command{Python}
environment.}

\item{clearEnv}{Set to \code{TRUE} to remove the virtual environment if it already exists (using
\link[reticulate:virtualenv_remove]{reticulate::virtualenv_remove}). Ignored if \code{envname} is \code{NULL}.}

\item{\dots}{Further arguments passed to \code{\link[reticulate:py_install]{reticulate::py_install}}.}
}
\description{
Automatically installs the \href{https://github.com/dylanhross/c3sdb}{C3SDB} \command{Python} package
}
\details{
This function uses \CRANpkg{reticulate} to install the \href{https://github.com/dylanhross/c3sdb}{C3SDB}
\command{Python} package in a virtual environment.
}
\references{
\insertRef{Ross2020}{patRoon}
}
