% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{installTIMSCONVERT}
\alias{installTIMSCONVERT}
\title{Automatically installs TIMSCONVERT}
\usage{
installTIMSCONVERT(envname = "patRoon-TIMSCONVERT", clearEnv = FALSE, ...)
}
\arguments{
\item{envname}{The name of the virtual Python environment to install \command{TIMSCONVERT} into. Passed to
\link[reticulate:py_install]{reticulate::py_install}. Set to \code{NULL} to not set any virtual Python environment.}

\item{clearEnv}{Set to \code{TRUE} to remove the virtual environment if it already exists (using
\link[reticulate:virtualenv_remove]{reticulate::virtualenv_remove}). Ignored if \code{envname} is \code{NULL}.}

\item{\dots}{Further arguments passed to \code{\link[reticulate:py_install]{reticulate::py_install}}.}
}
\description{
Automatically installs the \href{https://gtluu.github.io/timsconvert/}{TIMSCONVERT} \command{Python} package
}
\details{
This function uses \CRANpkg{reticulate} to install the \href{https://gtluu.github.io/timsconvert/}{TIMSCONVERT}
Python package in a virtual environment.
}
\references{
\insertRef{Luu2022}{patRoon}
}
