% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/features-kpic2.R
\name{kpic2-conv}
\alias{kpic2-conv}
\alias{getPICSet,features-method}
\alias{getPICSet}
\alias{getPICSet,featuresKPIC2-method}
\title{Conversion to KPIC2 objects}
\usage{
getPICSet(obj, ...)

\S4method{getPICSet}{features}(
  obj,
  loadRawData = TRUE,
  IMS = FALSE,
  EICParams = getDefEICParams(window = 0)
)

\S4method{getPICSet}{featuresKPIC2}(obj, ...)
}
\arguments{
\item{obj}{The \code{features} object that should be converted.}

\item{\dots}{Ignored}

\item{loadRawData}{Set to \code{TRUE} if analyses are available as \code{mzXML} or \code{mzML} files. Otherwise MS
data is not loaded, and some dummy data (\emph{e.g.} file paths) is used in the returned object.}

\item{IMS}{\IMSWF Specifies which feature groups are considered for export in IMS workflows. The following options
  are valid: \itemize{

  \item \code{"both"}: Selects IMS and non-IMS features.

  \item \code{"maybe"}: Selects non-IMS features and IMS features without assigned IMS parent.

  \item \code{FALSE}: Selects only non-IMS features.

  \item \code{TRUE}: Selects only IMS features.

  }

  This should be kept \code{FALSE} as \pkg{KPIC2} export currently does not support IMS features.}

\item{EICParams}{A named \code{list} with parameters used for extracted ion chromatogram (EIC) creation. See the
\link[=EIXParams]{EIC parameters} documentation for more details.}
}
\description{
Converts a \code{\link{features}} object to an \pkg{KPIC} object.
}
\details{
The conversion process will introduce some dummy values for metadata not present in \pkg{patRoon} objects. If the
\code{features} object was generated with \pkg{KPIC2} and no post mobility assignment was performed, then no
conversion is performed and the original \pkg{KPIC2} object will be returned.
}
\section{Use of raw HRMS data}{
 The \link[=msdata]{raw data interface} of \pkg{patRoon} is used by \code{getPICSet} to
  process HRMS (or IMS-HRMS) data. Please see \link[=msdata]{its documentation} for more information on the supported
  formats and available configuration options.
}

