% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic_eim_gui.R
\name{launchEICGUI}
\alias{launchEICGUI}
\alias{launchEICGUI,features-method}
\alias{launchEICGUI,featureGroups-method}
\title{Launch EIC/EIM GUI}
\usage{
\S4method{launchEICGUI}{data.frame}(obj, suspects = NULL, adduct = NULL)

\S4method{launchEICGUI}{features}(obj)

\S4method{launchEICGUI}{featureGroups}(obj)
}
\arguments{
\item{obj}{For data.frame method: analysis information data.frame. For other methods: features/featureGroups object.}

\item{suspects}{Optional suspect list data.frame (data.frame method only).}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}.}
}
\value{
A Shiny app object.
}
\description{
Launches a Shiny app for generating and plotting EICs and EIMs.
}
\details{
These tools were originally developed as internal tools to debug and explore IMS related algorithms. They were mostly
generated with LLMs, hence, may contain nonsense and bugs. Nevertheless, they may be useful for users to explore
EICs and EIMs interactively. Please use with caution and report any issues.
}
