% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{loadCacheData}
\alias{loadCacheData}
\title{loadCacheData}
\usage{
loadCacheData(category, hashes, dbArg = NULL)
}
\arguments{
\item{category}{The category of the objects to be loaded from cache.}

\item{hashes}{The hash strings of the objects to be loaded.}

\item{dbArg}{Alternative connection to database. Default is `NULL` and uses
the cache options as defined by "patRoon.cache.fileName".}
}
\description{
Loads cached objects.
}
