% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mslibrary.R
\name{loadMSLibrary}
\alias{loadMSLibrary}
\title{Loading of MS library data}
\usage{
loadMSLibrary(file, algorithm, ...)
}
\arguments{
\item{file}{A \code{character} string that specifies the the file path to the library.}

\item{algorithm}{A character string describing the algorithm that should be
used: \code{"msp"}, \code{"json"}}

\item{\dots}{Any parameters to be passed to the selected MS library loading algorithm.}
}
\value{
A \code{\link{MSLibrary}} object containing the loaded library data.
}
\description{
Loads, parses, verifies and curates MS library data, \emph{e.g.} obtained from MassBank.
}
\details{
\code{loadMSLibrary} is a generic function that will loads MS library data by one of the supported algorithms. The actual
  functionality is provided by algorithm specific functions such as \code{loadMSLibraryMSP} and \code{loadMSLibraryMoNAJSON}. While these
  functions may be called directly, \code{loadMSLibrary} provides a generic interface and is therefore usually preferred.
}
\seealso{
The \code{\link{MSLibrary}} output class and its methods and the algorithm specific functions:
  \code{\link{loadMSLibraryMSP}}, \code{\link{loadMSLibraryMoNAJSON}}
}
