% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{makeHash}
\alias{makeHash}
\title{makeHash}
\usage{
makeHash(..., checkDT = TRUE)
}
\arguments{
\item{...}{Arguments/objects to be used for hashing.}

\item{checkDT}{Logical, set to `TRUE` for striping DT self references as they
sometimes lead hashing issues.}
}
\description{
Make a hash string of given arguments.
}
