% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/features-set.R,
%   R/feature_groups-set.R
\name{makeSet}
\alias{makeSet}
\alias{makeSet,features-method}
\alias{makeSet,featuresSet-method}
\alias{makeSet,featureGroups-method}
\alias{makeSet,featureGroupsSet-method}
\title{Initiate sets workflows}
\usage{
\S4method{makeSet}{features}(obj, ..., adducts, labels = NULL)

\S4method{makeSet}{featuresSet}(obj, ...)

\S4method{makeSet}{featureGroups}(
  obj,
  ...,
  groupAlgo,
  groupArgs = NULL,
  verbose = TRUE,
  adducts = NULL,
  labels = NULL
)

\S4method{makeSet}{featureGroupsSet}(obj, ...)
}
\arguments{
\item{obj, \dots}{\code{\link{features}} or \code{\link{featureGroups}} objects that should be used for the
\link[=sets-workflow]{sets workflow}.}

\item{adducts}{The adduct assignments to each set. Should either be a \code{list} with \code{\link{adduct}} objects
  or a \code{character} vector (\emph{e.g.} \code{"[M+H]+"}). The order should follow that of the objects given to
  the \code{obj} and \code{\dots} arguments.

  For the \code{\link{featureGroups}} method: if \code{NULL} then adduct annotations are used.}

\item{labels}{The labels, or \emph{set names}, for each set to be created. The order should follow that of the
objects given to the \code{obj} and \code{\dots} arguments. If \code{NULL}, then labels are automatically generated
from the polarity of the specified \code{adducts} argument (\emph{e.g.} \code{"positive"}, \code{"negative"}).}

\item{groupAlgo}{groupAlgo The name of the feature grouping algorithm. See the \code{algorithm} argument of
\code{\link{groupFeatures}} for details.}

\item{groupArgs}{A \code{list} with arguments directly passed to \code{groupFeatures} (can be named). Example:
\code{groupArgs=list(maxAlignMZ=0.002)}.}

\item{verbose}{If set to \code{FALSE} then no text output is shown.}
}
\value{
Either a \code{\link{featuresSet}} object (\code{features} method) or \code{\link{featureGroupsSet}} object
  (\code{featureGroups} method).
}
\description{
Initiate sets workflows from specified feature data.
}
\details{
The \code{makeSet} method function is used to initiate a \link[=sets-workflow]{sets workflow}. The features from
input objects are combined and then \emph{neutralized} by replacing their \emph{m/z} values by neutral monoisotopic
masses. After neutralization features measured with \emph{e.g.} different ionization polarities can be grouped since
their neutral mass will be the same.

The \link[=analysis-information]{analysis information} for this object is updated with all analyses, and a \code{set}
column is added to designate the set of each analysis. Note that currently, all analyses names \strong{must} be
unique across different sets.

\code{makeSet} supports two types of input: \enumerate{

\item \code{\link{features}} objects: \code{makeSet} combines the input objects into a \code{featuresSet} object,
which is then grouped in the 'usual way' with \code{\link{groupFeatures}}.

\item \code{\link{featureGroups}} objects: In this case the features from the input objects are first neutralized and
feature groups between sets are then combined with \code{groupFeatures}.

}

The advantage of the \code{featureGroups} method is that it preserves any adduct annotations already present
(\emph{e.g.} as set by \code{selectIons} or \code{adducts<-}). Furthermore, this approach allows more advanced
workflows where the input \code{featureGroups} are first pre-treated with \emph{e.g.} filter before the sets object
is made. On the other hand, the \code{features} method is easier, as it doesn't require intermediate feature grouping
steps and is often sufficient since adduct annotations can be made afterwards with \code{selectIons}/\code{adducts<-}
and most \code{filter} operations do not need to be done per individual set.

The adduct information used for feature neutralization is specified through the \code{adducts} argument.
Alternatively, when the \code{featureGroups} method of \code{makeSet} is used, then the adduct annotations already
present in the input objects can also by used by setting \code{adducts=NULL}. The adduct information is also used to
add adduct annotations to the output of \code{makeSet}.
}
\note{
Initiating a sets workflow recursively, \emph{i.e.} with \code{featuresSet} or \code{featureGroupsSet} objects
  as input, is currently not supported.
}
