% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{MSConversion}
\alias{MSConversion}
\alias{getMSConversionTypes}
\alias{getMSConversionFormats}
\alias{convertMSFilesPWiz}
\alias{convertMSFilesOpenMS}
\alias{convertMSFilesBruker}
\alias{convertMSFilesIMSCollapse}
\alias{convertMSFilesTIMSCONVERT}
\alias{convertMSFilePaths}
\alias{convertMSFiles}
\title{MS data conversion}
\usage{
getMSConversionTypes(algorithm, direction)

getMSConversionFormats(algorithm, direction, type = NULL)

convertMSFilesPWiz(
  inFiles,
  outFiles,
  formatTo = "mzML",
  centroid = TRUE,
  IMS = FALSE,
  minIntensity = 0,
  filters = NULL,
  extraOpts = NULL,
  PWizBatchSize = 1
)

convertMSFilesOpenMS(inFiles, outFiles, formatTo = "mzML", extraOpts = NULL)

convertMSFilesBruker(inFiles, outFiles, formatTo = "mzML", centroid = TRUE)

convertMSFilesIMSCollapse(
  inFiles,
  outFiles,
  typeFrom,
  formatTo = "mzML",
  mzRange = NULL,
  mobilityRange = NULL,
  smoothWindow = 0,
  halfWindow = 2,
  maxGap = 0.005,
  clusterMethod = "distance_mean",
  mzWindow = defaultLim("mz", "medium"),
  minIntensityIMS = 0,
  includeMSMS = FALSE,
  ...
)

convertMSFilesTIMSCONVERT(
  inFiles,
  outFiles,
  formatTo = "mzML",
  centroid = TRUE,
  centroidRaw = FALSE,
  IMS = FALSE,
  extraOpts = NULL,
  virtualenv = "patRoon-TIMSCONVERT"
)

convertMSFilePaths(
  files,
  formatFrom,
  formatTo = "mzML",
  outPath = NULL,
  dirs = TRUE,
  overwrite = FALSE,
  algorithm = "pwiz",
  ...
)

convertMSFiles(
  anaInfo,
  typeFrom = "raw",
  typeTo = "centroid",
  formatFrom,
  formatTo = "mzML",
  overwrite = FALSE,
  algorithm = "pwiz",
  centroidVendor = TRUE,
  ...
)
}
\arguments{
\item{algorithm}{Either \code{"pwiz"} (ProteoWizard), \code{"openms"}, \code{"bruker"} (Bruker DataAnalysis) ,
\code{"imscollapse"} or \code{"timsconvert"}.}

\item{direction}{A \code{character} specifying the direction of conversion. Either \code{"input"} or \code{"output"}.}

\item{type, typeFrom, typeTo}{The type of the input or output files. See \code{getMSConversionTypes} for the supported
types.}

\item{inFiles, outFiles}{A \code{character} vector with input and output files, respectively. Lengths and order should
be the same.}

\item{centroid}{Set to \code{TRUE} to perform centroiding.

  For \code{convertMSFilesPWiz}: the value may be \code{"vendor"} to perform centroiding with the vendor algorithm or
  \code{"cwt"} to use ProteoWizard's wavelet algorithm.}

\item{IMS}{How to handle IMS data.

  For \code{convertMSFilesPWiz}: if \code{TRUE} then IMS data is exported and spectra for each IMS frame are combined
  into a single spectrum (using the \command{--combineIonMobilitySpectra} option), which is the format supported by
  \pkg{patRoon}. Set to \code{NA} to collapse the IMS data by scan summing, which mimics 'regular' HRMS data. Set to
  \code{FALSE} for non-IMS data. \strong{NOTE}: do not set \code{IMS=FALSE} if the data has IMS data. This will
  result in very large files where MS spectra are not combined by frame, which \strong{cannot} be properly read by
  \pkg{patRoon}.

  For \code{convertMSFilesTIMSCONVERT}: set to \code{TRUE} to keep IMS data or \code{FALSE} to exclude IMS data to
  mimic 'regular' LC-MS data.}

\item{minIntensity}{The minimum intensity of the mass peaks to be kept. Applying an intensity threshold is especially
beneficial to reduce export file size when there are a lot of zero or very low intensity mass peaks. \strong{NOTE}
this currently does \emph{not} work well with IMS data.}

\item{filters}{A \code{character} vector specifying one or more filters to \command{msconvert}. The elements of the
specified vector are directly passed to the \code{--filter} option (see
\href{http://proteowizard.sourceforge.net/tools/filters.html}{here})}

\item{extraOpts}{A \code{character} vector specifying any extra command line parameters passed to \command{msconvert}
or \command{FileConverter}. Set to \code{NULL} to ignore. For options: see
\href{https://abibuilder.cs.uni-tuebingen.de/archive/openms/Documentation/nightly/html/TOPP_FileConverter.html}{FileConverter}
and \href{http://proteowizard.sourceforge.net/tools/msconvert.html}{msconvert}.}

\item{PWizBatchSize}{The number of analyses to process by a single call to \command{msconvert}. Usually a value of
one is most efficient. Set to zero to run all analyses all at once from a single call.}

\item{mzRange, mobilityRange}{A two sized vector specifying the m/z and mobility range to be exported, respectively.
Set to \code{NULL} to export the full range.}

\item{smoothWindow, halfWindow, maxGap}{Centroiding parameters: see \code{\link{getDefAvgPListParams}} for details.
\strong{NOTE}: As described there, \code{maxGap} may need to be increased for Agilent instruments (\emph{e.g.}
\samp{0.01}).}

\item{clusterMethod, mzWindow}{The clustering method and window (see \link[=cluster-params]{clustering parameters})
used to find and combine MS/MS spectra of precursors with close \emph{m/z}.}

\item{minIntensityIMS}{The minimum intensity for MS peaks in raw data.}

\item{includeMSMS}{Set to \code{TRUE} to include MS/MS spectra in the output. For IMS workflows where IMS data is
only collapsed to produce compatible data files for feature detection, MS/MS data are not needed and can be
excluded to reduce computational times and file sizes. Setting \code{includeMSMS=TRUE} is primarily intended to
perform 'classical LC-MS workflows' with IMS data.}

\item{\dots}{For \code{convertMSFilesIMSCollapse}: further arguments passed to
  \code{\link[mzR:writeMSData]{mzR::writeMSData}}.

  For \code{convertMSFilePaths} and \code{convertMSFiles}: further arguments passed to algorithm specific conversion
  functions.}

\item{centroidRaw}{Only applicable if \code{IMS=FALSE}. Sets the \code{mode} parameter of \command{TIMSCONVERT}:
\code{raw} if \code{centroidRaw=TRUE} or \code{centroid} if \code{centroidRaw=FALSE}. See
\url{https://gtluu.github.io/timsconvert/local.html#notes-on-mode-parameter} for more details.}

\item{virtualenv}{The virtual Python environment in which \command{TIMSCONVERT} is installed. This is passed to
\code{\link[reticulate:use_virtualenv]{reticulate::use_virtualenv}}, which will ensure that the
\command{TIMSCONVERT} command line utility can be found by \pkg{patRoon}. Set to \code{NULL} to skip this step.}

\item{files, dirs}{The \code{files} argument should be a \code{character} vector with input files. If \code{files}
contains directories and \code{dirs=TRUE} then files from these directories are also considered.}

\item{formatFrom, formatTo}{The input or output format. See \code{getMSConversionFormats} for the supported formats.}

\item{outPath}{A character vector specifying directories that should be used for the output. Will be re-cycled if
necessary. If \code{NULL}, output directories will be kept the same as the input directories.}

\item{overwrite}{Should existing destination file be overwritten (\code{TRUE}) or not (\code{FALSE})?}

\item{anaInfo}{An \link[=analysis-information]{analysis info table} that is used to retrieve the input files. The
paths set by \code{path_centroid}, \code{path_profile} and \code{path_ims} are used to determine the output
directories. This function automatically determines if and how centroiding and IMS conversions should be applied.}

\item{centroidVendor}{Only for \code{algorithm="pwiz"}: whether centroiding should be performed with vendor
algorithms.}
}
\description{
Conversion of MS analysis files between several open and closed data formats.
}
\details{
\code{getMSConversionTypes} returns a \code{character} with all supported input or output conversion types
  for an algorithm.

\code{getMSConversionFormats} returns a \code{character} with all supported input or output conversion
  formats for an algorithm, optionally filtered by the given \code{type}.

\code{convertMSFilesPWiz} converts and pre-treats HRMS data with the \command{msconvert} tool from
  \href{http://proteowizard.sourceforge.net/}{ProteoWizard}.

\code{convertMSFilesOpenMS} converts HRMS data with the \command{FileConvert} tool of
  \href{http://www.openms.de/}{OpenMS}.

\code{convertMSFilesBruker} converts and pre-treats Bruker HRMS data with Bruker DataAnalysis. Note that
  TIMS data currently is not supported.

\code{convertMSFilesIMSCollapse} is used to convert IMS data to data that mimics 'regular' HRMS data by
  collapsing the IMS dimension. The raw data interface of \pkg{patRoon} first sums up all spectra within each IMS
  frame, performs centroiding and finally exports the resulting data with the
  \code{\link[mzR:writeMSData]{mzR::writeMSData}} function. Several thresholds can be set to speed up the conversion
  process and reduce noise, but care should be taken that no mass peaks of interest are lost.

\code{convertMSFilesTIMSCONVERT} converts and pre-treats TIMS data with
  \href{https://gtluu.github.io/timsconvert/}{TIMSCONVERT}. The \code{\link{installTIMSCONVERT}} function can be used
  to automatically install \command{TIMSCONVERT}.

\code{convertMSFilePaths} is a wrapper function that simplifies the use of algorithm specific MS conversion
  functions, such as \code{convertMSFilesPWiz}, and \code{convertMSFilesTIMSCONVERT}.

\code{convertMSFiles} is a wrapper function that simplifies the use of \code{convertMSFilePaths}.
}
\section{Parallelization}{
 \code{convertMSFilesPWiz}, \code{convertMSFilesOpenMS} and \code{convertMSFilesTIMSCONVERT} uses multiprocessing to parallelize
  computations. Please see the parallelization section in the handbook for
  more details and \link[=patRoon-package]{patRoon options} for configuration
  options.
}

\section{Use of raw HRMS data}{
 The \link[=msdata]{raw data interface} of \pkg{patRoon} is used by \code{convertMSFilesIMSCollapse} to
  process HRMS (or IMS-HRMS) data. Please see \link[=msdata]{its documentation} for more information on the supported
  formats and available configuration options.
}

\references{
\insertRef{Rst2016}{patRoon} \cr\cr

  \insertRef{Chambers2012}{patRoon} \cr\cr

  \insertRef{Luu2022}{patRoon} \cr\cr

  \addCitations{mzR}
}
