% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/mspeaklists.R,
%   R/mspeaklists-bruker.R, R/mspeaklists-mzr.R
\docType{methods}
\name{MSPeakLists-generation}
\alias{MSPeakLists-generation}
\alias{generateMSPeakLists,featureGroups-method}
\alias{generateMSPeakLists}
\alias{generateDAMSPeakLists}
\alias{generateDAFMFMSPeakLists}
\alias{generateMzRPeakLists}
\title{Generation of MS Peak Lists}
\usage{
\S4method{generateMSPeakLists}{featureGroups}(fGroups, algorithm, ...)

generateDAMSPeakLists(fGroups, bgsubtr = TRUE, maxRtMSWidth = NULL,
  clear = TRUE, save = TRUE, MSMSType = "MSMS")

generateDAFMFMSPeakLists(fGroups)

generateMzRPeakLists(fGroups, avgMzWindow = 0.001, maxRtMSWidth = NULL,
  avgTopPeaks = 50, avgMinIntensity = 500, avgMassFun = mean,
  avgMethod = "distance", precursorMzWindow = 8, topMost = NULL)
}
\arguments{
\item{fGroups}{The \code{\link{featureGroups}} object from which MS peak
lists should be extracted.}

\item{algorithm}{A character string describing the algorithm that should be
used: \code{"bruker"}, \code{"brukerfmf"}, \code{"mzr"}}

\item{\dots}{Any parameters to be passed to the selected MS peak lists
generation algorithm.}

\item{bgsubtr}{If \code{TRUE} background will be subtracted using the
'spectral' algorithm.}

\item{maxRtMSWidth}{Maximum total chromatographic peak width (seconds) used
for spectrum averaging. If \code{NULL} all spectra from a feature will be
taken into account. Lower to decrease processing time.}

\item{clear}{Remove any existing chromatogram traces/mass spectra prior to
making new ones.}

\item{save}{Save data files after completion.}

\item{MSMSType}{The type of MS/MS experiment performed: \code{"MSMS"} for
MRM/AutoMSMS or \code{"BBCID"} for broadband CID.}

\item{avgMzWindow}{\emph{m/z} window (in Da) used for clustering \emph{m/z}
values when spectra are averaged. Too small windows will prevent clustering
\emph{m/z} values (thus erroneously creating 'extra' values), whereas too
big windows may cluster unrelated \emph{m/z} values from different or even
the same spectrum together.}

\item{avgTopPeaks}{Only retain a maximum number of \code{avgTopPeaks} MS
peaks when generating averaged spectra. Lowering this number may exclude
more irrelevant (noisy) MS peaks and decrease processing time, whereas
higher values may avoid excluding lower intense MS peaks that may still be
of interest.}

\item{avgMinIntensity}{MS peaks with intensities below this value will be
completely excluded (overrides \code{avgTopPeaks}).}

\item{avgMassFun}{Function that is used to calculate average \emph{m/z}
values.}

\item{avgMethod}{Method used for producing averaged MS spectra. Valid values
are \code{"hclust"}, used for hierarchical clustering, and
\code{"distance"}, to use the between peak distance. The latter method
significantly reduces processing time and memory requirements, at the
(potential) cost of slightly reducing accuracy.}

\item{precursorMzWindow}{The \emph{m/z} window (in Da) to find MS/MS spectra
of a precursor. This is typically used for Data-Dependent like MS/MS data
and should correspond to the isolation \emph{m/z} width that was used to
collect the data. For Data-Independent MS/MS experiments, where precursor
ions are not isolated prior to fragmentation (\emph{e.g.} bbCID, MSe,
all-ion, ...) the value should be \code{NULL}.}

\item{topMost}{Only extract MS peak lists from a maximum of \code{topMost}
analyses with highest intensity. If \code{NULL} all analyses will be used.}
}
\value{
A \code{\link{MSPeakLists}} object that can be used for formulae
  calculation and compound identification.
}
\description{
Functionality to generate MS peak lists.
}
\details{
Formula calculation and identification tools rely on mass spectra that belong
to features of interest. For processing, MS (and MS/MS) spectra are typically
reduced to a table with a column containing measured \emph{m/z} values and a
column containing their intensities. These 'MS peak lists' can then be used
for \link[=formula-generation]{formula generation} and
\link[=compound-generation]{compound generation}. By default, peak lists are
generated from all analyses containing a particular feature group. The main
advantage is that this will allow subsequent processes (at the moment only
formulae calculation) to generate a consensus from feature group data
originating from multiple analyses. Nevertheless, it might be useful to limit
to maximum amount of analyses (\code{topMost} parameter) to signficantly
reduce the processing time required for MS peak list extraction.

Several functions exist to automatically extract MS peak lists for feature
groups.

\code{generateMSPeakLists} is a generic function that will generate MS peak lists
  using one of the supported algorithms. The actual functionality is provided
  by algorithm specific functions such as \code{generateFormulasDA} and
  \code{generateFormulasGenForm}. While these functions may be called directly,
  \code{generateMSPeakLists} provides a generic interface and is therefore usually
  preferred.

\code{generateDAMSPeakLists} uses Bruker DataAnalysis to generate MS
  peak lists. Naturally, this only works with analyses in the Bruker data
  format (\file{.d}). This function leverages DataAnalysis functionality to
  support averaging of spectra, background subtraction and identification of
  isotopes. In order to obtain mass spectra TICs will be added of the MS and
  relevant MS/MS signals.

\code{generateDAFMFMSPeakLists} is similar to \code{generateDAMSPeakLists},
  but uses compounds that were generated by the Find Molecular Features (FMF)
  algorithm to extract MS peak lists. This is generally much faster than
  \code{generateDAMSPeakLists}, however, it only works when features were obtained
  using the \code{\link{findFeaturesBruker}} function. Since all MS spectra
  are generated in advance by Bruker DataAnalysis, no further parameters
  exist to customize its operation.

\code{generateMzRPeakLists} uses the \pkg{\link{mzR}} package to
  extract MS peak lists. For this analyses should be either in \file{.mzXML}
  or \file{.mzML} format. This function can optionally average multiple
  spectra over a chromatgraphic peak to improve accuracy.
}
\note{
\code{generateDAMSPeakLists} requires that the \option{Component}
  column is active (Method-->Parameters-->Layouts-->Mass List Layout).
}
\references{
Averaging of mass spectra algorithms used by
  \code{generateMzRPeakLists} were based on the
  \href{https://github.com/zeehio/msProcess}{msProcess} R package (now
  archived on CRAN). \cr\cr
  \addCitations{mzR}{1} \cr\cr
  \addCitations{mzR}{2} \cr\cr
  \addCitations{mzR}{3} \cr\cr
  \addCitations{mzR}{4} \cr\cr
  \addCitations{mzR}{5}
}
\seealso{
\code{\link{MSPeakLists-class}}
}
