% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-tool.R
\name{newProject}
\alias{newProject}
\title{Easily create new \pkg{patRoon} projects}
\usage{
newProject(destPath = NULL)
}
\arguments{
\item{destPath}{Set destination path value to this value (useful for debugging). Set to \code{NULL} for a default
value.}
}
\description{
The \code{newProject} function is used to quickly generate a processing R script. This tool allows the user to
quickly select the targeted analyses, workflow steps and configuring some of their common parameters. This function
requires to be run within a \href{https://www.rstudio.com/}{RStudio} session. The resulting script is either added to
the current open file or to a new file. The \link[=analysis-information]{analysis information} will be written to a
\file{.csv} file so that it can easily be modified afterwards.
}
