% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe-optimizer.R
\docType{class}
\name{optimizationResult-class}
\alias{optimizationResult-class}
\alias{optimizationResult}
\alias{algorithm,optimizationResult-method}
\alias{length,optimizationResult-method}
\alias{lengths,optimizationResult-method}
\alias{show,optimizationResult-method}
\alias{plot,optimizationResult,missing-method}
\alias{optimizedParameters,optimizationResult-method}
\alias{optimizedParameters}
\alias{optimizedObject,optimizationResult-method}
\alias{optimizedObject}
\alias{scores,optimizationResult-method}
\alias{scores}
\alias{experimentInfo,optimizationResult-method}
\alias{experimentInfo}
\title{Class containing optimization results.}
\usage{
\S4method{algorithm}{optimizationResult}(obj)

\S4method{length}{optimizationResult}(x)

\S4method{lengths}{optimizationResult}(x, use.names = FALSE)

\S4method{show}{optimizationResult}(object)

\S4method{plot}{optimizationResult,missing}(
  x,
  paramSet,
  DoEIteration,
  paramsToPlot = NULL,
  maxCols = NULL,
  type = "contour",
  image = TRUE,
  contours = "colors",
  ...
)

\S4method{optimizedParameters}{optimizationResult}(object, paramSet = NULL, DoEIteration = NULL)

\S4method{optimizedObject}{optimizationResult}(object, paramSet = NULL)

\S4method{scores}{optimizationResult}(object, paramSet = NULL, DoEIteration = NULL)

\S4method{experimentInfo}{optimizationResult}(object, paramSet, DoEIteration)
}
\arguments{
\item{obj, x, object}{An \code{optimizationResult} object.}

\item{use.names}{Ignored.}

\item{paramSet}{Numeric index of the parameter set (\emph{i.e.} the first
parameter set gets index \samp{1}). For some methods optional: if
\code{NULL} the best will be selected.}

\item{DoEIteration}{Numeric index specifying the DoE iteration within the
specified \code{paramSet}. For some methods optional: if \code{NULL} the
best will be selected.}

\item{paramsToPlot}{Which parameters relations should be plot. If \code{NULL}
all will be plot. Alternatively, a \code{list} containing one or more
\code{character} vectors specifying each two parameters that should be
plotted. Finally, if only one pair should be plotted, can be a
\code{character} vector specifying both parameters.}

\item{maxCols}{Multiple parameter pairs are plotted in a grid. The maximum
number of columns can be set with this argument. Set to \code{NULL} for no
limit.}

\item{type}{The type of plots to be generated: \code{"contour"},
\code{"image"} or \code{"persp"}. The equally named functions will be
called for plotting.}

\item{image}{Passed to \code{\link{contour}} (if \code{type="contour"}).}

\item{contours}{Passed to \code{\link{persp}} (if \code{type="persp"}).}

\item{\dots}{Further arguments passed to \code{\link{contour}},
\code{\link{image}} or \code{\link{persp}} (depending on \code{type}).}
}
\description{
Objects from this class contain optimization results resulting from design of
experiment (DoE).
}
\details{
Objects from this class are returned by \code{\link{optimizeFeatureFinding}} and
\code{\link{optimizeFeatureGrouping}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{algorithm}: Returns the algorithm that was used for finding features.

\item \code{length}: Obtain total number of experimental design iterations performed.

\item \code{lengths}: Obtain number of experimental design iterations performed for each parameter set.

\item \code{show}: Shows summary information for this object.

\item \code{plot}: Generates response plots for all or a selected
set of parameters.

\item \code{optimizedParameters}: Returns parameter set yielding optimal
results. The \code{paramSet} and \code{DoEIteration} arguments can be
\code{NULL}.

\item \code{optimizedObject}: Returns the object (\emph{i.e.} a
\code{\link{features}} or \code{\link{featureGroups}} object) that was
generated with optimized parameters. The \code{paramSet} argument can be
\code{NULL}.

\item \code{scores}: Returns optimization scores. The
\code{paramSet} and \code{DoEIteration} arguments can be \code{NULL}.

\item \code{experimentInfo}: Returns a \code{list} with optimization
information from an DoE iteration.
}}

\section{Slots}{

\describe{
\item{\code{algorithm}}{A character specifying the algorithm that was optimized.}

\item{\code{paramSets}}{A \code{list} with detailed results from each parameter set
that was tested.}

\item{\code{bestParamSet}}{Numeric index of the parameter set yielding the best
response.}
}}

\examples{
\dontrun{
# ftOpt is an optimization object.

# plot contour of all parameter pairs from the first parameter set/iteration.
plot(ftOpt, paramSet = 1, DoEIteration = 1)

# as above, but only plot two parameter pairs
plot(ftOpt, paramSet = 1, DoEIteration = 1,
     paramsToPlot = list(c("mzPPM", "chromFWHM"), c("chromFWHM", "chromSNR")))

# plot 3d perspective plots
plot(ftOpt, paramSet = 1, DoEIteration = 1, type = "persp")
}

}
