% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{patRoon-defunct}
\alias{patRoon-defunct}
\alias{compoundViewer,featureGroups,MSPeakLists,compounds-method}
\alias{compoundViewer}
\title{Defunct functions.}
\usage{
\S4method{compoundViewer}{featureGroups,MSPeakLists,compounds}(fGroups, MSPeakLists, compounds)
}
\arguments{
\item{MSPeakLists}{A \code{\link{MSPeakLists}} object.}

\item{compounds}{A \code{\link{compounds}} object.}
}
\description{
These functions do not work anymore and may be updated in the future.
future.
}
\details{
The \code{compoundViewer} method is used to view compound
  identification results. It will display available candidate information
  such as scorings and identifiers, MS/MS spectra with explained peaks and
  chemical structures.
}
\keyword{internal}
