% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{patRoon-deprecated}
\alias{patRoon-deprecated}
\alias{reportHTML}
\alias{annotateSuspects}
\title{Deprecated and renamed functions.}
\usage{
reportHTML(...)

annotateSuspects(...)
}
\arguments{
\item{\dots}{Passed to successor function.}
}
\description{
Please do not use these functions anymore since they may be removed in the
future.
}
\details{
\code{reportHTML} performs HTML reporting, please use
  \code{\link{report}} instead.

Please use \code{\link{estimateIDConfidence}} instead.
}
\keyword{internal}
