% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{printPackageOpts}
\alias{printPackageOpts}
\title{Prints all the package options of \code{patRoon} and their currently set values.}
\usage{
printPackageOpts()
}
\description{
Prints all the package options of \code{patRoon} and their currently set values.
}
