% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/report-legacy.R
\name{reporting-legacy}
\alias{reporting-legacy}
\alias{reportCSV,featureGroups-method}
\alias{reportCSV}
\alias{reportPDF,featureGroups-method}
\alias{reportPDF}
\title{Report feature group data (legacy interface)}
\usage{
reportCSV(
  fGroups,
  path = "report",
  reportFeatures = FALSE,
  formulas = NULL,
  formulasNormalizeScores = "max",
  formulasExclNormScores = NULL,
  compounds = NULL,
  compoundsNormalizeScores = "max",
  compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount",
    "annotHitCount", "libMatch"),
  compsCluster = NULL,
  components = NULL,
  retMin = TRUE,
  clearPath = FALSE
)

reportPDF(
  fGroups,
  path = "report",
  reportFGroups = TRUE,
  formulas = NULL,
  formulasTopMost = 5,
  formulasNormalizeScores = "max",
  formulasExclNormScores = NULL,
  reportFormulaSpectra = TRUE,
  compounds = NULL,
  compoundsNormalizeScores = "max",
  compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount",
    "annotHitCount", "libMatch"),
  compoundsOnlyUsedScorings = TRUE,
  compoundsTopMost = 5,
  compsCluster = NULL,
  components = NULL,
  MSPeakLists = NULL,
  retMin = TRUE,
  EICGrid = c(2, 1),
  EICParams = getDefEICParams(window = 20, topMost = 1, topMostByReplicate = TRUE),
  clearPath = FALSE
)

\S4method{reportCSV}{featureGroups}(
  fGroups,
  path = "report",
  reportFeatures = FALSE,
  formulas = NULL,
  formulasNormalizeScores = "max",
  formulasExclNormScores = NULL,
  compounds = NULL,
  compoundsNormalizeScores = "max",
  compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount",
    "annotHitCount", "libMatch"),
  compsCluster = NULL,
  components = NULL,
  retMin = TRUE,
  clearPath = FALSE
)

\S4method{reportPDF}{featureGroups}(
  fGroups,
  path = "report",
  reportFGroups = TRUE,
  formulas = NULL,
  formulasTopMost = 5,
  formulasNormalizeScores = "max",
  formulasExclNormScores = NULL,
  reportFormulaSpectra = TRUE,
  compounds = NULL,
  compoundsNormalizeScores = "max",
  compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount",
    "annotHitCount", "libMatch"),
  compoundsOnlyUsedScorings = TRUE,
  compoundsTopMost = 5,
  compsCluster = NULL,
  components = NULL,
  MSPeakLists = NULL,
  retMin = TRUE,
  EICGrid = c(2, 1),
  EICParams = getDefEICParams(),
  clearPath = FALSE
)
}
\arguments{
\item{fGroups}{The \code{\link{featureGroups}} object that should be used for reporting data.}

\item{path}{The destination file path for files generated during reporting. Will be generated if needed.}

\item{reportFeatures}{If set to \code{TRUE} then for each analysis a
\file{.csv} file will be generated with information about its detected
features.}

\item{formulas, compounds, compsCluster, components}{Further objects (\code{\link{formulas}}, \code{\link{compounds}},
\code{\link{compoundsCluster}}, \code{\link{components}}) that should be reported. Specify \code{NULL} to skip
reporting a particular object.}

\item{compoundsNormalizeScores, formulasNormalizeScores}{A \code{character} that specifies how normalization of
annotation scorings occurs. Either
\code{"none"} (no normalization),
\code{"max"} (normalize to max value) or \code{"minmax"} (perform min-max
normalization). Note that normalization of negative scores (e.g. output by
\command{SIRIUS}) is always performed as min-max. Furthermore, currently
normalization for \code{compounds} takes the original min/max scoring
values into account when candidates were generated. Thus, for
\code{compounds} scoring, normalization is not affected when candidate
results were removed after they were generated (\emph{e.g.} by use of
\code{filter}).}

\item{compoundsExclNormScores, formulasExclNormScores}{A
  \code{character} vector specifying any compound scoring names that
  should \emph{not} be normalized. Set to \code{NULL} to normalize all
  scorings. Note that whether any normalization occurs is set by the
  \code{compoundsExclNormScores,formulasExclNormScores} argument.

  For \code{compounds}: By default \code{score} and
  \code{individualMoNAScore} are set to mimic the behavior of the
  \command{MetFrag} web interface.}

\item{retMin}{If \code{TRUE} then report retention times in minutes (otherwise seconds).}

\item{clearPath}{If \code{TRUE} then the destination path will be (recursively) removed prior to reporting.}

\item{reportFGroups}{If \code{TRUE} then feature group data will be reported.}

\item{formulasTopMost, compoundsTopMost}{Only this amount of top ranked candidate formulae/compounds are reported.
Lower values may significantly speed up reporting. Set to \code{NULL} to ignore.}

\item{reportFormulaSpectra}{If \code{TRUE} then explained MS/MS spectra (if available) for candidate formulae will be
reported. Specifying \code{formulas} and setting this argument to \code{FALSE} still allows further annotation of
compound MS/MS spectra.}

\item{compoundsOnlyUsedScorings}{If \code{TRUE} then only scorings are plotted that actually have been used to rank
data (see the \code{scoreTypes} argument to \code{\link{generateCompoundsMetFrag}} for more details).}

\item{MSPeakLists}{A \code{\link{MSPeakLists}} object that is \emph{mandatory} when spectra for formulae and/or
compounds will be reported.}

\item{EICGrid}{An integer vector in the form \code{c(columns, rows)} that is used to determine the plotting grid when
reporting EICs in PDF files.}

\item{EICParams}{A named \code{list} with parameters used for extracted ion chromatogram (EIC) creation. See the
\link[=EIXParams]{EIC parameters} documentation for more details.}
}
\description{
Functionality to report data produced by most workflow steps such as features, feature groups, calculated chemical
formulae and tentatively identified compounds. This is the legacy interface, for the updated interface see
\link{reporting}.
}
\details{
These functions are usually called at the very end of the workflow. It is used to report various data on features and
feature groups. In addition, these functions may be used for reporting formulae and/or compounds that were generated
for the specified feature groups. Data can be reported in tabular form (\emph{i.e.} \file{.csv} files) by
\code{reportCSV} or graphically by \code{reportPDF} and \code{reportHTML}. The latter functions will plot for
instance chromatograms and annotated mass spectra, which are useful to get a graphical overview of results.

All functions have a wide variety of arguments that influence the reporting process. Nevertheless, most parameters
are optional and only required to be given for fine tuning. In addition, only those objects (\emph{e.g.} formulae,
compounds, clustering) that are desired to be reported need to be specified.

\code{reportCSV} generates tabular data (\emph{i.e.} \file{.csv}
  files) for given data to be reported. This may also be useful to allow
  import by other tools for post processing.

\code{reportPDF} will report graphical data (\emph{e.g.} chromatograms and mass spectra) within PDF files.
  Compared to \code{reportHTML} this function may be faster and yield smaller report files, however, its
  functionality is a bit more basic and generated data is more 'scattered' around.
}
\note{
Any formulae and compounds for feature groups which are not present within \code{fGroups} (\emph{i.e.} because
  it has been subset afterwards) will not be reported.

  The \code{topMost}, \code{topMostByReplicate} and \code{onlyPresent} \link[=EIXParams]{EIC parameters} may be ignored,
  \emph{e.g.}, when generating overview plots.
}
\section{Use of raw HRMS data}{
 The \link[=msdata]{raw data interface} of \pkg{patRoon} is used by the report functions to
  process HRMS (or IMS-HRMS) data. Please see \link[=msdata]{its documentation} for more information on the supported
  formats and available configuration options.
}

\seealso{
reporting
}
