% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{saveCacheData}
\alias{saveCacheData}
\title{saveCacheData}
\usage{
saveCacheData(category, data, hash, dbArg = NULL)
}
\arguments{
\item{category}{The category of the object to be cached.}

\item{data}{The object to be cached.}

\item{hash}{The hash string of the object to be cached.}

\item{dbArg}{Alternative connection to database. Default is `NULL` and uses
the cache options as defined by "patRoon.cache.fileName".}
}
\description{
Caches objects.
}
