% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\docType{class}
\name{silhouetteInfo-class}
\alias{silhouetteInfo-class}
\alias{silhouetteInfo}
\title{Silhouette information}
\description{
Objects from this class hold silhouette information from a
\code{\link{clusterInfo}} object.
}
\details{
Objects from this class are returned by the \code{\link{getSilhouetteInfo}}
method.
}
\section{Slots}{

\describe{
\item{\code{ranges}}{Numeric vector of \code{k} values that were used to cut the
cluster and generate silhouette information from it.}

\item{\code{maxk}}{\code{k} value (from \code{ranges} parameter) that generated the
largest average silhouette width.}

\item{\code{maxmw}}{Largest average silhouette width from all tested \code{ranges}.}

\item{\code{silInfo}}{A \code{list} containing for each test iteration the fields
\code{k} (the tested \code{k} value), \code{meanw} (mean of all silhouette
widths) and \code{cmeans} (mean silhouette width for each cluster).}
}}

