% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{SIRIUSLogin}
\alias{SIRIUSLogin}
\title{Log in to SIRIUS}
\usage{
SIRIUSLogin(login = "check", alwaysLogin = FALSE, SIRIUSAPI = NULL)
}
\arguments{
\item{login, alwaysLogin}{Specifies if and how account logging of SIRIUS should be handled:

  \code{login=FALSE}: no automatic login is performed and the active login status is not checked.

  \code{login="check"}: aborts if no active login is present.

  \code{login="interactive"}: interactively ask for login (using \CRANpkg{getPass}).

  \code{login=c(username="...", password="...")}: perform the login with the given details. For security reasons,
  please do not enter the details directly, but use e.g. environment variables or store/retrieve them with the
  \CRANpkg{keyring} package.
  
  if \code{alwaysLogin=TRUE} then a login is always performed, otherwise only if SIRIUS reports no active login.

  See the \href{https://v6.docs.sirius-ms.io/account-and-license/}{SIRIUS website} and \pkg{patRoon} handbook for
  more information.
  
  \strong{NOTE}: By loggin in you will accept the terms of the Service and Privacy Policy of the SIRIUS Webservice.}

\item{SIRIUSAPI}{An \code{rsirius_api} object for connecting to the \command{SIRIUS} API. If \code{NULL}, a new
connection will be started automatically.}
}
\description{
Performs an account login to \command{SIRIUS}. This function starts the \command{SIRIUS} API (if needed) and performs
the login. It can be used to explicitly log in before running \code{\link{generateFormulasSIRIUS}} or
\code{\link{generateCompoundsSIRIUS}}.
}
