% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/features.R, R/feature_groups.R,
%   R/feature_groups-screening.R, R/feature_groups-tasq.R, R/features-tasq.R
\docType{methods}
\name{target-screening}
\alias{target-screening}
\alias{screenTargets}
\alias{screenTargets,features-method}
\alias{screenTargets,featureGroups-method}
\alias{featureGroupsScreening-class}
\alias{featureGroupsScreening}
\alias{groupFeaturesScreening,featureGroups-method}
\alias{groupFeaturesScreening}
\alias{featureGroupsBrukerTASQ-class}
\alias{featureGroupsBrukerTASQ}
\alias{importFeatureGroupsBrukerTASQ}
\alias{featuresBrukerTASQ-class}
\alias{featuresBrukerTASQ}
\title{Target and suspect screening}
\usage{
screenTargets(obj, targets, rtWindow = 12, mzWindow = 0.005)

\S4method{screenTargets}{features}(obj, targets, rtWindow = 12,
  mzWindow = 0.005)

\S4method{screenTargets}{featureGroups}(obj, targets, rtWindow = 12,
  mzWindow = 0.005)

\S4method{groupFeaturesScreening}{featureGroups}(fGroups, scr)

importFeatureGroupsBrukerTASQ(path, analysisInfo)
}
\arguments{
\item{obj}{The object that should be screened (\emph{i.e.}
\code{\link{features}} or \code{\link{featureGroups}}).}

\item{targets}{A \code{data.frame} consisting of mandatory columns
\code{"name"} (the targeted analyte name) and \code{"mz"} (the \emph{m/z}
value of the target). In addition, a column \code{"rt"} can be used to
specify the retention time (if unspecified no retention times are checked).}

\item{rtWindow, mzWindow}{The retention time window (in seconds) and
\emph{m/z} window that will be used for matching a target.}

\item{fGroups}{The \code{\link{featureGroups}} object that should be
transformed (and was used to obtain the screening results).}

\item{scr}{The screening results table returned by \code{screenTargets}.}

\item{path}{The file path to an Excel export of the Global results table from
TASQ, converted to \file{.csv} format.}

\item{analysisInfo}{A table with \link[=analysis-information]{analysis
information}.}
}
\value{
\code{screenTargets} will return a table (a \code{\link{data.table}})
  with detected targets and details such as retention and \emph{m/z}
  values. If a target is matched on multiple features/feature groups then
  each hit is reported as a separate row. In case concentration values were
  specified in the \link[=analysis-information]{analysis info} then the
  \code{featureGroups} method will also report correlation coefficients
  (\code{'RSQ'} column) calculated from the linear regression between
  intensity values and specified concentrations.

\code{groupFeaturesScreening} returns a modified \code{featureGroups}
  object in which those feature groups that were not matched by any targets
  are removed and others renamed by the respective target name. In case of
  duplicate target results, feature group names are made unique with
  \code{\link{make.unique}}.

\code{importFeatureGroupsBrukerTASQ} returns a new
  \code{featureGroups} object containing converted screening results from
  Bruker TASQ.
}
\description{
Utilities to screen for targets with known or suspected identity.
}
\details{
Besides 'full non-target analysis', where compounds may be identitified with
little to no prior knowledge, a common strategy is to screen for compounds
with known or suspected identity. This may be a generally favourable approach
if possible, as it can significantly reduce the load on data interpretation.

\code{screenTargets} will screen a set of targets (provided as a
  \code{data.frame}) within an \code{\link{features}} or
  \code{\link{featureGroups}} object.

\code{groupFeaturesScreening} uses results from \code{screenTargets}
  to transform an existing \code{\link{featureGroups}} object by (1) renaming
  any matched feature groups by the respective name of the target and (2)
  filtering out any feature groups that were not matched. A common workflow
  is to first obtain and group features (with \emph{e.g.}
  \code{\link{findFeatures}} and \code{\link{groupFeatures}}), screen them
  with \code{screenTargets}, convert the \code{featureGroups} object that was
  used for screening with this method function and continue any further
  workflow steps such as compound identification as with 'regular'
  \code{featureGroups}.

\code{importFeatureGroupsBrukerTASQ} will convert screening results
  from Bruker TASQ to a \code{\link{featureGroups}} object. Groups are made
  based on target analyte names and individual results per analysis are used
  to generate the features. The input for this function is obtained by
  generating an Excel export of the 'global' results and subsequently
  converting the file to \file{.csv} format. Similar to
  \code{groupFeaturesScreening}, this method will return an object that is
  suitable for any further workflow processes.
}
\note{
Both \code{groupFeaturesScreening} and
  \code{importFeatureGroupsBrukerTASQ} use names from targets as feature
  group names, therefore, it is important that these are file-compatible
  names when \link[=reporting]{reporting data}.

Please note that \code{groupFeaturesScreening} method can only
  transform the \code{featureGroups} object that was used to obtain the given
  screening results.

\code{importFeatureGroupsBrukerTASQ} will use dummy values for
  retention time and \emph{m/z} values for conversion to features, since this
  information is not available. Hence, when plotting, for instance, extracted
  ion chromatograms (with \code{\link{plotEIC}}) the integrated
  chromatographic peak range shown is incorrect.
}
