% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{TPLogicTransformations}
\alias{TPLogicTransformations}
\title{Obtain default rules for metabolic logic}
\usage{
TPLogicTransformations()
}
\value{
A \code{data.frame} with columns describing each transformation rule.
}
\description{
This function returns a \code{data.frame} with the default rules for metabolic logic, which can be used by
\code{\link{generateTPsLogic}} and \code{\link{genFormulaTPLibrary}}.
}
\section{Source}{
 The table is based on the work done by Schollee \emph{et al.} (see references).
}

\references{
\insertRef{Scholle2015}{patRoon}
}
