% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP.R
\docType{class}
\name{TPPredictions-class}
\alias{TPPredictions-class}
\alias{TPPredictions}
\alias{suspects,TPPredictions-method}
\alias{suspects}
\alias{predictions,TPPredictions-method}
\alias{predictions}
\alias{names,TPPredictions-method}
\alias{show,TPPredictions-method}
\alias{[,TPPredictions,ANY,missing,missing-method}
\alias{[[,TPPredictions,ANY,missing-method}
\alias{$,TPPredictions-method}
\alias{as.data.table,TPPredictions-method}
\title{Base transformation product (TP) predictions class}
\usage{
\S4method{suspects}{TPPredictions}(pred)

\S4method{predictions}{TPPredictions}(pred)

\S4method{names}{TPPredictions}(x)

\S4method{show}{TPPredictions}(object)

\S4method{[}{TPPredictions,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{TPPredictions,ANY,missing}(x, i, j)

\S4method{$}{TPPredictions}(x, name)

\S4method{as.data.table}{TPPredictions}(x)
}
\arguments{
\item{i}{A numeric or character value which is used to select suspects by
their index or name, respectively (for the order/names see \code{names()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all suspects are selected.\cr\cr For \code{[[}: should be a scalar value.}

\item{drop, j}{ignored.}

\item{name}{The suspect name (partially matched).}

\item{obj, x, object}{\code{TPPredictions} object to be accessed}
}
\description{
Holds information for all predicted TPs for a set of suspects.
}
\details{
This class holds all generated data for predicted transformation products for
a set of suspects. The class is \code{virtual} and derived objects are
created by \link[=TP-prediction]{TP predictors}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{suspects}: Accessor method for the \code{suspects} slot of a
\code{TPPredictions} class. This is a \code{data.table} with all suspects
used for predictions.

\item \code{predictions}: Accessor method for the \code{predictions} slot of
a \code{TPPredictions} class. Each TP result is stored as a
\code{\link{data.table}}.

\item \code{names}: Obtain the names of all suspects with predicted
TPs.

\item \code{show}: Show summary information for this object.

\item \code{[}: Subset on suspects.

\item \code{[[}: Extracts a table with TPs for a suspect.

\item \code{$}: Extracts a table with TPs for a suspect.

\item \code{as.data.table}: Returns all prediction data in a table.
}}

\section{Slots}{

\describe{
\item{\code{suspects}}{Table with all suspects with predictions. Use the
\code{suspects} method for access.}

\item{\code{predictions}}{List of predicted TPs for each suspect. Use the
\code{predictions} method for access.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{workflowStep}}}   \itemize{     \item{\strong{\code{\link{TPPredictions}}}}     \itemize{       \item{\code{\link{TPPredictionsBT}}}       \item{\code{\link{TPPredictionsComponents}}}       \item{\code{\link{TPPredictionsLogic}}}     }   } }
}

\seealso{
\code{\link{TP-prediction}}
}
