% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/TP.R, R/TP-logic.R
\docType{class}
\name{parents}
\alias{parents}
\alias{products}
\alias{transformationProducts-class}
\alias{transformationProducts}
\alias{parents,transformationProducts-method}
\alias{products,transformationProducts-method}
\alias{length,transformationProducts-method}
\alias{names,transformationProducts-method}
\alias{show,transformationProducts-method}
\alias{[,transformationProducts,ANY,missing,missing-method}
\alias{[[,transformationProducts,ANY,missing-method}
\alias{$,transformationProducts-method}
\alias{as.data.table,transformationProducts-method}
\alias{convertToSuspects,transformationProducts-method}
\alias{delete,transformationProducts-method}
\alias{filter,transformationProducts-method}
\alias{transformationProductsLogic-class}
\alias{transformationProductsLogic}
\title{Base transformation products (TP) class}
\usage{
parents(TPs)

products(TPs)

\S4method{parents}{transformationProducts}(TPs)

\S4method{products}{transformationProducts}(TPs)

\S4method{length}{transformationProducts}(x)

\S4method{names}{transformationProducts}(x)

\S4method{show}{transformationProducts}(object)

\S4method{[}{transformationProducts,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{transformationProducts,ANY,missing}(x, i, j)

\S4method{$}{transformationProducts}(x, name)

\S4method{as.data.table}{transformationProducts}(x)

\S4method{convertToSuspects}{transformationProducts}(obj, includeParents = FALSE)

\S4method{delete}{transformationProducts}(obj, i = NULL, j = NULL, ...)

\S4method{filter}{transformationProducts}(obj, properties = NULL, verbose = TRUE, negate = FALSE)
}
\arguments{
\item{TPs, x, obj, object}{\code{transformationProducts} object to be accessed}

\item{i, j}{For \code{[}/\code{[[}: A numeric or character value which is used to select parents by
their index or name, respectively (for the order/names see \code{names()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all parents are selected.\cr\cr For \code{[[}: should be a scalar value.\cr\cr For \code{delete}: The data to remove from. \code{i} are the
parents as numeric index, logical or character, \code{j} the transformation products as numeric index (row) or name of the TP. If either is
\code{NULL} then data for all is removed. \code{j} may also be a function: it will be called for each parent,
with the TP info table (a \code{data.table}), the parent name and any other arguments passed as \code{\dots} to \code{delete}. The return value of this
function specifies the TP indices (rows) (specified as an \code{integer} or \code{logical} vector) or names to be removed.}

\item{\dots}{For \code{delete}: passed to the function specified as \code{j}. Otherwise ignored.}

\item{drop}{ignored.}

\item{name}{The parent name (partially matched).}

\item{includeParents}{If \code{TRUE} then parents are also included in the returned suspect list.}

\item{properties}{A named \code{list} with properties to be filtered. Each item in the \code{list} should be named
with the name of the property, and should be a vector with allowed values. To obtain the possible properties, run
\emph{e.g.} \code{names(TPs)[[1]]}. Example: \code{properties=list(likelihood=c("LIKELY","PROBABLE"))}. Set to \code{NULL} to ignore.}

\item{verbose}{If set to \code{FALSE} then no text output is shown.}

\item{negate}{If \code{TRUE} then filters are performed in opposite manner.}
}
\value{
\code{delete} returns the object for which the specified data was removed.

\code{filter} returns a filtered \code{transformationProducts} object.
}
\description{
Holds information for all TPs for a set of parents.
}
\details{
This class holds all generated data for transformation products for a set of parents. The class is \code{virtual} and
derived objects are created by \link[=generateTPs]{TP generators}.

The TP data in objects from this class include a \code{retDir} column. These are \code{numeric} values that hint what
the the chromatographic retention order of a TP might be compared to its parent: a value of \samp{-1} means it will
elute earlier, \samp{1} it will elute later and \samp{0} that there is no significant difference or the direction is
unknown. These values are based on a typical reversed phase separation. When structural information is available
(\emph{e.g.} when \code{\link{generateTPsBioTransformer}} or \code{\link{generateTPsLibrary}} was used to generate
the data), the \code{retDir} values are based on calculated \code{log P} values of the parent and its TPs.
}
\section{Methods (by generic)}{
\itemize{
\item \code{parents(transformationProducts)}: Accessor method for the \code{parents} slot of a
\code{transformationProducts} class.

\item \code{products(transformationProducts)}: Accessor method for the \code{products} slot.

\item \code{length(transformationProducts)}: Obtain total number of transformation products.

\item \code{names(transformationProducts)}: Obtain the names of all parents in this object.

\item \code{show(transformationProducts)}: Show summary information for this object.

\item \code{x[i}: Subset on parents.

\item \code{x[[i}: Extracts a table with TPs for a parent.

\item \code{$}: Extracts a table with TPs for a parent.

\item \code{as.data.table(transformationProducts)}: Returns all TP data in a table.

\item \code{convertToSuspects(transformationProducts)}: Converts this object to a suspect list that can be used as input for
\code{\link{screenSuspects}}.

\item \code{delete(transformationProducts)}: Completely deletes specified transformation product data.

\item \code{filter(transformationProducts)}: Performs rule-based filtering. Useful to simplify and clean-up the data.

}}
\section{Slots}{

\describe{
\item{\code{parents}}{A \code{\link{data.table}} with metadata for all parents that have TPs in this object. Use the
\code{parents} method for access.}

\item{\code{products}}{A \code{list} with \code{\link{data.table}} entries with TP information for each parent. Use the
\code{products} method for access.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{workflowStep}}}   \itemize{     \item{\strong{\code{\link{transformationProducts}}}}     \itemize{       \item{\code{\link{transformationProductsStructure}}}       \itemize{         \item{\code{\link{transformationProductsStructureConsensus}}}         \item{\code{\link{transformationProductsCTS}}}         \item{\code{\link{transformationProductsBT}}}         \item{\code{\link{transformationProductsLibrary}}}       }       \item{\code{\link{transformationProductsFormula}}}       \itemize{         \item{\code{\link{transformationProductsLibraryFormula}}}       }       \item{\code{\link{transformationProductsLogic}}}     }   } }
}

\seealso{
The derived \code{\link{transformationProductsStructure}} class for more methods and
  \code{\link{generateTPs}}
}
