% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-ann_comp.R
\docType{class}
\name{transformationProductsAnnComp-class}
\alias{transformationProductsAnnComp-class}
\alias{transformationProductsAnnComp}
\alias{filter,transformationProductsAnnComp-method}
\title{Transformation products obtained from compound annotations}
\usage{
\S4method{filter}{transformationProductsAnnComp}(
  obj,
  ...,
  minFitFormula = 0,
  minFitCompound = 0,
  minSimSusp = 0,
  minFitCompOrSimSusp = c(0, 0),
  minTPScore = 0,
  topMost = NULL,
  verbose = TRUE,
  negate = FALSE
)
}
\arguments{
\item{obj}{The \code{transformationProductsAnnComp} object that should be filtered.}

\item{\dots}{Further arguments passed to the {\link[=filter,transformationProductsStructure-method]{parent filter
method}}.}

\item{minFitFormula, minFitCompound, minSimSusp, minFitCompOrSimSusp, minTPScore}{Thresholds related to TP scoring. See
\code{\link{generateTPsAnnComp}} for more details.}

\item{topMost}{Only keep this number of top-most TPs (based on \code{TPScore}) for each parent/feature group
combination. Set to \code{NULL} to skip this step.}

\item{verbose}{If set to \code{FALSE} then no text output is shown.}

\item{negate}{If \code{TRUE} then filters are performed in opposite manner.}
}
\description{
Class to store results transformation products (TPs) obtained from compound annotations.
}
\details{
This class is derived from the \code{\link{transformationProductsStructure}} base class, please see its documentation
for more details. Objects from this class are returned by \code{\link{generateTPsAnnComp}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{filter(transformationProductsAnnComp)}: Performs rule-based filtering. Useful to simplify and clean-up the data.

}}
\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{transformationProductsStructure}}}   \itemize{     \item{\strong{\code{\link{transformationProductsAnnComp}}}}   } }
}

\seealso{
The base class \code{\link{transformationProductsStructure}} for more relevant methods and
  \code{\link{generateTPsAnnComp}}
}
