% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-ann_form.R
\docType{class}
\name{transformationProductsAnnForm-class}
\alias{transformationProductsAnnForm-class}
\alias{transformationProductsAnnForm}
\alias{filter,transformationProductsAnnForm-method}
\title{Transformation products obtained from formula annotations}
\usage{
\S4method{filter}{transformationProductsAnnForm}(
  obj,
  ...,
  minFitFormula = 0,
  minTPScore = 0,
  topMost = NULL,
  verbose = TRUE,
  negate = FALSE
)
}
\arguments{
\item{obj}{The \code{transformationProductsAnnForm} object that should be filtered.}

\item{\dots}{Further arguments passed to the {\link[=filter,transformationProducts-method]{parent filter method}}.}

\item{minFitFormula, minTPScore}{Thresholds related to TP scoring. See \code{\link{generateTPsAnnForm}} for more
details.}

\item{topMost}{Only keep this number of top-most TPs (based on \code{TPScore}) for each parent/feature group
combination. Set to \code{NULL} to skip this step.}

\item{verbose}{If set to \code{FALSE} then no text output is shown.}

\item{negate}{If \code{TRUE} then filters are performed in opposite manner.}
}
\description{
Class to store results transformation products (TPs) obtained from formula annotations.
}
\details{
This class is derived from the \code{\link{transformationProductsFormula}} base class, please see its documentation
for more details. Objects from this class are returned by \code{\link{generateTPsAnnForm}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{filter(transformationProductsAnnForm)}: Performs rule-based filtering. Useful to simplify and clean-up the data.

}}
\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{transformationProductsFormula}}}   \itemize{     \item{\strong{\code{\link{transformationProductsAnnForm}}}}   } }
}

\seealso{
The base class \code{\link{transformationProductsFormula}} for more relevant methods and
  \code{\link{generateTPsAnnForm}}
}
