% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-biotransformer.R
\docType{class}
\name{transformationProductsBT-class}
\alias{transformationProductsBT-class}
\alias{transformationProductsBT}
\alias{convertToMFDB,transformationProductsBT-method}
\alias{filter,transformationProductsBT-method}
\title{Class to store transformation products (TPs) predicted by BioTransformer}
\usage{
\S4method{convertToMFDB}{transformationProductsBT}(TPs, out, includeParents = TRUE)

\S4method{filter}{transformationProductsBT}(obj, removeEqualFormulas = FALSE, minSimilarity = NULL, negate = FALSE)
}
\arguments{
\item{out}{The file name of the the output \command{MetFrag} database.}

\item{includeParents}{Set to \code{TRUE} to include the parents in the database.}

\item{obj, TPs}{\code{transformationProductsBTs} object to be accessed}

\item{removeEqualFormulas}{If \code{TRUE} then the TPs that have an equal formula as their parent are removed.}

\item{minSimilarity}{Minimum structure similarity (\samp{0-1}) that a TP should have relative to its parent. For
details on how these similarities are calculated, see the \code{\link{generateTPsBioTransformer}} function. May be
useful under the assumption that parents and TPs who have a high structural similarity, also likely have a high
MS/MS spectral similarity (which can be evaluated after componentization with \code{\link{generateComponentsTPs}}.}

\item{negate}{If \code{TRUE} then filters are performed in opposite manner.}
}
\value{
\code{filter} returns a filtered \code{transformationProductsBT} object.
}
\description{
This class is used to store prediction results that are generated with
\href{http://biotransformer.ca/}{BioTransformer}.
}
\details{
Objects from this class are generate with \code{\link{generateTPsBioTransformer}}. This class is derived from the
\code{\link{transformationProducts}} base class, please see its documentation for more details.
}
\section{Methods (by generic)}{
\itemize{
\item \code{convertToMFDB}: Exports this object as a \file{.csv} file that can be used as a \command{MetFrag} local
database.

\item \code{filter}: Performs rule-based filtering of the \command{BioTransformer} predictions.
Useful to simplify and clean-up the data.
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{transformationProducts}}}   \itemize{     \item{\strong{\code{\link{transformationProductsBT}}}}   } }
}

\references{
\insertRef{DjoumbouFeunang2019}{patRoon} \cr\cr \insertRef{Wicker2015}{patRoon}
}
\seealso{
The base class \code{\link{transformationProducts}} for more relevant methods and
  \code{\link{TP-generation}}
}
