% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-formula.R, R/TP-library_formula.R
\docType{class}
\name{transformationProductsFormula-class}
\alias{transformationProductsFormula-class}
\alias{transformationProductsFormula}
\alias{plotGraph,transformationProductsFormula-method}
\alias{transformationProductsLibraryFormula-class}
\alias{transformationProductsLibraryFormula}
\title{Base transformation products (TP) class with formula information}
\usage{
\S4method{plotGraph}{transformationProductsFormula}(
  obj,
  which,
  components = NULL,
  prune = TRUE,
  onlyCompletePaths = FALSE
)
}
\arguments{
\item{obj, TPs}{\code{transformationProductsFormula} derived object to be accessed}

\item{which}{Either a \code{character} or \code{integer} vector with one or more names/indices of the parents to
plot.}

\item{components}{If specified (\emph{i.e.} not \code{NULL}), a \code{\link{componentsTPs}} object that is used for
matching the graph with screening results. The TPs that were found will be marked. See also the \code{prune} and
\code{onlyCompletePaths} arguments.}

\item{prune}{If \code{TRUE} and \code{components} is set, then pathways without \emph{any} detected TPs are not shown
(pruned). See also the \code{onlyCompletePaths} and \code{components} arguments.}

\item{onlyCompletePaths}{If \code{TRUE} and \code{components} is set, then only pathways are shown for which
\emph{all} TPs were detected. See also the \code{prune} and \code{components} arguments.}
}
\value{
\code{plotGraph} returns the result of \code{\link{visNetwork}}.
}
\description{
Holds information for all TPs for a set of parents, including chemical formulae.
}
\details{
This (virtual) class is derived from the \code{\link{transformationProducts}} base class, please see its
documentation for more details. Objects from this class are returned by \link[=generateTPs]{TP generators}. More
specifically, algorithms that works with chemical formulae (\emph{e.g.} \code{library_formula}), uses this class to
store their results. The methods defined for this class extend the functionality for the base
\code{\link{transformationProducts}} class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plotGraph(transformationProductsFormula)}: Plots an interactive hierarchy graph of the transformation products. The
resulting graph can be browsed interactively and allows exploration of the different TP formation pathways.
Furthermore, results from \link[=generateComponentsTPs]{TP componentization} can be used to match the hierarchy
with screening results. The graph is rendered with \pkg{\link{visNetwork}}.

}}
\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{transformationProducts}}}   \itemize{     \item{\strong{\code{\link{transformationProductsFormula}}}}     \itemize{       \item{\code{\link{transformationProductsLibraryFormula}}}     }   } }
}

\seealso{
The base class \code{\link{transformationProducts}} for more relevant methods and \code{\link{generateTPs}}
}
