% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-library.R
\docType{class}
\name{transformationProductsLibrary-class}
\alias{transformationProductsLibrary-class}
\alias{transformationProductsLibrary}
\alias{convertToMFDB,transformationProductsLibrary-method}
\title{Class to store transformation products (TPs) obtained from a library}
\usage{
\S4method{convertToMFDB}{transformationProductsLibrary}(TPs, out, includeParents)
}
\arguments{
\item{out}{The file name of the the output \command{MetFrag} database.}

\item{includeParents}{Set to \code{TRUE} to include the parents in the database.}

\item{obj, TPs}{\code{transformationProductsLibrary} object to be accessed}
}
\description{
This class is used to store prediction results that are available in a TP library.
}
\details{
Objects from this class are generate with \code{\link{generateTPsLibrary}}. This class is derived from the
\code{\link{transformationProducts}} base class, please see its documentation for more details.
}
\section{Methods (by generic)}{
\itemize{
\item \code{convertToMFDB}: Exports this object as a \file{.csv} file that can be used as a \command{MetFrag} local
database.
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{transformationProducts}}}   \itemize{     \item{\strong{\code{\link{transformationProductsLibrary}}}}   } }
}

\seealso{
The base class \code{\link{transformationProducts}} for more relevant methods and
  \code{\link{TP-generation}}
}
