% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verifyDependencies}
\alias{verifyDependencies}
\title{Verifies if all dependencies are installed properly and instructs the user if
this is not the case.}
\usage{
verifyDependencies()
}
\description{
Verifies if all dependencies are installed properly and instructs the user if
this is not the case.
}
