% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{withOpt}
\alias{withOpt}
\title{Temporarily changes package options}
\usage{
withOpt(code, ..., prefix = "patRoon.")
}
\arguments{
\item{code}{The code to be executed.}

\item{\dots}{Named arguments with options to change.}

\item{prefix}{A \code{character} that will be used to prefix given option
names.}
}
\description{
This function is inspired by
\code{\link[withr:with_options]{withr::with_options}}: it can be used to
execute some code where package options are temporarily changed. This
function uses a shortened syntax, especially when changing options for
\code{patRoon}.
}
\examples{
\dontrun{
# Set max parallel processes to five while performing formula calculations
withOpt(MP.maxProcs = 5, {
    formulas <- generateFormulas(fGroups, "genform", ...)
})
}

}
