% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/utils.R
\name{analysis-information}
\alias{analysis-information}
\alias{generateAnalysisInfo}
\alias{generateAnalysisInfoFromEnviMass}
\title{Analysis information}
\usage{
generateAnalysisInfo(paths, groups = "", blanks = "",
  formats = MSFileFormats())

generateAnalysisInfoFromEnviMass(path)
}
\arguments{
\item{paths}{A character vector containing one or more file paths that should
be used for finding the analyses.}

\item{groups, blanks}{An (optional) character vector containing replicate groups
and references, respectively (will be recycled). If \code{groups} is an
empty character string (\code{""}) the analysis name will be set as
replicate group.}

\item{formats}{A character vector of analyses file types. Valid values are:
\code{Bruker}, \code{mzXML} and \code{mzML}.}

\item{path}{The path of the enviMass project.}
}
\description{
Required information for analyses that should be processed and utilities to
automatically generate this information.
}
\details{
Several properties need to be known about analyses that should be
  processed during various workflow steps such as
  \link[=feature-finding]{finding features}, averaging intensities of feature
  groups and blank subtraction. This information should be made available
  with an 'analysis info' object, which is a \code{data.frame} containing the
  following columns:

\itemize{
   \item \code{path} the full path to the analysis.
   \item \code{analysis} the filename \strong{without} extension. Must be \strong{unique},
   even if the \code{path} is different.
   \item \code{group} name of \emph{replicate group}. A replicate group is used
   to group analyses together that are replicates of each other. Thus, the
   \code{group} column for all  analyses considered to be belonging to the same
   replicate group should have an equal (but unique) value. Used for \emph{e.g.}
   avaraging and \code{\link[=filter,featureGroups-method]{filter}}.
   \item \code{blank}: all analyses within this replicate group are used by
   \code{\link[=filter,featureGroups-method]{filter}} for blank subtraction.
   Multiple entries can be entered by separation with a comma.
   \item \code{conc} a numeric value specifying the 'concentration' of the analysis.
   This can be actually any kind of quantitative value such as exposure time,
   dilution factor or anything else which may be used to form a linear relationship.
}

Most functionality requires the data files to be in either the
  \file{.mzXML} or \file{.mzML} format. Functionality that utilizes Bruker
  DataAnalysis (\emph{e.g.} \code{\link{findFeaturesBruker}}) may only work
  with data files in the proprietary \file{.d} format. Therefore, when tools
  with varying requirements are mixed (a common scenario), the data files
  also need to be present in the required formats. To deal with this
  situation the data files with varying formats should all be placed in the
  same path that was used to specify the location of the analysis.

  Whether analysis data files are present in multiple formats or not, each
  analysis should only be entered \emph{once}. The \code{analysis} column is
  used as a basename to automatically find back the data file with the
  required format, hence, analysis names should be specified as the file name
  without its file extension.

  The \code{group} column is \emph{mandatory} and needs to be filled in for
  each analysis. The \code{blank} column should also be present, however, these
  may contain empty character strings (\code{""}) for analyses where no blank
  subtraction should occur. The \code{conc} column is only required when
  obtaining regression information is desired with the
  \code{\link[=as.data.table,featureGroups-method]{as.data.table}} method.

\code{generateAnalysisInfo} is an utility function that automatically
  generates an analysis information object. It will collect all datafiles
  from given file paths and convert the filenames into valid analysis names
  (\emph{i.e.} without extensions such as \file{.d} and \file{.mzML}).
  Duplicate analyses, which may appear when datafiles with different file
  extension (\file{.d}, \file{.mzXML} and/or \file{.mzML}) are present, will
  be automatically removed.

\code{generateAnalysisInfoFromEnviMass} loads analysis information
  from an \pkg{enviMass} project. Note: this funtionality has only been
  tested with older versions of \pkg{enviMass}.
}
