% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisInfo.R
\name{analysisinfo-dataframe}
\alias{analysisinfo-dataframe}
\alias{getTICs,data.frame-method}
\alias{getBPCs,data.frame-method}
\alias{plotTICs,data.frame-method}
\alias{plotBPCs,data.frame-method}
\title{AnalysisInfo data.frame methods}
\usage{
\S4method{getTICs}{data.frame}(obj, retentionRange = NULL, MSLevel = 1)

\S4method{getBPCs}{data.frame}(obj, retentionRange = NULL, MSLevel = 1)

\S4method{plotTICs}{data.frame}(
  obj,
  retentionRange = NULL,
  MSLevel = 1,
  retMin = FALSE,
  title = NULL,
  colourBy = c("none", "analyses", "rGroups"),
  showLegend = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)

\S4method{plotBPCs}{data.frame}(
  obj,
  retentionRange = NULL,
  MSLevel = 1,
  retMin = FALSE,
  title = NULL,
  colourBy = c("none", "analyses", "rGroups"),
  showLegend = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{obj}{An \code{analysisInfo} data.frame object as obtained by \link[patRoon]{generateAnalysisInfo} function.}

\item{retentionRange}{Range of retention time (in seconds), m/z, respectively. Should be a numeric vector with 
length of two containing the min/max values. The maximum can be Inf to specify no maximum range. Set to NULL to skip 
this step.}

\item{MSLevel}{Integer vector with the ms levels (i.e., 1 for MS1 and 2 for MS2) 
to obtain traces.}

\item{retMin}{Plot retention time in minutes (instead of seconds).}

\item{title}{Character string used for title of the plot. If \code{NULL} a title will be automatically generated.}

\item{colourBy}{Sets the automatic colour selection: "none" for a single 
colour or "analyses"/"rGroups" for a distinct colour per analysis or analysis replicate group.}

\item{showLegend}{Plot a legend if TRUE.}

\item{xlim, ylim}{Sets the plot size limits used by
\code{\link[graphics]{plot}}. Set to \code{NULL} for automatic plot sizing.}
}
\description{
Various parsing and plotting functions for the analysisInfo data.frame.
}
\section{Functions}{
\itemize{
\item \code{getTICs(data.frame)}: Obtain the total ion chromatogram/s (TICs) of the analyses.

\item \code{getBPCs(data.frame)}: Obtain the base peak chromatogram/s (BPCs) of the analyses.

\item \code{plotTICs(data.frame)}: Plots the TICs of the analyses.

\item \code{plotBPCs(data.frame)}: Plots the BPCs of the analyses.

}}
\author{
Ricardo Cunha, \email{cunha@iuta.de}
}
