% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clearCache}
\alias{clearCache}
\title{Clearing of cached data}
\usage{
clearCache(what = NULL, file = NULL)
}
\arguments{
\item{what}{This argument describes what should be done. When \code{what =
NULL} this function will list which tables are present along with an
indication of their size (database rows). If \code{what = "all"} then the
complete file will be removed. Otherwise, \code{what} should be a character
string (a regular expression) that is used to match the table names that
should be removed.}

\item{file}{The cache file. If \code{NULL} then the value of the
\code{patRoon.cache.fileName} option is used.}
}
\description{
Remove (part of) the cache database used to store (intermediate) processing
results.
}
\details{
This function will either remove one or more tables within the cache
\code{sqlite} database or simply wipe the whole cache file. Removing tables
will \code{VACUUM} the database, which may take some time for large cache
files.
}
