% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-intclust.R
\docType{class}
\name{componentsIntClust-class}
\alias{componentsIntClust-class}
\alias{componentsIntClust}
\alias{plotHeatMap,componentsIntClust-method}
\alias{plotHeatMap}
\alias{plotInt,componentsIntClust-method}
\title{Components based on clustered intensity profiles.}
\usage{
\S4method{plotHeatMap}{componentsIntClust}(
  obj,
  interactive = FALSE,
  col = NULL,
  margins = c(6, 2),
  cexCol = 1,
  ...
)

\S4method{plotInt}{componentsIntClust}(obj, index, pch = 20, type = "b", lty = 3, col = NULL, ...)
}
\arguments{
\item{obj}{A \code{componentsIntClust} object.}

\item{interactive}{If \code{TRUE} an interactive heatmap will be drawn (with
\code{\link{heatmaply}}).}

\item{col}{The colour used for plotting. Set to \code{NULL} for automatic colours.}

\item{margins, cexCol}{Passed to \code{\link{heatmap.2}}}

\item{\dots}{Further options passed to \code{\link{heatmap.2}} / \code{\link{heatmaply}} (\code{plotHeatMap}),
\code{\link[graphics]{plot}} (\code{plotInt}).}

\item{index}{Numeric component/cluster index or component name.}

\item{pch, type, lty}{Passed to \code{\link{lines}}.}
}
\value{
\code{plotHeatMap} returns the same as \code{\link{heatmap.2}} or
  \code{\link{heatmaply}}.
}
\description{
This class is derived from \code{\link{componentsClust}} and is used to store hierarchical clustering information
from intensity profiles of feature groups.
}
\details{
Objects from this class are generated by \code{\link{generateComponentsIntClust}}
}
\section{Methods (by generic)}{
\itemize{
\item \code{plotHeatMap(componentsIntClust)}: draws a heatmap using the
\code{\link{heatmap.2}} or \code{\link{heatmaply}} function.

\item \code{plotInt(componentsIntClust)}: makes a plot for all (normalized) intensity
profiles of the feature groups within a given cluster.

}}
\section{Slots}{

\describe{
\item{\code{clusterm}}{Numeric matrix with normalized feature group intensities that was used for clustering.}
}}

\note{
When the object is altered (\emph{e.g.} by filtering or subsetting it), methods that need the original
  clustered data such as plotting methods do not work anymore and stop with an error.
}
\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{componentsClust}}}   \itemize{     \item{\strong{\code{\link{componentsIntClust}}}}   } }
}

\seealso{
\code{\link{componentsClust}} for other relevant methods and \code{\link{generateComponents}}
}
