% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-intclust.R
\docType{class}
\name{componentsIntClust-class}
\alias{componentsIntClust-class}
\alias{componentsIntClust}
\alias{clusters,componentsIntClust-method}
\alias{cutClusters,componentsIntClust-method}
\alias{clusterProperties,componentsIntClust-method}
\alias{treeCut,componentsIntClust-method}
\alias{treeCutDynamic,componentsIntClust-method}
\alias{plotHeatMap,componentsIntClust-method}
\alias{plotHeatMap}
\alias{plotInt,componentsIntClust-method}
\alias{plot,componentsIntClust,ANY-method}
\alias{plotSilhouettes,componentsIntClust-method}
\title{Components based on clustered intensity profiles.}
\usage{
\S4method{clusters}{componentsIntClust}(obj)

\S4method{cutClusters}{componentsIntClust}(obj)

\S4method{clusterProperties}{componentsIntClust}(obj)

\S4method{treeCut}{componentsIntClust}(obj, k = NULL, h = NULL)

\S4method{treeCutDynamic}{componentsIntClust}(obj, maxTreeHeight,
  deepSplit, minModuleSize)

\S4method{plotHeatMap}{componentsIntClust}(obj, interactive = FALSE,
  col = NULL, margins = c(6, 2), ...)

\S4method{plotInt}{componentsIntClust}(obj, index, pch = 20,
  type = "b", lty = 3, col = NULL, ...)

\S4method{plot}{componentsIntClust,ANY}(x, pal = "Paired",
  numericLabels = TRUE, colourBranches = length(x) < 50,
  showLegend = length(x) < 20, ...)

\S4method{plotSilhouettes}{componentsIntClust}(obj, kSeq, pch = 16,
  type = "b", ...)
}
\arguments{
\item{k, h}{Desired number of clusters or tree height to be used for cutting
the dendrogram, respectively. One or the other must be specified.
Analogous to \code{\link{cutree}}.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}

\item{interactive}{If \code{TRUE} an interactive heatmap will be drawn (with
\code{\link{d3heatmap}}).}

\item{margins}{Passed to \code{\link{heatmap.2}}}

\item{\dots}{Further options passed to \code{\link{heatmap.2}} /
\code{\link{d3heatmap}} (\code{plotHeatMap}), \code{\link{plot.dendrogram}}
(\code{plot}) or \code{\link[graphics]{plot}} (\code{plotInt}).}

\item{index}{Numeric component/cluster index.}

\item{pch, type}{Passed to \code{\link[graphics]{plot}}.}

\item{x, obj}{A \code{componentsIntClust} object.}

\item{pal}{Colour palette to be used from \pkg{\link{RColorBrewer}}.}

\item{numericLabels}{Set to \code{TRUE} to label with numeric indices instead
of (long) feature group names.}

\item{colourBranches}{Whether branches from cut clusters (and their labels)
should be coloured. Might be slow with large numbers of clusters, hence,
the default is only \code{TRUE} when this is not the case.}

\item{showLegend}{If \code{TRUE} and \code{colourBranches} is also
\code{TRUE} then a legend will be shown which outlines cluster numbers and
their colours. By default \code{TRUE} for small amount of clusters to avoid
overflowing the plot.}

\item{kSeq}{An integer vector containing the sequence that should be used for
average silhouette width calculation.}
}
\value{
\code{plotHeatMap} returns the same as \code{\link{heatmap.2}} or
  \code{\link{d3heatmap}}.
}
\description{
This class is derived from \code{\link{components}} and is used to store
hierarchical clustering information from intensity profiles of feature
groups.
}
\details{
Objects from this class are generated by
\code{\link{generateComponentsIntClust}}
}
\section{Methods (by generic)}{
\itemize{
\item \code{clusters}: Accessor method to the \code{clust} slot,
which was generated by \code{\link{hclust}}.

\item \code{cutClusters}: Accessor method to the \code{cutClusters}
slot. Returns a vector with cluster membership for each candidate (format
as \code{\link{cutree}}).

\item \code{clusterProperties}: Returns a list with properties on how the
clustering was performed.

\item \code{treeCut}: Manually (re-)cut the dendrogram.

\item \code{treeCutDynamic}: Automatically (re-)cut the dendrogram using
the \code{\link{cutreeDynamicTree}} function from
\pkg{\link{dynamicTreeCut}}.

\item \code{plotHeatMap}: draws a heatmap using the
\code{\link{heatmap.2}} or \code{\link{d3heatmap}} function.

\item \code{plotInt}: makes a plot for all (normalized) intensity
profiles of the feature groups within a given cluster.

\item \code{plot}: generates a dendrogram from a given cluster
object and optionally highlights resulting branches when the cluster is
cut.

\item \code{plotSilhouettes}: Plots the average silhouette width when the
clusters are cut by a sequence of k numbers. The k value with the highest
value (marked in the plot) may be considered as the optimal number of
clusters.
}}

\section{Slots}{

\describe{
\item{\code{clusterm}}{Numeric matrix with normalized feature group intensities that
was used for clustering.}

\item{\code{distm}}{Distance matrix that was used for clustering (obtained with
\code{\link{daisy}}).}

\item{\code{clust}}{Object returned by \code{\link{hclust}}.}

\item{\code{cutClusters}}{A \code{list} with assigned clusters (same format as what
\code{\link{cutree}} returns).}

\item{\code{gInfo}}{The \code{\link{groupInfo}} of the feature groups object that was
used.}

\item{\code{properties}}{A list containing general properties and parameters used for
clustering.}
}}

\note{
The intensity values for components (used by \code{plotSpec}) are set
  to a dummy value (1) as no single intensity value exists for this kind of
  components.
}
\references{
\insertRef{Scholle2018}{patRoon}
}
\seealso{
\code{\link{components}} and \link{component-generation}
}
