% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-net.R, R/components-net-set.R
\docType{class}
\name{componentsNet-class}
\alias{componentsNet-class}
\alias{componentsNet}
\alias{plotGraph,componentsNet-method}
\alias{componentsNetSet-class}
\alias{componentsNetSet}
\alias{plotGraph,componentsNetSet-method}
\alias{componentsNetUnset-class}
\alias{componentsNetUnset}
\alias{unset,componentsNetSet-method}
\title{Components class for network-based componentization.}
\usage{
\S4method{plotGraph}{componentsNet}(obj, analysis, width = NULL, height = NULL)

\S4method{plotGraph}{componentsNetSet}(obj, analysis, set, ...)

\S4method{unset}{componentsNetSet}(obj, set)
}
\arguments{
\item{obj}{The \code{componentsNet} object to plot.}

\item{analysis}{The name of the analysis to plot.}

\item{width, height}{Passed to \code{\link{visNetwork}}.}

\item{set}{\setsWF The name of the set.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}
}
\value{
\code{plotGraph} returns the result of \code{\link{visNetwork}}.
}
\description{
This class is derived from \code{\link{components}} and is used to store results from network-based componentization.
}
\details{
Objects from this class are generated by \code{\link{generateComponentsNet}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plotGraph(componentsNet)}: Plots an interactive network graph for the feature components of an analysis.

}}
\section{Slots}{

\describe{
\item{\code{featureComponents}}{A \code{list} with feature components for each analysis.}

\item{\code{featureGraphs}}{A \code{list} with feature network graphs (\CRANpkg{igraph}) for each analysis.}

\item{\code{annotationObjects}}{A \code{list} with annotation objects for each component (from \code{\link{nontarget}} or
\code{\link{InterpretMSSpectrum}}).}
}}

\section{Sets workflows}{
 \setsWFClass{componentsNetSet}{componentsNet}

  \setsWFNewMethodsSO{componentsNetUnset}{Only the components in the specified set are kept. Furthermore, the
  component names are restored to non-set specific names (see \code{\link{generateComponents}} for more details).}

  \setsWFChangedMethods{

  \item \code{plotGraph} Currently can only create graph networks from one set (specified by the \code{set}
  argument).

  }

  Note that the \code{componentsNetSet} class does not have \code{featureComponents} and \code{featureGraphs} slots.
  Instead, the \code{\link{setObjects}} method can be used to access this data for a specific set.
}

\references{
\addCitations{igraph}
}
\seealso{
\code{\link{components}}, \code{\link{generateComponents}}, \code{\link{generateComponentsNet}}
}
