% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/compounds-cluster.R
\name{makeHCluster}
\alias{makeHCluster}
\alias{compounds-cluster}
\alias{makeHCluster,compounds-method}
\title{Hierarchical clustering of compounds}
\usage{
makeHCluster(obj, method = "complete", ...)

\S4method{makeHCluster}{compounds}(
  obj,
  method,
  fpType = "extended",
  fpSimMethod = "tanimoto",
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1
)
}
\arguments{
\item{obj}{The \code{\link{compounds}} object to be clustered.}

\item{method}{The clustering method passed to \code{\link{hclust}}.}

\item{\dots}{further arguments specified to methods.}

\item{fpType}{The type of structural fingerprint that should be calculated. See the \code{type} argument of the
\code{\link{get.fingerprint}} function of \CRANpkg{rcdk}.}

\item{fpSimMethod}{The method for calculating similarities (i.e. not dissimilarity!). See the \code{method} argument
of the \code{\link{fp.sim.matrix}} function of the \CRANpkg{fingerprint} package.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}
}
\value{
\code{makeHCluster} returns an \code{\link{compoundsCluster}} object.
}
\description{
Perform hierarchical clustering of structure candidates based on chemical
similarity and obtain overall structural information based on the maximum
common structure (MCS).
}
\details{
Often many possible chemical structure candidates are found for each feature
group when performing \link[=generateCompounds]{compound annotation}.
Therefore, it may be useful to obtain an overview of their general structural
properties. One strategy is to perform hierarchical clustering based on their
chemical (dis)similarity, for instance, using the Tanimoto score. The
resulting clusters can then be characterized by evaluating their
\emph{maximum common substructure} (MCS).

\code{makeHCluster} performs hierarchical clustering of all
  structure candidates for each feature group within a
  \code{\link{compounds}} object. The resulting dendrograms are automatically
  cut using the \code{\link{cutreeDynamicTree}} function from the
  \pkg{\link{dynamicTreeCut}} package. The returned
  \code{\link{compoundsCluster}} object can then be used, for instance, for
  plotting dendrograms and MCS structures and manually re-cutting specific
  clusters.
}
\section{Source}{
 The methodology applied here has been largely derived from
  \file{chemclust.R} from the \pkg{metfRag} package and the package vignette
  of \CRANpkg{rcdk}.
}

\references{
\addCitations{rcdk}{1}
}
\seealso{
compoundsCluster
}
