% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-compounds.R
\name{compoundScorings}
\alias{compoundScorings}
\title{Scorings terms for compound candidates}
\usage{
compoundScorings(
  algorithm = NULL,
  database = NULL,
  includeSuspectLists = TRUE,
  onlyDefault = FALSE,
  includeNoDB = TRUE
)
}
\arguments{
\item{algorithm}{The algorithm: \code{"metfrag"} or \code{"sirius"}. Set to \code{NULL} to return all scorings.}

\item{database}{The database for which results should be returned (\emph{e.g.} \code{"pubchem"}). Set to \code{NULL}
to return all scorings.}

\item{includeSuspectLists, onlyDefault, includeNoDB}{A logical specifying whether scoring terms related to suspect
lists, default scoring terms and non-database specific scoring terms should be included in the output,
respectively.}
}
\value{
A \code{data.frame} with information on which scoring terms are used, what their algorithm specific name is
  and other information such as to which database they apply and short remarks.
}
\description{
Returns an overview of scorings may be applied to rank candidate compounds.
}
\seealso{
generateCompounds
}
