% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds-metfrag.R
\docType{class}
\name{compoundsMF-class}
\alias{compoundsMF-class}
\alias{compoundsMF}
\alias{settings,compoundsMF-method}
\alias{settings}
\title{Compounds list class for MetFrag results.}
\usage{
\S4method{settings}{compoundsMF}(compoundsMF)
}
\arguments{
\item{compoundsMF}{A \code{compoundsMF} object.}
}
\description{
This class is derived from \code{\link{compounds}} and contains additional
specific MetFrag data.
}
\details{
Objects from this class are generated by
\code{\link{generateCompoundsMetFrag}}
}
\section{Methods (by generic)}{
\itemize{
\item \code{settings}: Accessor method for the \code{settings} slot.
}}

\section{Slots}{

\describe{
\item{\code{settings}}{A list with all general configuration settings passed to
MetFrag. Feature specific items (\emph{e.g.} spectra and precursor masses)
are not contained in this list.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{compounds}}}   \itemize{     \item{\strong{\code{\link{compoundsMF}}}}   } }
}

\references{
\insertRef{Ruttkies2016}{patRoon}
}
\seealso{
\code{\link{compounds}} and \link{compound-generation}
}
