% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertMSFiles}
\alias{convertMSFiles}
\alias{MSFileFormats}
\title{MS data conversion}
\usage{
MSFileFormats(algorithm = "pwiz")

convertMSFiles(files = NULL, outPath = NULL, dirs = TRUE,
  anaInfo = NULL, from = MSFileFormats(algorithm), to = "mzML",
  overWrite = FALSE, algorithm = "pwiz", filters = NULL,
  extraOpts = NULL, logPath = file.path("log", "convert"),
  maxProcAmount = getOption("patRoon.maxProcAmount"))
}
\arguments{
\item{algorithm}{Either \code{"pwiz"} (uses \command{msConvert} of
ProteoWizard) or \code{"openms"} (uses \command{FileConverter} of OpenMS).}

\item{files, dirs}{The \code{files} argument should be a \code{character}
vector with input files. If \code{files} contains directories and
\code{dirs=TRUE} then files from these directories are also considered. An
alternative method to specify input files is by the \code{anaInfo}
argument. If the latter is specified \code{files} may be \code{NULL}.}

\item{outPath}{A character vector specifying directories that should be used
for the output. Will be re-cycled if necessary. If \code{NULL}, output
directories will be kept the same as the input directories.}

\item{anaInfo}{An \link[=analysis-information]{analysis info table} used to
retrieve input files. Either this argument or \code{files} (or both) should
be set (\emph{i.e.} not \code{NULL}).}

\item{from}{One or more input formats (see below). These are used to find
analyses when \code{dirs=TRUE}.}

\item{to}{Output format: \code{"mzXML"} or \code{"mzML"}.}

\item{overWrite}{Should existing destination file be overwritten
(\code{TRUE}) or not (\code{FALSE})?}

\item{filters}{When \code{algorithm="pwiz"}: a \code{character} vector
specifying one or more filters. Can be used for peak picking to obtain
centroided data (see examples). The elements of the specified vector are
directly passed to the \code{--filter} option (see
\href{http://proteowizard.sourceforge.net/tools/filters.html}{here})}

\item{extraOpts}{A \code{character} vector specifying any extra commandline
parameters passed to \command{msConvert} or \command{FileConverter}. Set to
\code{NULL} to ignore. For options: see
\href{http://ftp.mi.fu-berlin.de/pub/OpenMS/release-documentation/html/TOPP_FileConverter.html}{FileConverter}
 and
\href{http://proteowizard.sourceforge.net/tools/msconvert.html}{msConvert}.}

\item{logPath}{Destination directory for log files with output from executed
commands. Will be created if non-existant. Set to \code{NULL} to disable
logging.}

\item{maxProcAmount}{Maximum number of processes to run for parallelization.
Usually a number close to the amount of physical cores yields most
efficient results.}
}
\description{
Conversion of MS analysis files between several open and closed data formats.
}
\details{
\code{MSFileFormats} returns a \code{character} with all supported
  input formats (see below).

\code{convertMSFiles} converts the data format of an analysis to
  another. It either uses tools from
  \href{http://proteowizard.sourceforge.net/}{ProteoWizard}
  (\command{msConvert} command) or \href{http://www.openms.de/}{OpenMS}
  (\command{FileConverter} command) to perform the conversion. The supported
  input and output formats include \file{mzXML} and \file{.mzML}.
  Furthermore, when ProteoWizard is used for conversion, major (closed)
  vendor formats are supported for input files.
}
\section{Conversion formats}{
 Input and output formats include \code{mzXML}
  and \code{mzML}. In addition, when ProteoWizard is used, the following
  vendor input formats may be chosen (specified with the \code{from}
  argument): \itemize{

  \item \code{thermo}: Thermo \file{.RAW} files.

  \item \code{bruker}: Bruker \file{.d}, \file{.yep}, \file{.baf} and
  \file{.fid} files.

  \item \code{agilent}: Agilent \file{.d} files.

  \item \code{ab}: AB Sciex \file{.wiff} files.

  \item \code{waters} Waters \file{.RAW} files.

  }

  Note that the actual supported file formats depend on the ProteoWizard
  installation (see
  \href{http://proteowizard.sourceforge.net/formats/index.html}{here}).
}

\examples{
\dontrun{
# Use FileConverter of OpenMS to convert between open mzXML/mzML format
convertMSFiles("standard-1.mzXML", to = "mzML", algorithm = "openms")

# Convert all Thermo .RAW files in the analyses/raw directory to mzML and
# store the files in analyses/mzml. During conversion files are centroided by
# the peakPicking filter and only MS 1 data is kept.
convertMSFiles("analyses/raw", "analyses/mzml", dirs = TRUE, from = "thermo",
               filters = c("peakPicking vendor", "msLevel 1"))
}

}
\references{
\insertRef{Rst2016}{patRoon} \cr\cr
  \insertRef{Chambers2012}{patRoon}
}
