% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{EIXParams}
\alias{EIXParams}
\alias{getDefEICParams}
\alias{getDefEIMParams}
\title{Extracted Ion Chromatogram and Mobilogram parameters}
\usage{
getDefEICParams(...)

getDefEIMParams(..., IMS = getLimIMS())
}
\arguments{
\item{\dots}{optional named arguments that override defaults.}

\item{IMS}{A \code{character} that specifies for which IMS instrument defaults are returned. Should be
\code{"bruker"} or \code{"agilent"}. Defaults to what is specified in \link{limits}.}
}
\description{
Parameters for creation of extracted ion chromatograms and mobilograms.
}
\details{
The following parameters exist to configure the creation of extracted ion chromatograms (EICs) and extracted ion
mobilograms (EIMs):

\itemize{

\item \code{window} A value that is subtracted or added to the minimum and maximum retention time (EICs) or mobility
(EIMs) of the feature. Thus, setting this value to \samp{>0} will 'zoom out' on the x-axis of a chromatogram or
mobilogram. Defaults to \code{defaultLim("retention", "wide")} (EICs) or \code{defaultLim("mobility", "wide")} (EIMs)
(see \link{limits}). Set to \code{Inf} to use the full retention time or mobility range.

\item \code{topMost} Only create EICs/EIMs for this number of top most intense features. If \code{NULL} then
EICs/EIMs are created for all features.

\item \code{topMostByReplicate} If set to \code{TRUE} and \code{topMost} is set: only create EICs/EIMs for the top
most features in each replicate. For instance, when \code{topMost=1} and \code{topMostByReplicate=TRUE}, then only
the most intense feature of each replicate is considered.

\item \code{onlyPresent} If \code{TRUE} then EICs/EIMs are created only for analyses in which a feature was detected,
if \code{onlyPresent=FALSE} then data is generated for \strong{all} analyses. The latter is handy to evaluate if a
peak was 'missed' during peak detection or removed during \emph{e.g.} filtering.

\item \code{mzExpMobWindow} \IMSWF Additional \emph{m/z} tolerance on top of the feature limits. This is for IMS
workflows where features were detected from centroided LC-MS like data, while EICs/EIMs are generated from raw IMS
data. In this case the feature \emph{m/z} limits were derived from centroided data, which typically has smaller
\emph{m/z} deviations across scans compared to IMS data. The \code{mzExpMobWindow} parameter sets an additional
\emph{m/z} tolerance to specifically handle this case. Defaults to \code{defaultLim("mz", "default")} (see
\link{limits}).

\item \code{minIntensityIMS} \IMSWF Raw intensity threshold for IMS data. This is primarily intended to speed up raw
data processing.

}

if \code{onlyPresent=FALSE} then the following parameters are also relevant: \itemize{

\item \code{mzExpWindow},\code{mobExpWindow} To create EICs or EIMs for analyses in which no feature was found, the
\emph{m/z} or mobility value is derived from the min/max values of all features in the feature group. The value of
\code{mzExpWindow} and \code{mobExpWindow} further expands this window to allow a greater tolerance. Defaults to
\code{defaultLim("mz", "very_narrow")} and \code{defaultLim("mobility", "very_narrow")} (see \link{limits}).

\item \code{setsAdductPos},\code{setsAdductNeg} \setsWF In sets workflows the adduct must be known to calculate the
ionized \emph{m/z}. If a feature is completely absent in a particular set then it follows no adduct annotations are
available and the value of \code{setsAdductPos} (positive ionization data) or \code{setsAdductNeg} (negative
ionization data) will be used instead.

}

The following additional parameters exist specifically for EICs (\code{EICParams}): \itemize{

  \item \code{gapFactor} Bruker TIMS data (and maybe others?) seem to omit zero intensity scans, which will lead to
  time gaps between spectra and incorrect EICs. To determine a time gap, the \code{gapFactor} is multiplied with the
  median of time differences between scans. If a gap is detected, then appropriate zero intensity points are added to
  the EIC. Set to \code{0} to disable this.

}

The following additional parameters exist specifically for EIMs (\code{EIMParams}): \itemize{

  \item \code{maxRTWindow} Maximum retention time window (seconds, +/- feature retention time) in which mobilograms
  are collected and averaged. Defaults to \code{defaultLim("retention", "very_narrow")} (see \link{limits}).

  \item \code{smooth} The smoothing method that is applied to the EIM. Can be \code{"none"} for no smoothing,
  \code{"sg"} for Savitzky-Golay (using \CRANpkg{signal})) or \code{"ma"} for centered moving average (same algorithm
  as used by \code{\link{findFeaturesPiek}}).

  \item \code{smLength} The smoothing length. If \code{smooth="sg"} then this is passed as the \code{n} argument to
  the \code{\link[signal:sgolayfilt]{signal::sgolayfilt}} function.

  \item \code{sgOrder} The smoothing order for Savitzky-Golay. Passed as the \code{p} argument to the
  \code{\link[signal:sgolayfilt]{signal::sgolayfilt}} function.

}

These parameters are passed as a named \code{list} as the \code{EICParams} or \code{EIMParams} argument to functions
that work with EIC or EIM data. The \code{getDefEICParams} and \code{getDefEIMParams} functions generate such
parameter list with defaults.
}
