% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-filter.R
\docType{methods}
\name{feature-filtering}
\alias{feature-filtering}
\alias{filter,featureGroups-method}
\alias{replicateGroupSubtract,featureGroups-method}
\alias{replicateGroupSubtract}
\title{Filtering of grouped features}
\usage{
\S4method{filter}{featureGroups}(obj, intensityThreshold = NULL,
  relAbundance = NULL, absAbundance = NULL,
  interRelRGroupAbundance = NULL, interAbsRGroupAbundance = NULL,
  intraRGroupAbundance = NULL, minBlankThreshold = NULL,
  retentionRange = NULL, rGroups = NULL, formConsensus = NULL,
  compounds = NULL, repetitions = 2, negate = FALSE)

\S4method{replicateGroupSubtract}{featureGroups}(fGroups, rGroups,
  threshold = 0)
}
\arguments{
\item{intensityThreshold}{Minimum intensity. Set to 0 or \code{NULL} to skip
this step.}

\item{relAbundance, absAbundance, interRelRGroupAbundance, interAbsRGroupAbundance}{Minimal overall relative/absolute abundance for a feature group to be present
within all analyses/replicate groups.}

\item{intraRGroupAbundance}{Minimum relative abundance (0-1) that a feature
group should be present within a replicate group. Set to \code{NULL} to
skip this step.}

\item{minBlankThreshold}{Feature groups that are also present in reference
analyses (see \link[=analysis-information]{analysis info}) are filtered out
unless their relative intensity is above this threshold. Set to \code{NULL}
to skip this step.}

\item{retentionRange}{Retention time range (in seconds). Should be a numeric
vector with length of two containing the min/max values. If the max value
is set to a value below 0 then no maximum retention is assumed. Set to
\code{NULL} to skip this step.}

\item{rGroups}{A character vector of replicate groups that should be kept
(\code{filter}) or subtracted from (\code{replicateGroupSubtract}).}

\item{formConsensus, compounds}{Only keep feature groups which have results in
the given \code{\link{formulaConsensus}} and \code{\link{compounds}}
object, respectively.}

\item{repetitions}{Sometimes more feature groups may be removed by repeating
filtering steps after another. Usually a value of two is enough to filter
the maximum amount of feature groups.}

\item{negate}{If set to \code{TRUE} then filtering operations are performed
in opposite manner.}

\item{fGroups, obj}{\code{\link{featureGroups}} object to which the filter is
applied.}

\item{threshold}{Minimal relative threshold for a feature group to be
\emph{not} removed. When \samp{0} a feature group is always removed when
present in the given replicate groups.}
}
\value{
A filtered \code{\link{featureGroups}} object. Feature groups that
  are filtered away have their intensity set to zero. In case a feature group
  is not present in any of the analyses anymore it will be removed
  completely.
}
\description{
Basic rule based filtering of feature groups.
}
\details{
\code{filter} performs common rule based filtering of feature groups
  such as blank subtraction, minimum intensity and minimal replicate
  abundance.

\code{replicateGroupSubtract} removes feature groups present in a
  given set of replicate groups (unless intensities are above a given
  threshold).
}
\seealso{
\code{\link{featureGroups-class}}

\code{\link{feature-grouping}}
}
