% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-plot.R
\name{feature-plotting}
\alias{feature-plotting}
\alias{plot,featureGroups,missing-method}
\alias{plotInt,featureGroups-method}
\alias{plotInt,featureGroupsSet-method}
\alias{plotChord,featureGroups-method}
\alias{plotChroms,featureGroups-method}
\alias{plotVenn,featureGroups-method}
\alias{plotVenn,featureGroupsSet-method}
\alias{plotUpSet,featureGroups-method}
\alias{plotVolcano,featureGroups-method}
\alias{plotGraph,featureGroups-method}
\alias{plotGraph,featureGroupsSet-method}
\title{Plotting of grouped features}
\usage{
\S4method{plot}{featureGroups,missing}(
  x,
  colourBy = c("none", "rGroups", "fGroups"),
  onlyUnique = FALSE,
  retMin = FALSE,
  showLegend = TRUE,
  col = NULL,
  pch = NULL,
  ...
)

\S4method{plotInt}{featureGroups}(
  obj,
  average = FALSE,
  normalized = FALSE,
  xnames = TRUE,
  showLegend = FALSE,
  pch = 20,
  type = "b",
  lty = 3,
  col = NULL,
  ...
)

\S4method{plotInt}{featureGroupsSet}(
  obj,
  average = FALSE,
  normalized = FALSE,
  xnames = !sets,
  showLegend = sets,
  pch = 20,
  type = "b",
  lty = 3,
  col = NULL,
  ...,
  sets = FALSE
)

\S4method{plotChord}{featureGroups}(
  obj,
  addSelfLinks = FALSE,
  addRetMzPlots = TRUE,
  average = FALSE,
  outerGroups = NULL,
  addIntraOuterGroupLinks = FALSE,
  ...
)

\S4method{plotChroms}{featureGroups}(
  obj,
  rtWindow = 30,
  mzExpWindow = 0.001,
  retMin = FALSE,
  topMost = NULL,
  topMostByRGroup = FALSE,
  EICs = NULL,
  showPeakArea = FALSE,
  showFGroupRect = TRUE,
  title = NULL,
  colourBy = c("none", "rGroups", "fGroups"),
  showLegend = TRUE,
  onlyPresent = TRUE,
  annotate = c("none", "ret", "mz"),
  showProgress = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)

\S4method{plotVenn}{featureGroups}(obj, which = NULL, ...)

\S4method{plotVenn}{featureGroupsSet}(obj, which = NULL, ..., sets = FALSE)

\S4method{plotUpSet}{featureGroups}(obj, which = NULL, nsets = length(which), nintersects = NA, ...)

\S4method{plotVolcano}{featureGroups}(
  obj,
  FCParams,
  showLegend = TRUE,
  averageFunc = mean,
  normalized = FALSE,
  col = NULL,
  pch = 19,
  ...
)

\S4method{plotGraph}{featureGroups}(obj, onlyPresent = TRUE)

\S4method{plotGraph}{featureGroupsSet}(obj, onlyPresent = TRUE, set)
}
\arguments{
\item{colourBy}{Sets the automatic colour selection: \code{"none"} for a single colour or
\code{"rGroups"}/\code{"fGroups"} for a distinct colour per replicate/feature group.}

\item{onlyUnique}{If \code{TRUE} and \code{colourBy="rGroups"} then only
feature groups that are unique to a replicate group are plotted.}

\item{retMin}{Plot retention time in minutes (instead of seconds).}

\item{showLegend}{Plot a legend if \code{TRUE}.}

\item{col}{Colour(s) used. If \code{col=NULL} then colours are automatically generated.}

\item{pch, type, lty}{Common plotting parameters passed to \emph{e.g.} \code{\link[graphics]{plot}}. For \code{plot}:
if \code{pch=NULL} then values are automatically assigned.}

\item{\dots}{passed to \code{\link[base]{plot}} (\code{plot} and \code{plotChroms}), \code{\link[graphics]{lines}}
(\code{plotInt}), \pkg{\link{VennDiagram}} plotting functions (\code{plotVenn}), \code{\link{chordDiagram}}
(\code{plotChord}) or \code{\link[UpSetR]{upset}} (\code{plotUpSet}).}

\item{obj, x}{\code{featureGroups} object to be used for plotting.}

\item{average}{If \code{TRUE} then data within replicate groups are averaged.

  For \code{as.data.table}: if \code{features=TRUE} other feature properties are also averaged.}

\item{xnames}{Plot analysis (or replicate group if \code{average=TRUE}) names on the x axis.}

\item{sets}{\setsWF For \code{plotInt}: if \code{TRUE} then feature intensities are plot per set (order follows the
  \link[=analysis-information]{analysis information}).

  For \code{plotVenn}: If \code{TRUE} then the \code{which} argument changes its meaning and is used to specify the
  names of the sets to be compared.}

\item{addSelfLinks}{If \code{TRUE} then 'self-links' are added which
represent non-shared data.}

\item{addRetMzPlots}{Set to \code{TRUE} to enable \emph{m/z} \emph{vs}
retention time scatter plots.}

\item{outerGroups}{Character vector of names to be used as outer groups. The
values in the specified vector should be named by analysis names
(\code{average} set to \code{FALSE}) or replicate group names
(\code{average} set to \code{TRUE}), for instance: \code{c(analysis1 =
"group1", analysis2 = "group1", analysis3 = "group2")}. Set to \code{NULL}
to disable outer groups.}

\item{addIntraOuterGroupLinks}{If \code{TRUE} then links will be added within
outer groups.}

\item{rtWindow}{Retention time (in seconds) that will be subtracted/added to respectively the minimum and maximum
retention time of the plotted feature groups. Thus, setting this value to a positive value will 'zoom out' on the
retention time axis.}

\item{mzExpWindow}{In case the \emph{m/z} window to plot an EIC for a particular analysis is not known (\emph{i.e.}
no feature was detected of the feature group to be plot and \code{onlyPresent=FALSE}) then the EIC \emph{m/z} range
is estimated from the range for the complete feature group and expanded by the offset defined by
\code{mzExpWindow}.}

\item{topMost}{Only plot EICs from features within this number of top most intense analyses. If \code{NULL} then all
analyses are used for plotted.}

\item{topMostByRGroup}{If set to \code{TRUE} and \code{topMost} is set: only plot EICs for the top most features in
each replicate group. For instance, when \code{topMost=1} and \code{topMostByRGroup=TRUE}, then EICs will be
plotted for the most intense feature of each replicate group.}

\item{EICs}{Internal parameter for now and should be kept at \code{NULL} (default).}

\item{showPeakArea}{Set to \code{TRUE} to display integrated chromatographic peak ranges by filling (shading) their
areas.}

\item{showFGroupRect}{Set to \code{TRUE} to mark the full retention/intensity range of all features within a feature
group by drawing a rectangle around it.}

\item{title}{Character string used for title of the plot. If \code{NULL} a title will be automatically generated.}

\item{onlyPresent}{Only plot feature groups of internal standards that are still present in the \code{featureGroups}
input object (which may be otherwise be removed by \emph{e.g.} subsetting or
\code{\link[=filter,featureGroups-method]{filter}}).}

\item{annotate}{If set to \code{"ret"} and/or \code{"mz"} then retention and/or \emph{m/z} values will be drawn for
each plotted feature group.}

\item{showProgress}{if set to \code{TRUE} then a text progressbar will be displayed when all EICs are being plot. Set
to \code{"none"} to disable any annotation.}

\item{xlim, ylim}{Sets the plot size limits used by
\code{\link[graphics]{plot}}. Set to \code{NULL} for automatic plot sizing.}

\item{which}{A character vector with replicate groups used for comparison. Set to \code{NULL} to ignore.

  For \code{plotVenn}: alternatively a named \code{list} containing elements of \code{character} vectors with
  replicate groups to compare. For instance, \code{which=list(infl = c("influent-A", "influent-B"), effl =
  c("effluent-A", "effluent-B"))}, will compare the features in replicate groups \samp{"influent-A/B"} against those
  in \samp{"effluent-A/B"}. The names of the list are used for labelling in the plot, and will be made automatically
  if not specified.}

\item{nsets, nintersects}{See \code{\link[UpSetR]{upset}}.}

\item{FCParams}{A parameter list to calculate Fold change data. See \code{getFCParams} for more details.}

\item{averageFunc, normalized}{Used for intensity data treatment, see the documentation for the
\code{\link[=as.data.table,featureGroups-method]{as.data.table method}}.}

\item{set}{\setsWF The set for which data must be plotted.}
}
\value{
\code{plotVenn} (invisibly) returns a list with the following fields: \itemize{
\item \code{gList} the \code{gList} object that was returned by
  the utilized \pkg{\link{VennDiagram}} plotting function.
\item \code{areas} The total area for each plotted group.
\item \code{intersectionCounts} The number of intersections between groups.
}

The order for the \code{areas} and \code{intersectionCounts} fields is the same as the parameter order
from the used plotting function (see \emph{e.g.} \code{\link{draw.pairwise.venn}} and
\code{\link{draw.triple.venn}}).

\code{plotGraph} returns the result of \code{\link{visNetwork}}.
}
\description{
Various plotting functions for feature group data.
}
\details{
\code{plot} Generates an \emph{m/z} \emph{vs} retention time
  plot for all featue groups. Optionally highlights unique/overlapping
  presence amongst replicate groups.

\code{plotInt} Generates a line plot for the (averaged) intensity
  of feature groups within all analyses

\code{plotChord} Generates a chord diagram which can be used to
  visualize shared presence of feature groups between analyses or replicate
  groups. In addition, analyses/replicates sharing similar properties
  (\emph{e.g.} location, age, type) may be grouped to enhance visualization
  between these 'outer groups'.

\code{plotChroms} Plots extracted ion chromatograms (EICs) of feature groups.

\code{plotVenn} plots a Venn diagram (using \pkg{\link{VennDiagram}}) outlining unique and shared feature
  groups between up to five replicate groups.

\code{plotUpSet} plots an UpSet diagram (using the \code{\link[UpSetR]{upset}} function) outlining unique
  and shared feature groups between given replicate groups.

\code{plotVolcano} Plots Fold change data in a 'Volcano plot'.

\code{plotGraph} generates an interactive network plot which is used to explore internal standard (IS)
  assignments to each feature group. This requires the availability of IS assignments, see the documentation for
  \code{\link{normInts}} for details. The graph is rendered with \pkg{\link{visNetwork}}.
}
\section{Sets workflows}{
 \setsWFChangedMethods{

  \item \code{plotVenn} and \code{plotInt} allow to handle data per set. See the \code{sets} argument description.
  
  \item \code{plotGraph} only plots data per set, and requires the \code{set} argument to be set.

  }
}

\references{
\addCitations{circlize}{1}

\insertRef{Conway2017}{patRoon} \cr\cr \insertRef{Lex2014}{patRoon}
}
\seealso{
\code{\link{featureGroups-class}}, \code{\link{groupFeatures}}
}
