% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/feature_groups.R,
%   R/feature_groups-set.R, R/feature_groups-bruker.R,
%   R/feature_groups-envimass.R, R/feature_groups-greedy.R,
%   R/feature_groups-ims.R, R/feature_groups-kpic2.R,
%   R/feature_groups-openms.R, R/feature_groups-sirius.R,
%   R/feature_groups-table.R, R/feature_groups-xcms.R, R/feature_groups-xcms3.R
\docType{class}
\name{featureGroups-class}
\alias{featureGroups-class}
\alias{featureGroups}
\alias{names,featureGroups-method}
\alias{analyses,featureGroups-method}
\alias{replicates,featureGroups-method}
\alias{groupNames,featureGroups-method}
\alias{length,featureGroups-method}
\alias{hasIMS,featureGroups-method}
\alias{fromIMS,featureGroups-method}
\alias{show,featureGroups-method}
\alias{groupTable,featureGroups-method}
\alias{groupTable}
\alias{analysisInfo,featureGroups-method}
\alias{analysisInfo<-,featureGroups-method}
\alias{groupInfo,featureGroups-method}
\alias{groupInfo}
\alias{featureTable,featureGroups-method}
\alias{getFeatures,featureGroups-method}
\alias{groupFeatIndex,featureGroups-method}
\alias{groupFeatIndex}
\alias{groupQualities,featureGroups-method}
\alias{groupQualities}
\alias{groupScores,featureGroups-method}
\alias{groupScores}
\alias{getFeatureQualityNames,featureGroups-method}
\alias{annotations,featureGroups-method}
\alias{internalStandards,featureGroups-method}
\alias{internalStandards}
\alias{internalStandardAssignments,featureGroups-method}
\alias{internalStandardAssignments}
\alias{adducts,featureGroups-method}
\alias{adducts<-,featureGroups-method}
\alias{concentrations,featureGroups-method}
\alias{concentrations}
\alias{toxicities,featureGroups-method}
\alias{toxicities}
\alias{[,featureGroups,ANY,ANY,missing-method}
\alias{[[,featureGroups,ANY,ANY-method}
\alias{$,featureGroups-method}
\alias{delete,featureGroups-method}
\alias{export,featureGroups-method}
\alias{unique,featureGroups-method}
\alias{unique}
\alias{overlap,featureGroups-method}
\alias{overlap}
\alias{calculatePeakQualities,featureGroups-method}
\alias{selectIons,featureGroups-method}
\alias{selectIons}
\alias{normInts,featureGroups-method}
\alias{normInts}
\alias{getTICs,featureGroups-method}
\alias{getBPCs,featureGroups-method}
\alias{updateGroups,featureGroups-method}
\alias{updateGroups}
\alias{featureGroupsSet-class}
\alias{featureGroupsSet}
\alias{sets,featureGroupsSet-method}
\alias{internalStandardAssignments,featureGroupsSet-method}
\alias{adducts,featureGroupsSet-method}
\alias{adducts<-,featureGroupsSet-method}
\alias{delete,featureGroupsSet-method}
\alias{show,featureGroupsSet-method}
\alias{featureTable,featureGroupsSet-method}
\alias{[,featureGroupsSet,ANY,ANY,missing-method}
\alias{export,featureGroupsSet-method}
\alias{selectIons,featureGroupsSet-method}
\alias{normInts,featureGroupsSet-method}
\alias{featureGroupsUnset-class}
\alias{featureGroupsUnset}
\alias{unset,featureGroupsSet-method}
\alias{featureGroupsBruker-class}
\alias{featureGroupsBruker}
\alias{featureGroupsEnviMass-class}
\alias{featureGroupsEnviMass}
\alias{featureGroupsGreedy-class}
\alias{featureGroupsGreedy}
\alias{featureGroupsIMS-class}
\alias{featureGroupsIMS}
\alias{featureGroupsKPIC2-class}
\alias{featureGroupsKPIC2}
\alias{delete,featureGroupsKPIC2-method}
\alias{featureGroupsOpenMS-class}
\alias{featureGroupsOpenMS}
\alias{featureGroupsSIRIUS-class}
\alias{featureGroupsSIRIUS}
\alias{delete,featureGroupsSIRIUS-method}
\alias{featureGroupsTable-class}
\alias{featureGroupsTable}
\alias{featureGroupsXCMS-class}
\alias{featureGroupsXCMS}
\alias{analysisInfo<-,featureGroupsXCMS-method}
\alias{delete,featureGroupsXCMS-method}
\alias{featureGroupsXCMS3-class}
\alias{featureGroupsXCMS3}
\alias{delete,featureGroupsXCMS3-method}
\title{Base class for grouped features.}
\usage{
groupTable(object, ...)

groupFeatIndex(fGroups)

groupInfo(fGroups)

unique(x, incomparables = FALSE, ...)

overlap(fGroups, which = NULL, aggregate = TRUE, exclusive = FALSE, ...)

selectIons(fGroups, components, prefAdduct, ...)

groupQualities(fGroups)

groupScores(fGroups)

internalStandards(fGroups)

internalStandardAssignments(fGroups, ...)

normInts(
  fGroups,
  featNorm = "none",
  groupNorm = FALSE,
  normFunc = max,
  standards = NULL,
  ISTDRTWindow = 120,
  ISTDMZWindow = 300,
  minISTDs = 3,
  ...
)

concentrations(fGroups, ...)

toxicities(fGroups, ...)

updateGroups(fGroups, ...)

\S4method{names}{featureGroups}(x)

\S4method{analyses}{featureGroups}(obj)

\S4method{replicates}{featureGroups}(obj)

\S4method{groupNames}{featureGroups}(obj)

\S4method{length}{featureGroups}(x)

\S4method{hasIMS}{featureGroups}(obj)

\S4method{fromIMS}{featureGroups}(obj)

\S4method{show}{featureGroups}(object)

\S4method{groupTable}{featureGroups}(object, areas = FALSE, normalized = FALSE)

\S4method{analysisInfo}{featureGroups}(obj, df = FALSE)

\S4method{analysisInfo}{featureGroups}(obj) <- value

\S4method{groupInfo}{featureGroups}(fGroups)

\S4method{featureTable}{featureGroups}(obj)

\S4method{getFeatures}{featureGroups}(obj)

\S4method{groupFeatIndex}{featureGroups}(fGroups)

\S4method{groupQualities}{featureGroups}(fGroups)

\S4method{groupScores}{featureGroups}(fGroups)

\S4method{getFeatureQualityNames}{featureGroups}(
  obj,
  feat = TRUE,
  group = TRUE,
  scores = FALSE,
  totScore = TRUE
)

\S4method{annotations}{featureGroups}(obj)

\S4method{internalStandards}{featureGroups}(fGroups)

\S4method{internalStandardAssignments}{featureGroups}(fGroups)

\S4method{adducts}{featureGroups}(obj)

\S4method{adducts}{featureGroups}(obj) <- value

\S4method{concentrations}{featureGroups}(fGroups)

\S4method{toxicities}{featureGroups}(fGroups)

\S4method{[}{featureGroups,ANY,ANY,missing}(x, i, j, ..., ni, replicates, IMS, results, reorder = FALSE, drop = TRUE)

\S4method{[[}{featureGroups,ANY,ANY}(x, i, j)

\S4method{$}{featureGroups}(x, name)

\S4method{delete}{featureGroups}(obj, i = NULL, j = NULL, ...)

\S4method{export}{featureGroups}(obj, type, out, IMS = FALSE)

\S4method{unique}{featureGroups}(
  x,
  incomparables = FALSE,
  which,
  aggregate = TRUE,
  relativeTo = NULL,
  outer = FALSE
)

\S4method{overlap}{featureGroups}(fGroups, which, aggregate, exclusive)

\S4method{calculatePeakQualities}{featureGroups}(
  obj,
  weights,
  flatnessFactor,
  featureQualities = NULL,
  featureGroupQualities = NULL,
  avgFunc = mean,
  EICParams = getDefEICParams(window = 0),
  parallel = TRUE
)

\S4method{selectIons}{featureGroups}(
  fGroups,
  components,
  prefAdduct,
  onlyMonoIso = TRUE,
  chargeMismatch = "adduct"
)

\S4method{normInts}{featureGroups}(
  fGroups,
  featNorm = "none",
  groupNorm = FALSE,
  normFunc = max,
  standards = NULL,
  ISTDRTWindow = 120,
  ISTDMZWindow = 300,
  minISTDs = 3,
  ...
)

\S4method{getTICs}{featureGroups}(obj, retentionRange = NULL, MSLevel = 1)

\S4method{getBPCs}{featureGroups}(obj, retentionRange = NULL, MSLevel = 1)

\S4method{updateGroups}{featureGroups}(
  fGroups,
  what = c("ret", "mz", "mobility", "CCS"),
  intWeight = FALSE
)

\S4method{sets}{featureGroupsSet}(obj)

\S4method{internalStandardAssignments}{featureGroupsSet}(fGroups, set = NULL)

\S4method{adducts}{featureGroupsSet}(obj, set, ...)

\S4method{adducts}{featureGroupsSet}(obj, set, reGroup = TRUE) <- value

\S4method{delete}{featureGroupsSet}(obj, i = NULL, j = NULL, ...)

\S4method{show}{featureGroupsSet}(object)

\S4method{featureTable}{featureGroupsSet}(obj)

\S4method{[}{featureGroupsSet,ANY,ANY,missing}(x, i, j, ..., sets = NULL, reorder = FALSE, drop = TRUE)

\S4method{export}{featureGroupsSet}(obj, type, out, ..., set)

\S4method{selectIons}{featureGroupsSet}(fGroups, components, prefAdduct, ...)

\S4method{normInts}{featureGroupsSet}(
  fGroups,
  featNorm = "none",
  groupNorm = FALSE,
  normFunc = max,
  standards = NULL,
  ISTDRTWindow = 120,
  ISTDMZWindow = 300,
  minISTDs = 3,
  ...
)

\S4method{unset}{featureGroupsSet}(obj, set)

\S4method{delete}{featureGroupsKPIC2}(obj, ...)

\S4method{delete}{featureGroupsSIRIUS}(obj, ...)

\S4method{analysisInfo}{featureGroupsXCMS}(obj) <- value

\S4method{delete}{featureGroupsXCMS}(obj, ...)

\S4method{delete}{featureGroupsXCMS3}(obj, ...)
}
\arguments{
\item{\dots}{For the \code{"["} operator: ignored.

  For \code{delete}: passed to the function specified as \code{j}.

  For \code{normInts}: passed to \code{\link{screenSuspects}} if \code{featNorm="istd"}.

  \setsPassedArgs1{featureGroups}}

\item{fGroups, obj, x, object}{\code{featureGroups} object to be accessed.}

\item{incomparables}{Not used. Included for compatibility with the generic.}

\item{which}{A character vector with the selection to compare (\emph{e.g.} replicates, as set by the \code{aggregate}
  argument).

  For \code{overlap}: can also be a \code{NULL} to compare all elements.}

\item{aggregate}{Specifies how data should be aggregated prior to comparison. Set to \code{FALSE} to compare
analyses, \code{TRUE} to compare replicates or to the name of a column in the \link[=analysis-information]{analysis
information} to compare by a custom grouping of analyses.}

\item{exclusive}{If \code{TRUE} then all feature groups are removed that are not unique to the given replicates.}

\item{components}{The \code{components} object that was generated for the given \code{featureGroups} object.
Obviously, the components must be created with algorithms that support adduct/isotope annotations, such as those
from \pkg{RAMClustR} and \pkg{cliqueMS}.}

\item{prefAdduct}{The 'preferred adduct' (see method description). This is often \code{"[M+H]+"} or \code{"[M-H]-"}.}

\item{featNorm}{The method applied for feature normalization: \code{"istd"}, \code{"tic"}, \code{"conc"} or
\code{"none"}. See the \verb{Feature intensity normalization} section for details.}

\item{groupNorm}{If \code{TRUE} then group normalization is performed. See the \verb{Feature intensity normalization}
section for details.}

\item{normFunc}{A \code{function} to combine data for normalization. See the \verb{Feature intensity normalization}
section for details.}

\item{standards}{A \code{data.table} (or \code{data.frame}) with all internal standards. Should follow the format of
  a \link[=suspect-screening]{suspect list}. Only used if \code{featNorm="istd"}. See the \verb{Feature intensity
  normalization} section for details.

  \setsWF Can also be a \code{list} with internal standard lists.

  See the \code{suspects} argument to \code{\link{screenSuspects}} for more details.}

\item{ISTDRTWindow, ISTDMZWindow}{The retention time and \emph{m/z} windows for IS selection. Only used if
\code{featNorm="istd"}. See the \verb{Feature intensity normalization} section for details.}

\item{minISTDs}{The minimum number of IS that should be assigned to each feature (if possible). Only used if
\code{featNorm="istd"}. See the \verb{Feature intensity normalization} section for details.}

\item{areas}{If set to \code{TRUE} then areas are considered instead of peak intensities.}

\item{normalized}{If \code{TRUE} then normalized intensity data is used (see the \verb{Feature intensity
normalization} section.}

\item{df}{If \code{TRUE} then the returned value is a \code{data.frame}, otherwise a \code{data.table}.}

\item{value}{For \code{analysisInfo<-}: A \code{data.frame} or \code{data.table} with updated analysis information.

  For \code{adducts<-}: A \code{character} with adduct annotations assigned to each feature group. The length should
  equal the number of feature groups. Can be named with feature group names to customize the assignment order.}

\item{feat, group}{If \code{TRUE} then names specific to features and feature groups are returned, respectively.}

\item{scores}{If \code{TRUE} the score names are returned, otherwise the quality names.}

\item{totScore}{If \code{TRUE} (and \code{scores=TRUE}) then the name of the total score is included.}

\item{i, j}{For \code{[}/\code{[[}: A numeric or character value which is used to select analyses/feature groups by
their index or name, respectively (for the order/names see \code{analyses()/names()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all analyses/feature groups are selected.\cr\cr For \code{[[}: should be a scalar value. If \code{j} is not specified, \code{i} selects by feature groups instead.\cr\cr For \code{delete}: The data to remove from. \code{i} are the
analyses as numeric index, logical or character, \code{j} the feature groups as numeric index, logical or character. If either is
\code{NULL} then data for all is removed. \code{j} may also be a function: it will be called for each feature group,
with a vector of the group intensities, the group name and any other arguments passed as \code{\dots} to \code{delete}. The return value of this
function specifies the analyses of the features in the group to be removed (same format as \code{i}).}

\item{ni}{Optional argument. An expression used for subsetting the analyses. The
\link[=analysis-information]{analysis information} is first subset and the remaining rows are used to determine for
which analyses the results should be kept. The unevaluated \code{ni} expression is used to set the \code{i}
argument of the subset operator of \CRANpkg{data.table}, which therefore brings the advanced subsetting
capabilities of \CRANpkg{data.table} (see the
\code{\link[data.table]{data.table}} documentation for more details). For instance, \code{fGroups[replicate ==
"standard"]} would subset all analyses assigned with the replicate \code{"standard"}.}

\item{replicates}{An optional \code{character} vector: if specified only keep results for the given replicates
(equivalent to the \code{replicates} argument to \code{\link[=filter,featureGroups-method]{filter}}).}

\item{IMS}{\IMSWF Specifies which feature groups are considered to be kept (\code{"["}) or exported (\code{export}) in IMS workflows. The following options
  are valid: \itemize{

  \item \code{"both"}: Selects IMS and non-IMS features.

  \item \code{"maybe"}: Selects non-IMS features and IMS features without assigned IMS precursor.

  \item \code{FALSE}: Selects only non-IMS features.

  \item \code{TRUE}: Selects only IMS features.

  }

  For \code{"["}: Leave unassigned to perform no IMS selection.
  For \code{export}: keep \code{FALSE} as the export formats currently do not support IMS data.}

\item{results}{Optional argument. If specified only feature groups with results in the specified object are kept. The
class of \code{results} should be \code{\link{featureAnnotations}} or \code{\link{components}}. Multiple objects
can be specified in a \code{list}: in this case a feature group is kept if it has a result in \emph{any} of the
objects (equivalent to the \code{results} argument to \code{\link[=filter,featureGroups-method]{filter}}).}

\item{reorder}{If \code{TRUE} then the order of the analyses is changed to match the order of the \code{i} argument.

  \setsWF If the \code{sets} argument is specified (and \code{i} is not) then the order of sets is changed instead.}

\item{drop}{Ignored.}

\item{name}{The feature group name (partially matched).}

\item{type}{The export type: \code{"brukerpa"} (Bruker ProfileAnalysis), \code{"brukertasq"} (Bruker TASQ) or
\code{"mzmine"} (MZmine).}

\item{out}{The destination file for the exported data.}

\item{relativeTo}{A character vector of groupings that should be used for unique comparison. The groupings
(\emph{e.g.} replicates) are configured by the \code{aggregate} argument. If \code{NULL} then all groupings are
used for comparison. Data specified in \code{which} are ignored.}

\item{outer}{If \code{TRUE} then only feature groups are kept which do not overlap among those present in the
groupings specified by the \code{which} parameter.}

\item{weights}{A named \code{numeric} vector that defines the weight for each score to calculate the
\verb{totalScore}. The names of the vector follow the score names. Unspecified weights are defaulted to \samp{1}.
Example: \code{weights=c(ApexBoundaryRatioScore=0.5, GaussianSimilarityScore=2)}.}

\item{flatnessFactor}{Passed to \pkg{MetaClean} as the \code{flatness.factor} argument to
\code{\link[MetaClean]{calculateJaggedness}} and \code{\link[MetaClean]{calculateModality}}.}

\item{featureQualities}{Specifies which feature qualities to calculate. Can be \code{NULL} (default, calculates all
qualities), a \code{character} vector with names of qualities to calculate (e.g., \code{c("FWHM2Base",
"Symmetry")}), or a \code{list} of custom quality definitions. See the
\code{\link[=feature-quality]{featureQualities}} function for more details.}

\item{featureGroupQualities}{Analogous to \code{featureQualities} for feature groups metrics. See the
\code{\link[=feature-quality]{featureGroupQualities}} function for more details.}

\item{avgFunc}{The function used to average the peak qualities and scores for each feature group.}

\item{EICParams}{A named \code{list} with parameters used for extracted ion chromatogram (EIC) creation. See the
\link[=EIXParams]{EIC parameters} documentation for more details.}

\item{parallel}{If set to \code{TRUE} then code is executed in parallel through the \CRANpkg{future} package. Please
see the parallelization section in the handbook for more details.}

\item{onlyMonoIso}{Set to \code{TRUE} to only keep feature groups that were annotated as monoisotopic. Feature groups
are never removed by this setting if no isotope annotations are available.}

\item{chargeMismatch}{Specifies how to deal with a mismatch in charge between adduct and isotope annotations. Valid
values are: \code{"adduct"} (ignore isotope annotation), \code{"isotope"} (ignore adduct annotation), \code{"none"}
(ignore both annotations) and \code{"ignore"} (don't check for charge mismatches). \emph{Important}: when
\command{OpenMS} is used to find features, it already removes any detected non-monoisotopic features by default.
Hence, in such case setting \code{chargeMismatch="adduct"} is more appropriate.}

\item{retentionRange}{Range of retention time (in seconds) to collect TIC traces. Should be a numeric vector with
length of two containing the min/max values. Set to NULL to ignore.}

\item{MSLevel}{Integer vector with the ms levels (i.e., 1 for MS1 and 2 for MS2) to obtain TIC traces.}

\item{what}{A \code{character} vector specifying which group-wise values to update. Valid values are \code{"ret"}
(retention time, in seconds), \code{"mz"} and \code{"mobility"}. At least one must be specified. If
\code{"mobility"} is specified an no IMS information is present, it will be ignored.}

\item{intWeight}{If \code{TRUE}, calculate intensity-weighted means per group; otherwise use simple arithmetic means.}

\item{set}{\setsWF The name of the set.}

\item{reGroup}{\setsWF Set to \code{TRUE} to re-group the features after the adduct annotations are changed. See the
\verb{Sets workflow} section for more details.}

\item{sets}{\setsWF A \code{character} with name(s) of the sets to keep.}
}
\value{
\code{delete} returns the object for which the specified data was removed.

\code{calculatePeakQualities} returns a modified object amended with peak qualities and scores.

\code{selectIons} returns a \code{featureGroups} object with only the selected feature groups and amended
  with adduct annotations.

\code{normInts} returns a \code{featureGroups} object, amended with data in the \code{ISTDs} and
  \code{ISTDAssignments} slots if \code{featNorm="istd"}.
}
\description{
This class holds all the information for grouped features.
}
\details{
The \code{featureGroup} class is the workhorse of \pkg{patRoon}: almost all functionality operate on its instantiated
objects. The class holds all information from grouped features (obtained from \code{\link{features}}). This class
itself is \code{virtual}, hence, objects are not created directly from it. Instead, 'feature groupers' such as
\code{\link{groupFeaturesXCMS}} return a \code{featureGroups} derived object after performing the actual grouping of
features across analyses.
}
\section{Methods (by generic)}{
\itemize{
\item \code{names(featureGroups)}: Obtain feature group names.

\item \code{analyses(featureGroups)}: returns a \code{character} vector with the names of the
analyses for which data is present in this object.

\item \code{replicates(featureGroups)}: returns a \code{character} vector with the names of the
replicates for which data is present in this object.

\item \code{groupNames(featureGroups)}: Same as \code{names}. Provided for consistency to other classes.

\item \code{length(featureGroups)}: Obtain number of feature groups.

\item \code{hasIMS(featureGroups)}: Returns \code{TRUE} if the feature groups object has ion mobility information.

\item \code{fromIMS(featureGroups)}: Returns \code{TRUE} if the features were directly generated from IMS data.

\item \code{show(featureGroups)}: Shows summary information for this object.

\item \code{groupTable(featureGroups)}: Accessor for \code{groups} slot.

\item \code{analysisInfo(featureGroups)}: Obtain analysisInfo (see analysisInfo slot in \code{\link{features}}).

\item \code{analysisInfo(featureGroups) <- value}: Modifies the \link[=analysis-information]{analysis information} of this \code{features} object.
This is primarily intended to change or add analysis metadata columns or can be used to re-order analysis. The
removal or addition of analyses and changes to the \code{"analysis"} column are not supported. This function
performs several internal updates after analysis information modifications. Hence, never attempt to change the
\code{analysisInfo} slot directly.

\item \code{groupInfo(featureGroups)}: Accessor for \code{groupInfo} slot.

\item \code{featureTable(featureGroups)}: Obtain feature information (see \code{\link{features}}).

\item \code{getFeatures(featureGroups)}: Accessor for \code{features} slot.

\item \code{groupFeatIndex(featureGroups)}: Accessor for \code{ftindex} slot.

\item \code{groupQualities(featureGroups)}: Accessor for \code{groupQualities} slot.

\item \code{groupScores(featureGroups)}: Accessor for \code{groupScores} slot.

\item \code{getFeatureQualityNames(featureGroups)}: Returns feature quality names that were calculated for this object.

\item \code{annotations(featureGroups)}: Accessor for \code{annotations} slot.

\item \code{internalStandards(featureGroups)}: Accessor for \code{ISTDs} slot.

\item \code{internalStandardAssignments(featureGroups)}: Accessor for \code{ISTDAssignments} slot.

\item \code{adducts(featureGroups)}: Returns a named \code{character} with adduct annotations assigned to each feature group (if
available).

\item \code{adducts(featureGroups) <- value}: Sets adduct annotations for feature groups.

\item \code{concentrations(featureGroups)}: Accessor for \code{concentrations} slot.

\item \code{toxicities(featureGroups)}: Accessor for \code{toxicities} slot.

\item \code{x[i}: Subset on analyses/feature groups.

\item \code{x[[i}: Extract intensity values.

\item \code{$}: Extract intensity values for a feature group.

\item \code{delete(featureGroups)}: Completely deletes specified feature groups.

\item \code{export(featureGroups)}: Exports feature groups to a \file{.csv} file that is readable to Bruker ProfileAnalysis (a
'bucket table'), Bruker TASQ (an analyte database) or that is suitable as input for the \verb{Targeted peak
detection} functionality of \href{http://mzmine.github.io/}{MZmine}.

\item \code{unique(featureGroups)}: Obtain a subset with unique feature groups present in one or more analyses, replicates etc.

\item \code{overlap(featureGroups)}: Obtain a subset with feature groups that overlap between a set of specified replicate(s).

\item \code{calculatePeakQualities(featureGroups)}: Calculates peak and group qualities for all features and feature groups. The peak qualities
(and scores) are calculated with the \link[=calculatePeakQualities,features-method]{features method of this
function}, and subsequently averaged per feature group. Group metrics are then calculated and scored and
scaled by normalizing qualities among all groups and scaling them from \samp{0} (worst) to \samp{1} (best). The
\verb{totalScore} for each group is then calculated as the weighted sum from all feature (group) scores. The
\code{\link{getMCTrainData}} and \code{\link{predictCheckFeaturesSession}} functions can be used to train and apply
Pass/Fail ML models from \pkg{MetaClean}.

\item \code{selectIons(featureGroups)}: uses \link[=generateComponents]{componentization} results to select feature groups with
preferred adduct ion and/or isotope annotation. Typically, this means that only feature groups are kept if they are
(de-)protonated adducts and are monoisotopic. The adduct annotation assignments for the selected feature groups are
copied from the components to the \code{annotations} slot. If the adduct for a feature group is unknown, its
annotation is defaulted to the 'preferred' adduct, and hence, the feature group will never be removed. Furthermore,
if a component does not contain an annotation with the preferred adduct, the most intense feature group is selected
instead. Similarly, if no isotope annotation is available, the feature group is assumed to be monoisotopic and thus
not removed. An important advantage of \code{selectIons} is that it may considerably simplify your dataset.
Furthermore, the adduct assignments allow formula/compound annotation steps later in the workflow to improve their
annotation accuracy. On the other hand, it is important the componentization results are reliable. Hence, it is
highly recommended that, prior to calling \code{selectIons}, the settings to \code{\link{generateComponents}} are
optimized and its results are reviewed with \code{\link{checkComponents}}. Finally, the \code{adducts<-} method can
be used to manually correct adduct assignments afterwards if necessary.

\item \code{normInts(featureGroups)}: Provides various methods to normalizes feature intensities for each sample analysis or of
all features within a feature group. See the \verb{Feature intensity normalization} section below.

\item \code{getTICs(featureGroups)}: Obtain the total ion chromatogram/s (TICs) of the analyses.

\item \code{getBPCs(featureGroups)}: Obtain the base peak chromatogram/s (BPCs) of the analyses.

\item \code{updateGroups(featureGroups)}: Recalculate group information from feature data.

}}
\section{Slots}{

\describe{
\item{\code{groups}}{Matrix (\code{\link{data.table}}) with intensities for each feature group (columns) per analysis (rows).
Access with \code{groups} method.}

\item{\code{features}}{\code{\link{features}} class associated with this object. Access with\code{featureTable} methods.}

\item{\code{groupInfo}}{\code{data.table} with retention time (\code{ret} column, in seconds) and \emph{m/z} (\code{mz}
column) for each feature group. Access with \code{groupInfo} method.}

\item{\code{ftindex}}{Matrix (\code{\link{data.table}}) with feature indices for each feature group (columns) per analysis
(rows). Each index corresponds to the row within the feature table of the analysis (see
\code{\link{featureTable}}).}

\item{\code{groupQualities,groupScores}}{A \code{\link{data.table}} with qualities/scores for each feature group (see the
\code{calculatePeakQualities} method).}

\item{\code{annotations}}{A \code{\link{data.table}} with adduct annotations for each group (see the \code{selectIons}
method).}

\item{\code{ISTDs}}{A \code{data.table} with screening results for internal standards (filled in by the \code{normInts}
method).}

\item{\code{ISTDAssignments}}{A \code{list}, where each item is named by a feature group and consists of a vector with
feature group names of the internal standards assigned to it (filled in by the \code{normInts} method).}

\item{\code{concentrations,toxicities}}{A \code{data.table} with predicted concentrations/toxicities for each feature group.
Assigned by the \code{\link{calculateConcs}}/\code{\link{calculateTox}} methods. Use the
\code{concentratrions}/\code{toxicities} methods for access.}

\item{\code{groupAlgo,groupArgs,groupVerbose}}{\setsWF Grouping parameters that were used when this object was created. Used
by \code{adducts<-} and \code{selectIons} when these methods perform a re-grouping of features.}

\item{\code{annotations,ISTDAssignments}}{\setsWF As the \code{featureGroups} slots, but contains the data per set.}

\item{\code{annotationsChanged}}{Set internally by \code{adducts()<-} and applied as soon as \code{reGroup=TRUE}.}
}}

\section{Use of raw HRMS data}{
 The \link[=msdata]{raw data interface} of \pkg{patRoon} is used by \code{calculatePeakQualities} to
  process HRMS (or IMS-HRMS) data. Please see \link[=msdata]{its documentation} for more information on the supported
  formats and available configuration options.
}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{workflowStep}}}   \itemize{     \item{\strong{\code{\link{featureGroups}}}}     \itemize{       \item{\code{\link{featureGroupsSet}}}       \itemize{         \item{\code{\link{featureGroupsScreeningSet}}}       }       \item{\code{\link{featureGroupsUnset}}}       \item{\code{\link{featureGroupsScreening}}}       \itemize{         \item{\code{\link{featureGroupsSetScreeningUnset}}}       }       \item{\code{\link{featureGroupsBruker}}}       \item{\code{\link{featureGroupsConsensus}}}       \item{\code{\link{featureGroupsEnviMass}}}       \item{\code{\link{featureGroupsGreedy}}}       \item{\code{\link{featureGroupsIMS}}}       \item{\code{\link{featureGroupsKPIC2}}}       \item{\code{\link{featureGroupsOpenMS}}}       \item{\code{\link{featureGroupsSIRIUS}}}       \item{\code{\link{featureGroupsTable}}}       \item{\code{\link{featureGroupsBrukerTASQ}}}       \item{\code{\link{featureGroupsXCMS}}}       \item{\code{\link{featureGroupsXCMS3}}}     }   } }
}

\section{Feature intensity normalization}{
 The \code{normInts} method performs normalization of feature intensities
  (and areas). These values are amended in the \code{features} slot, while the original intensities/areas are kept.
  To use the normalized intensities set \code{normalized=TRUE} to methods such as \code{\link{plotInt}},
  \code{\link{generateComponentsIntClust}} and \code{\link[=feature-table]{as.data.table}}. Please see the
  \code{normalized} argument documentation for these methods for more details.

  The \code{normInts} method supports several methods to normalize intensities/areas of features within the same
  analysis. Most methods are influenced by the \emph{normalization concentration} (\code{norm_conc} in the
  \link[=analysis-information]{analysis information}) set for each sample analysis. For \code{NA} or zero values the
  output will be zero. If \code{norm_conc} is completely absent from the analysis information or all values are
  \code{NA}, the normalization concentration is defaulted to one.

  The different normalization methods are:

  \enumerate{

  \item \code{featNorm="istd"} Uses \emph{internal standards} (IS) for normalization. The IS are screened internally
  by the \code{\link{screenSuspects}} function. Hence, the IS specified by the \code{standards} argument should
  follow the format of a \link[=suspect-screening]{suspect list}. Note that labelled elements in IS formulae should
  be specified with the \CRANpkg{rcdk} format, \emph{e.g.} \code{"[13]C"} for 13C, \code{"[2]H"} for a deuterium etc.
  Example IS lists are provided with the \pkg{patRoonData} and \pkg{patRoonDataIMS} packages.

  The assignment of IS to features is automatically performed, using the following criteria: \enumerate{

  \item Only analyses are considered with a defined normalization concentration.

  \item The IS must be detected in all of the analyses in which the feature was detected.

  \item The retention time and \emph{m/z} are reasonably close (\code{ISTDRTWindow}/\code{ISTDMZWindow} arguments).
  However, additional IS candidates outside these windows will be chosen if the number of candidates is less than the
  \code{minISTDs} argument. In this case the next close(st) candidate(s) will be chosen.

  }

  Normalization of features within the same feature group always occur with the same IS. If multiple IS are assigned
  to a feature then normalization occurs with the combined intensity (area), which is calculated with the function
  defined by the \code{normFunc} argument. The (combined) IS intensity is then normalized by the normalization
  concentration, and finally used for feature normalization.

  \item \code{featNorm="tic"} Uses the Total Ion Current (TIC) to normalize intensities. The TIC is calculated by
  combining all intensities with the function defined by the \code{normFunc} argument. For this reason, you may need
  to take care to perform normalization before \emph{e.g.} suspect screening or other prioritization techniques. The
  TIC normalized intensities are finally divided by the normalization concentration.

  \item \code{featNorm="conc"} Simply divides all intensities (areas) with the normalization concentration defined
  for the sample.

  \item \code{featNorm="none"} Performs no normalization. The raw intensity values are simply copied. This is mainly
  useful if you only want to do group normalization (described below).

  }

  The meaning of the normalization concentration differs for each method: for \code{"istd"} it resembles the IS
  concentration of a sample analysis, whereas for \code{"tic"} and \code{"conc"} it is used to normalize different
  sample amounts (\emph{e.g.} injection volume).

  If \code{groupNorm=TRUE} then feature intensities (areas) will be normalized by the combined values for its feature
  group (again, combination occurs with \code{normFunc}). This \emph{group normalization} always occurs \emph{after}
  aforementioned normalization methods. Group normalization was the only method with \pkg{patRoon} \samp{<2.1}, and
  still occurs automatically if \code{normInts} was not called when a method is executed that requests normalized
  data.

  In IMS workflows with post mobility assignment (see \link[=assignMobilities_feat]{assignMobilities}), any IMS
  features are excluded for the assignment of internal standards (\code{featNorm="istd"}) or calculation of TICs
  (\code{featNorm="tic"}). Furthermore, the normalized intensities and areas for IMS features are copied from
  their IMS precursors.
}

\section{Sets workflows}{
 \setsWFClass{featureGroupsSet}{featureGroups}

  \setsWFNewMethodsFeat{featureGroupsUnset}{The adduct annotations for the selected set are used to convert all
  feature (group) masses to ionic \emph{m/z} values. The annotations persist in the converted object. }

  \setsWFChangedMethods{

  \item \code{adducts}, \code{adducts<-} require the \code{set} argument. The order of the data that is
  returned/changed follows that of the \code{annotations} slot. Furthermore, \code{adducts<-} will perform a
  re-grouping of features when its \code{reGroup} parameter is set to \code{TRUE}. The implications for this are
  discussed below. Note that no adducts are changed \emph{until} \code{reGroup=TRUE}.

  \item the subset operator (\code{[}) has specific arguments to choose (feature presence in) sets. See the argument
  descriptions.

  \item \code{export} Only allows to export data from one set. The \code{unset} method is used prior to exporting the
  data.

  \item \code{overlap} and \code{unique} allow to handle data per set. See the \code{sets} argument description.

  \item \code{selectIons} Will perform a re-grouping of features. The implications of this are discussed below.

  \item \code{normInts} Performs normalization for each set \emph{independently}.

  }

  A re-grouping of features occurs if \code{selectIons} is called or \code{adducts<-} is used with
  \code{reGroup=TRUE}. Afterwards, it is very likely that feature group names are changed. Since data generated later
  in the workflow (\emph{e.g.} annotation steps) rely on feature group names, these objects are \strong{not valid}
  anymore, and \strong{must} be re-generated.
}

\references{
\insertRef{Chetnik2020}{patRoon}
}
\seealso{
\code{\link{groupFeatures}} for generating feature groups, \link{feature-filtering}, \link{feature-table}
  and \link{feature-plotting} for more advanced \code{featureGroups} methods.
}
\author{
Rick Helmus <\email{r.helmus@uva.nl}> and Ricardo Cunha <\email{cunha@iuta.de}> (\code{getTICs} and
  \code{getBPCs} functions)
}
