% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-comparison.R
\docType{class}
\name{featureGroupsComparison-class}
\alias{featureGroupsComparison-class}
\alias{featureGroupsComparison}
\alias{names,featureGroupsComparison-method}
\alias{length,featureGroupsComparison-method}
\alias{[,featureGroupsComparison,ANY,ANY,ANY-method}
\title{Feature groups comparison class}
\usage{
\S4method{names}{featureGroupsComparison}(x)

\S4method{length}{featureGroupsComparison}(x)

\S4method{[}{featureGroupsComparison,ANY,ANY,ANY}(x, i, j = "missing",
  drop = "missing")
}
\arguments{
\item{x}{A \code{featureGroupsComparison} object.}

\item{i}{Character, numeric or logical vector used for subsetting (analogous
to a \code{list}).}

\item{j, drop}{Not used. Should be ignored.}
}
\description{
This class is used for comparing different \code{\link{featureGroups}}
objects.
}
\details{
Objects from this class are returned by \code{\link{comparison}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{names}: Obtain the labels that were given to each compared feature group.

\item \code{length}: Number of feature groups objects that were compared.

\item \code{[}: Creates a subset of the compared feature groups.
}}

\section{Slots}{

\describe{
\item{\code{fGroupsList}}{A \code{list} of \code{\link{featureGroups}} object that
were compared}

\item{\code{comparedFGroups}}{A \emph{pseudo} \code{featureGroups} object containing
grouped feature groups.}
}}

