% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-comparison.R
\docType{class}
\name{featureGroupsComparison-class}
\alias{featureGroupsComparison-class}
\alias{featureGroupsComparison}
\alias{names,featureGroupsComparison-method}
\alias{length,featureGroupsComparison-method}
\alias{[,featureGroupsComparison,ANY,missing,missing-method}
\alias{[[,featureGroupsComparison,ANY,missing-method}
\alias{$,featureGroupsComparison-method}
\title{Feature groups comparison class}
\usage{
\S4method{names}{featureGroupsComparison}(x)

\S4method{length}{featureGroupsComparison}(x)

\S4method{[}{featureGroupsComparison,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{featureGroupsComparison,ANY,missing}(x, i, j)

\S4method{$}{featureGroupsComparison}(x, name)
}
\arguments{
\item{x}{A \code{featureGroupsComparison} object.}

\item{i}{A numeric or character value which is used to select labels by
their index or name, respectively (for the order/names see \code{names()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all labels are selected.\cr\cr For \code{[[}: should be a scalar value.}

\item{\dots}{Ignored.}

\item{drop, j}{ignored.}

\item{name}{The label name (partially matched).}
}
\description{
This class is used for comparing different \code{\link{featureGroups}}
objects.
}
\details{
Objects from this class are returned by \code{\link{comparison}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{names}: Obtain the labels that were given to each compared feature group.

\item \code{length}: Number of feature groups objects that were compared.

\item \code{[}: Subset on labels that were assigned to compared feature groups.

\item \code{[[}: Extract a \code{\link{featureGroups}} object by its label.

\item \code{$}: Extract a compound table for a feature group.
}}

\section{Slots}{

\describe{
\item{\code{fGroupsList}}{A \code{list} of \code{\link{featureGroups}} object that
were compared}

\item{\code{comparedFGroups}}{A \emph{pseudo} \code{featureGroups} object containing
grouped feature groups.}
}}

