% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-screening.R
\docType{class}
\name{featureGroupsScreening-class}
\alias{featureGroupsScreening-class}
\alias{featureGroupsScreening}
\alias{screenInfo,featureGroupsScreening-method}
\alias{show,featureGroupsScreening-method}
\alias{[,featureGroupsScreening,ANY,ANY,missing-method}
\alias{as.data.table,featureGroupsScreening-method}
\alias{annotateSuspects,featureGroupsScreening-method}
\alias{filter,featureGroupsScreening-method}
\title{Class for suspect screened feature groups.}
\usage{
\S4method{screenInfo}{featureGroupsScreening}(obj)

\S4method{show}{featureGroupsScreening}(object)

\S4method{[}{featureGroupsScreening,ANY,ANY,missing}(x, i, j, ..., rGroups, suspects = NULL, drop = TRUE)

\S4method{as.data.table}{featureGroupsScreening}(x, ..., collapseSuspects = ",", onlyHits = FALSE)

\S4method{annotateSuspects}{featureGroupsScreening}(
  fGroups,
  MSPeakLists,
  formulas,
  compounds,
  absMzDev = 0.005,
  relMinMSMSIntensity = 0.05,
  simMSMSMethod = "cosine",
  checkFragments = c("mz", "formula", "compound"),
  formulasNormalizeScores = "max",
  compoundsNormalizeScores = "max",
  IDFile = system.file("misc", "IDLevelRules.yml", package = "patRoon")
)

\S4method{filter}{featureGroupsScreening}(
  obj,
  ...,
  onlyHits = NULL,
  selectHitsBy = NULL,
  selectBestFGroups = FALSE,
  maxLevel = NULL,
  maxFormRank = NULL,
  maxCompRank = NULL,
  minAnnSimForm = NULL,
  minAnnSimComp = NULL,
  minAnnSimBoth = NULL,
  absMinFragMatches = NULL,
  relMinFragMatches = NULL,
  negate = FALSE
)
}
\arguments{
\item{obj, object, x, fGroups}{The \code{featureGroupsScreening} object.}

\item{i, j, rGroups}{Used for subsetting data analyses, feature groups and
replicate groups, see \code{\link{featureGroups}}.}

\item{\dots}{Passed to the base \code{\link{featureGroups}} method.}

\item{suspects}{An optional \code{character} vector with suspect names. If
specified, only \code{featureGroups} will be kept that are assigned to
these suspects.}

\item{collapseSuspects}{If a \code{character} then any suspects that were
matched to the same feature group are collapsed to a single row and suspect
names are separated by the value of \code{collapseSuspects}. If \code{NULL}
then no collapsing occurs, and each suspect match is reported on a single
row.}

\item{onlyHits}{For \code{as.data.table}: if \code{TRUE} then only feature
  groups with suspect hits are reported.

  For \code{filter} \itemize{

  \item if \code{negate=FALSE} and \code{onlyHits=TRUE} then all feature
  groups without suspect hits will be removed. Otherwise nothing will be
  done.

  \item if \code{negate=TRUE} then \code{onlyHits=TRUE} will select feature
  groups without suspect hits, \code{onlyHits=FALSE} will only retain feature
  groups with suspect matches and this filter is ignored if
  \code{onlyHits=NULL}.

  }}

\item{MSPeakLists, formulas, compounds}{Annotation data
(\code{\link{MSPeakLists}}, \code{\link{formulas}} and
\code{\code{compounds}}) obtained for this \code{featureGroupsScreening}
object. All arguments can be \code{NULL} to exclude it from the annotation.}

\item{absMzDev}{Maximum absolute \emph{m/z} deviation.}

\item{relMinMSMSIntensity}{Minimum relative intensity (\samp{0-1}) threshold
applied when calculating annotation similarities.}

\item{simMSMSMethod}{Either \code{"cosine"} or \code{"jaccard"}: used to
compare MS/MS peak lists for annotation similarity calculation.}

\item{checkFragments}{Which type(s) of MS/MS fragments from workflow data
should be checked to evaluate the number of suspect fragment matches
(\emph{i.e.} from the \code{fragments_mz}/\code{fragments_formula} columns
in the suspect list). Valid values are: \code{"mz"}, \code{"formula"},
\code{"compounds"}. The former uses \emph{m/z} values in the specified
\code{MSPeakLists} object, whereas the others use the formulae that were
annotated to MS/MS peaks in the given \code{formulas} or \code{compounds}
objects. Multiple values are possible: in this case the maximum number of
fragment matches will be reported.}

\item{compoundsNormalizeScores, formulasNormalizeScores}{A \code{character} that specifies how normalization of
annotation scorings occurs. Either

\code{"max"} (normalize to max value) or \code{"minmax"} (perform min-max
normalization). Note that normalization of negative scores (e.g. output by
\command{SIRIUS}) is always performed as min-max. Furthermore, currently
normalization for \code{compounds} takes the original min/max scoring
values into account when candidates were generated. Thus, for
\code{compounds} scoring, normalization is not affected when candidate
results were removed after they were generated (\emph{e.g.} by use of
\code{filter}).}

\item{IDFile}{A file path to a YAML file with rules used for estimation of
identification levels. See the \verb{Suspect annotation} section for more
details. If not specified then a default rules file will be used.}

\item{selectHitsBy}{Should be \code{"intensity"} or \code{"level"}. For cases
where the same suspect is matched to multiple feature groups, only the
suspect to the feature group with highest mean intensity
(\code{selectHitsBy="intensity"}) or best identification level
(\code{selectHitsBy="level"}) is kept. In case of ties only the first hit
is kept. Set to \code{NULL} to ignore this filter. If \code{negate=TRUE}
then only those hits with lowest mean intensity/poorest identification
level are kept.}

\item{selectBestFGroups}{If \code{TRUE} then for any cases where a single
feature group is matched to several suspects only the suspect assigned to
the feature group with best identification score is kept. In case of ties
only the first is kept.}

\item{maxLevel, maxFormRank, maxCompRank, minAnnSimForm, minAnnSimComp, minAnnSimBoth}{Filter suspects by maximum identification level (\emph{e.g.} \code{"3a"}),
formula/compound rank or with minimum formula/compound/combined annotation
similarity. Set to \code{NULL} to ignore.}

\item{absMinFragMatches, relMinFragMatches}{Only retain suspects with this
minimum number MS/MS matches with the fragments specified in the suspect
list (\emph{i.e.} \code{fragments_mz}/\code{fragments_formula}).
\code{relMinFragMatches} sets the minimum that is relative (\samp{0-1}) to
the maximum number of MS/MS fragments specified in the \code{fragments_*}
columns of the suspect list. Set to \code{NULL} to ignore.}

\item{negate}{If set to \code{TRUE} then filtering operations are performed
in opposite manner.}
}
\value{
\code{annotateSuspects} returns a \code{featureGroupsScreening}
  object amended with annotation data.

\code{filter} returns a filtered \code{featureGroupsScreening}
  object.
}
\description{
This class derives from \code{\link{featureGroups}} and adds suspect
screening information.
}
\section{Methods (by generic)}{
\itemize{
\item \code{screenInfo}: Returns a table with screening information
(see \code{screenInfo} slot).

\item \code{show}: Shows summary information for this object.

\item \code{[}: Subset on analyses, feature groups and/or
suspects.

\item \code{as.data.table}: Obtain a summary table (a
\code{\link{data.table}}) with retention, \emph{m/z}, intensity and
optionally other feature data. Furthermore, the output table will be merged
with information from \code{screenInfo}, such as suspect names and other
properties and annotation data.

\item \code{annotateSuspects}: Incorporates annotation data obtained
during the workflow to annotate suspects with matched known MS/MS
fragments, formula/candidate ranks and automatic estimation of
identification levels. See the \verb{Suspect annotation} section for more
details.

\item \code{filter}: Performs rule based filtering. This method
builds on the comprehensive filter functionality from the base
\code{\link{filter,featureGroups-method}}. It adds several filters to
select \emph{e.g.} the best ranked suspects or those with a minimum
estimated identification level. \strong{NOTE}: most filters \emph{only}
affect suspect hits, not feature groups. Set \code{onlyHits=TRUE} to
subsequently remove any feature groups that lost any suspect matches due to
other filter steps.
}}

\section{Slots}{

\describe{
\item{\code{screenInfo}}{A (\code{\link{data.table}}) with results from suspect
screening. This table will be amended with annotation data when
\code{annotateSuspects} is run.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{featureGroups}}}   \itemize{     \item{\strong{\code{\link{featureGroupsScreening}}}}   } }
}

\seealso{
\code{\link{featureGroups}}
}
