% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\docType{class}
\name{features-class}
\alias{features-class}
\alias{features}
\alias{length,features-method}
\alias{show,features-method}
\alias{featureTable,features-method}
\alias{analysisInfo,features-method}
\alias{analyses,features-method}
\alias{filter,features-method}
\alias{[,features,ANY,missing,missing-method}
\alias{[[,features,ANY,missing-method}
\alias{$,features-method}
\title{Base features class}
\usage{
\S4method{length}{features}(x)

\S4method{show}{features}(object)

\S4method{featureTable}{features}(obj)

\S4method{analysisInfo}{features}(obj)

\S4method{analyses}{features}(obj)

\S4method{filter}{features}(obj, intensityThreshold = NULL,
  retentionRange = NULL, mzRange = NULL, chromWidthRange = NULL,
  negate = FALSE)

\S4method{[}{features,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{features,ANY,missing}(x, i)

\S4method{$}{features}(x, name)
}
\arguments{
\item{obj, x, object}{\code{features} object to be accessed}

\item{intensityThreshold}{Minimum intensity of a feature. Set to \code{NULL}
to ignore.}

\item{retentionRange, mzRange, chromWidthRange}{Range of retention time (in
seconds), \emph{m/z} or chromatographic peak width (in seconds),
respectively. Features outside this range will be removed. Should be a
numeric vector with length of two containing the min/max values. If the max
value is set to a value below 0 then no maximum is assumed. Set to
\code{NULL} to skip this step.}

\item{negate}{If set to \code{TRUE} then filtering operations are performed
in opposite manner.}

\item{i}{A numeric or character value which is used to select analyses by
their index or name, respectively (for the order/names see \code{analyses()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all analyses are selected.\cr\cr For \code{[[}: should be a skalar value.}

\item{drop, j}{ignored.}

\item{name}{The analysis name (partially matched).}
}
\value{
\code{featureTable}: A \code{list} containing a
  \code{\link{data.table}} for each analysis with feature data

\code{analysisInfo}: A \code{data.frame} containing a column with
  analysis name (\code{analysis}), its path (\code{path}), and other columns
  such as replicate group name (\code{group}) and blank reference
  (\code{ref}).
}
\description{
Holds information for all features present within a set of analysis.
}
\details{
This class provides a way to store intensity, retention times, \emph{m/z} and
other data for all features in a set of analyses. The class is \code{virtual}
and derived objects are created by 'feature finders' such as
\code{findFeaturesOpenMS}, \code{findFeaturesXCMS} and
\code{findFeaturesBruker}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: Obtain total number of features.

\item \code{show}: Shows summary information for this object.

\item \code{featureTable}: Get table with feature information

\item \code{analysisInfo}: Get analysis information

\item \code{analyses}: returns a \code{character} vector with the names of the
analyses from which data is present in this object.

\item \code{filter}: Performs common rule based filtering of features.

\item \code{[}: Subset on analyses.

\item \code{[[}: Extract a feature table for an analysis.

\item \code{$}: Extract a feature table for an analysis.
}}

\section{Slots}{

\describe{
\item{\code{features}}{List of features per analysis file. Use the
\code{featureTable} method for access.}

\item{\code{analysisInfo}}{Analysis group information. Use the \code{analysisInfo} method
for access.}
}}

\seealso{
\code{\link{feature-finding}}
}
