% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R, R/features-bruker.R,
%   R/features-envipick.R, R/features-openms.R, R/features-safd.R,
%   R/features-xcms.R, R/features-xcms3.R
\docType{class}
\name{features-class}
\alias{features-class}
\alias{features}
\alias{length,features-method}
\alias{show,features-method}
\alias{featureTable,features-method}
\alias{analysisInfo,features-method}
\alias{analyses,features-method}
\alias{replicateGroups,features-method}
\alias{as.data.table,features-method}
\alias{filter,features-method}
\alias{[,features,ANY,missing,missing-method}
\alias{[[,features,ANY,missing-method}
\alias{$,features-method}
\alias{featuresBruker-class}
\alias{featuresBruker}
\alias{featuresEnviPick-class}
\alias{featuresEnviPick}
\alias{featuresOpenMS-class}
\alias{featuresOpenMS}
\alias{featuresSAFD-class}
\alias{featuresSAFD}
\alias{featuresXCMS-class}
\alias{featuresXCMS}
\alias{[,featuresXCMS,ANY,missing,missing-method}
\alias{filter,featuresXCMS-method}
\alias{featuresXCMS3-class}
\alias{featuresXCMS3}
\alias{[,featuresXCMS3,ANY,missing,missing-method}
\alias{filter,featuresXCMS3-method}
\title{Base features class}
\usage{
\S4method{length}{features}(x)

\S4method{show}{features}(object)

\S4method{featureTable}{features}(obj)

\S4method{analysisInfo}{features}(obj)

\S4method{analyses}{features}(obj)

\S4method{replicateGroups}{features}(obj)

\S4method{as.data.table}{features}(x)

\S4method{filter}{features}(
  obj,
  absMinIntensity = NULL,
  relMinIntensity = NULL,
  retentionRange = NULL,
  mzRange = NULL,
  mzDefectRange = NULL,
  chromWidthRange = NULL,
  negate = FALSE
)

\S4method{[}{features,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{features,ANY,missing}(x, i)

\S4method{$}{features}(x, name)

\S4method{[}{featuresXCMS,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{filter}{featuresXCMS}(obj, ...)

\S4method{[}{featuresXCMS3,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{filter}{featuresXCMS3}(obj, ...)
}
\arguments{
\item{obj, x, object}{\code{features} object to be accessed}

\item{absMinIntensity, relMinIntensity}{Minimum absolute/relative intensity
for features to be kept. The relative intensity is determined from the
feature with highest intensity (within the same analysis).
Set to \samp{0} or \code{NULL}
to skip this step.}

\item{retentionRange, mzRange, mzDefectRange, chromWidthRange}{Range of
retention time (in seconds), \emph{m/z}, mass defect (defined as the
decimal part of \emph{m/z} values) or chromatographic peak width (in
seconds), respectively. Features outside this range will be removed. Should
be a numeric vector with length of two containing the min/max values. The
maximum can be \code{Inf} to specify no maximum range. Set to \code{NULL}
to skip this step.}

\item{negate}{If set to \code{TRUE} then filtering operations are performed
in opposite manner.}

\item{i}{A numeric or character value which is used to select analyses by
their index or name, respectively (for the order/names see \code{analyses()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all analyses are selected.\cr\cr For \code{[[}: should be a scalar value.}

\item{\dots}{Ignored.}

\item{drop, j}{ignored.}

\item{name}{The analysis name (partially matched).}
}
\value{
\code{featureTable}: A \code{list} containing a
  \code{\link{data.table}} for each analysis with feature data

\code{analysisInfo}: A \code{data.frame} containing a column with
  analysis name (\code{analysis}), its path (\code{path}), and other columns
  such as replicate group name (\code{group}) and blank reference
  (\code{blank}).
}
\description{
Holds information for all features present within a set of analysis.
}
\details{
This class provides a way to store intensity, retention times, \emph{m/z} and
other data for all features in a set of analyses. The class is \code{virtual}
and derived objects are created by 'feature finders' such as
\code{findFeaturesOpenMS}, \code{findFeaturesXCMS} and
\code{findFeaturesBruker}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: Obtain total number of features.

\item \code{show}: Shows summary information for this object.

\item \code{featureTable}: Get table with feature information

\item \code{analysisInfo}: Get analysis information

\item \code{analyses}: returns a \code{character} vector with the names of the
analyses for which data is present in this object.

\item \code{replicateGroups}: returns a \code{character} vector with the names of the
replicate groups for which data is present in this object.

\item \code{as.data.table}: Returns all feature data in a table.

\item \code{filter}: Performs common rule based filtering of features. Note
that this (and much more) functionality is also provided by the
\code{filter} method defined for \code{\link{featureGroups}}. However,
filtering a \code{features} object may be useful to avoid grouping large
amounts of features.

\item \code{[}: Subset on analyses.

\item \code{[[}: Extract a feature table for an analysis.

\item \code{$}: Extract a feature table for an analysis.
}}

\section{Slots}{

\describe{
\item{\code{features}}{List of features per analysis file. Use the
\code{featureTable} method for access.}

\item{\code{analysisInfo}}{Analysis group information. Use the \code{analysisInfo} method
for access.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{workflowStep}}}   \itemize{     \item{\strong{\code{\link{features}}}}     \itemize{       \item{\code{\link{featuresFromFeatGroups}}}       \item{\code{\link{featuresConsensus}}}       \item{\code{\link{featuresBruker}}}       \item{\code{\link{featuresEnviPick}}}       \item{\code{\link{featuresOpenMS}}}       \item{\code{\link{featuresSAFD}}}       \item{\code{\link{featuresBrukerTASQ}}}       \item{\code{\link{featuresXCMS}}}       \item{\code{\link{featuresXCMS3}}}     }   } }
}

\seealso{
\code{\link{feature-finding}}
}
