% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R, R/formulas-set.R
\docType{class}
\name{formulas-class}
\alias{formulas-class}
\alias{formulas}
\alias{formulasConsensus-class}
\alias{formulasConsensus}
\alias{annotations,formulas-method}
\alias{analyses,formulas-method}
\alias{defaultExclNormScores,formulas-method}
\alias{show,formulas-method}
\alias{[[,formulas,ANY,ANY-method}
\alias{delete,formulas-method}
\alias{as.data.table,formulas-method}
\alias{annotatedPeakList,formulas-method}
\alias{plotSpectrum,formulas-method}
\alias{plotScores,formulas-method}
\alias{consensus,formulas-method}
\alias{formulasSet-class}
\alias{formulasSet}
\alias{formulasConsensusSet-class}
\alias{formulasConsensusSet}
\alias{show,formulasSet-method}
\alias{delete,formulasSet-method}
\alias{[,formulasSet,ANY,missing,missing-method}
\alias{filter,formulasSet-method}
\alias{plotSpectrum,formulasSet-method}
\alias{annotatedPeakList,formulasSet-method}
\alias{consensus,formulasSet-method}
\alias{formulasUnset-class}
\alias{formulasUnset}
\alias{unset,formulasSet-method}
\alias{unset,formulasConsensusSet-method}
\title{Formula annotations class}
\usage{
\S4method{annotations}{formulas}(obj, features = FALSE)

\S4method{analyses}{formulas}(obj)

\S4method{defaultExclNormScores}{formulas}(obj)

\S4method{show}{formulas}(object)

\S4method{[[}{formulas,ANY,ANY}(x, i, j)

\S4method{delete}{formulas}(obj, i = NULL, j = NULL, ...)

\S4method{as.data.table}{formulas}(
  x,
  fGroups = NULL,
  fragments = FALSE,
  countElements = NULL,
  countFragElements = NULL,
  OM = FALSE,
  normalizeScores = "none",
  excludeNormScores = defaultExclNormScores(x),
  average = FALSE
)

\S4method{annotatedPeakList}{formulas}(
  obj,
  index,
  groupName,
  analysis = NULL,
  MSPeakLists,
  onlyAnnotated = FALSE
)

\S4method{plotSpectrum}{formulas}(
  obj,
  index,
  groupName,
  analysis = NULL,
  MSPeakLists,
  title = NULL,
  specSimParams = getDefSpecSimParams(),
  mincex = 0.9,
  xlim = NULL,
  ylim = NULL,
  ...
)

\S4method{plotScores}{formulas}(
  obj,
  index,
  groupName,
  analysis = NULL,
  normalizeScores = "max",
  excludeNormScores = defaultExclNormScores(obj)
)

\S4method{consensus}{formulas}(
  obj,
  ...,
  absMinAbundance = NULL,
  relMinAbundance = NULL,
  uniqueFrom = NULL,
  uniqueOuter = FALSE,
  rankWeights = 1,
  labels = NULL
)

\S4method{show}{formulasSet}(object)

\S4method{delete}{formulasSet}(obj, i, j, ...)

\S4method{[}{formulasSet,ANY,missing,missing}(x, i, j, ..., sets = NULL, updateConsensus = FALSE, drop = TRUE)

\S4method{filter}{formulasSet}(obj, ..., sets = NULL, updateConsensus = FALSE, negate = FALSE)

\S4method{plotSpectrum}{formulasSet}(
  obj,
  index,
  groupName,
  analysis = NULL,
  MSPeakLists,
  title = NULL,
  specSimParams = getDefSpecSimParams(),
  mincex = 0.9,
  xlim = NULL,
  ylim = NULL,
  perSet = TRUE,
  mirror = TRUE,
  ...
)

\S4method{annotatedPeakList}{formulasSet}(obj, index, groupName, analysis = NULL, MSPeakLists, ...)

\S4method{consensus}{formulasSet}(
  obj,
  ...,
  absMinAbundance = NULL,
  relMinAbundance = NULL,
  uniqueFrom = NULL,
  uniqueOuter = FALSE,
  rankWeights = 1,
  labels = NULL,
  filterSets = FALSE,
  setThreshold = 0,
  setThresholdAnn = 0
)

\S4method{unset}{formulasSet}(obj, set)

\S4method{unset}{formulasConsensusSet}(obj, set)
}
\arguments{
\item{obj, x, object}{The \code{formulas} object.}

\item{features}{If \code{TRUE} returns formula data for features, otherwise
for feature groups.}

\item{i, j}{For \code{[[}: If both \code{i} and \code{j} are specified then \code{i} specifies the analysis and
\code{j} the feature group of the feature for which annotations should be returned. Otherwise \code{i} specifies
the feature group for which group annotations should be returned. \code{i}/\code{j} can be specified as
\code{integer} index or as a \code{character} name.

Otherwise passed to the \code{\link[=filter,featureAnnotations-method]{featureAnnotations}} method.}

\item{\dots}{For \code{plotSpectrum}: Further arguments passed to \code{\link[graphics]{plot}}.

  For \code{delete}: passed to the function specified as \code{j}.

  For \code{consensus}: Any further (and unique) \code{formulas} objects.

  \setsPassedArgs1{formulas}}

\item{fGroups, fragments, countElements, countFragElements, OM}{Passed to the
\code{\link[=as.data.table,featureAnnotations-method]{featureAnnotations}} method.}

\item{normalizeScores}{A \code{character} that specifies how normalization of
annotation scorings occurs. Either
\code{"none"} (no normalization),
\code{"max"} (normalize to max value) or \code{"minmax"} (perform min-max
normalization). Note that normalization of negative scores (e.g. output by
\command{SIRIUS}) is always performed as min-max. Furthermore, currently
normalization for \code{compounds} takes the original min/max scoring
values into account when candidates were generated. Thus, for
\code{compounds} scoring, normalization is not affected when candidate
results were removed after they were generated (\emph{e.g.} by use of
\code{filter}).}

\item{excludeNormScores}{A
  \code{character} vector specifying any compound scoring names that
  should \emph{not} be normalized. Set to \code{NULL} to normalize all
  scorings. Note that whether any normalization occurs is set by the
  \code{excludeNormScores} argument.

  For \code{compounds}: By default \code{score} and
  \code{individualMoNAScore} are set to mimic the behavior of the
  \command{MetFrag} web interface.}

\item{average}{If set to \code{TRUE} an 'average formula' is generated for each feature group by combining all
elements from all candidates and averaging their amounts. This obviously leads to non-existing formulae, however,
this data may be useful to deal with multiple candidate formulae per feature group when performing elemental
characterization. Setting this to \code{TRUE} disables reporting of most other data.}

\item{index}{The candidate index (row). For \code{plotSpectrum} two indices can be specified to compare spectra. In
this case \code{groupName} and \code{analysis} (if not \code{NULL}) should specify values for the spectra to
compare.}

\item{groupName}{The name of the feature group (or feature groups when comparing spectra) to which the candidate
belongs.}

\item{analysis}{A \code{character} specifying the analysis (or analyses when comparing spectra) for which the
annotated spectrum should be plotted. If \code{NULL} then annotation results for the complete feature group will be
plotted.}

\item{MSPeakLists}{The \code{\link{MSPeakLists}} object that was used to generate the candidate}

\item{onlyAnnotated}{Set to \code{TRUE} to filter out any peaks that could
not be annotated.}

\item{title}{The title of the plot. Set to \code{NULL} for an automatically generated title.}

\item{specSimParams}{A named \code{list} with parameters that influence the calculation of MS spectra similarities.
See the \link[=specSimParams]{spectral similarity parameters} documentation for more details.}

\item{mincex}{The formula annotation labels are automatically scaled. The \code{mincex} argument forces a minimum
\code{cex} value for readability.}

\item{xlim, ylim}{Sets the plot size limits used by
\code{\link[graphics]{plot}}. Set to \code{NULL} for automatic plot sizing.}

\item{absMinAbundance, relMinAbundance}{Minimum absolute or relative
(\samp{0-1}) abundance across objects for a result to be kept. For
instance, \code{relMinAbundance=0.5} means that a result should be present
in at least half of the number of compared objects. Set to \samp{NULL} to
ignore and keep all results. Limits cannot be set when \code{uniqueFrom} is
not \code{NULL}.}

\item{uniqueFrom}{Set this argument to only retain formulas that are unique
within one or more of the objects for which the consensus is made.
Selection is done by setting the value of \code{uniqueFrom} to a
\code{logical} (values are recycled), \code{numeric} (select by index) or a
\code{character} (as obtained with \code{algorithm(obj)}). For
\code{logical} and \code{numeric} values the order corresponds to the order
of the objects given for the consensus. Set to \code{NULL} to ignore.}

\item{uniqueOuter}{If \code{uniqueFrom} is not \code{NULL} and if
\code{uniqueOuter=TRUE}: only retain data that are also unique between
objects specified in \code{uniqueFrom}.}

\item{rankWeights}{A numeric vector with weights of to calculate the mean
ranking score for each candidate. The value will be re-cycled if necessary,
hence, the default value of \samp{1} means equal weights for all considered
objects.}

\item{labels}{A \code{character} with names to use for labelling. If \code{NULL} labels are automatically generated.}

\item{sets}{\setsWF A \code{character} with name(s) of the sets to keep (or remove if \code{negate=TRUE}). Note: if
\code{updateConsensus=FALSE} then the \code{setCoverage} column of the annotation results is not updated.}

\item{updateConsensus}{\setsWF If \code{TRUE} then the annonation consensus among set results is updated. See the
\verb{Sets workflows} section for more details.}

\item{drop}{Passed to the \code{\link[=filter,featureAnnotations-method]{featureAnnotations}} method.}

\item{negate}{Passed to the \code{\link[=filter,featureAnnotations-method]{featureAnnotations}} method.}

\item{perSet, mirror}{\setsWF If \code{perSet=TRUE} then the set specific mass peaks are annotated separately.
Furthermore, if \code{mirror=TRUE} (and there are two sets in the object) then a mirror plot is generated.}

\item{filterSets}{\setsWF Controls how algorithms concensus abundance filters are applied. See the \verb{Sets
workflows} section below.}

\item{setThreshold, setThresholdAnn}{\setsWF Thresholds used to create the annotation set consensus. See
\code{\link{generateFormulas}}.}

\item{set}{\setsWF The name of the set.}
}
\value{
\code{annotations} returns a \code{list} containing for each feature
  group (or feature if \code{features=TRUE}) a \code{\link{data.table}}
  with an overview of all generated formulae and other data such as candidate
  scoring and MS/MS fragments.

\code{consensus} returns a \code{formulas} object that is produced by
  merging results from multiple \code{formulas} objects.
}
\description{
Contains data of generated chemical formulae for given feature groups.
}
\details{
\code{formulas} objects are obtained from \link[=formula-generation]{formula generators}. This class is derived from
the \code{\link{featureAnnotations}} class, please see its documentation for more methods and other details.
}
\section{Methods (by generic)}{
\itemize{
\item \code{annotations}: Accessor method to obtain generated formulae.

\item \code{analyses}: returns a \code{character} vector with the names of the
analyses for which data is present in this object.

\item \code{defaultExclNormScores}: Returns default scorings that are excluded from normalization.

\item \code{show}: Show summary information for this object.

\item \code{[[}: Extracts a formula table, either for a feature group or for features in an analysis.

\item \code{as.data.table}: Generates a table with all candidate formulae for each feature group and other information such
as element counts.

\item \code{annotatedPeakList}: Returns an MS/MS peak list annotated with data from a
given candidate formula.

\item \code{plotSpectrum}: Plots an annotated spectrum for a given candidate formula of a feature or feature group. Two
spectra can be compared by specifying a two-sized vector for the \code{index}, \code{groupName} and (if desired)
\code{analysis} arguments.

\item \code{plotScores}: Plots a barplot with scoring of a candidate formula.

\item \code{consensus}: Generates a consensus of results from multiple
objects. In order to rank the consensus candidates, first
each of the candidates are scored based on their original ranking
(the scores are normalized and the highest ranked candidate gets value
\samp{1}). The (weighted) mean is then calculated for all scorings of each
candidate to derive the final ranking (if an object lacks the candidate its
score will be \samp{0}). The original rankings for each object is stored in
the \code{rank} columns.
}}

\section{Slots}{

\describe{
\item{\code{featureFormulas}}{A \code{list} with all generated formulae for each analysis/feature group. Use the
\code{annotations} method for access.}

\item{\code{setThreshold,setThresholdAnn}}{\setsWF A copy of the equally named arguments that were passed when this object
was created by \code{\link{generateFormulas}}.}

\item{\code{origFGNames}}{\setsWF The original (order of) names of the \code{\link{featureGroups}} object that was used to
create this object.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{featureAnnotations}}}   \itemize{     \item{\strong{\code{\link{formulas}}}}     \itemize{       \item{\code{\link{formulasConsensus}}}       \item{\code{\link{formulasSet}}}       \itemize{         \item{\code{\link{formulasConsensusSet}}}       }       \item{\code{\link{formulasUnset}}}     }   } }
}

\section{Source}{
 Subscripting of formulae for plots generated by
  \code{plotSpectrum} is based on the \code{chemistry2expression} function
  from the \href{https://github.com/schymane/ReSOLUTION}{ReSOLUTION} package.
}

\section{Sets workflows}{
 \setsWFClass{formulasSet}{formulas}

  \setsWFNewMethodsSO{formulasUnset}{Only the annotation results that are present in the specified set are kept
  (based on the set consensus, see below for implications).}

  \setsWFChangedMethods{

  \item \code{filter} and the subset operator (\code{[}) Can be used to select data that is only present for selected
  sets. Depending on the \code{updateConsenus}, both either operate on set consensus or original data (see below for
  implications).

  \item \code{annotatedPeakList} Returns a combined annotation table with all sets.

  \item \code{plotSpectrum} Is able to highlight set specific mass peaks (\code{perSet} and \code{mirror} arguments).

  \item \code{consensus} Creates the algorithm consensus based on the original annotation data (see below for
  implications). Then, like the sets workflow method for \code{\link{generateFormulas}}, a consensus is made for all
  sets, which can be controlled with the \code{setThreshold} and \code{setThresholdAnn} arguments. The candidate
  coverage among the different algorithms is calculated for each set (\emph{e.g.} \code{coverage-positive} column)
  and for all sets (\code{coverage} column), which is based on the presence of a candidate in all the algorithms from
  all sets data. The \code{consensus} method for sets workflow data supports the \code{filterSets} argument. This
  controls how the algorithm consensus abundance filters (\code{absMinAbundance}/\code{relMinAbundance}) are applied:
  if \code{filterSets=TRUE} then the minimum of all \code{coverage} set specific columns is used to obtain the
  algorithm abundance. Otherwise the overall \code{coverage} column is used. For instance, consider a consensus
  object to be generated from two objects generated by different algorithms (\emph{e.g.} \command{SIRIUS} and
  \command{GenForm}), which both have a positive and negative set. Then, if a candidate occurs with both
  algorithms for the positive mode set, but only with the first algorithm in the negative mode set,
  \code{relMinAbundance=1} will remove the candidate if \code{filterSets=TRUE} (because the minimum relative
  algorithm abundance is \samp{0.5}), while \code{filterSets=FALSE} will not remove the candidate (because based on
  all sets data the candidate occurs in both algorithms).

  

  }

  Two types of annotation data are stored in a \code{formulasSet} object: \enumerate{

  \item Annotations that are produced from a consensus between set results (see \code{generateFormulas}).

  \item The 'original' annotation data per set, prior to when the set consensus was made. This includes candidates
  that were filtered out because of the thresholds set by \code{setThreshold} and \code{setThresholdAnn}. However,
  when \code{filter} or subsetting (\code{[}) operations are performed, the original data is also updated.

  }

  In most cases the first data is used. However, in a few cases the original annotation data is used (as indicated
  above), for instance, to re-create the set consensus. It is important to realize that the original annotation data
  may have \emph{additional} candidates, and a newly created set consensus may therefore have 'new' candidates. For
  instance, when the object consists of the sets \code{"positive"} and \code{"negative"} and \code{setThreshold=1}
  was used to create it, then \code{formulas[, sets = "positive", updateConsensus = TRUE]} may now have additional
  candidates, \emph{i.e.} those that were not present in the \code{"negative"} set and were previously removed due to
  the consensus threshold filter.
}

\seealso{
The \code{\link{featureAnnotations}} base class for more relevant methods and \link{formula-generation}.
}
