% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\docType{class}
\name{formulas-class}
\alias{formulas-class}
\alias{formulas}
\alias{formulasConsensus-class}
\alias{formulasConsensus}
\alias{annotations,formulas-method}
\alias{analyses,formulas-method}
\alias{show,formulas-method}
\alias{[[,formulas,ANY,ANY-method}
\alias{as.data.table,formulas-method}
\alias{annotatedPeakList,formulas-method}
\alias{plotSpectrum,formulas-method}
\alias{plotScores,formulas-method}
\alias{consensus,formulas-method}
\title{Formula annotations class}
\usage{
\S4method{annotations}{formulas}(obj, features = FALSE)

\S4method{analyses}{formulas}(obj)

\S4method{show}{formulas}(object)

\S4method{[[}{formulas,ANY,ANY}(x, i, j)

\S4method{as.data.table}{formulas}(
  x,
  fGroups = NULL,
  fragments = FALSE,
  countElements = NULL,
  countFragElements = NULL,
  OM = FALSE,
  normalizeScores = "none",
  excludeNormScores = defaultExclNormScores(x),
  average = FALSE
)

\S4method{annotatedPeakList}{formulas}(
  obj,
  index,
  groupName,
  analysis = NULL,
  MSPeakLists,
  onlyAnnotated = FALSE
)

\S4method{plotSpectrum}{formulas}(
  obj,
  index,
  groupName,
  analysis = NULL,
  MSPeakLists,
  title = NULL,
  specSimParams = getDefSpecSimParams(),
  useGGPlot2 = FALSE,
  mincex = 0.9,
  xlim = NULL,
  ylim = NULL,
  ...
)

\S4method{plotScores}{formulas}(
  obj,
  index,
  groupName,
  analysis = NULL,
  normalizeScores = "max",
  excludeNormScores = defaultExclNormScores(obj),
  useGGPlot2 = FALSE
)

\S4method{consensus}{formulas}(
  obj,
  ...,
  absMinAbundance = NULL,
  relMinAbundance = NULL,
  uniqueFrom = NULL,
  uniqueOuter = FALSE,
  rankWeights = 1,
  labels = NULL
)
}
\arguments{
\item{obj, x, object}{The \code{formulas} object.}

\item{features}{If \code{TRUE} returns formula data for features, otherwise
for feature groups.}

\item{i, j}{If both \code{i} and \code{j} are specified then \code{i} specifies the analysis and \code{j} the feature
group of the feature for which annotations should be returned. Otherwise \code{i} specifies the feature group for
which group annotations should be returned. \code{i}/\code{j} can be specified as \code{integer} index or as a
\code{character} name.}

\item{fGroups, fragments, countElements, countFragElements, OM}{Passed to the
\code{\link[=as.data.table,featureAnnotations-method]{featureAnnotations}} method.}

\item{normalizeScores}{A \code{character} that specifies how normalization of
annotation scorings occurs. Either
\code{"none"} (no normalization),
\code{"max"} (normalize to max value) or \code{"minmax"} (perform min-max
normalization). Note that normalization of negative scores (e.g. output by
\command{SIRIUS}) is always performed as min-max. Furthermore, currently
normalization for \code{compounds} takes the original min/max scoring
values into account when candidates were generated. Thus, for
\code{compounds} scoring, normalization is not affected when candidate
results were removed after they were generated (\emph{e.g.} by use of
\code{filter}).}

\item{excludeNormScores}{A
  \code{character} vector specifying any compound scoring names that
  should \emph{not} be normalized. Set to \code{NULL} to normalize all
  scorings. Note that whether any normalization occurs is set by the
  \code{excludeNormScores} argument.

  For \code{compounds}: By default \code{score} and
  \code{individualMoNAScore} are set to mimic the behavior of the
  \command{MetFrag} web interface.}

\item{average}{If set to \code{TRUE} an 'average formula' is generated for each feature group by combining all
elements from all candidates and averaging their amounts. This obviously leads to non-existing formulae, however,
this data may be useful to deal with multiple candidate formulae per feature group when performing elemental
characterization. Setting this to \code{TRUE} disables reporting of most other data.}

\item{index}{The candidate index (row). For \code{plotSpectrum} two indices can be specified to compare spectra. In
this case \code{groupName} and \code{analysis} (if not \code{NULL}) should specify values for the spectra to
compare.}

\item{groupName}{The name of the feature group (or feature groups when comparing spectra) to which the candidate
belongs.}

\item{analysis}{A \code{character} specifying the analysis (or analyses when comparing spectra) for which the
annotated spectrum should be plotted. If \code{NULL} then annotation results for the complete feature group will be
plotted.}

\item{MSPeakLists}{The \code{\link{MSPeakLists}} object that was used to generate the candidate}

\item{onlyAnnotated}{Set to \code{TRUE} to filter out any peaks that could
not be annotated.}

\item{title}{The title of the plot. Set to \code{NULL} for an automatically generated title.}

\item{specSimParams}{A named \code{list} with parameters that influence the calculation of MS spectra similarities.
See the \link[=specSimParams]{spectral similarity parameters} documentation for more details.}

\item{useGGPlot2}{If \code{TRUE} then \pkg{\link{ggplot2}} is used for
plotting, otherwise base plot used. For \code{plotSpectrum}, \code{ggplot2}
allows nicely repelled text for annotation. However, base plot is generally
faster.}

\item{mincex}{The formula annotation labels are automatically scaled. The \code{mincex} argument forces a minimum
\code{cex} value for readability.}

\item{xlim, ylim}{Sets the plot size limits used by
\code{\link[graphics]{plot}}. Set to \code{NULL} for automatic plot sizing.}

\item{\dots}{For \code{plotSpectrum}: Further arguments passed to \code{\link[graphics]{plot}}.

  For \code{delete}: passed to the function specified as \code{j}.

  For \code{consensus}: Any further (and unique) \code{formulas} objects.}

\item{absMinAbundance, relMinAbundance}{Minimum absolute or relative
(\samp{0-1}) abundance across objects for a result to be kept. For
instance, \code{relMinAbundance=0.5} means that a result should be present
in at least half of the number of compared objects. Set to \samp{NULL} to
ignore and keep all results. Limits cannot be set when \code{uniqueFrom} is
not \code{NULL}.}

\item{uniqueFrom}{Set this argument to only retain formulas that are unique
within one or more of the objects for which the consensus is made.
Selection is done by setting the value of \code{uniqueFrom} to a
\code{logical} (values are recycled), \code{numeric} (select by index) or a
\code{character} (as obtained with \code{algorithm(obj)}). For
\code{logical} and \code{numeric} values the order corresponds to the order
of the objects given for the consensus. Set to \code{NULL} to ignore.}

\item{uniqueOuter}{If \code{uniqueFrom} is not \code{NULL} and if
\code{uniqueOuter=TRUE}: only retain data that are also unique between
objects specified in \code{uniqueFrom}.}

\item{rankWeights}{A numeric vector with weights of to calculate the mean
ranking score for each candidate. The value will be re-cycled if necessary,
hence, the default value of \samp{1} means equal weights for all considered
objects.}

\item{labels}{A \code{character} with names to use for labelling. If \code{NULL} labels are automatically generated.}
}
\value{
\code{annotations} returns a \code{list} containing for each feature
  group (or feature if \code{features=TRUE}) a \code{\link{data.table}}
  with an overview of all generated formulae and other data such as candidate
  scoring and MS/MS fragments.

\code{plotSpectrum} will return a \code{\link[=ggplot2]{ggplot object}} if \code{useGGPlot2} is \code{TRUE}.

\code{consensus} returns a \code{formulas} object that is produced by
  merging results from multiple \code{formulas} objects.
}
\description{
Contains data of generated chemical formulae for given feature groups.
}
\details{
\code{formulas} objects are obtained from \link[=formula-generation]{formula generators}. This class is derived from
the \code{\link{featureAnnotations}} class, please see its documentation for more methods and other details.
}
\section{Methods (by generic)}{
\itemize{
\item \code{annotations}: Accessor method to obtain generated formulae.

\item \code{analyses}: returns a \code{character} vector with the names of the
analyses for which data is present in this object.

\item \code{show}: Show summary information for this object.

\item \code{[[}: Extracts a formula table, either for a feature group or for features in an analysis.

\item \code{as.data.table}: Generates a table with all candidate formulae for each feature group and other information such
as element counts.

\item \code{annotatedPeakList}: Returns an MS/MS peak list annotated with data from a
given candidate formula.

\item \code{plotSpectrum}: Plots an annotated spectrum for a given candidate formula of a feature or feature group. Two
spectra can be compared by specifying a two-sized vector for the \code{index}, \code{groupName} and (if desired)
\code{analysis} arguments.

\item \code{plotScores}: Plots a barplot with scoring of a candidate formula.

\item \code{consensus}: Generates a consensus of results from multiple
objects. In order to rank the consensus candidates, first
each of the candidates are scored based on their original ranking
(the scores are normalized and the highest ranked candidate gets value
\samp{1}). The (weighted) mean is then calculated for all scorings of each
candidate to derive the final ranking (if an object lacks the candidate its
score will be \samp{0}). The original rankings for each object is stored in
the \code{rank} columns.
}}

\section{Slots}{

\describe{
\item{\code{featureFormulas}}{A \code{list} with all generated formulae for each analysis/feature group. Use the
\code{annotations} method for access.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{featureAnnotations}}}   \itemize{     \item{\strong{\code{\link{formulas}}}}     \itemize{       \item{\code{\link{formulasConsensus}}}       \item{\code{\link{formulasSet}}}       \itemize{         \item{\code{\link{formulasConsensusSet}}}       }       \item{\code{\link{formulasUnset}}}     }   } }
}

\section{Source}{
 Subscripting of formulae for plots generated by
  \code{plotSpectrum} is based on the \code{chemistry2expression} function
  from the \href{https://github.com/schymane/ReSOLUTION}{ReSOLUTION} package.
}

\seealso{
The \code{\link{featureAnnotations}} base class for more relevant methods and \link{formula-generation}.
}
