% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\docType{class}
\name{formulas-class}
\alias{formulas-class}
\alias{formulas}
\alias{formulaTable,formulas-method}
\alias{algorithm,formulas-method}
\alias{analyses,formulas-method}
\alias{groupNames,formulas-method}
\alias{length,formulas-method}
\alias{show,formulas-method}
\alias{[,formulas,ANY,missing,missing-method}
\alias{[[,formulas,ANY,ANY-method}
\alias{$,formulas-method}
\alias{makeTable,formulas-method}
\alias{plotSpec,formulas-method}
\alias{consensus,formulas-method}
\title{Formula lists class}
\usage{
\S4method{formulaTable}{formulas}(obj)

\S4method{algorithm}{formulas}(obj)

\S4method{analyses}{formulas}(obj)

\S4method{groupNames}{formulas}(obj)

\S4method{length}{formulas}(x)

\S4method{show}{formulas}(object)

\S4method{[}{formulas,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{formulas,ANY,ANY}(x, i, j)

\S4method{$}{formulas}(x, name)

\S4method{makeTable}{formulas}(obj, fGroups = NULL, average = FALSE,
  countElements = NULL, countFragElements = NULL, OM = FALSE,
  maxFormulas = NULL, maxFragFormulas = NULL)

\S4method{plotSpec}{formulas}(obj, precursor, groupName, analysis = NULL,
  MSPeakLists, useGGPlot2 = FALSE)

\S4method{consensus}{formulas}(obj, ..., formThreshold = 0)
}
\arguments{
\item{obj, x, object, formulas}{The \code{formulas} object.}

\item{i, j}{A numeric or character value which is used to select analyses/feature groups by
their index or name, respectively (for the order/names see \code{analyses()/groupNames()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all analyses/feature groups are selected.\cr\cr For \code{[[}: should be a skalar value. If \code{j} is not specified, \code{i} selects by feature groups instead.}

\item{\dots}{Any further \code{formulas} objects on top of the object given
for the \code{obj} parameter to which a consensus should be generated.}

\item{drop}{ignored.}

\item{fGroups}{The \code{\link{featureGroups}} object that was used to
generate this \code{formulas} object. If not \code{NULL} it is used to add
feature group information (retention and \emph{m/z} values).}

\item{countElements, countFragElements}{A \code{character} vector with
elements that should be counted for each MS(/MS) formula candidate. For
instance, \code{c("C", "H")} adds columns for both carbon and hydrogen
amounts in each formula. Set to \code{NULL} to not count any elements.}

\item{OM}{If set to \code{TRUE} several columns with information relevant for
organic matter (OM) characterization will be added (e.g. elemental ratios,
classification). This will also make sure that \code{countElements}
contains at least C, H, N, O, P and S.}

\item{maxFormulas, maxFragFormulas}{Maximum amount of unique candidate
formulae (or fragment formulae) per feature group. Set to \code{NULL} to
ignore.}

\item{precursor}{The formula of the precursor (in ionic form, \emph{i.e.} as
detected by the MS).}

\item{groupName}{The name of the feature group to which the candidate
belongs.}

\item{analysis}{A \code{character} specifying the analysis for which the
annotated spectrum should be plotted. If \code{NULL} then annotation
results for the complete feature group will be plotted.}

\item{MSPeakLists}{The \code{\link{MSPeakLists}} object that was used to
generate the candidate}

\item{useGGPlot2}{If \code{TRUE} then \pkg{\link{ggplot2}} is used for
plotting, otherwise base plot used. For \code{plotSpec}, \code{ggplot2}
allows nicely repelled text for annotation. However, base plot is generally
faster.}

\item{formThreshold}{Fractional minimum amount (0-1) to
which a generated precursor formula should be present in all given formula objects
(\code{formListThreshold}) containing the related feature group. For
instance, a value of \samp{0.5} for \code{formAnaThreshold} means that a
particular formula should be present in at least \samp{50\%} of all
analyses containing the feature group that was used to generate the
formula.}
}
\value{
\code{formulas} returns a \code{list} containing for each analysis
  and each feature group a \code{\link{data.table}} with an overview of all
  generated formulae and other data such as candidate scoring and MS/MS
  fragments.

\code{makeTable} returns a \code{\link{data.table}}.

\code{plotSpec} will return a \code{\link[=ggplot2]{ggplot object}}
  if \code{useGGPlot2} is \code{TRUE}.

\code{consensus} returns a \code{formulas} object that is produced
  by merging multiple \code{formulas} objects.
}
\description{
Contains data of generated chemical formulae for given feature groups.
}
\details{
\code{formulas} objects are obtained from \link[=formula-generation]{formula
generators}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{formulaTable}: Accessor method to obtain generated formulae.

\item \code{algorithm}: Accessor method for the algorithm (a character
string) used to generate formulae.

\item \code{analyses}: returns a \code{character} vector with the names of the
analyses for which data is present in this object.

\item \code{groupNames}: returns a \code{character} vector with the names of the
feature groups for which data is present in this object.

\item \code{length}: Obtain total number of formulae entries.

\item \code{show}: Show summary information for this object.

\item \code{[}: Subset on feature groups.

\item \code{[[}: Extract a formula table. If both arguments (\code{i} and
\code{j}) are specified, the feature specific formula table belonging to
the analysis (\code{i})/feature group (\code{j}) is returned. Otherwise the
formula table for the feature group specified by \code{j} is returned.

\item \code{$}: Extract a formula table for a feature group.

\item \code{makeTable}: Generates a table with all candidate formulae for each
feature group and other information such as element counts.

\item \code{plotSpec}: Plots an annotated spectrum for a given candidate
formula of a feature or feature group.

\item \code{consensus}: Generates a consensus and other summarizing data for
formulae of feature groups that are present in multiple analyses.
}}

\section{Slots}{

\describe{
\item{\code{formulas}}{Lists of all generated formulae. Use the \code{formulaTable}
method for access.}

\item{\code{algorithm}}{The algorithm that was used for generation of formulae. Use
the \code{algorithm} method for access.}
}}

