% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\docType{class}
\name{formulas-class}
\alias{formulas-class}
\alias{formulas}
\alias{formulaTable,formulas-method}
\alias{algorithm,formulas-method}
\alias{analyses,formulas-method}
\alias{groupNames,formulas-method}
\alias{length,formulas-method}
\alias{show,formulas-method}
\alias{[,formulas,ANY,ANY,missing-method}
\alias{[[,formulas,ANY,ANY-method}
\alias{consensus,formulas-method}
\title{Formula lists class}
\usage{
\S4method{formulaTable}{formulas}(obj)

\S4method{algorithm}{formulas}(obj)

\S4method{analyses}{formulas}(obj)

\S4method{groupNames}{formulas}(obj)

\S4method{length}{formulas}(x)

\S4method{show}{formulas}(object)

\S4method{[}{formulas,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{formulas,ANY,ANY}(x, i, j)

\S4method{consensus}{formulas}(obj, ..., fGroups,
  formAnaThreshold = 0.75, formListThreshold = 0, maxFormulas = 10,
  maxFragFormulas = 10, minIntensity = NULL, maxIntensity = NULL,
  minPreferredFormulas = 1, minPreferredIntensity = NULL,
  elements = c(), fragElements = c())
}
\arguments{
\item{obj, x, object, formulas}{The \code{formulas} object.}

\item{i, j}{A numeric or character value which is used to select analyses/feature groups by
their index or name, respectively (for the order/names see \code{analyses()/groupNames()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all analyses/feature groups are selected.\cr\cr For \code{[[}: should be a skalar value.}

\item{\dots}{Any further \code{formulas} objects on top of the object given
for the \code{obj} parameter to which a consensus should be generated.
Using the \dots parameter formulae from various
\link[=formula-generation]{formula generators} can be merged.}

\item{drop}{ignored.}

\item{fGroups}{The \code{\link{featureGroups}} object that was used when
formulae were generated.}

\item{formAnaThreshold, formListThreshold}{Fractional minimum amount (0-1) to
which a generated formula should be present in all the analyses
(\code{formAnaThreshold}) or in all given formula lists
(\code{formListThreshold}) containing the related feature group. For
instance, a value of \samp{0.5} for \code{formAnaThreshold} means that a
particular formula should be present in at least \samp{50\%} of all
analyses containing the feature group that was used to generate the
formula.}

\item{maxFormulas, maxFragFormulas}{Maximum amount of candidate formulae (or
fragment formulae) per feature group.}

\item{minIntensity}{Results from features below this intensity will be
excluded.}

\item{maxIntensity}{Results from features above this intensity will be
excluded.}

\item{minPreferredFormulas}{Preferred minimum amount of candidate formulae
per feature group. See \code{minPreferredIntensity}.}

\item{minPreferredIntensity}{Minimum preferred intensity of formulae
candidates per feature group. Results from features below this intensity
will be removed unless the total count of formulae of the related feature
group is below \code{minPreferredFormulas}. When a higher
\code{minIntensity} is given this parameter has no effect.}

\item{elements}{An optional character vector containing elements for which
the total count present in each formula should be calculated. For instance,
\code{c("C", "H")} will count all carbon and hydrogen atoms present in each
formula.}

\item{fragElements}{Same as \code{elements} parameter, but for MS/MS
formulae.}
}
\value{
\code{formulas} returns a \code{list} containing for each analysis
  and each feature group a \code{\link{data.table}} with an overview of all
  generated formulae and other data such as candidate scoring and MS/MS
  fragments.

\code{consensus} returns a \code{\link{formulaConsensus}} object.
}
\description{
Contains data of generated chemical formulae for given feature groups.
}
\details{
\code{formulas} objects are obtained from \link[=formula-generation]{formula
generators}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{formulaTable}: Accessor method to obtain generated formulae.

\item \code{algorithm}: Accessor method for the algorithm (a character
string) used to generate formulae.

\item \code{analyses}: returns a \code{character} vector with the names of the
analyses for which data is present in this object.

\item \code{groupNames}: returns a \code{character} vector with the names of the
feature groups for which data is present in this object.

\item \code{length}: Obtain total number of formulae entries.

\item \code{show}: Show summary information for this object.

\item \code{[}: Subset on analyses/feature groups.

\item \code{[[}: Extract a formula table from a specified analysis/feature group.

\item \code{consensus}: Generates a consensus and other summarizing data for
formulae of feature groups that are present in multiple analyses.
}}

\section{Slots}{

\describe{
\item{\code{formulas}}{Lists of all generated formulae. Use the \code{formulaTable}
method for access.}

\item{\code{algorithm}}{The algorithm that was used for generation of formulae. Use
the \code{algorithm} method for access.}
}}

\seealso{
\code{\link{formulaConsensus}}
}
