% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/components.R
\name{generateComponents}
\alias{generateComponents}
\alias{generateComponents,featureGroups-method}
\title{Grouping feature groups in components}
\usage{
generateComponents(fGroups, algorithm, ...)

\S4method{generateComponents}{featureGroups}(fGroups, algorithm, ...)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which components should be generated.}

\item{algorithm}{A character string describing the algorithm that should be
used: \code{"ramclustr"}, \code{"camera"}, \code{"nontarget"}, \code{"intclust"}, \code{"openms"}, \code{"cliquems"}, \code{"specclust"}, \code{"tp"}}

\item{\dots}{Any parameters to be passed to the selected component generation algorithm.}
}
\value{
A \code{\link{components}} (derived) object containing all generated components.
}
\description{
Functionality to automatically group related feature groups (\emph{e.g.} isotopes, adducts and homologues) to assist
and simplify annotation.
}
\details{
Several algorithms are provided to group feature groups that are related in some (chemical) way to each other. How
feature groups are related depends on the algorithm: examples include adducts, statistics and parents/transformation
products. The linking of this data is generally useful for annotation purposes and reducing data complexity.

\code{generateComponents} is a generic function that will generateComponents by one of the supported algorithms. The actual
  functionality is provided by algorithm specific functions such as \code{generateComponentsRAMClustR} and \code{generateComponentsNontarget}. While these
  functions may be called directly, \code{generateComponents} provides a generic interface and is therefore usually preferred.
}
\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} the componentization data is generated differently
  depending on the used algorithm. Please see the details in the algorithm specific functions linked in the \verb{See Also} section.
}

\seealso{
The \code{\link{components}} output class and its methods and the algorithm specific functions:
  \code{\link{generateComponentsRAMClustR}}, \code{\link{generateComponentsCAMERA}}, \code{\link{generateComponentsNontarget}}, \code{\link{generateComponentsIntClust}}, \code{\link{generateComponentsOpenMS}}, \code{\link{generateComponentsCliqueMS}}, \code{\link{generateComponentsSpecClust}}, \code{\link{generateComponentsTPs}}
}
