% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/components-intclust.R
\name{generateComponentsIntClust}
\alias{generateComponentsIntClust}
\alias{generateComponentsIntClust,featureGroups-method}
\title{Generate components based on intensity profiles}
\usage{
generateComponentsIntClust(fGroups, ...)

\S4method{generateComponentsIntClust}{featureGroups}(
  fGroups,
  method = "complete",
  metric = "euclidean",
  normalized = TRUE,
  average = TRUE,
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which components should be generated.}

\item{...}{Any parameters to be passed to the selected component generation algorithm.}

\item{method}{Clustering method that should be applied (passed to
\code{\link[fastcluster:hclust]{fastcluster::hclust}}).}

\item{metric}{Distance metric used to calculate the distance matrix (passed to \code{\link{daisy}}).}

\item{normalized, average}{Passed to \code{\link[=as.data.table,featureGroups-method]{as.data.table}} to perform
normalization and averaging of data.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}
}
\value{
The components are stored in objects derived from \code{\link{componentsIntClust}}.
}
\description{
Generates components based on intensity profiles of feature groups.
}
\details{
This function uses hierarchical clustering of intensity profiles to generate components. This function is called when calling \code{generateComponents} with
  \code{algorithm="intclust"}.

Hierarchical clustering is performed on normalized (and optionally replicate averaged) intensity data and
  the resulting dendrogram is automatically cut with \code{\link{cutreeDynamicTree}}. The distance matrix is
  calculated with \code{\link{daisy}} and clustering is performed with
  \code{\link[fastcluster:hclust]{fastcluster::hclust}}. The clustering of the resulting components can be further
  visualized and modified using the methods defined for \code{\link{componentsIntClust}}.
}
\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} normalization of feature intensities occur per
  set.
}

\references{
\addCitations{fastcluster}{1}

\insertRef{Scholle2018}{patRoon}
}
\seealso{
\code{\link{generateComponents}} for more details and other algorithms.
}
