% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/components-net.R
\name{generateComponentsNet}
\alias{generateComponentsNet}
\alias{generateComponentsNet,featureGroups-method}
\alias{generateComponentsNet,featureGroupsSet-method}
\title{Network-based componentization}
\source{
The componentization approach was inspired by \pkg{CAMERA} and \pkg{cliqueMS}.
}
\usage{
generateComponentsNet(fGroups, ...)

\S4method{generateComponentsNet}{featureGroups}(
  fGroups,
  ionization = NULL,
  minSize = 2,
  mzWindow = defaultLim("mz", "medium"),
  componSim = "pearson",
  componMinSim = 0.95,
  componMaxP = 0.05,
  componMethod = "community",
  componArgs = list(),
  groupClust = "complete",
  groupClustH = 0.5,
  annotAlgo = "imss",
  annotAdducts = c("[M+H]+", "[M+Na]+", "[M+K]+", "[M+NH4]+", "[M-H]-", "[M-H2O-H]-"),
  annotPrefAdducts = c("[M+H]+", "[M-H]-"),
  annotArgs = list()
)

\S4method{generateComponentsNet}{featureGroupsSet}(fGroups, ionization = NULL, ...)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which components should be generated.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{ionization}{Which ionization polarity was used to generate the data: should be \code{"positive"}
  or \code{"negative"}. If the \code{featureGroups} object has adduct annotations, and \code{ionization=NULL}, the
  ionization will be detected automatically.

  \setsWF This parameter is not supported for sets workflows, as the ionization will always be detected
  automatically.}

\item{minSize}{The minimum size of a component. Smaller components than this size will be removed. See note below.}

\item{componSim}{The similarity metric to use: \code{"pearson"} for Pearson correlation or \code{"cosine"} for cosine
similarity.}

\item{componMinSim}{The minimum similarity threshold (\samp{0-1}). Feature pairs with a lower similarity are not
connected in the network.}

\item{componMaxP}{The maximum \emph{p}-value threshold (\samp{0-1}) for Pearson correlation. Only applicable when
\code{componSim="pearson"}. Feature pairs with a higher \emph{p}-value are not connected.}

\item{componMethod}{The network componentization method: \code{"community"} for community detection with
\CRANpkg{igraph}, \code{"cliques"} for maximal cliques (using
\code{\link[igraph:max_cliques]{igraph::max_cliques}}), \code{"hcs"} for highly connected subgraphs (using
\code{\link[RBGL:highlyConnSG]{RBGL::highlyConnSG}}), or \code{"hclust"} for hierarchical clustering (using
\code{\link[fastcluster:hclust]{fastcluster::hclust}}).}

\item{componArgs}{A \code{list} with additional arguments passed to the network componentization function.

  For \code{componMethod="community"}, arguments are passed to the \CRANpkg{igraph} clustering function. The default
  function is \code{\link[igraph:cluster_walktrap]{igraph::cluster_walktrap}} and can be changed by setting the
  \code{func} argument.

  For \code{componMethod="cliques"}: arguments are passed to \code{\link[igraph:max_cliques]{igraph::max_cliques}}.

  For \code{componMethod="hcs"}: arguments are passed to \code{\link[RBGL:highlyConnSG]{RBGL::highlyConnSG}}.

  For \code{componMethod="hclust"}: arguments \code{method} to set the clustering method (see
  \code{\link[fastcluster:hclust]{fastcluster::hclust}}) (\code{"complete" by default}) and \code{h} to set the
  clustering height threshold (\samp{0.95} by default).}

\item{groupClust, groupClustH}{Clustering method (see \code{\link[fastcluster:hclust]{fastcluster::hclust}}) and
height at which to cut the tree for the consensus clustering step across analyses.}

\item{annotAlgo}{Annotation algorithm: \code{"imss"} for annotation using
\code{\link[InterpretMSSpectrum:findMAIN]{InterpretMSSpectrum::findMAIN}} or \code{"nontarget"} for annotation
using \code{\link[nontarget:pattern.search]{nontarget::pattern.search}} and
\code{\link[nontarget:adduct.search]{nontarget::adduct.search}}.}

\item{annotAdducts}{A \code{character} vector or \code{adduct} object vector with adducts to use for annotation. At
least two adducts are required. Adducts incompatible with the ionization mode are automatically removed.}

\item{annotPrefAdducts}{A \code{character} vector with preferential adducts. Only relevant if
\code{annotAlgo="nontarget"}, see the \verb{nontarget annotation} section.}

\item{annotArgs}{A \code{list} with additional arguments passed to the annotation function. For
\code{annotAlgo="imss"}, arguments are passed to
\code{\link[InterpretMSSpectrum:findMAIN]{InterpretMSSpectrum::findMAIN}}. For \code{annotAlgo="nontarget"}, the
list should contain \code{pattern} and \code{adduct} elements, which are themselves lists with arguments passed to
\code{\link[nontarget:pattern.search]{nontarget::pattern.search}} and
\code{\link[nontarget:adduct.search]{nontarget::adduct.search}}, respectively.}
}
\value{
A \code{\link{componentsNet}} object (derived from \code{\link{components}}). The
  \code{\link{componentTable}} of each component contains the following columns:
  \describe{
  \item{\code{group}}{The feature group name.}
  \item{\code{ret}}{The retention time.}
  \item{\code{mz}}{The mass-to-charge ratio.}
  \item{\code{degreeMin}, \code{degreeMax}, \code{degreeMean}}{The minimum, maximum and mean normalized degree of
    the feature in the network across all analyses.}
  \item{\code{corMin}, \code{corMax}, \code{corMean}}{The minimum, maximum and mean correlation (or similarity) of
    the feature with other features in the component across all analyses.}
  \item{\code{intensity}}{The mean intensity of the feature group across analyses where it was detected.}
  \item{\code{intensity_rel}}{The relative intensity within the component (\samp{0-1}).}
  }

  When \code{annotAlgo="imss"}, the following annotation columns are added:
  \describe{
  \item{\code{isogroup}}{The isotope group.}
  \item{\code{isonr}}{The isotope number within the group.}
  \item{\code{charge}}{The charge.}
  \item{\code{adduct_ion}}{The assigned adduct.}
  \item{\code{ppm}}{The mass error in ppm.}
  \item{\code{neutralMass}}{The calculated neutral mass.}
  }

  When \code{annotAlgo="nontarget"}, the output from \pkg{nontarget}'s \code{pattern.search} and \code{adduct.search}
  is significantly restructured to simplify interpretation. Isotope and adduct information is
  collapsed to one row per feature, the "best" grouping is selected in case of conflicts (see the \strong{nontarget
  annotation} section), and links to related features are expressed as feature group names. The following annotation
  columns are added:
  \describe{
  \item{\code{isogroup}}{The isotope group ID.}
  \item{\code{iso_interaction}}{The isotope interaction level(s), i.e. the distance from the monoisotope. Multiple
    values are slash-separated.}
  \item{\code{isotope}}{The isotope label(s) (e.g. \code{"13C"}). \code{"mono"} indicates a monoisotope peak.
    Multiple values are slash-separated.}
  \item{\code{iso_mz_tol}}{The mass tolerance(s) for the isotope assignment. Multiple values are slash-separated.}
  \item{\code{charge}}{The charge level.}
  \item{\code{iso_link}}{The feature group name(s) of the origin (monoisotope) peak(s) that this feature is an
    isotope of. \code{NA} for monoisotope peaks or features not part of an isotope pattern.}
  \item{\code{addgroup}}{The adduct group ID, grouping features that share the same neutral mass.}
  \item{\code{adduct_ion}}{The assigned adduct.}
  \item{\code{neutralMass}}{The calculated neutral mass.}
  \item{\code{add_link}}{The feature group name(s) of the peak(s) linked to this feature via an adduct
    relationship.}
  \item{\code{add_link_adduct_ion}}{The adduct(s) of the linked peak(s).}
  \item{\code{add_link_mz_tol}}{The mass tolerance(s) for the adduct assignment.}
  }

  The \code{\link{componentInfo}} table contains the columns \code{name} (component name), \code{cmp_ret} (mean
  retention time), \code{cmp_retsd} (retention time standard deviation), \code{neutral_mass} (mean neutral mass) and
  \code{size} (number of features).
}
\description{
Uses networking to generate components by grouping features based on similarity of their chromatographic elution
profiles.
}
\details{
This function uses network to generate components. This function is called when calling \code{generateComponents} with
  \code{algorithm="network"}.

Features are first grouped per analysis into feature components by constructing a network where nodes
  represent features and edges represent similarity between their elution profiles. Similarity is calculated using
  either Pearson correlation or cosine similarity. The network is then clustered using one of several methods (see
  \code{componMethod} argument). The resulting feature components are then merged across analyses using a consensus
  clustering approach, based on the pairwise presence similarity across analyses. Finally, components are annotated
  using either the \pkg{InterpretMSSpectrum} or \pkg{nontarget} package to identify isotopes, in-source fragments and
  adducts.

  Compared to other componentization algorithms supported in \pkg{patRoon}, \code{generateComponentsNet()} \itemize{

    \item does not need rely on \pkg{XCMS} and uses \link{msdata} for EIC extraction, which avoids the need of object
    conversions and is generally faster.

    \item feature based: this is more sensible where samples are not expected to be very similar, e.g. in
    environmental studies.

    \item uses a different approach to convert feature components across analyses into consensus components, which
    should (hopefully) perform better than other feature-based algorithms (\code{\link{generateComponentsOpenMS}} and
    \code{\link{generateComponentsCliqueMS}})).
    
    \item supports different and configurable componentization methods (see \code{componMethod} argument).
    
    \item supports isotope annotation that is not based on 13C differences thanks to the \pkg{nontarget} package, eg
    useful for halogenated compounds.

  }
  
  It is recommended to use the \code{\link[=plotGraph,componentsNet]{plotGraph()}} method to evaluate the feature
  componentization, e.g. to evaluate different methods (\code{componMethod} argument) or relevant parameters.
}
\note{
The \code{generateComponentsNet} function is still experimental. Any feedback is welcome!
}
\section{\code{nontarget} annotation}{
 When \code{annotAlgo="nontarget"}, annotation is performed using
  \pkg{nontarget}'s \code{\link[nontarget:pattern.search]{pattern.search}} and
  \code{\link[nontarget:adduct.search]{adduct.search}} functions.

  \strong{Isotope annotation}: \code{pattern.search} identifies isotope patterns. Peaks are grouped by isotope
  patterns and charge levels. When multiple charge levels are found for the same isotope group, the "best" grouping
  is selected using the following priority: (1) groups containing a 13C isotope, (2) the largest group, or (3) the
  lowest charge level.

  \strong{Adduct annotation}: \code{adduct.search} identifies adduct relationships between peaks. Adducts are grouped
  by their calculated neutral mass. If there is a conflict in neutral mass assignment, the "best" adduct group is
  selected using the following priority: (1) groups containing preferred adducts (see \code{annotPrefAdducts}), (2)
  the largest group, or (3) the first group encountered.
}

\section{IMS workflows}{
 The componentization algorithm is not aware of the IMS dimension. For this reason, no
  IMS feature groups will be considered for componentization, and direct IMS workflows (see
  \code{\link[=assignMobilities_feat]{assignMobilitities}}) are currently not supported.
}

\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} the componentization is first performed for each
  set independently. The resulting components are then all combined in a \code{\link{componentsNetSet}} object. Note that
  the components themselves are never merged. The components are renamed to include the set name from which they were
  generated (\emph{e.g.} \code{"CMP1"} becomes \code{"CMP1-positive"}).
}

\references{
\addCitations{CAMERA} \cr\cr \insertRef{Senan2019}{patRoon} \cr\cr \insertRef{Loos2017}{patRoon} \cr\cr
  \addCitations{enviPat} \cr\cr \insertRef{Jaeger2016}{patRoon} \cr\cr \insertRef{Jaeger2017}{patRoon} \cr\cr
  \addCitations{igraph} \cr\cr \addCitations{fastcluster}
}
\seealso{
\code{\link{generateComponents}} for more details and other algorithms.

\code{\link{componentsNet}}
}
