% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/components-specclust.R
\name{generateComponentsSpecClust}
\alias{generateComponentsSpecClust}
\alias{generateComponentsSpecClust,featureGroups-method}
\title{Generate components based on MS/MS similarity}
\usage{
generateComponentsSpecClust(fGroups, ...)

\S4method{generateComponentsSpecClust}{featureGroups}(
  fGroups,
  MSPeakLists,
  method = "complete",
  specSimParams = getDefSpecSimParams(),
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which components should be generated.}

\item{...}{Any parameters to be passed to the selected component generation algorithm.}

\item{MSPeakLists}{The \code{\link{MSPeakLists}} object for the given feature groups that should be used for MS
spectral similarity calculations.}

\item{method}{Clustering method that should be applied (passed to
\code{\link[fastcluster:hclust]{fastcluster::hclust}}).}

\item{specSimParams}{A named \code{list} with parameters that influence the calculation of MS spectra similarities.
See the \link[=specSimParams]{spectral similarity parameters} documentation for more details.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}
}
\value{
The components are stored in objects derived from \code{\link{componentsSpecClust}}.
}
\description{
Generates components based on MS/MS similarity between feature groups.
}
\details{
This function uses hierarchical clustering of MS/MS spectra to generate components. This function is called when calling \code{generateComponents} with
  \code{algorithm="specclust"}.

The similarities are converted to a distance matrix and used as input for hierarchical clustering, and the
  resulting dendrogram is automatically cut with \code{\link{cutreeDynamicTree}}. The clustering is performed with
  \code{\link[fastcluster:hclust]{fastcluster::hclust}}.
}
\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} the spectral similarities for each set are
  combined as is described for the \code{\link[=spectrumSimilarity,MSPeakListsSet-method]{spectrumSimilarity}} method
  for sets workflows.
}

\references{
\addCitations{fastcluster}{1}
}
\seealso{
\code{\link{generateComponents}} for more details and other algorithms.
}
\author{
Rick Helmus <\email{r.helmus@uva.nl}> and Bas van de Velde (major contributions to spectral binning and
  similarity calculation).
}
