% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-tps.R
\name{generateComponentsTPs}
\alias{generateComponentsTPs}
\alias{generateComponentsTPs,featureGroups-method}
\alias{generateComponentsTPs,featureGroupsSet-method}
\title{Generate components of transformation products}
\usage{
\S4method{generateComponentsTPs}{featureGroups}(
  fGroups,
  fGroupsTPs = fGroups,
  ignoreParents = FALSE,
  TPs = NULL,
  MSPeakLists = NULL,
  formulas = NULL,
  compounds = NULL,
  minRTDiff = 20,
  specSimParams = getDefSpecSimParams()
)

\S4method{generateComponentsTPs}{featureGroupsSet}(
  fGroups,
  fGroupsTPs = fGroups,
  ignoreParents = FALSE,
  TPs = NULL,
  MSPeakLists = NULL,
  formulas = NULL,
  compounds = NULL,
  minRTDiff = 20,
  specSimParams = getDefSpecSimParams()
)
}
\arguments{
\item{fGroups}{The input \code{\link{featureGroups}} for componentization. See \code{fGroupsTPs}.}

\item{fGroupsTPs}{A \code{\link{featureGroups}} object containing the feature groups that are expected to be
transformation products. If a distinction between parents and TPs is not yet known, \code{fGroupsTPs} should equal
the \code{fGroups} argument. Otherwise, \code{fGroups} should only contain the parent feature groups, and both
\code{fGroups} and \code{fGroupsTPs} \emph{must} be a subset of the same \code{\link{featureGroups}} object.}

\item{ignoreParents}{If \code{TRUE} then feature groups present in both \code{fGroups} and \code{fGroupsTPs} are not
considered as TPs.}

\item{TPs}{A \code{\link{transformationProducts}} object. Set to \code{NULL} to perform linking without this data.}

\item{MSPeakLists, formulas, compounds}{A \code{\link{MSPeakLists}}/\code{\link{formulas}}/\code{\link{compounds}}
object to calculate MS/MS or annotation similarities between parents and TPs. If \code{NULL} then this data is not
calculated. For more details see the \verb{Linking parents and transformation products} section below.}

\item{minRTDiff}{Minimum retention time (in seconds) difference between the parent and a TP to determine whether a TP
elutes prior/after the parent (to calculate \code{retDir} values, see Details in \link{componentsTPs}))}

\item{specSimParams}{A named \code{list} with parameters that influence the calculation of MS spectra similarities.
See the \link[=specSimParams]{spectral similarity parameters} documentation for more details.}
}
\value{
The components are stored in objects derived from \code{\link{componentsTPs}}.
}
\description{
Generates components by linking feature groups of transformation products and their parents.
}
\details{
This function uses transformation product screening to generate components. This function is called when calling \code{generateComponents} with
  \code{algorithm="tp"}.

This method typically employs data from \link[=generateTPs]{generated transformation products} to find
  parents and their TPs. However, this data is not necessary, and components can also be made based on MS/MS
  similarity and/or other annotation similarities between the parent and its TPs. For more details see the
  \verb{Linking parents and transformation products} section below.
}
\note{
The \code{shift} parameter of \code{specSimParams} is ignored by \code{generateComponentsTPs}, since it always
  calculates similarities with all supported options.
}
\section{Linking parents and transformation products}{
 Each component consists of feature groups that are considered
  to be transformation products for one parent (the parent that 'belongs' to the component can be retrieved with the
  \code{\link{componentInfo}} method). The parent feature groups are taken from the \code{fGroups} parameter, while
  the feature groups for TPs are taken from \code{fGroupsTPs}. If a feature group occurs in both variables, it may
  therefore be considered as both a parent or TP.

  If transformation product data is given, \emph{i.e.} the \code{TPs} argument is set, then a suspect screening of
  the TPs must be performed in advance (see \code{\link{screenSuspects}} and \code{\link{convertToSuspects}} to
  create the suspect list). Furthermore, if TPs were generated with \code{\link{generateTPsBioTransformer}} or
  \code{\link{generateTPsLibrary}} then the suspect screening must also include the parents (\emph{e.g.} by setting
  \code{includeParents=TRUE} when calling \code{convertToSuspects} or by amending results by setting
  \code{amend=TRUE} to \code{screenSuspects}). The suspect screening is necessary for the componentization algorithm
  to map the feature groups of the parent or TP. If the the suspect screening yields multiple TP hits, all will be
  reported. Similarly, if the suspect screening contains multiple hits for a parent, a component is made for each of
  the parent hits.

  In case no transformation product data is provided (\code{TPs=NULL}), the componentization algorithm simply assumes
  that each feature group from \code{fGroupsTPs} is a potential TP for every parent feature group in \code{fGroups}.
  For this reason, it is highly recommended to specify which feature groups are parents/TPs (see the
  \code{fGroupsTPs} argument description above) and \emph{crucial} that the data is post-processed, for instance by
  only retaining TPs that have high annotation similarity with their parents (see the
  \code{\link[=filter,componentsTPs-method]{filter}} method for \code{\link{componentsTPs}}).

  A typical way to distinguish which feature groups are parents or TPs from two different (groups of) samples is by
  calculating Fold Changes (see the \code{\link[=as.data.table,featureGroups-method]{as.data.table}} method for
  feature groups and \code{\link{plotVolcano}}). Of course, other statistical techniques from \R are also suitable.

  During componentization, several characteristics are calculated which may be useful for post-processing: \itemize{

  \item \code{specSimilarity}: the MS/MS spectral similarity between the feature groups of the TP and its parent
  (\samp{0-1}).

  \item \code{specSimilarityPrec},\code{specSimilarityBoth}: as \code{specSimilarity}, but calculated with binned
  data using the \code{"precursor"} and \code{"both"} method, respectively (see \link[=specSimParams]{MS spectral
  similarity parameters} for more details).

  \item \code{fragmentMatches} The number of MS/MS fragment formula annotations that overlap between the TP and
  parent. If both the \code{formulas} and \code{compounds} arguments are specified then the annotation data is pooled
  prior to calculation. Note that only unique matches are counted. Furthermore, note that annotations from \emph{all}
  candidates are considered, even if the formula/structure of the parent/TP is known. Hence, \code{fragmentMatches}
  is mainly useful when little or no chemical information is known on the parents/TPs, \emph{i.e.}, when
  \code{TPs=NULL} or originates from \code{\link{generateTPsLogic}}. Since annotations for all candidates are used,
  it is highly recommended that the annotation objects are first processed with the \code{\link{filter}} method, for
  instance, to select only the top ranked candidates.

  \item \code{neutralLossMatches} As \code{fragmentMatches}, but counting overlapping neutral loss formulae.

  \item \code{retDir} The retention time direction of the TP relative to its parent. See Details in
  \link{componentsTPs}. If TP data was specified, the expected direction is stored in \code{TP_retDir}.

  \item \code{retDiff},\code{mzDiff},\code{formulaDiff} The retention time, \emph{m/z} and formula difference between
  the parent and TP (latter only available if data TP formula is available).

  }
}

\section{Sets workflows}{
 In a \link[=sets-workflow]{sets workflow} the component tables are amended with extra
  information such as overall/specific set spectrum similarities. As sets data is mixed, transformation products are
  able to be linked with a parent, even if they were not measured in the same set.
}

\seealso{
\code{\link{generateComponents}} for more details and other algorithms.
}
