% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/compounds-sirius.R
\name{generateCompoundsSIRIUS}
\alias{generateCompoundsSIRIUS}
\alias{generateCompoundsSIRIUS,featureGroups-method}
\alias{generateCompoundsSIRIUS,featureGroupsSet-method}
\title{Compound annotation with SIRIUS}
\usage{
generateCompoundsSIRIUS(fGroups, ...)

\S4method{generateCompoundsSIRIUS}{featureGroups}(
  fGroups,
  MSPeakLists,
  specSimParams = getDefSpecSimParams(removePrecursor = TRUE),
  adduct = NULL,
  config = NULL,
  topMost = 100,
  login = "check",
  alwaysLogin = FALSE,
  minIMSSpecSim = 0,
  projectPath = NULL,
  runMode = "execute",
  SIRIUSAPI = NULL,
  SIRIUSPath = NULL,
  verbose = TRUE
)

\S4method{generateCompoundsSIRIUS}{featureGroupsSet}(
  fGroups,
  MSPeakLists,
  specSimParams = getDefSpecSimParams(removePrecursor = TRUE),
  adduct = NULL,
  config = NULL,
  login = "check",
  alwaysLogin = FALSE,
  minIMSSpecSim = 0,
  projectPath = NULL,
  ...,
  setThreshold = 0,
  setThresholdAnn = 0,
  setAvgSpecificScores = FALSE
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object which should be annotated. This should be the same or a subset of
the object that was used to create the specified \code{MSPeakLists}. In the case of a subset only the remaining
feature groups in the subset are considered.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{MSPeakLists}{A \code{\link{MSPeakLists}} object that was generated for the supplied \code{fGroups}.}

\item{specSimParams}{A named \code{list} with parameters that influence the calculation of the
\link[=id-conf]{annotation similarity}. See the \link[=specSimParams]{spectral similarity parameters} documentation
for more details.}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
  Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}. If the \code{featureGroups} object has
  adduct annotations then these are used if \code{adducts=NULL}.

  \setsWF The \code{adduct} argument is not supported for sets workflows, since the
  adduct annotations will then always be used.}

\item{config}{A \code{RSirius::JobSubmission} configuration object, typically obtained with
\code{\link{getSIRIUSConfig}}. If \code{NULL}, the default \command{SIRIUS} configuration is used.}

\item{topMost}{Only keep this number of candidates (per feature group) with highest score. Setting this to a high
number may result in significant usage of CPU/RAM resources for large numbers of candidates.}

\item{login, alwaysLogin}{Specifies if and how account logging of SIRIUS should be handled:

  \code{login=FALSE}: no automatic login is performed and the active login status is not checked.

  \code{login="check"}: aborts if no active login is present.

  \code{login="interactive"}: interactively ask for login (using \CRANpkg{getPass}).

  \code{login=c(username="...", password="...")}: perform the login with the given details. For security reasons,
  please do not enter the details directly, but use e.g. environment variables or store/retrieve them with the
  \CRANpkg{keyring} package.
  
  if \code{alwaysLogin=TRUE} then a login is always performed, otherwise only if SIRIUS reports no active login.

  See the \href{https://v6.docs.sirius-ms.io/account-and-license/}{SIRIUS website} and \pkg{patRoon} handbook for
  more information.}

\item{minIMSSpecSim}{\IMSWF If the spectrum similarity of an IMS feature group compared to its IMS precursor (see
\code{\link[=assignMobilities_feat]{assignMobilities}}) is at least this value, then the IMS feature group will not
be subjected to the annotation algorithm and all feature annotation properties will be copied from its precursor.
This assumes that feature annotation is primarily influenced by the MS/MS spectrum, and can be used to speed up the
feature annotation process. All scorings, annotation similarities etc. are copied from the IMS precursor. The
fragment annotations are also copied (\code{fragInfo} result column), however, these are adjusted based on the peak
list data of the IMS feature group.}

\item{runMode, projectPath}{Whether to execute a \command{SIRIUS} processing job (\code{runMode="execute"}) or load
  results from an existing \command{SIRIUS} project (\code{runMode"read"}). If \code{runMode="execute"} then
  \code{projectPath} can be \code{NULL} and a temporary project will be used, otherwise \code{projectPath} must point
  to an existing project.

  \strong{NOTE:} if \code{runMode="execute"} then any existing project at \code{projectPath} will be removed.

  \strong{NOTE:} This is primarily intended for internal purposes, but may be of interest to e.g. re-import SIRIUS
  results.

  \setsWF \code{projectPath} should be a \code{character} specifying the paths for each set.}

\item{SIRIUSAPI}{An \code{rsirius_api} object for connecting to the \command{SIRIUS} API. If \code{NULL}, a new
connection will be started automatically.}

\item{SIRIUSPath}{The full path to the \command{SIRIUS} command-line interface executable. Only used when
\code{SIRIUSAPI} is not provided.}

\item{verbose}{If \code{TRUE} then more output is shown in the terminal.}

\item{setThreshold}{\setsWF Minimum abundance for a candidate among all sets (\samp{0-1}). For instance, a value of
\samp{1} means that the candidate needs to be present in all the set data.}

\item{setThresholdAnn}{\setsWF As \code{setThreshold}, but only taking into account the set data that contain
annotations for the feature group of the candidate.}

\item{setAvgSpecificScores}{\setsWF If \code{TRUE} then set specific scorings (\emph{e.g.} MS/MS match) are also
averaged.}
}
\value{
A \code{\link{compoundsSIRIUS}} object.
}
\description{
Uses \href{https://bright-giant.com/sirius-features/}{SIRIUS} for compound annotation.
}
\details{
This function uses SIRIUS to generate compound candidates. This function is called when calling \code{generateCompounds} with
  \code{algorithm="sirius"}.

Similar to \code{\link{generateFormulasSIRIUS}}, candidate formulae are generated with SIRIUS. These results
  are then fed to \command{CSI:FingerID} to acquire candidate structures. Candidate formulae without any assigned
  structure will be removed (unlike \code{\link{generateFormulasSIRIUS}}). This method requires the availability of
  MS/MS data, and feature groups without will be ignored.
}
\section{SIRIUS 6 functionality}{
 The interface to \command{SIRIUS 6} is still in development and may be extended in
  the future. There is a vast amount of functionality available, which will require quite some effort to support all.
  However, the current functionality in \pkg{patRoon} is mostly equal to what was supported with previous
  \command{SIRIUS} releases. Any feedback on the inclusion of specific functionality is welcome!
}

\references{
\insertRef{Hoffmann2021}{patRoon} \cr\cr \insertRef{Dhrkop2020}{patRoon} \cr\cr
  \insertRef{Dhrkop2019}{patRoon} \cr\cr \insertRef{Duhrkop2015}{patRoon} \cr\cr
  \insertRef{Duhrkop2015-2}{patRoon} \cr\cr \insertRef{Bcker2008}{patRoon}
}
\seealso{
\code{\link{generateCompounds}} for more details and other algorithms.
}
