% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds-sirius.R
\name{generateCompoundsSIRIUS}
\alias{generateCompoundsSIRIUS}
\alias{generateCompoundsSIRIUS,featureGroups-method}
\alias{generateCompoundsSIRIUS,featureGroupsSet-method}
\title{Compound annotation with SIRIUS}
\usage{
\S4method{generateCompoundsSIRIUS}{featureGroups}(
  fGroups,
  MSPeakLists,
  relMzDev = 5,
  adduct = NULL,
  projectPath = NULL,
  elements = "CHNOP",
  profile = "qtof",
  formulaDatabase = NULL,
  fingerIDDatabase = "pubchem",
  noise = NULL,
  cores = NULL,
  topMost = 100,
  topMostFormulas = 5,
  extraOptsGeneral = NULL,
  extraOptsFormula = NULL,
  verbose = TRUE,
  splitBatches = FALSE,
  dryRun = FALSE
)

\S4method{generateCompoundsSIRIUS}{featureGroupsSet}(
  fGroups,
  MSPeakLists,
  relMzDev = 5,
  adduct = NULL,
  projectPath = NULL,
  ...,
  setThreshold = 0,
  setThresholdAnn = 0
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object which should be annotated. This should be the same or a subset of
the object that was used to create the specified \code{MSPeakLists}. In the case of a subset only the remaining
feature groups in the subset are considered.}

\item{MSPeakLists}{A \code{\link{MSPeakLists}} object that was generated for the supplied \code{fGroups}.}

\item{relMzDev}{Maximum relative deviation between the measured and candidate formula \emph{m/z} values (in ppm).
Sets the \option{--ppm-max} command line option.}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
  Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}. If the \code{featureGroups} object has
  adduct annotations then these are used if \code{adducts=NULL}.

  \setsWF The \code{adduct} argument is not supported for sets workflows, since the
  adduct annotations will then always be used.}

\item{projectPath, dryRun}{These are mainly for internal purposes. \code{projectPath} sets the output directory for
the \command{SIRIUS} output (a temporary directory if \code{NULL}). If \code{dryRun} is \code{TRUE} then no
computations are done and only the results from \code{projectPath} are processed.

\setsWF \code{projectPath} should be a \code{character} specifying the paths for each set.}

\item{elements}{Elements to be considered for formulae calculation. This will heavily affects the number of
candidates! Always try to work with a minimal set by excluding elements you don't expect. The minimum/maximum
number of elements can also be specified, for example: a value of \code{"C[5]H[10-15]O"} will only consider
formulae with up to five carbon atoms, between ten and fifteen hydrogen atoms and any amount of oxygen atoms. Sets
the \option{--elements} command line  option.}

\item{profile}{Name of the configuration profile, for example: \option{"qtof"}, \option{"orbitrap"},
\option{"fticr"}. Sets the \option{--profile} commandline option.}

\item{formulaDatabase}{If not \code{NULL}, use a database for retrieval of formula
candidates. Possible values are: \option{"pubchem"}, \option{"bio"}, \option{"kegg"}, \option{"hmdb"}. Sets the
\option{--database} commandline option.}

\item{fingerIDDatabase}{Database specifically used for \command{CSI:FingerID}. If \code{NULL}, the value of the
\code{formulaDatabase} parameter will be used or \code{"pubchem"} when that is also \code{NULL}. Sets the
\option{--fingerid-db} option.}

\item{noise}{Median intensity of the noise (\code{NULL} ignores this parameter). Sets the \option{--noise}
commandline option.}

\item{cores}{The number of cores \command{SIRIUS} will use. If \code{NULL} then the default of all cores will be
used.}

\item{topMost}{Only keep this number of candidates (per feature group) with highest score. Set to \code{NULL} to
always keep all candidates, however, please note that this may result in significant usage of CPU/RAM resources for
large numbers of candidates.}

\item{topMostFormulas}{Do not return more than this number of candidate formulae. Note that only compounds for these
formulae will be searched. Sets the \option{--candidates} commandline option.}

\item{extraOptsGeneral, extraOptsFormula}{a \code{character} vector with any extra commandline parameters for
\command{SIRIUS}. For \command{SIRIUS} versions \code{<4.4} there is no distinction between general and formula
options. Otherwise commandline options specified in \code{extraOptsGeneral} are added prior to the \code{formula}
command, while options specified in \code{extraOptsFormula} are added in afterwards. See the \command{SIRIUS}
manual for more details. Set to \code{NULL} to ignore.}

\item{verbose}{If \code{TRUE} then more output is shown in the terminal.}

\item{splitBatches}{If \code{TRUE} then the calculations done by \command{SIRIUS} will be evenly split over multiple
\command{SIRIUS} calls (which may be run in parallel depending on the \link[=patRoon-package]{set package
options}). If \code{splitBatches=FALSE} then all feature calculations are performed from a single \command{SIRIUS}
execution, which is often the fastest if calculations are performed on a single computer.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{setThreshold}{\setsWF Minimum abundance for a candidate among all sets (\samp{0-1}). For instance, a value of
\samp{1} means that the candidate needs to be present in all the set data.}

\item{setThresholdAnn}{\setsWF As \code{setThreshold}, but only taking into account the set data that contain
annotations for the feature group of the candidate.}
}
\value{
A \code{\link{compounds}} derived object containing all compound annotations.
}
\description{
Uses \href{https://bio.informatik.uni-jena.de/software/sirius/}{SIRIUS} in combination with
\href{https://www.csi-fingerid.uni-jena.de/}{CSI:FingerID} for compound annotation.
}
\details{
This function uses SIRIUS to generate compound candidates. This function is called when calling \code{generateCompounds} with
  \code{algorithm="sirius"}.

Similar to \code{\link{generateFormulasSIRIUS}}, candidate formulae are generated with SIRIUS. These results
  are then feed to CSI:FingerID to acquire candidate structures. This method requires the availability of MS/MS data,
  and feature groups without it will be ignored.
}
\note{
For annotations performed with \command{SIRIUS} it is often the fastest to keep the default
  \code{splitBatches=FALSE}. In this case, all \command{SIRIUS} output will be printed to the terminal (unless
  \code{verbose=FALSE} or \option{patRoon.MP.method="future"}). Furthermore, please note that only annotations to be
  performed for the same adduct are grouped in a single batch execution.
}
\section{Parallelization}{
 generateCompoundsSIRIUS uses multiprocessing to parallelize
  computations. Please see the parallelization section in the handbook for
  more details and \link[=patRoon-package]{patRoon options} for configuration
  options.
}

\references{
\insertRef{Dhrkop2019}{patRoon} \cr\cr \insertRef{Duhrkop2015}{patRoon} \cr\cr
  \insertRef{Duhrkop2015-2}{patRoon} \cr\cr \insertRef{Bcker2008}{patRoon}
}
\seealso{
\code{\link{generateCompounds}} for more details and other algorithms.
}
