% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/formulas-sirius.R
\name{generateFormulasSIRIUS}
\alias{generateFormulasSIRIUS}
\alias{generateFormulasSIRIUS,featureGroups-method}
\alias{generateFormulasSIRIUS,featureGroupsSet-method}
\title{Generate formula with SIRIUS}
\usage{
generateFormulasSIRIUS(fGroups, ...)

\S4method{generateFormulasSIRIUS}{featureGroups}(
  fGroups,
  MSPeakLists,
  specSimParams = getDefSpecSimParams(removePrecursor = TRUE),
  adduct = NULL,
  config = NULL,
  getFingerprints = FALSE,
  login = "check",
  alwaysLogin = FALSE,
  calculateFeatures = FALSE,
  featThreshold = 0,
  featThresholdAnn = 0.75,
  absAlignMzDev = defaultLim("mz", "narrow"),
  minIMSSpecSim = 0,
  projectPath = NULL,
  runMode = "execute",
  SIRIUSAPI = NULL,
  verbose = TRUE
)

\S4method{generateFormulasSIRIUS}{featureGroupsSet}(
  fGroups,
  MSPeakLists,
  specSimParams = getDefSpecSimParams(removePrecursor = TRUE),
  adduct = NULL,
  config = NULL,
  login = "check",
  alwaysLogin = FALSE,
  calculateFeatures = FALSE,
  featThreshold = 0,
  featThresholdAnn = 0.75,
  minIMSSpecSim = 0,
  projectPath = NULL,
  ...,
  setThreshold = 0,
  setThresholdAnn = 0,
  setAvgSpecificScores = FALSE
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which formulae should be generated. This should be the same or
a subset of the object that was used to create the specified \code{MSPeakLists}. In the case of a subset only the
remaining feature groups in the subset are considered.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{MSPeakLists}{An \code{\link{MSPeakLists}} object that was generated for the supplied \code{fGroups}.}

\item{specSimParams}{A named \code{list} with parameters that influence the calculation of the
\link[=id-conf]{annotation similarity}. See the \link[=specSimParams]{spectral similarity parameters} documentation
for more details.}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
  Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}. If the \code{featureGroups} object has
  adduct annotations then these are used if \code{adducts=NULL}.

  \setsWF The \code{adduct} argument is not supported for sets workflows, since the
  adduct annotations will then always be used.}

\item{config}{A \code{RSirius::JobSubmission} configuration object, typically obtained with
\code{\link{getSIRIUSConfig}}. If \code{NULL}, the default \command{SIRIUS} configuration is used.}

\item{login, alwaysLogin}{Specifies if and how account logging of SIRIUS should be handled:

  \code{login=FALSE}: no automatic login is performed and the active login status is not checked.

  \code{login="check"}: aborts if no active login is present.

  \code{login="interactive"}: interactively ask for login (using \CRANpkg{getPass}).

  \code{login=c(username="...", password="...")}: perform the login with the given details. For security reasons,
  please do not enter the details directly, but use e.g. environment variables or store/retrieve them with the
  \CRANpkg{keyring} package.
  
  if \code{alwaysLogin=TRUE} then a login is always performed, otherwise only if SIRIUS reports no active login.

  See the \href{https://v6.docs.sirius-ms.io/account-and-license/}{SIRIUS website} and \pkg{patRoon} handbook for
  more information.
  
  \strong{NOTE}: By loggin in you will accept the terms of the Service and Privacy Policy of the SIRIUS Webservice.}

\item{calculateFeatures}{If \code{TRUE} fomulae are first calculated for all features
prior to feature group assignment (see \verb{Candidate assignment} in \code{\link{generateFormulas}}).}

\item{featThreshold}{If \code{calculateFeatures=TRUE}: minimum presence (\samp{0-1}) of a formula in all features
before it is considered as a candidate for a feature group. For instance, \code{featThreshold=0.75} dictates that a
formula should be present in at least 75\% of the features inside a feature group.}

\item{featThresholdAnn}{As \code{featThreshold}, but only considers features with annotations. For instance,
\code{featThresholdAnn=0.75} dictates that a formula should be present in at least 75\% of the features with
annotations inside a feature group.}

\item{absAlignMzDev}{When the group formula annotation consensus is made from feature annotations, the \emph{m/z}
values of annotated MS/MS fragments may slightly deviate from those of the corresponding group MS/MS peak list. The
\code{absAlignMzDev} argument specifies the maximum \emph{m/z} window used to re-align the mass peaks.}

\item{minIMSSpecSim}{\IMSWF If the spectrum similarity of an IMS feature group compared to its IMS precursor (see
\code{\link[=assignMobilities_feat]{assignMobilities}}) is at least this value, then the IMS feature group will not
be subjected to the annotation algorithm and all feature annotation properties will be copied from its precursor.
This assumes that feature annotation is primarily influenced by the MS/MS spectrum, and can be used to speed up the
feature annotation process. All scorings, annotation similarities etc. are copied from the IMS precursor. The
fragment annotations are also copied (\code{fragInfo} result column), however, these are adjusted based on the peak
list data of the IMS feature group.}

\item{runMode, projectPath}{Whether to execute a \command{SIRIUS} processing job (\code{runMode="execute"}) or load
  results from an existing \command{SIRIUS} project (\code{runMode"read"}). If \code{runMode="execute"} then
  \code{projectPath} can be \code{NULL} and a temporary project will be used, otherwise \code{projectPath} must point
  to an existing project.

  \strong{NOTE:} if \code{runMode="execute"} then any existing project at \code{projectPath} will be removed.

  \strong{NOTE:} This is primarily intended for internal purposes, but may be of interest to e.g. re-import SIRIUS
  results.

  \setsWF \code{projectPath} should be a \code{character} specifying the paths for each set.}

\item{SIRIUSAPI}{An \code{rsirius_api} object for connecting to the \command{SIRIUS} API. If \code{NULL}, a new
connection will be started automatically.}

\item{verbose}{If \code{TRUE} then more output is shown.}

\item{setThreshold}{\setsWF Minimum abundance for a candidate among all sets (\samp{0-1}). For instance, a value of
\samp{1} means that the candidate needs to be present in all the set data.}

\item{setThresholdAnn}{\setsWF As \code{setThreshold}, but only taking into account the set data that contain
annotations for the feature group of the candidate.}

\item{setAvgSpecificScores}{\setsWF If \code{TRUE} then set specific scorings (\emph{e.g.} MS/MS match) are also
averaged.}
}
\value{
A \code{\link{formulasSIRIUS}} object.
}
\description{
Uses \href{https://bio.informatik.uni-jena.de/software/sirius/}{SIRIUS} to generate chemical formulae candidates.
}
\details{
This function uses sirius to generate formula candidates. This function is called when calling \code{generateFormulas} with
  \code{algorithm="sirius"}.

Note that \command{SIRIUS} requires availability of MS/MS data.
}
\section{Running SIRIUS}{
 By default, \pkg{patRoon} tries to connect to a running instance of \command{SIRIUS}. This
  is generally faster and may be useful for debugging by \emph{e.g.} checking the logs in \command{SIRIUS}.
  Otherwise, an attempt will be made to start \command{SIRIUS} automatically. The binaries are searched from the
  \option{patRoon.path.SIRIUS} package option, \pkg{patRoonExt} package or the system \code{PATH} environment
  variable. Any automatically started \command{SIRIUS} instances are automatically closed if jobs are finished. By
  default, a temporary \command{SIRIUS} project is made for \command{SIRIUS} data processing and removed afterwards.
  See the \code{projectPath} to change this.
}

\section{SIRIUS 6 functionality}{
 The interface to \command{SIRIUS 6} is still in development and may be extended in
  the future. There is a vast amount of functionality available, which will require quite some effort to support all.
  However, the current functionality in \pkg{patRoon} is mostly equal to what was supported with previous
  \command{SIRIUS} releases. Any feedback on the inclusion of specific functionality is welcome!
}

\references{
\insertRef{Hoffmann2021}{patRoon} \cr\cr \insertRef{Dhrkop2020}{patRoon} \cr\cr
  \insertRef{Dhrkop2019}{patRoon} \cr\cr \insertRef{Duhrkop2015}{patRoon} \cr\cr
  \insertRef{Duhrkop2015-2}{patRoon} \cr\cr \insertRef{Bcker2008}{patRoon}
}
\seealso{
\code{\link{generateFormulas}} for more details and other algorithms.
}
