% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/formulas-sirius.R
\name{generateFormulasSIRIUS}
\alias{generateFormulasSIRIUS}
\alias{generateFormulasSIRIUS,featureGroups-method}
\alias{generateFormulasSIRIUS,featureGroupsSet-method}
\title{Generate formula with SIRIUS}
\usage{
generateFormulasSIRIUS(fGroups, ...)

\S4method{generateFormulasSIRIUS}{featureGroups}(
  fGroups,
  MSPeakLists,
  relMzDev = 5,
  adduct = NULL,
  projectPath = NULL,
  elements = "CHNOP",
  profile = "qtof",
  database = NULL,
  noise = NULL,
  cores = NULL,
  getFingerprints = FALSE,
  topMost = 100,
  login = FALSE,
  alwaysLogin = FALSE,
  extraOptsGeneral = NULL,
  extraOptsFormula = NULL,
  calculateFeatures = TRUE,
  featThreshold = 0,
  featThresholdAnn = 0.75,
  absAlignMzDev = 0.002,
  verbose = TRUE,
  splitBatches = FALSE,
  dryRun = FALSE
)

\S4method{generateFormulasSIRIUS}{featureGroupsSet}(
  fGroups,
  MSPeakLists,
  relMzDev = 5,
  adduct = NULL,
  projectPath = NULL,
  ...,
  setThreshold = 0,
  setThresholdAnn = 0,
  setAvgSpecificScores = FALSE
)
}
\arguments{
\item{fGroups}{\code{\link{featureGroups}} object for which formulae should be generated. This should be the same or
a subset of the object that was used to create the specified \code{MSPeakLists}. In the case of a subset only the
remaining feature groups in the subset are considered.}

\item{\dots}{\setsWF Further arguments passed to the non-sets workflow method.}

\item{MSPeakLists}{An \code{\link{MSPeakLists}} object that was generated for the supplied \code{fGroups}.}

\item{relMzDev}{Maximum relative deviation between the measured and candidate formula \emph{m/z} values (in ppm).
Sets the \option{--ppm-max} command line option.}

\item{adduct}{An \code{\link{adduct}} object (or something that can be converted to it with \code{\link{as.adduct}}).
  Examples: \code{"[M-H]-"}, \code{"[M+Na]+"}. If the \code{featureGroups} object has
  adduct annotations then these are used if \code{adducts=NULL}.

  \setsWF The \code{adduct} argument is not supported for sets workflows, since the
  adduct annotations will then always be used.}

\item{projectPath, dryRun}{These are mainly for internal purposes. \code{projectPath} sets the output directory for
the \command{SIRIUS} output (a temporary directory if \code{NULL}). If \code{dryRun} is \code{TRUE} then no
computations are done and only the results from \code{projectPath} are processed.

\setsWF \code{projectPath} should be a \code{character} specifying the paths for each set.}

\item{elements}{Elements to be considered for formulae calculation. This will heavily affects the number of
candidates! Always try to work with a minimal set by excluding elements you don't expect. The minimum/maximum
number of elements can also be specified, for example: a value of \code{"C[5]H[10-15]O"} will only consider
formulae with up to five carbon atoms, between ten and fifteen hydrogen atoms and any amount of oxygen atoms. Sets
the \option{--elements} command line  option.}

\item{profile}{Name of the configuration profile, for example: \option{"qtof"}, \option{"orbitrap"},
\option{"fticr"}. Sets the \option{--profile} commandline option.}

\item{database}{If not \code{NULL}, use a database for retrieval of formula
candidates. Possible values are: \option{"pubchem"}, \option{"bio"}, \option{"kegg"}, \option{"hmdb"}. Sets the
\option{--database} commandline option.}

\item{noise}{Median intensity of the noise (\code{NULL} ignores this parameter). Sets the \option{--noise}
commandline option.}

\item{cores}{The number of cores \command{SIRIUS} will use. If \code{NULL} then the default of all cores will be
used.}

\item{getFingerprints}{Set to \code{TRUE} to load \command{SIRIUS-CSI:FingerID} MS/MS fingerprints for the formula
candidates. This is currently only supported with \code{calculateFeatures=FALSE} to avoid heavy server traffic. The
fingerprints are stored in the \code{fingerprints} slot of the returned \code{\link{formulasSIRIUS}} object, and
are used by the \code{\link{predictTox}} and \code{\link{predictRespFactors}} methods.}

\item{topMost}{Only keep this number of candidates (per feature group) with highest
score. Sets the \option{--candidates} command line option.}

\item{login, alwaysLogin}{Specifies if and how account logging of SIRIUS should be handled:

  \code{login=FALSE}: no automatic login is performed and the active login status is not checked.

  \code{login=check}: aborts if no active login is present.

  \code{login=interactive}: interactively ask for login (using \CRANpkg{getPass}).

  \code{login=c(username="...", password="...")}: perform the login with the given details. For security reasons,
  please do not enter the details directly, but use e.g. environment variables or store/retrieve them with the
  \CRANpkg{keyring} package.
  
  if \code{alwaysLogin=TRUE} then a login is always performed, otherwise only if SIRIUS reports no active login.

  See the \href{https://boecker-lab.github.io/docs.sirius.github.io/account-and-license/}{SIRIUS website} for more
  information.}

\item{extraOptsGeneral, extraOptsFormula}{a \code{character} vector with any extra commandline parameters for
\command{SIRIUS}. For \command{SIRIUS} versions \code{<4.4} there is no distinction between general and formula
options. Otherwise commandline options specified in \code{extraOptsGeneral} are added prior to the \code{formula}
command, while options specified in \code{extraOptsFormula} are added in afterwards. See the \command{SIRIUS}
manual for more details. Set to \code{NULL} to ignore.}

\item{calculateFeatures}{If \code{TRUE} fomulae are first calculated for all features
prior to feature group assignment (see \verb{Candidate assignment} in \code{\link{generateFormulas}}).}

\item{featThreshold}{If \code{calculateFeatures=TRUE}: minimum presence (\samp{0-1}) of a formula in all features
before it is considered as a candidate for a feature group. For instance, \code{featThreshold=0.75} dictates that a
formula should be present in at least 75\% of the features inside a feature group.}

\item{featThresholdAnn}{As \code{featThreshold}, but only considers features with annotations. For instance,
\code{featThresholdAnn=0.75} dictates that a formula should be present in at least 75\% of the features with
annotations inside a feature group. @param topMost Only keep this number of candidates
(per feature group) with highest score. Sets the \option{--candidates} command line
option.}

\item{absAlignMzDev}{When the group formula annotation consensus is made from feature annotations, the \emph{m/z}
values of annotated MS/MS fragments may slightly deviate from those of the corresponding group MS/MS peak list. The
\code{absAlignMzDev} argument specifies the maximum \emph{m/z} window used to re-align the mass peaks.}

\item{verbose}{If \code{TRUE} then more output is shown in the terminal.}

\item{splitBatches}{If \code{TRUE} then the calculations done by \command{SIRIUS} will be evenly split over multiple
\command{SIRIUS} calls (which may be run in parallel depending on the \link[=patRoon-package]{set package
options}). If \code{splitBatches=FALSE} then all feature calculations are performed from a single \command{SIRIUS}
execution, which is often the fastest if calculations are performed on a single computer.}

\item{setThreshold}{\setsWF Minimum abundance for a candidate among all sets (\samp{0-1}). For instance, a value of
\samp{1} means that the candidate needs to be present in all the set data.}

\item{setThresholdAnn}{\setsWF As \code{setThreshold}, but only taking into account the set data that contain
annotations for the feature group of the candidate.}

\item{setAvgSpecificScores}{\setsWF If \code{TRUE} then set specific scorings (\emph{e.g.} MS/MS match) are also
averaged.}
}
\value{
A \code{\link{formulasSIRIUS}} object.
}
\description{
Uses \href{https://bio.informatik.uni-jena.de/software/sirius/}{SIRIUS} to generate chemical formulae candidates.
}
\details{
This function uses sirius to generate formula candidates. This function is called when calling \code{generateFormulas} with
  \code{algorithm="sirius"}.

Similarity of measured and theoretical isotopic patterns will be used for scoring candidates. Note that
  \command{SIRIUS} requires availability of MS/MS data.
}
\note{
For annotations performed with \command{SIRIUS} it is often the fastest to keep the default
  \code{splitBatches=FALSE}. In this case, all \command{SIRIUS} output will be printed to the terminal (unless
  \code{verbose=FALSE} or \option{patRoon.MP.method="future"}). Furthermore, please note that only annotations to be
  performed for the same adduct are grouped in a single batch execution.
}
\section{Parallelization}{
 generateFormulasSIRIUS uses multiprocessing to parallelize
  computations. Please see the parallelization section in the handbook for
  more details and \link[=patRoon-package]{patRoon options} for configuration
  options.
}

\references{
\insertRef{Dhrkop2019}{patRoon} \cr\cr \insertRef{Duhrkop2015}{patRoon} \cr\cr
  \insertRef{Duhrkop2015-2}{patRoon} \cr\cr \insertRef{Bcker2008}{patRoon}
}
\seealso{
\code{\link{generateFormulas}} for more details and other algorithms.
}
