% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-biotransformer.R
\name{generateTPsBioTransformer}
\alias{generateTPsBioTransformer}
\title{Obtain transformation products (TPs) with BioTransformer}
\usage{
generateTPsBioTransformer(
  parents,
  type = "env",
  steps = 2,
  extraOpts = NULL,
  skipInvalid = TRUE,
  fpType = "extended",
  fpSimMethod = "tanimoto",
  MP = FALSE
)
}
\arguments{
\item{parents}{The parents for which transformation products should be obtained. This can be (1) a suspect list (see
\link[=suspect-screening]{suspect screening} for more information), (2) the resulting output
\code{\link{screenSuspects}} or (3) a \code{\link{compounds}} annotation object. In the former two cases, the
suspect (hits) are used as parents, whereas in the latter case all candidates are used as parents.}

\item{type}{The type of prediction. Valid values are: \code{"env"}, \code{"ecbased"}, \code{"cyp450"},
\code{"phaseII"}, \code{"hgut"}, \code{"superbio"}, \code{"allHuman"}. Sets the \command{-b} command line option.}

\item{steps}{The number of steps for the predictions. Sets the \command{-s} command line option.}

\item{extraOpts}{A \code{character} with extra command line options passed to the \command{biotransformer.jar} tool.}

\item{skipInvalid}{If set to \code{TRUE} then the parents will be skipped (with a warning) for which insufficient
information (\emph{e.g.} SMILES) is available.}

\item{fpType}{The type of structural fingerprint that should be calculated. See the \code{type} argument of the
\code{\link{get.fingerprint}} function of \CRANpkg{rcdk}.}

\item{fpSimMethod}{The method for calculating similarities (i.e. not dissimilarity!). See the \code{method} argument
of the \code{\link{fp.sim.matrix}} function of the \CRANpkg{fingerprint} package.}

\item{MP}{If \code{TRUE} then multiprocessing is enabled. Since \command{BioTransformer} supports native
parallelization, additional multiprocessing generally doesn't lead to significant reduction in computational times.
Furthermore, enabling multiprocessing can lead to very high CPU/RAM usage.}
}
\value{
The TPs are stored in an object from the \code{\link{transformationProductsBT}} class.
}
\description{
Uses \href{http://biotransformer.ca/}{BioTransformer} to predict TPs
}
\details{
This function uses BioTransformer to obtain transformation products. This function is called when calling \code{generateTPs} with
  \code{algorithm="biotransformer"}.

Structural similarities between the parent and its TPs are calculated, which can be used to
  \link[=filter,transformationProductsBT-method]{filter} the results.

  In order to use this function the \file{.jar} command line utility should be installed and specified in the
  \code{\link[=patRoon-package]{patRoon.path.BioTransformer}} option. The \file{.jar} file can be obtained via
  \url{https://bitbucket.org/djoumbou/biotransformer/src/master}.

An important advantage of this algorithm is that it provides structural information for generated TPs.
  However, this also means that if the input is from a parent suspect list or screening then either \acronym{SMILES}
  or \acronym{InChI} information must be available for the parents.
}
\note{
When the \code{parents} argument is a \code{\link{compounds}} object, the candidate library \code{identifier}
  is used in case the candidate has no defined \code{compoundName}.
}
\section{Parallelization}{
 \code{generateTPsBioTransformer} uses multiprocessing to parallelize
  computations. Please see the parallelization section in the handbook for
  more details and \link[=patRoon-package]{patRoon options} for configuration
  options.
}

\references{
\insertRef{DjoumbouFeunang2019}{patRoon} \cr\cr \insertRef{Wicker2015}{patRoon} \cr\cr
  \addCitations{rcdk}{1}
}
\seealso{
\code{\link{generateTPs}} for more details and other algorithms.
}
