% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP-library.R
\name{generateTPsLibrary}
\alias{generateTPsLibrary}
\title{Obtain transformation products (TPs) from a library}
\usage{
generateTPsLibrary(
  parents = NULL,
  TPLibrary = NULL,
  skipInvalid = TRUE,
  matchParentsBy = "InChIKey"
)
}
\arguments{
\item{parents}{The parents for which transformation products should be obtained. This can be (1) a suspect list (see
\link[=suspect-screening]{suspect screening} for more information), (2) the resulting output
\code{\link{screenSuspects}} or (3) a \code{\link{compounds}} annotation object. In the former two cases, the
suspect (hits) are used as parents, whereas in the latter case all candidates are used as parents.
If \code{NULL} then TPs for all parents in the library are obtained.}

\item{TPLibrary}{If \code{NULL}, a default \href{https://doi.org/10.5281/zenodo.5644560}{PubChem} based library is
used. Otherwise, \code{TPLibrary} should be a \code{data.frame}. See the details below.}

\item{skipInvalid}{If set to \code{TRUE} then the parents will be skipped (with a warning) for which insufficient
information (\emph{e.g.} SMILES) is available.}

\item{matchParentsBy}{A \code{character} that specifies how the input parents are matched with the data from the TP
library. Valid options are: \code{"InChIKey"}, \code{"InChIKey1"}, \code{"InChI"}, \code{"SMILES"}.}
}
\value{
The TPs are stored in an object from the \code{\link{transformationProductsLibrary}} class.
}
\description{
Automatically obtains transformation products from a library.
}
\details{
This function uses a library to obtain transformation products. This function is called when calling \code{generateTPs} with
  \code{algorithm="library"}.

By default, a library is used that is based on data from
  \href{https://doi.org/10.5281/zenodo.5644560}{PubChem}. However, it also possible to use your own library.

An important advantage of this algorithm is that it provides structural information for generated TPs.
  However, this also means that if the input is from a parent suspect list or screening then either \acronym{SMILES}
  or \acronym{InChI} information must be available for the parents.
}
\note{
When the \code{parents} argument is a \code{\link{compounds}} object, the candidate library \code{identifier}
  is used in case the candidate has no defined \code{compoundName}.
}
\section{Custom TP libraries}{
 The \code{TPLibrary} argument is used to specify a custom TP library. This should be a
  \code{data.frame} where each row specifies a TP for a parent, with the following columns: \itemize{

  \item \code{parent_name} and \code{TP_name}: The name of the parent/TP.

  \item \code{parent_SMILES} and \code{TP_SMILES} The \acronym{SMILES} of the parent structure.

  \item \code{parent_LogP} and \code{TP_LogP} The \code{log P} values for the parent/TP. (\strong{optional})

  \item \code{LogPDiff} The difference between parent and TP \code{Log P} values. Ignored if \emph{both}
  \code{parent_LogP} and \code{TP_LogP} are specified. (\strong{optional})

  }

  Other columns are allowed, and will be included in the final object. Multiple TPs for a single parent are specified
  by repeating the value within \code{parent_} columns.
}

\seealso{
\code{\link{generateTPs}} for more details and other algorithms.
}
