% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics}
\alias{generics}
\alias{adducts}
\alias{adducts<-}
\alias{algorithm}
\alias{analysisInfo}
\alias{analysisInfo<-}
\alias{analyses}
\alias{annotatedPeakList}
\alias{annotations}
\alias{assignMobilities}
\alias{calculatePeakQualities}
\alias{clusterProperties}
\alias{clusters}
\alias{consensus}
\alias{convertToMFDB}
\alias{convertToSuspects}
\alias{cutClusters}
\alias{defaultExclNormScores}
\alias{export}
\alias{featureTable}
\alias{filter}
\alias{fromIMS}
\alias{getBPCs}
\alias{getEICs}
\alias{getFeatures}
\alias{getFeatureQualityNames}
\alias{getMCS}
\alias{getTICs}
\alias{groupNames}
\alias{hasIMS}
\alias{plotBPCs}
\alias{plotChord}
\alias{plotChroms}
\alias{plotChroms3D}
\alias{plotGraph}
\alias{plotInt}
\alias{plotScores}
\alias{plotSilhouettes}
\alias{plotSpectrum}
\alias{plotStructure}
\alias{plotTICs}
\alias{plotVenn}
\alias{plotUpSet}
\alias{predictRespFactors}
\alias{predictTox}
\alias{delete}
\alias{plotVolcano}
\alias{replicates}
\alias{setObjects}
\alias{sets}
\alias{treeCut}
\alias{treeCutDynamic}
\alias{unset}
\alias{[}
\alias{[[}
\alias{$}
\alias{as.data.table}
\alias{as.data.frame}
\alias{length}
\alias{lengths}
\alias{names}
\alias{plot}
\alias{show}
\title{Miscellaneous generics}
\usage{
adducts(obj, ...)

adducts(obj, ...) <- value

algorithm(obj)

analysisInfo(obj, df = FALSE)

analysisInfo(obj) <- value

analyses(obj)

annotatedPeakList(obj, ...)

annotations(obj, ...)

assignMobilities(obj, ...)

calculatePeakQualities(
  obj,
  weights = NULL,
  flatnessFactor = 0.05,
  featureQualities = NULL,
  featureGroupQualities = NULL,
  ...
)

clusterProperties(obj)

clusters(obj)

consensus(obj, ...)

convertToMFDB(TPs, out, ...)

convertToSuspects(obj, ...)

cutClusters(obj)

defaultExclNormScores(obj)

export(obj, type, out, ...)

featureTable(obj, ...)

filter(obj, ...)

fromIMS(obj)

getBPCs(obj, ...)

getEICs(obj, ...)

getFeatures(obj)

getFeatureQualityNames(obj, ...)

getMCS(obj, ...)

getTICs(obj, ...)

groupNames(obj)

hasIMS(obj)

plotBPCs(obj, ...)

plotChord(obj, addSelfLinks = FALSE, addRetMzPlots = TRUE, ...)

plotChroms(obj, ...)

plotChroms3D(obj, ...)

plotGraph(obj, ...)

plotInt(obj, ...)

plotScores(obj, ...)

plotSilhouettes(obj, kSeq, ...)

plotSpectrum(obj, ...)

plotStructure(obj, ...)

plotTICs(obj, ...)

plotVenn(obj, ...)

plotUpSet(obj, ...)

predictRespFactors(obj, ...)

predictTox(obj, ...)

delete(obj, ...)

plotVolcano(obj, ...)

replicates(obj)

setObjects(obj)

sets(obj)

treeCut(obj, k = NULL, h = NULL, ...)

treeCutDynamic(
  obj,
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1,
  ...
)

unset(obj, set)
}
\arguments{
\item{obj}{The object the generic should be applied to.}

\item{\dots}{Any further method specific arguments. See method documentation
for details.}

\item{value}{The replacement value.}

\item{df}{If \code{TRUE} then a \code{data.frame} is returned, otherwise a \code{data.table} is returned.}

\item{weights, flatnessFactor, featureQualities, featureGroupQualities}{See method documentation.}

\item{TPs}{The \code{\link{transformationProducts}} derived object.}

\item{out}{Output file.}

\item{type}{The export type.}

\item{addSelfLinks}{If \code{TRUE} then 'self-links' are added which
represent non-shared data.}

\item{addRetMzPlots}{Set to \code{TRUE} to enable \emph{m/z} \emph{vs}
retention time scatter plots.}

\item{kSeq}{An integer vector containing the sequence that should be used for
average silhouette width calculation.}

\item{k, h}{Desired numbers of clusters. See \code{\link{cutree}}.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}

\item{set}{The name of the set.}
}
\description{
Various (S4) generic functions providing a common interface for common tasks
such as plotting and filtering data. The actual functionality and function
arguments are often specific for the implemented methods, for this reason,
please refer to the linked method documentation for each generic.
}
\details{
\code{adducts} returns assigned adducts of the object.

\itemize{
    \item Methods are defined for: \code{\link[=adducts,featureGroups-method]{featureGroups}}; \code{\link[=adducts,featureGroupsSet-method]{featureGroupsSet}}.
}

\code{adducts<-} sets adducts of the object.

\itemize{
    \item Methods are defined for: \code{\link[=adducts<-,featureGroups-method]{featureGroups}}; \code{\link[=adducts<-,featureGroupsSet-method]{featureGroupsSet}}.
}

\code{algorithm} returns the algorithm that was used to generate the object.

\itemize{
    \item Methods are defined for: \code{\link[=algorithm,optimizationResult-method]{optimizationResult}}; \code{\link[=algorithm,workflowStep-method]{workflowStep}}.
}

\code{analysisInfo} returns the \link[=analysis-information]{analysis information} of an object.

\itemize{
    \item Methods are defined for: \code{\link[=analysisInfo,featureGroups-method]{featureGroups}}; \code{\link[=analysisInfo,features-method]{features}}; \code{\link[=analysisInfo,MSPeakListsSet-method]{MSPeakListsSet}}.
}

\code{analysisInfo<-} modifies the \link[=analysis-information]{analysis information} of an object.

\itemize{
    \item Methods are defined for: \code{\link[=analysisInfo<-,featureGroups-method]{featureGroups}}; \code{\link[=analysisInfo<-,featureGroupsXCMS-method]{featureGroupsXCMS}}; \code{\link[=analysisInfo<-,features-method]{features}}; \code{\link[=analysisInfo<-,featuresXCMS-method]{featuresXCMS}}.
}

\code{analyses} returns a \code{character} vector with the analyses for which data is present in this object.

\itemize{
    \item Methods are defined for: \code{\link[=analyses,featureGroups-method]{featureGroups}}; \code{\link[=analyses,features-method]{features}}; \code{\link[=analyses,formulas-method]{formulas}}; \code{\link[=analyses,MSPeakLists-method]{MSPeakLists}}.
}

\code{annotatedPeakList} returns an annotated MS peak list.

\itemize{
    \item Methods are defined for: \code{\link[=annotatedPeakList,compounds-method]{compounds}}; \code{\link[=annotatedPeakList,compoundsSet-method]{compoundsSet}}; \code{\link[=annotatedPeakList,formulas-method]{formulas}}; \code{\link[=annotatedPeakList,formulasSet-method]{formulasSet}}.
}

\code{annotations} returns annotations.

\itemize{
    \item Methods are defined for: \code{\link[=annotations,featureAnnotations-method]{featureAnnotations}}; \code{\link[=annotations,featureGroups-method]{featureGroups}}; \code{\link[=annotations,formulas-method]{formulas}}.
}

\code{assignMobilities} assigns ion mobility and/or \acronym{CCS} values to workflow data.

\itemize{
    \item Methods are defined for: \code{\link[=assignMobilities,compounds-method]{compounds}}; \code{\link[=assignMobilities,compoundsSet-method]{compoundsSet}}; \code{\link[=assignMobilities,data.frame-method]{data.frame}}; \code{\link[=assignMobilities,data.table-method]{data.table}}; \code{\link[=assignMobilities,featureGroups-method]{featureGroups}}; \code{\link[=assignMobilities,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=assignMobilities,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=assignMobilities,featureGroupsSet-method]{featureGroupsSet}}.
}

\code{calculatePeakQualities} calculates chromatographic peak qualities and scores.

\itemize{
    \item Methods are defined for: \code{\link[=calculatePeakQualities,featureGroups-method]{featureGroups}}; \code{\link[=calculatePeakQualities,features-method]{features}}.
}

\code{clusterProperties} Obtain a list with properties of the generated cluster(s).

\itemize{
    \item Methods are defined for: \code{\link[=clusterProperties,componentsClust-method]{componentsClust}}; \code{\link[=clusterProperties,compoundsCluster-method]{compoundsCluster}}.
}

\code{clusters} Obtain clustering object(s).

\itemize{
    \item Methods are defined for: \code{\link[=clusters,componentsClust-method]{componentsClust}}; \code{\link[=clusters,compoundsCluster-method]{compoundsCluster}}.
}

\code{consensus} combines and merges data from various algorithms to generate a consensus.

\itemize{
    \item Methods are defined for: \code{\link[=consensus,components-method]{components}}; \code{\link[=consensus,componentsSet-method]{componentsSet}}; \code{\link[=consensus,compounds-method]{compounds}}; \code{\link[=consensus,compoundsSet-method]{compoundsSet}}; \code{\link[=consensus,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=consensus,featureGroupsComparisonSet-method]{featureGroupsComparisonSet}}; \code{\link[=consensus,formulas-method]{formulas}}; \code{\link[=consensus,formulasSet-method]{formulasSet}}; \code{\link[=consensus,transformationProductsStructure-method]{transformationProductsStructure}}.
}

\code{convertToMFDB} Exports the object to a local database that can be used with \command{MetFrag}.

\itemize{
    \item Methods are defined for: \code{\link[=convertToMFDB,transformationProductsStructure-method]{transformationProductsStructure}}.
}

\code{convertToSuspects} Converts an object to a suspect list.

\itemize{
    \item Methods are defined for: \code{\link[=convertToSuspects,MSLibrary-method]{MSLibrary}}; \code{\link[=convertToSuspects,transformationProducts-method]{transformationProducts}}.
}

\code{cutClusters} Returns assigned cluster indices of a cut cluster.

\itemize{
    \item Methods are defined for: \code{\link[=cutClusters,componentsClust-method]{componentsClust}}; \code{\link[=cutClusters,compoundsCluster-method]{compoundsCluster}}.
}

\code{defaultExclNormScores} Returns default scorings that are excluded from normalization.

\itemize{
    \item Methods are defined for: \code{\link[=defaultExclNormScores,compounds-method]{compounds}}; \code{\link[=defaultExclNormScores,formulas-method]{formulas}}.
}

\code{export} exports workflow data to a given format.

\itemize{
    \item Methods are defined for: \code{\link[=export,featureGroups-method]{featureGroups}}; \code{\link[=export,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=export,MSLibrary-method]{MSLibrary}}.
}

\code{featureTable} returns feature information.

\itemize{
    \item Methods are defined for: \code{\link[=featureTable,featureGroups-method]{featureGroups}}; \code{\link[=featureTable,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=featureTable,features-method]{features}}.
}

\code{filter} provides various functionality to do post-filtering of data.

\itemize{
    \item Methods are defined for: \code{\link[=filter,components-method]{components}}; \code{\link[=filter,componentsSet-method]{componentsSet}}; \code{\link[=filter,componentsTPs-method]{componentsTPs}}; \code{\link[=filter,compounds-method]{compounds}}; \code{\link[=filter,compoundsSet-method]{compoundsSet}}; \code{\link[=filter,featureAnnotations-method]{featureAnnotations}}; \code{\link[=filter,featureGroups-method]{featureGroups}}; \code{\link[=filter,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=filter,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=filter,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=filter,features-method]{features}}; \code{\link[=filter,featuresSet-method]{featuresSet}}; \code{\link[=filter,formulasSet-method]{formulasSet}}; \code{\link[=filter,MSLibrary-method]{MSLibrary}}; \code{\link[=filter,MSPeakLists-method]{MSPeakLists}}; \code{\link[=filter,MSPeakListsSet-method]{MSPeakListsSet}}; \code{\link[=filter,transformationProducts-method]{transformationProducts}}; \code{\link[=filter,transformationProductsAnnComp-method]{transformationProductsAnnComp}}; \code{\link[=filter,transformationProductsAnnForm-method]{transformationProductsAnnForm}}; \code{\link[=filter,transformationProductsStructure-method]{transformationProductsStructure}}.
}

\code{fromIMS} returns \code{TRUE} if the object was directly created from IMS data.

\itemize{
    \item Methods are defined for: \code{\link[=fromIMS,featureGroups-method]{featureGroups}}; \code{\link[=fromIMS,features-method]{features}}.
}

\code{getBPCs} gets base peak chromatogram(s).

\itemize{
    \item Methods are defined for: \code{\link[=getBPCs,data.frame-method]{data.frame}}; \code{\link[=getBPCs,featureGroups-method]{featureGroups}}; \code{\link[=getBPCs,features-method]{features}}.
}

\code{getEICs} obtains extracted ion chromatogram(s) (EICs).

\itemize{
    \item Methods are defined for: \code{\link[=getEICs,data.frame-method]{data.frame}}; \code{\link[=getEICs,data.table-method]{data.table}}; \code{\link[=getEICs,featureGroups-method]{featureGroups}}; \code{\link[=getEICs,features-method]{features}}.
}

\code{getFeatures} returns the object's \code{\link{features}} object.

\itemize{
    \item Methods are defined for: \code{\link[=getFeatures,featureGroups-method]{featureGroups}}.
}

\code{getFeatureQualityNames} returns the object's feature quality names.

\itemize{
    \item Methods are defined for: \code{\link[=getFeatureQualityNames,featureGroups-method]{featureGroups}}; \code{\link[=getFeatureQualityNames,features-method]{features}}.
}

\code{getMCS} Calculates the maximum common substructure.

\itemize{
    \item Methods are defined for: \code{\link[=getMCS,compounds-method]{compounds}}; \code{\link[=getMCS,compoundsCluster-method]{compoundsCluster}}.
}

\code{getTICs} gets total ion chromatogram(s).

\itemize{
    \item Methods are defined for: \code{\link[=getTICs,data.frame-method]{data.frame}}; \code{\link[=getTICs,featureGroups-method]{featureGroups}}; \code{\link[=getTICs,features-method]{features}}.
}

\code{groupNames} returns a \code{character} vector with the names of the feature groups for which data is present in this object.

\itemize{
    \item Methods are defined for: \code{\link[=groupNames,components-method]{components}}; \code{\link[=groupNames,compoundsCluster-method]{compoundsCluster}}; \code{\link[=groupNames,featureAnnotations-method]{featureAnnotations}}; \code{\link[=groupNames,featureGroups-method]{featureGroups}}; \code{\link[=groupNames,MSPeakLists-method]{MSPeakLists}}.
}

\code{hasIMS} returns \code{TRUE} if the object has ion mobility values

\itemize{
    \item Methods are defined for: \code{\link[=hasIMS,featureGroups-method]{featureGroups}}; \code{\link[=hasIMS,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=hasIMS,features-method]{features}}.
}

\code{plotBPCs} plots base peak chromatogram(s).

\itemize{
    \item Methods are defined for: \code{\link[=plotBPCs,data.frame-method]{data.frame}}; \code{\link[=plotBPCs,data.table-method]{data.table}}; \code{\link[=plotBPCs,featureGroups-method]{featureGroups}}; \code{\link[=plotBPCs,features-method]{features}}.
}

\code{plotChord} plots a Chord diagram to assess overlapping data.

\itemize{
    \item Methods are defined for: \code{\link[=plotChord,featureGroups-method]{featureGroups}}; \code{\link[=plotChord,featureGroupsComparison-method]{featureGroupsComparison}}.
}

\code{plotChroms} plots extracted ion chromatogram(s).

\itemize{
    \item Methods are defined for: \code{\link[=plotChroms,components-method]{components}}; \code{\link[=plotChroms,featureGroups-method]{featureGroups}}.
}

\code{plotChroms3D} plots a three dimensional chromatogram.

\itemize{
    \item Methods are defined for: \code{\link[=plotChroms3D,featureGroups-method]{featureGroups}}; \code{\link[=plotChroms3D,featureGroupsSet-method]{featureGroupsSet}}.
}

\code{plotGraph} Plots an interactive network graph.

\itemize{
    \item Methods are defined for: \code{\link[=plotGraph,componentsNet-method]{componentsNet}}; \code{\link[=plotGraph,componentsNetSet-method]{componentsNetSet}}; \code{\link[=plotGraph,componentsNT-method]{componentsNT}}; \code{\link[=plotGraph,componentsNTSet-method]{componentsNTSet}}; \code{\link[=plotGraph,componentsTPs-method]{componentsTPs}}; \code{\link[=plotGraph,featureGroups-method]{featureGroups}}; \code{\link[=plotGraph,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=plotGraph,transformationProductsFormula-method]{transformationProductsFormula}}; \code{\link[=plotGraph,transformationProductsStructure-method]{transformationProductsStructure}}.
}

\code{plotInt} plots the intensity of all contained features.

\itemize{
    \item Methods are defined for: \code{\link[=plotInt,componentsIntClust-method]{componentsIntClust}}; \code{\link[=plotInt,featureGroups-method]{featureGroups}}.
}

\code{plotScores} plots candidate scorings.

\itemize{
    \item Methods are defined for: \code{\link[=plotScores,compounds-method]{compounds}}; \code{\link[=plotScores,formulas-method]{formulas}}.
}

\code{plotSilhouettes} plots silhouette widths to evaluate the desired cluster size.

\itemize{
    \item Methods are defined for: \code{\link[=plotSilhouettes,componentsClust-method]{componentsClust}}; \code{\link[=plotSilhouettes,compoundsCluster-method]{compoundsCluster}}.
}

\code{plotSpectrum} plots a (annotated) spectrum.

\itemize{
    \item Methods are defined for: \code{\link[=plotSpectrum,components-method]{components}}; \code{\link[=plotSpectrum,compounds-method]{compounds}}; \code{\link[=plotSpectrum,compoundsSet-method]{compoundsSet}}; \code{\link[=plotSpectrum,formulas-method]{formulas}}; \code{\link[=plotSpectrum,formulasSet-method]{formulasSet}}; \code{\link[=plotSpectrum,MSPeakLists-method]{MSPeakLists}}; \code{\link[=plotSpectrum,MSPeakListsSet-method]{MSPeakListsSet}}.
}

\code{plotStructure} plots a chemical structure.

\itemize{
    \item Methods are defined for: \code{\link[=plotStructure,compounds-method]{compounds}}; \code{\link[=plotStructure,compoundsCluster-method]{compoundsCluster}}.
}

\code{plotTICs} plots total ion chromatogram(s).

\itemize{
    \item Methods are defined for: \code{\link[=plotTICs,data.frame-method]{data.frame}}; \code{\link[=plotTICs,data.table-method]{data.table}}; \code{\link[=plotTICs,featureGroups-method]{featureGroups}}; \code{\link[=plotTICs,features-method]{features}}.
}

\code{plotVenn} plots a Venn diagram to assess unique and overlapping data.

\itemize{
    \item Methods are defined for: \code{\link[=plotVenn,featureAnnotations-method]{featureAnnotations}}; \code{\link[=plotVenn,featureGroups-method]{featureGroups}}; \code{\link[=plotVenn,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=plotVenn,transformationProductsStructure-method]{transformationProductsStructure}}.
}

\code{plotUpSet} plots an UpSet diagram to assess unique and overlapping data.

\itemize{
    \item Methods are defined for: \code{\link[=plotUpSet,featureAnnotations-method]{featureAnnotations}}; \code{\link[=plotUpSet,featureGroups-method]{featureGroups}}; \code{\link[=plotUpSet,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=plotUpSet,transformationProductsStructure-method]{transformationProductsStructure}}.
}

\code{predictRespFactors} Prediction of response factors.

\itemize{
    \item Methods are defined for: \code{\link[=predictRespFactors,compounds-method]{compounds}}; \code{\link[=predictRespFactors,compoundsSet-method]{compoundsSet}}; \code{\link[=predictRespFactors,compoundsSIRIUS-method]{compoundsSIRIUS}}; \code{\link[=predictRespFactors,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=predictRespFactors,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=predictRespFactors,formulasSet-method]{formulasSet}}; \code{\link[=predictRespFactors,formulasSIRIUS-method]{formulasSIRIUS}}.
}

\code{predictTox} Prediction of toxicity values.

\itemize{
    \item Methods are defined for: \code{\link[=predictTox,compounds-method]{compounds}}; \code{\link[=predictTox,compoundsSet-method]{compoundsSet}}; \code{\link[=predictTox,compoundsSIRIUS-method]{compoundsSIRIUS}}; \code{\link[=predictTox,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=predictTox,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=predictTox,formulasSet-method]{formulasSet}}; \code{\link[=predictTox,formulasSIRIUS-method]{formulasSIRIUS}}.
}

\code{delete} Deletes results.

\itemize{
    \item Methods are defined for: \code{\link[=delete,components-method]{components}}; \code{\link[=delete,componentsClust-method]{componentsClust}}; \code{\link[=delete,componentsSet-method]{componentsSet}}; \code{\link[=delete,compoundsSet-method]{compoundsSet}}; \code{\link[=delete,compoundsSIRIUS-method]{compoundsSIRIUS}}; \code{\link[=delete,featureAnnotations-method]{featureAnnotations}}; \code{\link[=delete,featureGroups-method]{featureGroups}}; \code{\link[=delete,featureGroupsKPIC2-method]{featureGroupsKPIC2}}; \code{\link[=delete,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=delete,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=delete,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=delete,featureGroupsXCMS-method]{featureGroupsXCMS}}; \code{\link[=delete,featureGroupsXCMS3-method]{featureGroupsXCMS3}}; \code{\link[=delete,features-method]{features}}; \code{\link[=delete,featuresKPIC2-method]{featuresKPIC2}}; \code{\link[=delete,featuresPiek-method]{featuresPiek}}; \code{\link[=delete,featuresXCMS-method]{featuresXCMS}}; \code{\link[=delete,featuresXCMS3-method]{featuresXCMS3}}; \code{\link[=delete,formulas-method]{formulas}}; \code{\link[=delete,formulasSet-method]{formulasSet}}; \code{\link[=delete,formulasSIRIUS-method]{formulasSIRIUS}}; \code{\link[=delete,MSLibrary-method]{MSLibrary}}; \code{\link[=delete,MSPeakLists-method]{MSPeakLists}}; \code{\link[=delete,MSPeakListsSet-method]{MSPeakListsSet}}; \code{\link[=delete,transformationProducts-method]{transformationProducts}}.
}

\code{plotVolcano} plots a volcano plot.

\itemize{
    \item Methods are defined for: \code{\link[=plotVolcano,featureGroups-method]{featureGroups}}.
}

\code{replicates} returns a \code{character} vector with the replicates for which data is present in this object.

\itemize{
    \item Methods are defined for: \code{\link[=replicates,featureGroups-method]{featureGroups}}; \code{\link[=replicates,features-method]{features}}.
}

\code{setObjects} returns the \emph{set objects} of this object. See the documentation of \code{\link{workflowStepSet}}.

\itemize{
    \item Methods are defined for: \code{\link[=setObjects,workflowStepSet-method]{workflowStepSet}}.
}

\code{sets} returns the names of the sets inside this object. See the documentation for \link[=sets-workflow]{sets workflows}.

\itemize{
    \item Methods are defined for: \code{\link[=sets,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=sets,featuresSet-method]{featuresSet}}; \code{\link[=sets,workflowStepSet-method]{workflowStepSet}}.
}

\code{treeCut} Manually cut a cluster.

\itemize{
    \item Methods are defined for: \code{\link[=treeCut,componentsClust-method]{componentsClust}}; \code{\link[=treeCut,compoundsCluster-method]{compoundsCluster}}.
}

\code{treeCutDynamic} Automatically cut a cluster.

\itemize{
    \item Methods are defined for: \code{\link[=treeCutDynamic,componentsClust-method]{componentsClust}}; \code{\link[=treeCutDynamic,compoundsCluster-method]{compoundsCluster}}.
}

\code{unset} Converts this object to a regular non-set object. See the documentation for \link[=sets-workflow]{sets workflows}.

\itemize{
    \item Methods are defined for: \code{\link[=unset,componentsNetSet-method]{componentsNetSet}}; \code{\link[=unset,componentsNTSet-method]{componentsNTSet}}; \code{\link[=unset,componentsSet-method]{componentsSet}}; \code{\link[=unset,compoundsConsensusSet-method]{compoundsConsensusSet}}; \code{\link[=unset,compoundsSet-method]{compoundsSet}}; \code{\link[=unset,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=unset,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=unset,featuresSet-method]{featuresSet}}; \code{\link[=unset,formulasConsensusSet-method]{formulasConsensusSet}}; \code{\link[=unset,formulasSet-method]{formulasSet}}; \code{\link[=unset,MSPeakListsSet-method]{MSPeakListsSet}}.
}
}
\section{Other generics}{
 Below are methods that are defined for existing
  generics (\emph{e.g.} defined in \code{base}). Please see method specific
  documentation for more details.

 \code{[} Subsets data within an object.

\itemize{
    \item Methods are defined for: \code{\link[=[,components,ANY,ANY,missing-method]{components,ANY,ANY,missing}}; \code{\link[=[,componentsSet,ANY,ANY,missing-method]{componentsSet,ANY,ANY,missing}}; \code{\link[=[,compoundsCluster,ANY,missing,missing-method]{compoundsCluster,ANY,missing,missing}}; \code{\link[=[,compoundsSet,ANY,missing,missing-method]{compoundsSet,ANY,missing,missing}}; \code{\link[=[,featureAnnotations,ANY,missing,missing-method]{featureAnnotations,ANY,missing,missing}}; \code{\link[=[,featureGroups,ANY,ANY,missing-method]{featureGroups,ANY,ANY,missing}}; \code{\link[=[,featureGroupsComparison,ANY,missing,missing-method]{featureGroupsComparison,ANY,missing,missing}}; \code{\link[=[,featureGroupsScreening,ANY,ANY,missing-method]{featureGroupsScreening,ANY,ANY,missing}}; \code{\link[=[,featureGroupsScreeningSet,ANY,ANY,missing-method]{featureGroupsScreeningSet,ANY,ANY,missing}}; \code{\link[=[,featureGroupsSet,ANY,ANY,missing-method]{featureGroupsSet,ANY,ANY,missing}}; \code{\link[=[,features,ANY,missing,missing-method]{features,ANY,missing,missing}}; \code{\link[=[,featuresSet,ANY,missing,missing-method]{featuresSet,ANY,missing,missing}}; \code{\link[=[,formulasSet,ANY,missing,missing-method]{formulasSet,ANY,missing,missing}}; \code{\link[=[,MSLibrary,ANY,missing,missing-method]{MSLibrary,ANY,missing,missing}}; \code{\link[=[,MSPeakLists,ANY,ANY,missing-method]{MSPeakLists,ANY,ANY,missing}}; \code{\link[=[,MSPeakListsSet,ANY,ANY,missing-method]{MSPeakListsSet,ANY,ANY,missing}}; \code{\link[=[,transformationProducts,ANY,missing,missing-method]{transformationProducts,ANY,missing,missing}}.
}

 \code{[[} Extract data from an object.

\itemize{
    \item Methods are defined for: \code{\link[=[[,components,ANY,ANY-method]{components,ANY,ANY}}; \code{\link[=[[,featureAnnotations,ANY,missing-method]{featureAnnotations,ANY,missing}}; \code{\link[=[[,featureGroups,ANY,ANY-method]{featureGroups,ANY,ANY}}; \code{\link[=[[,featureGroupsComparison,ANY,missing-method]{featureGroupsComparison,ANY,missing}}; \code{\link[=[[,features,ANY,missing-method]{features,ANY,missing}}; \code{\link[=[[,formulas,ANY,ANY-method]{formulas,ANY,ANY}}; \code{\link[=[[,MSLibrary,ANY,missing-method]{MSLibrary,ANY,missing}}; \code{\link[=[[,MSPeakLists,ANY,ANY-method]{MSPeakLists,ANY,ANY}}; \code{\link[=[[,transformationProducts,ANY,missing-method]{transformationProducts,ANY,missing}}.
}

 \code{$} Extract data from an object.

\itemize{
    \item Methods are defined for: \code{\link[=$,components-method]{components}}; \code{\link[=$,featureAnnotations-method]{featureAnnotations}}; \code{\link[=$,featureGroups-method]{featureGroups}}; \code{\link[=$,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=$,features-method]{features}}; \code{\link[=$,MSLibrary-method]{MSLibrary}}; \code{\link[=$,MSPeakLists-method]{MSPeakLists}}; \code{\link[=$,transformationProducts-method]{transformationProducts}}.
}

 \code{as.data.table} Converts an object to a table (\code{\link{data.table}}).

\itemize{
    \item Methods are defined for: \code{\link[=as.data.table,components-method]{components}}; \code{\link[=as.data.table,componentsTPs-method]{componentsTPs}}; \code{\link[=as.data.table,featureAnnotations-method]{featureAnnotations}}; \code{\link[=as.data.table,featureGroups-method]{featureGroups}}; \code{\link[=as.data.table,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=as.data.table,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=as.data.table,features-method]{features}}; \code{\link[=as.data.table,featuresSet-method]{featuresSet}}; \code{\link[=as.data.table,formulas-method]{formulas}}; \code{\link[=as.data.table,MSLibrary-method]{MSLibrary}}; \code{\link[=as.data.table,MSPeakLists-method]{MSPeakLists}}; \code{\link[=as.data.table,MSPeakListsSet-method]{MSPeakListsSet}}; \code{\link[=as.data.table,transformationProducts-method]{transformationProducts}}; \code{\link[=as.data.table,workflowStep-method]{workflowStep}}.
}

 \code{as.data.frame} Converts an object to a table (\code{data.frame}).

\itemize{
    \item Methods are defined for: \code{\link[=as.data.frame,workflowStep-method]{workflowStep}}.
}

 \code{length} Returns the length of an object.

\itemize{
    \item Methods are defined for: \code{\link[=length,components-method]{components}}; \code{\link[=length,compoundsCluster-method]{compoundsCluster}}; \code{\link[=length,featureAnnotations-method]{featureAnnotations}}; \code{\link[=length,featureGroups-method]{featureGroups}}; \code{\link[=length,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=length,features-method]{features}}; \code{\link[=length,MSLibrary-method]{MSLibrary}}; \code{\link[=length,MSPeakLists-method]{MSPeakLists}}; \code{\link[=length,optimizationResult-method]{optimizationResult}}; \code{\link[=length,transformationProducts-method]{transformationProducts}}.
}

 \code{lengths} Returns the lengths of elements within this object.

\itemize{
    \item Methods are defined for: \code{\link[=lengths,compoundsCluster-method]{compoundsCluster}}; \code{\link[=lengths,optimizationResult-method]{optimizationResult}}.
}

 \code{names} Return names for this object.

\itemize{
    \item Methods are defined for: \code{\link[=names,components-method]{components}}; \code{\link[=names,featureGroups-method]{featureGroups}}; \code{\link[=names,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=names,MSLibrary-method]{MSLibrary}}; \code{\link[=names,transformationProducts-method]{transformationProducts}}.
}

 \code{plot} Generates a plot for an object.

\itemize{
    \item Methods are defined for: \code{\link[=plot,componentsClust,missing-method]{componentsClust,missing}}; \code{\link[=plot,compoundsCluster,missing-method]{compoundsCluster,missing}}; \code{\link[=plot,featureGroups,missing-method]{featureGroups,missing}}; \code{\link[=plot,featureGroupsComparison,missing-method]{featureGroupsComparison,missing}}; \code{\link[=plot,optimizationResult,missing-method]{optimizationResult,missing}}.
}

 \code{show} Prints information about this object.

\itemize{
    \item Methods are defined for: \code{\link[=show,adduct-method]{adduct}}; \code{\link[=show,components-method]{components}}; \code{\link[=show,componentsFeatures-method]{componentsFeatures}}; \code{\link[=show,componentsSet-method]{componentsSet}}; \code{\link[=show,compounds-method]{compounds}}; \code{\link[=show,compoundsCluster-method]{compoundsCluster}}; \code{\link[=show,compoundsSet-method]{compoundsSet}}; \code{\link[=show,featureGroups-method]{featureGroups}}; \code{\link[=show,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=show,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=show,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=show,features-method]{features}}; \code{\link[=show,featuresSet-method]{featuresSet}}; \code{\link[=show,formulas-method]{formulas}}; \code{\link[=show,formulasSet-method]{formulasSet}}; \code{\link[=show,MSLibrary-method]{MSLibrary}}; \code{\link[=show,MSPeakLists-method]{MSPeakLists}}; \code{\link[=show,MSPeakListsSet-method]{MSPeakListsSet}}; \code{\link[=show,optimizationResult-method]{optimizationResult}}; \code{\link[=show,transformationProducts-method]{transformationProducts}}; \code{\link[=show,workflowStep-method]{workflowStep}}; \code{\link[=show,workflowStepSet-method]{workflowStepSet}}.
}
}

