% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics}
\alias{generics}
\alias{adducts}
\alias{adducts<-}
\alias{algorithm}
\alias{analysisInfo}
\alias{analyses}
\alias{annotatedPeakList}
\alias{annotations}
\alias{calculatePeakQualities}
\alias{clusterProperties}
\alias{clusters}
\alias{consensus}
\alias{convertToMFDB}
\alias{convertToSuspects}
\alias{cutClusters}
\alias{defaultExclNormScores}
\alias{export}
\alias{featureTable}
\alias{featureTable<-}
\alias{filter}
\alias{getFeatures}
\alias{getMCS}
\alias{groupNames}
\alias{makeSet}
\alias{plotChord}
\alias{plotChroms}
\alias{plotGraph}
\alias{plotInt}
\alias{plotScores}
\alias{plotSilhouettes}
\alias{plotSpectrum}
\alias{plotStructure}
\alias{plotVenn}
\alias{plotUpSet}
\alias{delete}
\alias{plotVolcano}
\alias{replicateGroups}
\alias{treeCut}
\alias{treeCutDynamic}
\alias{[}
\alias{[[}
\alias{$}
\alias{as.data.table}
\alias{as.data.frame}
\alias{length}
\alias{lengths}
\alias{names}
\alias{plot}
\alias{show}
\title{Miscellaneous generics}
\usage{
adducts(obj, ...)

adducts(obj, ...) <- value

algorithm(obj)

analysisInfo(obj)

analyses(obj)

annotatedPeakList(obj, ...)

annotations(obj, ...)

calculatePeakQualities(obj, weights = NULL, flatnessFactor = 0.05, ...)

clusterProperties(obj)

clusters(obj)

consensus(obj, ...)

convertToMFDB(TPs, out, ...)

convertToSuspects(TPs, ...)

cutClusters(obj)

defaultExclNormScores(obj)

export(obj, type, out, ...)

featureTable(obj, ...)

featureTable(obj) <- value

filter(obj, ...)

getFeatures(obj)

getMCS(obj, ...)

groupNames(obj)

makeSet(obj, ...)

plotChord(obj, addSelfLinks = FALSE, addRetMzPlots = TRUE, ...)

plotChroms(obj, ...)

plotGraph(obj, onlyLinked = TRUE, ...)

plotInt(obj, ...)

plotScores(obj, ...)

plotSilhouettes(obj, kSeq, ...)

plotSpectrum(obj, ...)

plotStructure(obj, ...)

plotVenn(obj, ...)

plotUpSet(obj, ...)

delete(obj, ...)

plotVolcano(obj, ...)

replicateGroups(obj)

treeCut(obj, k = NULL, h = NULL, ...)

treeCutDynamic(
  obj,
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1,
  ...
)
}
\arguments{
\item{obj}{The object the generic should be applied to.}

\item{\dots}{Any further method specific arguments. See method documentation
for details.}

\item{weights, flatnessFactor}{See method documentation.}

\item{addSelfLinks}{If \code{TRUE} then 'self-links' are added which
represent non-shared data.}

\item{addRetMzPlots}{Set to \code{TRUE} to enable \emph{m/z} \emph{vs}
retention time scatter plots.}

\item{onlyLinked}{Only plots linked objects if \code{TRUE}.}

\item{kSeq}{An integer vector containing the sequence that should be used for
average silhouette width calculation.}

\item{k, h}{Desired numbers of clusters. See \code{\link{cutree}}.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}

\item{fTable}{New value for the \code{features} slot: a \code{list} with \code{data.table} items for each analysis.}
}
\description{
Various (S4) generic functions providing a common interface for common tasks
such as plotting and filtering data. The actual functionality and function
arguments are often specific for the implemented methods, for this reason,
please refer to the linked method documentation for each generic.
}
\details{
\code{adducts} returns assigned adducts of the object.

\itemize{
    \item Methods are defined for: \code{\link[=adducts,featureGroups-method]{featureGroups}}; \code{\link[=adducts,featureGroupsSet-method]{featureGroupsSet}}.
}

\code{adducts<-} sets adducts of the object.

\itemize{
    \item Methods are defined for: \code{\link[=adducts<-,featureGroups-method]{featureGroups}}; \code{\link[=adducts<-,featureGroupsSet-method]{featureGroupsSet}}.
}

\code{algorithm} returns the algorithm that was used to generate the object.

\itemize{
    \item Methods are defined for: \code{\link[=algorithm,optimizationResult-method]{optimizationResult}}; \code{\link[=algorithm,workflowStep-method]{workflowStep}}.
}

\code{analysisInfo} returns the \link[=analysis-information]{analysis information} from an object.

\itemize{
    \item Methods are defined for: \code{\link[=analysisInfo,featureGroups-method]{featureGroups}}; \code{\link[=analysisInfo,features-method]{features}}; \code{\link[=analysisInfo,MSPeakListsSet-method]{MSPeakListsSet}}.
}

\code{analyses} returns a \code{character} vector with the analyses for which data is present in this object.

\itemize{
    \item Methods are defined for: \code{\link[=analyses,featureGroups-method]{featureGroups}}; \code{\link[=analyses,features-method]{features}}; \code{\link[=analyses,formulas-method]{formulas}}; \code{\link[=analyses,MSPeakLists-method]{MSPeakLists}}.
}

\code{annotatedPeakList} returns an annotated MS peak list.

\itemize{
    \item Methods are defined for: \code{\link[=annotatedPeakList,compounds-method]{compounds}}; \code{\link[=annotatedPeakList,compoundsSet-method]{compoundsSet}}; \code{\link[=annotatedPeakList,formulas-method]{formulas}}; \code{\link[=annotatedPeakList,formulasSet-method]{formulasSet}}.
}

\code{annotations} returns annotations.

\itemize{
    \item Methods are defined for: \code{\link[=annotations,featureAnnotations-method]{featureAnnotations}}; \code{\link[=annotations,featureGroups-method]{featureGroups}}; \code{\link[=annotations,formulas-method]{formulas}}.
}

\code{calculatePeakQualities} calculates chromatographic peak qualities and scores.

\itemize{
    \item Methods are defined for: \code{\link[=calculatePeakQualities,featureGroups-method]{featureGroups}}; \code{\link[=calculatePeakQualities,features-method]{features}}.
}

\code{clusterProperties} Obtain a list with properties of the generated cluster(s).

\itemize{
    \item Methods are defined for: \code{\link[=clusterProperties,componentsClust-method]{componentsClust}}; \code{\link[=clusterProperties,compoundsCluster-method]{compoundsCluster}}.
}

\code{clusters} Obtain clustering object(s).

\itemize{
    \item Methods are defined for: \code{\link[=clusters,componentsClust-method]{componentsClust}}; \code{\link[=clusters,compoundsCluster-method]{compoundsCluster}}.
}

\code{consensus} combines and merges data from various algorithms to generate a consensus.

\itemize{
    \item Methods are defined for: \code{\link[=consensus,components-method]{components}}; \code{\link[=consensus,componentsSet-method]{componentsSet}}; \code{\link[=consensus,compounds-method]{compounds}}; \code{\link[=consensus,compoundsSet-method]{compoundsSet}}; \code{\link[=consensus,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=consensus,featureGroupsComparisonSet-method]{featureGroupsComparisonSet}}; \code{\link[=consensus,formulas-method]{formulas}}; \code{\link[=consensus,formulasSet-method]{formulasSet}}.
}

\code{convertToMFDB} Exports the object to a local database that can be used with \command{MetFrag}.

\itemize{
    \item Methods are defined for: .
}

\code{convertToSuspects} Converts an object to a suspect list.

\itemize{
    \item Methods are defined for: .
}

\code{cutClusters} Returns assigned cluster indices of a cut cluster.

\itemize{
    \item Methods are defined for: \code{\link[=cutClusters,componentsClust-method]{componentsClust}}; \code{\link[=cutClusters,compoundsCluster-method]{compoundsCluster}}.
}

\code{defaultExclNormScores} Returns default scorings that are excluded from normalization.

\itemize{
    \item Methods are defined for: \code{\link[=defaultExclNormScores,compounds-method]{compounds}}; \code{\link[=defaultExclNormScores,formulas-method]{formulas}}.
}

\code{export} exports workflow data to a given format.

\itemize{
    \item Methods are defined for: \code{\link[=export,featureGroups-method]{featureGroups}}; \code{\link[=export,featureGroupsSet-method]{featureGroupsSet}}.
}

\code{featureTable} returns feature information.

\itemize{
    \item Methods are defined for: \code{\link[=featureTable,featureGroups-method]{featureGroups}}; \code{\link[=featureTable,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=featureTable,features-method]{features}}.
}

\code{featureTable} updates feature information.

\itemize{
    \item Methods are defined for: \code{\link[=featureTable,featureGroups-method]{featureGroups}}; \code{\link[=featureTable,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=featureTable,features-method]{features}}.
}

\code{filter} provides various functionality to do post-filtering of data.

\itemize{
    \item Methods are defined for: \code{\link[=filter,components-method]{components}}; \code{\link[=filter,componentsSet-method]{componentsSet}}; \code{\link[=filter,componentsTPs-method]{componentsTPs}}; \code{\link[=filter,compounds-method]{compounds}}; \code{\link[=filter,compoundsSet-method]{compoundsSet}}; \code{\link[=filter,featureAnnotations-method]{featureAnnotations}}; \code{\link[=filter,featureGroups-method]{featureGroups}}; \code{\link[=filter,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=filter,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=filter,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=filter,features-method]{features}}; \code{\link[=filter,featuresSet-method]{featuresSet}}; \code{\link[=filter,formulasSet-method]{formulasSet}}; \code{\link[=filter,MSPeakLists-method]{MSPeakLists}}; \code{\link[=filter,MSPeakListsSet-method]{MSPeakListsSet}}; \code{\link[=filter,transformationProductsBT-method]{transformationProductsBT}}.
}

\code{getFeatures} returns the object's \code{\link{features}} object.

\itemize{
    \item Methods are defined for: \code{\link[=getFeatures,featureGroups-method]{featureGroups}}.
}

\code{getMCS} Calculcates the maximum common substructure.

\itemize{
    \item Methods are defined for: \code{\link[=getMCS,compounds-method]{compounds}}; \code{\link[=getMCS,compoundsCluster-method]{compoundsCluster}}.
}

\code{groupNames} returns a \code{character} vector with the names of the feature groups for which data is present in this object.

\itemize{
    \item Methods are defined for: \code{\link[=groupNames,components-method]{components}}; \code{\link[=groupNames,compoundsCluster-method]{compoundsCluster}}; \code{\link[=groupNames,featureAnnotations-method]{featureAnnotations}}; \code{\link[=groupNames,featureGroups-method]{featureGroups}}; \code{\link[=groupNames,MSPeakLists-method]{MSPeakLists}}.
}

\code{makeSet} creates a set from given workflow objects.

\itemize{
    \item Methods are defined for: \code{\link[=makeSet,featureGroups-method]{featureGroups}}; \code{\link[=makeSet,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=makeSet,features-method]{features}}.
}

\code{plotChord} plots a Chord diagram to assess overlapping data.

\itemize{
    \item Methods are defined for: \code{\link[=plotChord,featureGroups-method]{featureGroups}}; \code{\link[=plotChord,featureGroupsComparison-method]{featureGroupsComparison}}.
}

\code{plotChroms} plots extracted ion chromatogram(s).

\itemize{
    \item Methods are defined for: \code{\link[=plotChroms,components-method]{components}}; \code{\link[=plotChroms,featureGroups-method]{featureGroups}}.
}

\code{plotGraph} Plots an interactive network graph.

\itemize{
    \item Methods are defined for: \code{\link[=plotGraph,componentsNT-method]{componentsNT}}; \code{\link[=plotGraph,componentsNTSet-method]{componentsNTSet}}; \code{\link[=plotGraph,componentsTPs-method]{componentsTPs}}.
}

\code{plotInt} plots the intensity of all contained features.

\itemize{
    \item Methods are defined for: \code{\link[=plotInt,componentsIntClust-method]{componentsIntClust}}; \code{\link[=plotInt,featureGroups-method]{featureGroups}}; \code{\link[=plotInt,featureGroupsSet-method]{featureGroupsSet}}.
}

\code{plotScores} plots candidate scorings.

\itemize{
    \item Methods are defined for: \code{\link[=plotScores,compounds-method]{compounds}}; \code{\link[=plotScores,formulas-method]{formulas}}.
}

\code{plotSilhouettes} plots silhouette widths to evaluate the desired cluster size.

\itemize{
    \item Methods are defined for: \code{\link[=plotSilhouettes,componentsClust-method]{componentsClust}}; \code{\link[=plotSilhouettes,compoundsCluster-method]{compoundsCluster}}.
}

\code{plotSpectrum} plots a (annotated) spectrum.

\itemize{
    \item Methods are defined for: \code{\link[=plotSpectrum,components-method]{components}}; \code{\link[=plotSpectrum,compounds-method]{compounds}}; \code{\link[=plotSpectrum,compoundsSet-method]{compoundsSet}}; \code{\link[=plotSpectrum,formulas-method]{formulas}}; \code{\link[=plotSpectrum,formulasSet-method]{formulasSet}}; \code{\link[=plotSpectrum,MSPeakLists-method]{MSPeakLists}}; \code{\link[=plotSpectrum,MSPeakListsSet-method]{MSPeakListsSet}}.
}

\code{plotStructure} plots a chemical structure.

\itemize{
    \item Methods are defined for: \code{\link[=plotStructure,compounds-method]{compounds}}; \code{\link[=plotStructure,compoundsCluster-method]{compoundsCluster}}.
}

\code{plotVenn} plots a Venn diagram to assess unique and overlapping data.

\itemize{
    \item Methods are defined for: \code{\link[=plotVenn,featureAnnotations-method]{featureAnnotations}}; \code{\link[=plotVenn,featureGroups-method]{featureGroups}}; \code{\link[=plotVenn,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=plotVenn,featureGroupsSet-method]{featureGroupsSet}}.
}

\code{plotUpSet} plots an UpSet diagram to assess unique and overlapping data.

\itemize{
    \item Methods are defined for: \code{\link[=plotUpSet,featureAnnotations-method]{featureAnnotations}}; \code{\link[=plotUpSet,featureGroups-method]{featureGroups}}; \code{\link[=plotUpSet,featureGroupsComparison-method]{featureGroupsComparison}}.
}

\code{delete} Deletes results.

\itemize{
    \item Methods are defined for: \code{\link[=delete,components-method]{components}}; \code{\link[=delete,componentsClust-method]{componentsClust}}; \code{\link[=delete,componentsSet-method]{componentsSet}}; \code{\link[=delete,compoundsSet-method]{compoundsSet}}; \code{\link[=delete,featureAnnotations-method]{featureAnnotations}}; \code{\link[=delete,featureGroups-method]{featureGroups}}; \code{\link[=delete,featureGroupsKPIC2-method]{featureGroupsKPIC2}}; \code{\link[=delete,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=delete,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=delete,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=delete,featureGroupsXCMS-method]{featureGroupsXCMS}}; \code{\link[=delete,featureGroupsXCMS3-method]{featureGroupsXCMS3}}; \code{\link[=delete,features-method]{features}}; \code{\link[=delete,featuresKPIC2-method]{featuresKPIC2}}; \code{\link[=delete,featuresXCMS-method]{featuresXCMS}}; \code{\link[=delete,featuresXCMS3-method]{featuresXCMS3}}; \code{\link[=delete,formulas-method]{formulas}}; \code{\link[=delete,formulasSet-method]{formulasSet}}.
}

\code{plotVolcano} plots a volcano plot.

\itemize{
    \item Methods are defined for: \code{\link[=plotVolcano,featureGroups-method]{featureGroups}}.
}

\code{replicateGroups} returns a \code{character} vector with the analyses for which data is present in this object.

\itemize{
    \item Methods are defined for: \code{\link[=replicateGroups,featureGroups-method]{featureGroups}}; \code{\link[=replicateGroups,features-method]{features}}.
}

\code{treeCut} Manually cut a cluster.

\itemize{
    \item Methods are defined for: \code{\link[=treeCut,componentsClust-method]{componentsClust}}; \code{\link[=treeCut,compoundsCluster-method]{compoundsCluster}}.
}

\code{treeCutDynamic} Automatically cut a cluster.

\itemize{
    \item Methods are defined for: \code{\link[=treeCutDynamic,componentsClust-method]{componentsClust}}; \code{\link[=treeCutDynamic,compoundsCluster-method]{compoundsCluster}}.
}
}
\section{Other generics}{
 Below are methods that are defined for existing
  generics (\emph{e.g.} defined in \code{base}). Please see method specific
  documentation for more details.

 \code{[} Subsets data within an object.

\itemize{
    \item Methods are defined for: \code{\link[=[,components,ANY,ANY,missing-method]{components,ANY,ANY,missing}}; \code{\link[=[,componentsSet,ANY,ANY,missing-method]{componentsSet,ANY,ANY,missing}}; \code{\link[=[,compoundsCluster,ANY,missing,missing-method]{compoundsCluster,ANY,missing,missing}}; \code{\link[=[,compoundsSet,ANY,ANY,ANY-method]{compoundsSet,ANY,ANY,ANY}}; \code{\link[=[,featureAnnotations,ANY,missing,missing-method]{featureAnnotations,ANY,missing,missing}}; \code{\link[=[,featureGroups,ANY,ANY,missing-method]{featureGroups,ANY,ANY,missing}}; \code{\link[=[,featureGroupsComparison,ANY,missing,missing-method]{featureGroupsComparison,ANY,missing,missing}}; \code{\link[=[,featureGroupsScreening,ANY,ANY,missing-method]{featureGroupsScreening,ANY,ANY,missing}}; \code{\link[=[,featureGroupsScreeningSet,ANY,ANY,missing-method]{featureGroupsScreeningSet,ANY,ANY,missing}}; \code{\link[=[,featureGroupsSet,ANY,ANY,missing-method]{featureGroupsSet,ANY,ANY,missing}}; \code{\link[=[,features,ANY,missing,missing-method]{features,ANY,missing,missing}}; \code{\link[=[,featuresSet,ANY,missing,missing-method]{featuresSet,ANY,missing,missing}}; \code{\link[=[,formulasSet,ANY,ANY,ANY-method]{formulasSet,ANY,ANY,ANY}}; \code{\link[=[,MSPeakLists,ANY,ANY,missing-method]{MSPeakLists,ANY,ANY,missing}}; \code{\link[=[,MSPeakListsSet,ANY,ANY,missing-method]{MSPeakListsSet,ANY,ANY,missing}}; \code{\link[=[,transformationProducts,ANY,missing,missing-method]{transformationProducts,ANY,missing,missing}}.
}

 \code{[[} Extract data from an object.

\itemize{
    \item Methods are defined for: \code{\link[=[[,components,ANY,ANY-method]{components,ANY,ANY}}; \code{\link[=[[,featureAnnotations,ANY,missing-method]{featureAnnotations,ANY,missing}}; \code{\link[=[[,featureGroups,ANY,ANY-method]{featureGroups,ANY,ANY}}; \code{\link[=[[,featureGroupsComparison,ANY,missing-method]{featureGroupsComparison,ANY,missing}}; \code{\link[=[[,features,ANY,missing-method]{features,ANY,missing}}; \code{\link[=[[,formulas,ANY,ANY-method]{formulas,ANY,ANY}}; \code{\link[=[[,MSPeakLists,ANY,ANY-method]{MSPeakLists,ANY,ANY}}; \code{\link[=[[,transformationProducts,ANY,missing-method]{transformationProducts,ANY,missing}}.
}

 \code{$} Extract data from an object.

\itemize{
    \item Methods are defined for: \code{\link[=$,components-method]{components}}; \code{\link[=$,featureAnnotations-method]{featureAnnotations}}; \code{\link[=$,featureGroups-method]{featureGroups}}; \code{\link[=$,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=$,features-method]{features}}; \code{\link[=$,MSPeakLists-method]{MSPeakLists}}; \code{\link[=$,transformationProducts-method]{transformationProducts}}.
}

 \code{as.data.table} Converts an object to a table (\code{\link{data.table}}).

\itemize{
    \item Methods are defined for: \code{\link[=as.data.table,components-method]{components}}; \code{\link[=as.data.table,componentsTPs-method]{componentsTPs}}; \code{\link[=as.data.table,featureAnnotations-method]{featureAnnotations}}; \code{\link[=as.data.table,featureGroups-method]{featureGroups}}; \code{\link[=as.data.table,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=as.data.table,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=as.data.table,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=as.data.table,features-method]{features}}; \code{\link[=as.data.table,featuresSet-method]{featuresSet}}; \code{\link[=as.data.table,formulas-method]{formulas}}; \code{\link[=as.data.table,MSPeakLists-method]{MSPeakLists}}; \code{\link[=as.data.table,MSPeakListsSet-method]{MSPeakListsSet}}; \code{\link[=as.data.table,transformationProducts-method]{transformationProducts}}; \code{\link[=as.data.table,workflowStep-method]{workflowStep}}.
}

 \code{as.data.frame} Converts an object to a table (\code{data.frame}).

\itemize{
    \item Methods are defined for: \code{\link[=as.data.frame,workflowStep-method]{workflowStep}}.
}

 \code{length} Returns the length of an object.

\itemize{
    \item Methods are defined for: \code{\link[=length,components-method]{components}}; \code{\link[=length,compoundsCluster-method]{compoundsCluster}}; \code{\link[=length,featureAnnotations-method]{featureAnnotations}}; \code{\link[=length,featureGroups-method]{featureGroups}}; \code{\link[=length,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=length,features-method]{features}}; \code{\link[=length,MSPeakLists-method]{MSPeakLists}}; \code{\link[=length,optimizationResult-method]{optimizationResult}}; \code{\link[=length,transformationProducts-method]{transformationProducts}}.
}

 \code{lengths} Returns the lengths of elements within this object.

\itemize{
    \item Methods are defined for: \code{\link[=lengths,compoundsCluster-method]{compoundsCluster}}; \code{\link[=lengths,optimizationResult-method]{optimizationResult}}.
}

 \code{names} Return names for this object.

\itemize{
    \item Methods are defined for: \code{\link[=names,components-method]{components}}; \code{\link[=names,featureGroups-method]{featureGroups}}; \code{\link[=names,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=names,transformationProducts-method]{transformationProducts}}.
}

 \code{plot} Generates a plot for an object.

\itemize{
    \item Methods are defined for: \code{\link[=plot,componentsClust,ANY-method]{componentsClust,ANY}}; \code{\link[=plot,compoundsCluster,ANY-method]{compoundsCluster,ANY}}; \code{\link[=plot,featureGroups,ANY-method]{featureGroups,ANY}}; \code{\link[=plot,featureGroupsComparison,ANY-method]{featureGroupsComparison,ANY}}; \code{\link[=plot,optimizationResult,ANY-method]{optimizationResult,ANY}}.
}

 \code{show} Prints information about this object.

\itemize{
    \item Methods are defined for: \code{\link[=show,adduct-method]{adduct}}; \code{\link[=show,components-method]{components}}; \code{\link[=show,componentsFeatures-method]{componentsFeatures}}; \code{\link[=show,componentsSet-method]{componentsSet}}; \code{\link[=show,compounds-method]{compounds}}; \code{\link[=show,compoundsCluster-method]{compoundsCluster}}; \code{\link[=show,compoundsSet-method]{compoundsSet}}; \code{\link[=show,featureGroups-method]{featureGroups}}; \code{\link[=show,featureGroupsScreening-method]{featureGroupsScreening}}; \code{\link[=show,featureGroupsScreeningSet-method]{featureGroupsScreeningSet}}; \code{\link[=show,featureGroupsSet-method]{featureGroupsSet}}; \code{\link[=show,features-method]{features}}; \code{\link[=show,featuresSet-method]{featuresSet}}; \code{\link[=show,formulas-method]{formulas}}; \code{\link[=show,formulasSet-method]{formulasSet}}; \code{\link[=show,MSPeakLists-method]{MSPeakLists}}; \code{\link[=show,MSPeakListsSet-method]{MSPeakListsSet}}; \code{\link[=show,optimizationResult-method]{optimizationResult}}; \code{\link[=show,transformationProducts-method]{transformationProducts}}; \code{\link[=show,workflowStep-method]{workflowStep}}; \code{\link[=show,workflowStepSet-method]{workflowStepSet}}.
}
}

