% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics}
\alias{generics}
\alias{analysisInfo}
\alias{analyses}
\alias{groupNames}
\alias{algorithm}
\alias{clusterProperties}
\alias{clusters}
\alias{cutClusters}
\alias{consensus}
\alias{featureTable}
\alias{filter}
\alias{formulaTable}
\alias{getMCS}
\alias{plotChord}
\alias{plotEIC}
\alias{plotInt}
\alias{plotSpec}
\alias{plotStructure}
\alias{plotVenn}
\alias{treeCut}
\alias{treeCutDynamic}
\title{Miscellaneous generics}
\usage{
analysisInfo(obj)

analyses(obj)

groupNames(obj)

algorithm(obj)

clusterProperties(obj)

clusters(obj)

cutClusters(obj)

consensus(obj, ...)

featureTable(obj)

filter(obj, ...)

formulaTable(obj)

getMCS(obj, ...)

plotChord(obj, addSelfLinks = FALSE, addRetMzPlots = TRUE, ...)

plotEIC(obj, ...)

plotInt(obj, ...)

plotSpec(obj, ...)

plotStructure(obj, ...)

plotVenn(obj, which = NULL, ...)

treeCut(obj, k = NULL, h = NULL, ...)

treeCutDynamic(obj, maxTreeHeight = 1, deepSplit = TRUE,
  minModuleSize = 1, ...)
}
\arguments{
\item{obj}{The object the generic should be applied to.}

\item{\dots}{Any further method specific arguments. See method documentation
for details.}

\item{addSelfLinks}{If \code{TRUE} then 'self-links' are added which
represent non-shared data.}

\item{addRetMzPlots}{Set to \code{TRUE} to enable \emph{m/z} \emph{vs}
retention time scatter plots.}

\item{which}{What should be plotted. See method documentation for specifics.}

\item{k, h}{Desired numbers of clusters. See \code{\link{cutree}}.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}
}
\description{
Various (S4) generic functions provide a common interface for common tasks
such as plotting and filtering data. The actual functionality and function
arguments are often specific for the implemented methods, for this reason,
please refer to the linked method documentation for each generic.
}
\details{
\code{analysisInfo} returns the \link[=analysis-information]{analysis information} from an object.






\itemize{
    \item Methods are defined for: \code{\link[=analysisInfo,featureGroups-method]{featureGroups}}, \code{\link[=analysisInfo,features-method]{features}}.
}

\code{analyses} returns a \code{character} vector with the analyses from






\itemize{
    \item Methods are defined for: \code{\link[=analyses,MSPeakLists-method]{MSPeakLists}}, \code{\link[=analyses,featureGroups-method]{featureGroups}}, \code{\link[=analyses,features-method]{features}}, \code{\link[=analyses,formulas-method]{formulas}}.
}

\code{groupNames} returns a \code{character} vector with the names of the






\itemize{
    \item Methods are defined for: \code{\link[=groupNames,MSPeakLists-method]{MSPeakLists}}, \code{\link[=groupNames,components-method]{components}}, \code{\link[=groupNames,compounds-method]{compounds}}, \code{\link[=groupNames,compoundsCluster-method]{compoundsCluster}}, \code{\link[=groupNames,featureGroups-method]{featureGroups}}, \code{\link[=groupNames,formulaConsensus-method]{formulaConsensus}}, \code{\link[=groupNames,formulas-method]{formulas}}.
}

\code{algorithm} returns the algorithm that was used to generate an object.






\itemize{
    \item Methods are defined for: \code{\link[=algorithm,MSPeakLists-method]{MSPeakLists}}, \code{\link[=algorithm,components-method]{components}}, \code{\link[=algorithm,compounds-method]{compounds}}, \code{\link[=algorithm,formulaConsensus-method]{formulaConsensus}}, \code{\link[=algorithm,formulas-method]{formulas}}.
}

\code{clusterProperties} Obtain a list with properties of the generated cluster(s).






\itemize{
    \item Methods are defined for: \code{\link[=clusterProperties,componentsIntClust-method]{componentsIntClust}}, \code{\link[=clusterProperties,compoundsCluster-method]{compoundsCluster}}.
}

\code{clusters} Obtain clustering object(s).






\itemize{
    \item Methods are defined for: \code{\link[=clusters,componentsIntClust-method]{componentsIntClust}}, \code{\link[=clusters,compoundsCluster-method]{compoundsCluster}}.
}

\code{cutClusters} Returns assigned cluster indices of a cut cluster.






\itemize{
    \item Methods are defined for: \code{\link[=cutClusters,componentsIntClust-method]{componentsIntClust}}, \code{\link[=cutClusters,compoundsCluster-method]{compoundsCluster}}.
}

\code{consensus} combines and merges data from various algorithms to generate a consensus.






\itemize{
    \item Methods are defined for: \code{\link[=consensus,components-method]{components}}, \code{\link[=consensus,compounds-method]{compounds}}, \code{\link[=consensus,featureGroupsComparison-method]{featureGroupsComparison}}, \code{\link[=consensus,formulas-method]{formulas}}.
}

\code{featureTable} returns feature information.






\itemize{
    \item Methods are defined for: \code{\link[=featureTable,featureGroups-method]{featureGroups}}, \code{\link[=featureTable,features-method]{features}}.
}

\code{filter} provides various functionality to do post-filtering of data.






\itemize{
    \item Methods are defined for: \code{\link[=filter,MSPeakLists-method]{MSPeakLists}}, \code{\link[=filter,compounds-method]{compounds}}, \code{\link[=filter,featureGroups-method]{featureGroups}}, \code{\link[=filter,features-method]{features}}.
}

\code{formulaTable} returns formula data.






\itemize{
    \item Methods are defined for: \code{\link[=formulaTable,formulaConsensus-method]{formulaConsensus}}, \code{\link[=formulaTable,formulas-method]{formulas}}.
}

\code{getMCS} Calculcates the maximum common substructure.






\itemize{
    \item Methods are defined for: \code{\link[=getMCS,compounds-method]{compounds}}, \code{\link[=getMCS,compoundsCluster-method]{compoundsCluster}}.
}

\code{plotChord} plots a Chord diagram to assess overlapping data.






\itemize{
    \item Methods are defined for: \code{\link[=plotChord,featureGroups-method]{featureGroups}}, \code{\link[=plotChord,featureGroupsComparison-method]{featureGroupsComparison}}.
}

\code{plotEIC} plots extracted ion chromatogram(s).






\itemize{
    \item Methods are defined for: \code{\link[=plotEIC,components-method]{components}}, \code{\link[=plotEIC,featureGroups-method]{featureGroups}}.
}

\code{plotInt} plots the intensity of all contained features.






\itemize{
    \item Methods are defined for: \code{\link[=plotInt,componentsIntClust-method]{componentsIntClust}}, \code{\link[=plotInt,featureGroups-method]{featureGroups}}.
}

\code{plotSpec} plots a (annotated) spectrum.






\itemize{
    \item Methods are defined for: \code{\link[=plotSpec,components-method]{components}}, \code{\link[=plotSpec,compounds-method]{compounds}}, \code{\link[=plotSpec,formulaConsensus-method]{formulaConsensus}}.
}

\code{plotStructure} plots a chemical structure.






\itemize{
    \item Methods are defined for: \code{\link[=plotStructure,compounds-method]{compounds}}, \code{\link[=plotStructure,compoundsCluster-method]{compoundsCluster}}.
}

\code{plotVenn} plots a Venn diagram to assess unique and overlapping data.






\itemize{
    \item Methods are defined for: \code{\link[=plotVenn,featureGroups-method]{featureGroups}}, \code{\link[=plotVenn,featureGroupsComparison-method]{featureGroupsComparison}}.
}

\code{treeCut} Manually cut a cluster.






\itemize{
    \item Methods are defined for: \code{\link[=treeCut,componentsIntClust-method]{componentsIntClust}}, \code{\link[=treeCut,compoundsCluster-method]{compoundsCluster}}.
}

\code{treeCutDynamic} Automatically cut a cluster.






\itemize{
    \item Methods are defined for: \code{\link[=treeCutDynamic,componentsIntClust-method]{componentsIntClust}}, \code{\link[=treeCutDynamic,compoundsCluster-method]{compoundsCluster}}.
}
}
