% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics}
\alias{generics}
\alias{analysisInfo}
\alias{featureTable}
\alias{algorithm}
\alias{filter}
\alias{plotInt}
\alias{plotVenn}
\alias{plotChord}
\alias{plotSpec}
\alias{plotEIC}
\alias{consensus}
\alias{formulaTable}
\title{Miscellaneous generics}
\usage{
analysisInfo(obj)

featureTable(obj)

algorithm(obj)

filter(obj, ...)

plotInt(obj, ...)

plotVenn(obj, which = NULL, ...)

plotChord(obj, addSelfLinks = FALSE, addRetMzPlots = TRUE, ...)

plotSpec(obj, ...)

plotEIC(obj, ...)

consensus(obj, ...)

formulaTable(obj)
}
\arguments{
\item{obj}{The object the generic should be applied to.}

\item{\dots}{Any further method specific arguments. See method documentation
for details.}

\item{which}{What should be plotted. See method documentation for specifics.}

\item{addSelfLinks}{If \code{TRUE} then 'self-links' are added which
represent non-shared data.}

\item{addRetMzPlots}{Set to \code{TRUE} to enable \emph{m/z} \emph{vs}
retention time scatter plots.}
}
\description{
Various (S4) generic functions provide a common interface for common tasks
such as plotting and filtering data. The actual functionality and function
arguments are often specific for the implemented methods, for this reason,
please refer to the linked method documentation for each generic.
}
\details{
\code{analysisInfo} returns the \link[=analysis-information]{analysis information} from an object.





\itemize{
    \item Methods are defined for: \code{\link[=analysisInfo,featureGroups-method]{featureGroups}}, \code{\link[=analysisInfo,features-method]{features}}.
}

\code{featureTable} returns feature information.





\itemize{
    \item Methods are defined for: \code{\link[=featureTable,featureGroups-method]{featureGroups}}, \code{\link[=featureTable,features-method]{features}}.
}

\code{algorithm} returns the algorithm that was used to generate an object.





\itemize{
    \item Methods are defined for: \code{\link[=algorithm,MSPeakLists-method]{MSPeakLists}}, \code{\link[=algorithm,components-method]{components}}, \code{\link[=algorithm,compounds-method]{compounds}}, \code{\link[=algorithm,formulaConsensus-method]{formulaConsensus}}, \code{\link[=algorithm,formulas-method]{formulas}}.
}

\code{filter} provides various functionality to do post-filtering of data.





\itemize{
    \item Methods are defined for: \code{\link[=filter,MSPeakLists-method]{MSPeakLists}}, \code{\link[=filter,compounds-method]{compounds}}, \code{\link[=filter,featureGroups-method]{featureGroups}}.
}

\code{plotInt} plots the intensity of all contained features.





\itemize{
    \item Methods are defined for: \code{\link[=plotInt,clusterInfo-method]{clusterInfo}}, \code{\link[=plotInt,featureGroups-method]{featureGroups}}.
}

\code{plotVenn} plots a Venn diagram to assess unique and overlapping data.





\itemize{
    \item Methods are defined for: \code{\link[=plotVenn,featureGroups-method]{featureGroups}}, \code{\link[=plotVenn,featureGroupsComparison-method]{featureGroupsComparison}}.
}

\code{plotChord} plots a Chord diagram to assess overlapping data.





\itemize{
    \item Methods are defined for: \code{\link[=plotChord,featureGroups-method]{featureGroups}}, \code{\link[=plotChord,featureGroupsComparison-method]{featureGroupsComparison}}.
}

\code{plotSpec} plots a (annotated) spectrum.





\itemize{
    \item Methods are defined for: \code{\link[=plotSpec,components-method]{components}}, \code{\link[=plotSpec,compounds-method]{compounds}}, \code{\link[=plotSpec,formulaConsensus-method]{formulaConsensus}}.
}

\code{plotEIC} plots extracted ion chromatogram(s).





\itemize{
    \item Methods are defined for: \code{\link[=plotEIC,components-method]{components}}, \code{\link[=plotEIC,featureGroups-method]{featureGroups}}.
}

\code{consensus} combines and merges data from various algorithms to





\itemize{
    \item Methods are defined for: \code{\link[=consensus,components-method]{components}}, \code{\link[=consensus,compounds-method]{compounds}}, \code{\link[=consensus,featureGroupsComparison-method]{featureGroupsComparison}}, \code{\link[=consensus,formulas-method]{formulas}}.
}

\code{formulaTable} returns formula data.





\itemize{
    \item Methods are defined for: \code{\link[=formulaTable,formulaConsensus-method]{formulaConsensus}}, \code{\link[=formulaTable,formulas-method]{formulas}}.
}
}
