% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics}
\alias{generics}
\alias{algorithm}
\alias{analysisInfo}
\alias{analyses}
\alias{annotatedPeakList}
\alias{clusterProperties}
\alias{clusters}
\alias{cutClusters}
\alias{consensus}
\alias{featureTable}
\alias{filter}
\alias{getFeatures}
\alias{getMCS}
\alias{groupNames}
\alias{plotChord}
\alias{plotChroms}
\alias{plotInt}
\alias{plotScores}
\alias{plotSilhouettes}
\alias{plotSpectrum}
\alias{plotStructure}
\alias{plotVenn}
\alias{plotUpSet}
\alias{replicateGroups}
\alias{treeCut}
\alias{treeCutDynamic}
\alias{[}
\alias{[[}
\alias{$}
\alias{as.data.table}
\alias{as.data.frame}
\alias{length}
\alias{lengths}
\alias{names}
\alias{plot}
\alias{show}
\title{Miscellaneous generics}
\usage{
algorithm(obj)

analysisInfo(obj)

analyses(obj)

annotatedPeakList(obj, ...)

clusterProperties(obj)

clusters(obj)

cutClusters(obj)

consensus(obj, ...)

featureTable(obj, ...)

filter(obj, ...)

getFeatures(obj)

getMCS(obj, ...)

groupNames(obj)

plotChord(obj, addSelfLinks = FALSE, addRetMzPlots = TRUE, ...)

plotChroms(obj, ...)

plotInt(obj, ...)

plotScores(obj, ...)

plotSilhouettes(obj, kSeq, ...)

plotSpectrum(obj, ...)

plotStructure(obj, ...)

plotVenn(obj, ...)

plotUpSet(obj, ...)

replicateGroups(obj)

treeCut(obj, k = NULL, h = NULL, ...)

treeCutDynamic(
  obj,
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1,
  ...
)
}
\arguments{
\item{obj}{The object the generic should be applied to.}

\item{\dots}{Any further method specific arguments. See method documentation
for details.}

\item{addSelfLinks}{If \code{TRUE} then 'self-links' are added which
represent non-shared data.}

\item{addRetMzPlots}{Set to \code{TRUE} to enable \emph{m/z} \emph{vs}
retention time scatter plots.}

\item{kSeq}{An integer vector containing the sequence that should be used for
average silhouette width calculation.}

\item{k, h}{Desired numbers of clusters. See \code{\link{cutree}}.}

\item{maxTreeHeight, deepSplit, minModuleSize}{Arguments used by
\code{\link{cutreeDynamicTree}}.}
}
\description{
Various (S4) generic functions providing a common interface for common tasks
such as plotting and filtering data. The actual functionality and function
arguments are often specific for the implemented methods, for this reason,
please refer to the linked method documentation for each generic.
}
\details{
\code{algorithm} returns the algorithm that was used to generate the object.

\itemize{
    \item Methods are defined for: \code{\link[=algorithm,compounds-method]{compounds}}; \code{\link[=algorithm,formulas-method]{formulas}}; \code{\link[=algorithm,optimizationResult-method]{optimizationResult}}; \code{\link[=algorithm,workflowStep-method]{workflowStep}}.
}

\code{analysisInfo} returns the \link[=analysis-information]{analysis information} from an object.

\itemize{
    \item Methods are defined for: \code{\link[=analysisInfo,featureGroups-method]{featureGroups}}; \code{\link[=analysisInfo,features-method]{features}}.
}

\code{analyses} returns a \code{character} vector with the analyses for which data is present in this object.

\itemize{
    \item Methods are defined for: \code{\link[=analyses,featureGroups-method]{featureGroups}}; \code{\link[=analyses,features-method]{features}}; \code{\link[=analyses,formulas-method]{formulas}}; \code{\link[=analyses,MSPeakLists-method]{MSPeakLists}}.
}

\code{annotatedPeakList} returns an annotated MS peak list.

\itemize{
    \item Methods are defined for: \code{\link[=annotatedPeakList,compounds-method]{compounds}}; \code{\link[=annotatedPeakList,formulas-method]{formulas}}.
}

\code{clusterProperties} Obtain a list with properties of the generated cluster(s).

\itemize{
    \item Methods are defined for: \code{\link[=clusterProperties,componentsIntClust-method]{componentsIntClust}}; \code{\link[=clusterProperties,compoundsCluster-method]{compoundsCluster}}.
}

\code{clusters} Obtain clustering object(s).

\itemize{
    \item Methods are defined for: \code{\link[=clusters,componentsIntClust-method]{componentsIntClust}}; \code{\link[=clusters,compoundsCluster-method]{compoundsCluster}}.
}

\code{cutClusters} Returns assigned cluster indices of a cut cluster.

\itemize{
    \item Methods are defined for: \code{\link[=cutClusters,componentsIntClust-method]{componentsIntClust}}; \code{\link[=cutClusters,compoundsCluster-method]{compoundsCluster}}.
}

\code{consensus} combines and merges data from various algorithms to generate a consensus.

\itemize{
    \item Methods are defined for: \code{\link[=consensus,components-method]{components}}; \code{\link[=consensus,compounds-method]{compounds}}; \code{\link[=consensus,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=consensus,formulas-method]{formulas}}.
}

\code{featureTable} returns feature information.

\itemize{
    \item Methods are defined for: \code{\link[=featureTable,featureGroups-method]{featureGroups}}; \code{\link[=featureTable,features-method]{features}}; \code{\link[=featureTable,featuresSet-method]{featuresSet}}.
}

\code{filter} provides various functionality to do post-filtering of data.

\itemize{
    \item Methods are defined for: \code{\link[=filter,components-method]{components}}; \code{\link[=filter,compounds-method]{compounds}}; \code{\link[=filter,featureGroups-method]{featureGroups}}; \code{\link[=filter,features-method]{features}}; \code{\link[=filter,featuresSet-method]{featuresSet}}; \code{\link[=filter,featuresXCMS-method]{featuresXCMS}}; \code{\link[=filter,featuresXCMS3-method]{featuresXCMS3}}; \code{\link[=filter,formulas-method]{formulas}}; \code{\link[=filter,MSPeakLists-method]{MSPeakLists}}.
}

\code{getFeatures} returns the object's \code{\link{features}} object.

\itemize{
    \item Methods are defined for: \code{\link[=getFeatures,featureGroups-method]{featureGroups}}.
}

\code{getMCS} Calculcates the maximum common substructure.

\itemize{
    \item Methods are defined for: \code{\link[=getMCS,compounds-method]{compounds}}; \code{\link[=getMCS,compoundsCluster-method]{compoundsCluster}}.
}

\code{groupNames} returns a \code{character} vector with the names of the feature groups for which data is present in this object.

\itemize{
    \item Methods are defined for: \code{\link[=groupNames,components-method]{components}}; \code{\link[=groupNames,compounds-method]{compounds}}; \code{\link[=groupNames,compoundsCluster-method]{compoundsCluster}}; \code{\link[=groupNames,featureGroups-method]{featureGroups}}; \code{\link[=groupNames,formulas-method]{formulas}}; \code{\link[=groupNames,MSPeakLists-method]{MSPeakLists}}.
}

\code{plotChord} plots a Chord diagram to assess overlapping data.

\itemize{
    \item Methods are defined for: \code{\link[=plotChord,featureGroups-method]{featureGroups}}; \code{\link[=plotChord,featureGroupsComparison-method]{featureGroupsComparison}}.
}

\code{plotChroms} plots extracted ion chromatogram(s).

\itemize{
    \item Methods are defined for: \code{\link[=plotChroms,components-method]{components}}; \code{\link[=plotChroms,featureGroups-method]{featureGroups}}.
}

\code{plotInt} plots the intensity of all contained features.

\itemize{
    \item Methods are defined for: \code{\link[=plotInt,componentsIntClust-method]{componentsIntClust}}; \code{\link[=plotInt,featureGroups-method]{featureGroups}}.
}

\code{plotScores} plots candidate scorings.

\itemize{
    \item Methods are defined for: \code{\link[=plotScores,compounds-method]{compounds}}; \code{\link[=plotScores,formulas-method]{formulas}}.
}

\code{plotSilhouettes} plots silhouette widths to evaluate the desired cluster size.

\itemize{
    \item Methods are defined for: \code{\link[=plotSilhouettes,componentsIntClust-method]{componentsIntClust}}; \code{\link[=plotSilhouettes,compoundsCluster-method]{compoundsCluster}}.
}

\code{plotSpectrum} plots a (annotated) spectrum.

\itemize{
    \item Methods are defined for: \code{\link[=plotSpectrum,components-method]{components}}; \code{\link[=plotSpectrum,compounds-method]{compounds}}; \code{\link[=plotSpectrum,formulas-method]{formulas}}; \code{\link[=plotSpectrum,MSPeakLists-method]{MSPeakLists}}.
}

\code{plotStructure} plots a chemical structure.

\itemize{
    \item Methods are defined for: \code{\link[=plotStructure,compounds-method]{compounds}}; \code{\link[=plotStructure,compoundsCluster-method]{compoundsCluster}}.
}

\code{plotVenn} plots a Venn diagram to assess unique and overlapping data.

\itemize{
    \item Methods are defined for: \code{\link[=plotVenn,compounds-method]{compounds}}; \code{\link[=plotVenn,featureGroups-method]{featureGroups}}; \code{\link[=plotVenn,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=plotVenn,formulas-method]{formulas}}.
}

\code{plotUpSet} plots an UpSet diagram to assess unique and overlapping data.

\itemize{
    \item Methods are defined for: \code{\link[=plotUpSet,compounds-method]{compounds}}; \code{\link[=plotUpSet,featureGroups-method]{featureGroups}}; \code{\link[=plotUpSet,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=plotUpSet,formulas-method]{formulas}}.
}

\code{replicateGroups} returns a \code{character} vector with the analyses for which data is present in this object.

\itemize{
    \item Methods are defined for: \code{\link[=replicateGroups,featureGroups-method]{featureGroups}}; \code{\link[=replicateGroups,features-method]{features}}.
}

\code{treeCut} Manually cut a cluster.

\itemize{
    \item Methods are defined for: \code{\link[=treeCut,componentsIntClust-method]{componentsIntClust}}; \code{\link[=treeCut,compoundsCluster-method]{compoundsCluster}}.
}

\code{treeCutDynamic} Automatically cut a cluster.

\itemize{
    \item Methods are defined for: \code{\link[=treeCutDynamic,componentsIntClust-method]{componentsIntClust}}; \code{\link[=treeCutDynamic,compoundsCluster-method]{compoundsCluster}}.
}
}
\section{Other generics}{
 Below are methods that are defined for existing
  generics (\emph{e.g.} defined in \code{base}). Please see method specific
  documentation for more details.

 \code{[} Subsets data within an object.

\itemize{
    \item Methods are defined for: \code{\link[=[,components,ANY,ANY,missing-method]{components,ANY,ANY,missing}}; \code{\link[=[,compounds,ANY,missing,missing-method]{compounds,ANY,missing,missing}}; \code{\link[=[,compoundsCluster,ANY,missing,missing-method]{compoundsCluster,ANY,missing,missing}}; \code{\link[=[,featureGroups,ANY,ANY,missing-method]{featureGroups,ANY,ANY,missing}}; \code{\link[=[,featureGroupsComparison,ANY,missing,missing-method]{featureGroupsComparison,ANY,missing,missing}}; \code{\link[=[,features,ANY,missing,missing-method]{features,ANY,missing,missing}}; \code{\link[=[,featuresSet,ANY,missing,missing-method]{featuresSet,ANY,missing,missing}}; \code{\link[=[,featuresXCMS,ANY,missing,missing-method]{featuresXCMS,ANY,missing,missing}}; \code{\link[=[,featuresXCMS3,ANY,missing,missing-method]{featuresXCMS3,ANY,missing,missing}}; \code{\link[=[,formulas,ANY,missing,missing-method]{formulas,ANY,missing,missing}}; \code{\link[=[,MSPeakLists,ANY,ANY,missing-method]{MSPeakLists,ANY,ANY,missing}}.
}

 \code{[[} Extract data from an object.

\itemize{
    \item Methods are defined for: \code{\link[=[[,components,ANY,ANY-method]{components,ANY,ANY}}; \code{\link[=[[,compounds,ANY,missing-method]{compounds,ANY,missing}}; \code{\link[=[[,featureGroups,ANY,ANY-method]{featureGroups,ANY,ANY}}; \code{\link[=[[,featureGroupsComparison,ANY,missing-method]{featureGroupsComparison,ANY,missing}}; \code{\link[=[[,features,ANY,missing-method]{features,ANY,missing}}; \code{\link[=[[,featuresSet,ANY,missing-method]{featuresSet,ANY,missing}}; \code{\link[=[[,formulas,ANY,ANY-method]{formulas,ANY,ANY}}; \code{\link[=[[,MSPeakLists,ANY,ANY-method]{MSPeakLists,ANY,ANY}}.
}

 \code{$} Extract data from an object.

\itemize{
    \item Methods are defined for: \code{\link[=$,components-method]{components}}; \code{\link[=$,compounds-method]{compounds}}; \code{\link[=$,featureGroups-method]{featureGroups}}; \code{\link[=$,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=$,features-method]{features}}; \code{\link[=$,formulas-method]{formulas}}; \code{\link[=$,MSPeakLists-method]{MSPeakLists}}.
}

 \code{as.data.table} Converts an object to a table (\code{\link{data.table}}).

\itemize{
    \item Methods are defined for: \code{\link[=as.data.table,components-method]{components}}; \code{\link[=as.data.table,compounds-method]{compounds}}; \code{\link[=as.data.table,featureGroups-method]{featureGroups}}; \code{\link[=as.data.table,features-method]{features}}; \code{\link[=as.data.table,featuresSet-method]{featuresSet}}; \code{\link[=as.data.table,formulas-method]{formulas}}; \code{\link[=as.data.table,MSPeakLists-method]{MSPeakLists}}; \code{\link[=as.data.table,workflowStep-method]{workflowStep}}.
}

 \code{as.data.frame} Converts an object to a table (\code{data.frame}).

\itemize{
    \item Methods are defined for: \code{\link[=as.data.frame,workflowStep-method]{workflowStep}}.
}

 \code{length} Returns the length of an object.

\itemize{
    \item Methods are defined for: \code{\link[=length,components-method]{components}}; \code{\link[=length,compounds-method]{compounds}}; \code{\link[=length,compoundsCluster-method]{compoundsCluster}}; \code{\link[=length,featureGroups-method]{featureGroups}}; \code{\link[=length,featureGroupsComparison-method]{featureGroupsComparison}}; \code{\link[=length,features-method]{features}}; \code{\link[=length,formulas-method]{formulas}}; \code{\link[=length,MSPeakLists-method]{MSPeakLists}}; \code{\link[=length,optimizationResult-method]{optimizationResult}}.
}

 \code{lengths} Returns the lengths of elements within this object.

\itemize{
    \item Methods are defined for: \code{\link[=lengths,compoundsCluster-method]{compoundsCluster}}; \code{\link[=lengths,optimizationResult-method]{optimizationResult}}.
}

 \code{names} Return names for this object.

\itemize{
    \item Methods are defined for: \code{\link[=names,components-method]{components}}; \code{\link[=names,featureGroups-method]{featureGroups}}; \code{\link[=names,featureGroupsComparison-method]{featureGroupsComparison}}.
}

 \code{plot} Generates a plot for an object.

\itemize{
    \item Methods are defined for: \code{\link[=plot,componentsIntClust,ANY-method]{componentsIntClust,ANY}}; \code{\link[=plot,compoundsCluster,ANY-method]{compoundsCluster,ANY}}; \code{\link[=plot,featureGroups,ANY-method]{featureGroups,ANY}}; \code{\link[=plot,featureGroupsComparison,ANY-method]{featureGroupsComparison,ANY}}; \code{\link[=plot,optimizationResult,ANY-method]{optimizationResult,ANY}}.
}

 \code{show} Prints information about this object.

\itemize{
    \item Methods are defined for: \code{\link[=show,adduct-method]{adduct}}; \code{\link[=show,components-method]{components}}; \code{\link[=show,compounds-method]{compounds}}; \code{\link[=show,compoundsCluster-method]{compoundsCluster}}; \code{\link[=show,featureGroups-method]{featureGroups}}; \code{\link[=show,features-method]{features}}; \code{\link[=show,featuresSet-method]{featuresSet}}; \code{\link[=show,formulas-method]{formulas}}; \code{\link[=show,MSPeakLists-method]{MSPeakLists}}; \code{\link[=show,optimizationResult-method]{optimizationResult}}; \code{\link[=show,workflowStep-method]{workflowStep}}.
}
}

