% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mspeaklists.R
\name{getDefAvgPListParams}
\alias{getDefAvgPListParams}
\title{Parameters for averaging MS peak list data}
\usage{
getDefAvgPListParams(..., IMS = getLimIMS())
}
\arguments{
\item{\dots}{Optional named arguments that override defaults.}

\item{IMS}{A \code{character} that specifies for which IMS instrument defaults are returned. Should be
\code{"bruker"} or \code{"agilent"}. Defaults to what is specified in \link{limits}.}
}
\value{
\code{getDefAvgPListParams} returns a \code{list} with the peak list averaging parameters.
}
\description{
Create parameter lists for averaging MS peak list data.
}
\details{
The parameters set used for averaging peak lists are set by the \code{avgFeatParams} and \code{avgFGroupParams}
arguments to \code{\link{generateMSPeakLists}} and its related algorithm specific functions. The parameters are
specified as a named \code{list} with the following values: \itemize{

\item \code{method},\code{clusterMzWindow} The cluster method and window (see \link[=cluster-params]{clustering
parameters}) used to average mass spectra. \code{clusterMzWindow} is defaulted as \code{defaultLim("mz", "medium")} (see
\link{limits}).

\item \code{topMost} Only retain this maximum number of MS peaks when generating averaged spectra. Lowering this
number may exclude more irrelevant (noisy) MS peaks and decrease processing time, whereas higher values may avoid
excluding lower intense MS peaks that may still be of interest.

\item \code{minIntensityPre} MS peaks with intensities below this value will be removed (applied prior to selection
by \code{topMost} and averaging).

\item \code{minIntensityPost} MS peaks with intensities below this value will be removed (after averaging).

\item \code{minIntensityIMS} MS peaks in spectra of raw IMS frames with intensities below this value will be removed
(applied prior to any other treatment steps).

\item \code{absMinAbundance},\code{relMinAbundance} Minimum absolute/relative abundance of an MS peak across the
spectra that are averaged. If \code{absMinAbundance} exceeds the number of spectra then the threshold is
automatically lowered to the number of spectra.

\item \code{minRelCumIntensity} Minimum relative cumulative intensity of an MS peak in the averaged spectrum.

\item \code{smoothWindowIMS},\code{halfWindowIMS},\code{maxGapIMS} Parameters used for centroiding \emph{m/z} peaks
from IMS-HRMS data. See \verb{Centroiding IMS data} for more details.

\item \code{withPrecursorMS} For MS data only: ignore any spectra that do not contain the precursor peak.

  For IMS data this excludes MS spectra within an IMS frame that do not contain the precursor peak, typically due to
  mobility separation. Hence, setting this option performs some crude cleanup of MS spectra, even for features for
  which no mobilities were assigned (\emph{e.g.} non-IMS workflows).

\item \code{pruneMissingPrecursorMS} For MS data only: if \code{TRUE} then peak lists without a precursor peak are
removed. Note that even when this is set to \code{FALSE}, functionality that relies on MS (not MS/MS) peak lists
(\emph{e.g.} formulae calculation) will still skip calculation if a precursor is not found.

\item \code{retainPrecursorMSMS} For MS/MS data only: if \code{TRUE} then always retain the precursor mass peak even
if is not amongst the \code{topMost} peaks. Note that MS precursor mass peaks are always kept. Furthermore, note that
precursor peaks in both MS and MS/MS data may still be removed by intensity thresholds (this is unlike the
\code{\link[=filter,MSPeakLists-method]{filter}} method function).

}

The \code{getDefAvgPListParams} function can be used to generate a default parameter list. The defaults are (with
\code{IMS="bruker"}):

\preformatted{list(
  method = "distance_mean",
  clusterMzWindow = 0.005,
  topMost = 50,
  minIntensityPre = 500,
  minIntensityPost = 500,
  minIntensityIMS = 25,
  absMinAbundance = 0,
  relMinAbundance = 0,
  minRelCumIntensity = 0,
  smoothWindowIMS = 0,
  halfWindowIMS = 2,
  maxGapIMS = 0.005,
  withPrecursorMS = TRUE,
  pruneMissingPrecursorMS = TRUE,
  retainPrecursorMSMS = TRUE
)}
}
\section{Centroiding IMS data}{
 With IMS-HRMS data the \emph{m/z} peaks are often not or partially centroided. The
  following steps are performed to centroid the data: \enumerate{

  \item Sum up mass spectra within an IMS frame. If the feature has mobility data, only spectra within its mobility
  boundaries are considered.
  
  \item Use point-distance clustering (see \link[=cluster-params]{clustering parameters}) with a window defined by
  \code{maxGapIMS} to find related mass signals. This is primarily meant for non-continuous data, \emph{e.g.} due to
  intensity thresholding. The \code{maxGapIMS} parameter should be set to a value that represents the maximum
  expected distance between two \emph{m/z} datapoints. For some instruments, such as Agilent IMS-QTOF, this value may
  be higher than expected. For that reason, if \code{IMS="agilent"} then the default is set to \code{0.01}.
  
  \item Smooth the intensity data using a centered moving average with window size \code{smoothWindowIMS} (set to
  zero to disable smoothing).
  
  \item Find local maxima within sliding window with +/- \code{halfWindowIMS} points and eliminate non-centroids.
  This algorithm is based on the \code{C_localMaxima} function from \CRANpkg{MALDIquant}.

  }
}

\references{
\insertRef{Gibb2012}{patRoon}
}
