% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/features.R, R/feature_groups.R,
%   R/analysisInfo.R
\name{getEICs-methods}
\alias{getEICs-methods}
\alias{getEICs,features-method}
\alias{getEICs,featureGroups-method}
\alias{getEICs,data.table-method}
\alias{getEICs,data.frame-method}
\title{Obtains extracted ion chromatograms (EICs)}
\usage{
\S4method{getEICs}{features}(
  obj,
  analysis = analyses(obj),
  EICParams = getDefEICParams(),
  output = "fill"
)

\S4method{getEICs}{featureGroups}(
  obj,
  analysis = analyses(obj),
  groupName = names(obj),
  EICParams = getDefEICParams(),
  output = "fill"
)

\S4method{getEICs}{data.table}(obj, ranges, gapFactor = 3, output = "fill", minIntensityIMS = 25)

\S4method{getEICs}{data.frame}(obj, ...)
}
\arguments{
\item{obj}{For the \code{data.table} and \code{data.frame} methods: a table with
\link[=analysis-information]{analysis information}. For the \code{features} and \code{featureGroups} methods: the
object for which EICs should be generated.}

\item{analysis}{A \code{character} vector with the analyses for which EICs should be generated.}

\item{EICParams}{A named \code{list} with parameters used for extracted ion chromatogram (EIC) creation. See the
\link[=EIXParams]{EIC parameters} documentation for more details.}

\item{output}{Should be \code{"fill"}, \code{"pad"} or \code{"raw"}. Internally, EIC data is compressed by omitting
any zero intensity data points. If \code{output="fill"} then the zero intensity points are re-added to obtain
continuous chromatograms. If \code{output="pad"} then zero intensity points are only re-added that surround others,
which is sufficient for \emph{e.g.} plotting. If \code{output="raw"} then the original compressed data is returned.}

\item{groupName}{A \code{character} vector with the names of the feature groups for which EICs should be generated.}

\item{ranges}{A \code{list} with for each analysis a \code{data.frame} with \code{numeric} columns \code{"retmin"},
\code{"retmax"}, \code{"mzmin"}, \code{"mzmax"} with the lower/upper ranges of the retention time and \emph{m/z}.
Furthermore, columns \code{"mobmin"} and \code{"mobmax"} can be added for mobility lower/upper ranges in IMS data.}

\item{gapFactor}{A \code{numeric} that configures gap filling. See \code{\link{getDefEICParams}} for more details.}

\item{minIntensityIMS}{\IMSWF Raw intensity threshold for IMS data. This is primarily intended to speed up raw data
processing.}

\item{\dots}{For the \code{data.frame} method: further arguments passed to the \code{data.table} method.}
}
\value{
A \code{list} with for each analysis a \code{list} with EIC data. For the \code{data.table} and
  \code{data.frame} methods the EICs are ordered according to the rows in \code{ranges}. For the \code{features} and
  \code{featureGroups} methods the EICs are named after the feature IDs and feature group names, respectively, and
  analyses without any EIC data are omitted.

  If \code{output="raw"} then additional columns with \emph{e.g.} mean-averaged and base peak \emph{m/z} values for
  each data point are returned. Furthermore, the \code{allXValues} attribute is set that can be used to obtain the
  original retention time values to reconstruct the original complete chromatogram.
}
\description{
These methods generate one or more EIC(s). The \code{data.table} and \code{data.frame} methods generate EIC(s) for
given retention time, \emph{m/z} and optionally mobility ranges. The \code{features} and \code{featureGroups}
methods generate EICs for all (or selected) features and feature groups, respectively, whereby the ranges are
automatically determined from the feature data (see the \link[=EIXParams]{EIC parameters} for configuration
options).
}
\section{Functions}{
\itemize{
\item \code{getEICs(features)}: Generates EICs for all (or selected) features (method for \code{features}).

\item \code{getEICs(featureGroups)}: Generates EICs for all (or selected) feature groups (method for
\code{featureGroups}).

\item \code{getEICs(data.table)}: Generates one or more EIC(s) for given retention time, \emph{m/z} and optionally
mobility ranges (method for \code{data.table}).

\item \code{getEICs(data.frame)}: Wrapper for the \code{data.table} method (method for \code{data.frame}).

}}
\section{Use of raw HRMS data}{
 The \link[=msdata]{raw data interface} of \pkg{patRoon} is used by \code{getEICs} to
  process HRMS (or IMS-HRMS) data. Please see \link[=msdata]{its documentation} for more information on the supported
  formats and available configuration options.
}

