% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-exported.R
\name{getEICs}
\alias{getEICs}
\title{Obtains extracted ion chromatograms (EICs)}
\usage{
getEICs(file, ranges)
}
\arguments{
\item{file}{The file path to the sample analysis data file (\file{.mzXML} or \file{.mzML}).}

\item{ranges}{A \code{data.frame} with \code{numeric} columns \code{"retmin"}, \code{"retmin"}, \code{"mzmin"},
\code{"mzmax"} with the lower/upper ranges of the retention time and \emph{m/z}.}
}
\value{
A \code{list} with EIC data for each of the rows in \code{ranges}.
}
\description{
This function generates one or more EIC(s) for given retention time and \emph{m/z} ranges.
}
