% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils-xcms.R
\docType{methods}
\name{getXCMSSet}
\alias{getXCMSSet}
\alias{getXCMSSet,features-method}
\alias{getXCMSSet,featuresOpenMS-method}
\alias{getXCMSSet,featuresXCMS-method}
\alias{getXCMSSet,featureGroups-method}
\alias{getXCMSSet,featureGroupsXCMS-method}
\title{Conversion to xcmsSet objects}
\usage{
getXCMSSet(obj, ..., verbose = TRUE)

\S4method{getXCMSSet}{features}(obj, exportedData, verbose)

\S4method{getXCMSSet}{featuresOpenMS}(obj, exportedData, verbose)

\S4method{getXCMSSet}{featuresXCMS}(obj, exportedData, verbose)

\S4method{getXCMSSet}{featureGroups}(obj, exportedData, verbose)

\S4method{getXCMSSet}{featureGroupsXCMS}(obj, exportedData, verbose)
}
\arguments{
\item{obj}{The object that should be converted.}

\item{verbose}{If \code{FALSE} then no text output is shown.}

\item{exportedData, \dots}{Set to \code{TRUE} if analyses were exported as
\code{mzXML} or \code{mzML} files (ignored for \code{featuresOpenMS} and
\code{featuresXCMS} methods).}
}
\description{
Converts a \code{\link{features}} or \code{\link{featureGroups}} object to an
\code{\link{xcmsSet}} object.
}
