% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-sirius.R
\name{groupFeaturesSIRIUS}
\alias{groupFeaturesSIRIUS}
\title{Group features using SIRIUS}
\usage{
groupFeaturesSIRIUS(analysisInfo, verbose = TRUE)
}
\arguments{
\item{analysisInfo}{A \code{data.frame} with \link[=analysis-information]{Analysis information}.}

\item{verbose}{if \code{FALSE} then no text output will be shown.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.

The \code{featuresSet} method (for \link[=sets-workflow]{sets workflows}) returns a
  \code{\link{featureGroupsSet}} object.
}
\description{
Uses \href{https://bio.informatik.uni-jena.de/software/sirius/}{SIRIUS} to find \emph{and} group features.
}
\details{
This function uses SIRIUS to group features. This function is called when calling \code{groupFeatures} with
  \code{algorithm="sirius"}.

Finding and grouping features is done by running the \command{lcms-align} command on every analyses at once.
  For this reason, grouping feature data from other algorithms than \command{SIRIUS} is not supported.
  
  The MS files should be in the \file{mzML} or \file{mzXML} format.

The input MS data files need to be centroided. The \code{\link{convertMSFiles}} function can be used to
  centroid data.
}
\references{
\insertRef{Dhrkop2019}{patRoon}
}
\seealso{
\code{\link{groupFeatures}} for more details and other algorithms.
}
