% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups.R
\name{importFeatureGroups}
\alias{importFeatureGroups}
\title{Import feature groups from files}
\usage{
importFeatureGroups(path, type, ...)
}
\arguments{
\item{path}{The path that should be used for importing. See the algorithm specific functions for more details.}

\item{type}{Which file type should be imported: \code{"brukerpa"} (Bruker ProfileAnalysis), \code{"brukertasq"}
(Bruker TASQ) or \code{"envimass"} (\pkg{enviMass}).}

\item{\dots}{Further arguments passed to the selected import algorithm function.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.

The \code{featuresSet} method (for \link[=sets-workflow]{sets workflows}) returns a
  \code{\link{featureGroupsSet}} object.
}
\description{
Generic function to import feature groups produced by other software from files.
}
\details{
\code{importFeatureGroups} is a generic function that will import feature groups from files by one of the supported algorithms. The actual
  functionality is provided by algorithm specific functions such as \code{importFeatureGroupsBrukerTASQ} and \code{importFeatureGroupsBrukerPA}. While these
  functions may be called directly, \code{importFeatureGroups} provides a generic interface and is therefore usually preferred.
}
\seealso{
The \code{\link{featureGroups}} output class and its methods and the algorithm specific functions:
  \code{\link{importFeatureGroupsBrukerPA}}, \code{\link{importFeatureGroupsBrukerTASQ}}, \code{\link{importFeatureGroupsEnviMass}}

\code{\link{groupFeatures}} to group features. Other import functions:
  \code{\link{importFeatureGroupsXCMS}}, \code{\link{importFeatureGroupsXCMS3}} and
  \code{\link{importFeatureGroupsKPIC2}}.
}
