% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_groups-xcms.R
\name{importFeatureGroupsXCMS}
\alias{importFeatureGroupsXCMS}
\title{Imports feature groups from XCMS (old interface)}
\usage{
importFeatureGroupsXCMS(xs, analysisInfo)
}
\arguments{
\item{xs}{An \code{\link{xcmsSet}} object.}

\item{analysisInfo}{A \code{data.frame} with \link[=analysis-information]{Analysis information}.}
}
\value{
An object of a class which is derived from \code{\link{featureGroups}}.

The \code{featuresSet} method (for \link[=sets-workflow]{sets workflows}) returns a
  \code{\link{featureGroupsSet}} object.
}
\description{
Imports grouped features from a legacy \code{\link{xcmsSet}} object from the \pkg{xcms} package.
}
\references{
\addCitations{xcms}{1} \cr\cr \addCitations{xcms}{2} \cr\cr \addCitations{xcms}{3}
}
\seealso{
\code{\link{importFeaturesXCMS3}} and \code{\link{groupFeatures}}
}
