% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-xcms3.R
\name{importFeaturesXCMS3}
\alias{importFeaturesXCMS3}
\title{Imports features from XCMS (new interface)}
\usage{
importFeaturesXCMS3(xdata, analysisInfo)
}
\arguments{
\item{xdata}{An \code{\link{XCMSnExp}} object.}

\item{analysisInfo}{A \code{data.frame} with \link[=analysis-information]{Analysis information}.}
}
\value{
An object of a class which is derived from \code{\link{features}}.
}
\description{
Imports feature data generated from an existing \code{\link{XCMSnExp}} object generated by the \pkg{xcms} package.
}
\details{
This function imports data from XCMS3. This function is called when calling \code{importFeatures} with
  \code{type="xcms3"}.
}
\references{
\addCitations{xcms}{1} \cr\cr \addCitations{xcms}{2} \cr\cr \addCitations{xcms}{3}
}
\seealso{
\code{\link{importFeatures}} for more details and other algorithms.
}
