% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/features-kpic2.R
\name{getPICSet}
\alias{getPICSet}
\alias{getPICSet,features-method}
\alias{getPICSet,featuresKPIC2-method}
\title{Conversion to KPIC2 objects}
\usage{
getPICSet(obj, ...)

\S4method{getPICSet}{features}(obj, loadRawData = TRUE)

\S4method{getPICSet}{featuresKPIC2}(obj, ...)
}
\arguments{
\item{obj}{The \code{features} object that should be converted.}

\item{\dots}{Ignored}

\item{loadRawData}{Set to \code{TRUE} if analyses are available as \code{mzXML} or \code{mzML} files. Otherwise MS
data is not loaded, and some dummy data (\emph{e.g.} file paths) is used in the returned object.}
}
\description{
Converts a \code{\link{features}} object to an \pkg{KPIC} object.
}
