% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-kpic2.R
\name{getPICSet,features-method}
\alias{getPICSet,features-method}
\alias{getPICSet}
\alias{getPICSet,featuresKPIC2-method}
\title{Conversion to KPIC2 objects}
\usage{
\S4method{getPICSet}{features}(obj, loadRawData = TRUE)

\S4method{getPICSet}{featuresKPIC2}(obj, ...)
}
\arguments{
\item{obj}{The \code{features} object that should be converted.}

\item{loadRawData, \dots}{Set to \code{TRUE} if analyses were exported as
\code{mzXML} or \code{mzML} files (ignored by \code{featuresKPIC2} method).}
}
\description{
Converts a \code{\link{features}} object to an \pkg{KPIC} object.
}
