% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspeaklists.R
\docType{class}
\name{MSPeakLists-class}
\alias{MSPeakLists-class}
\alias{MSPeakLists}
\alias{peakLists,MSPeakLists-method}
\alias{peakLists}
\alias{analyses,MSPeakLists-method}
\alias{groupNames,MSPeakLists-method}
\alias{algorithm,MSPeakLists-method}
\alias{length,MSPeakLists-method}
\alias{show,MSPeakLists-method}
\alias{[,MSPeakLists,ANY,ANY,missing-method}
\alias{filter,MSPeakLists-method}
\alias{filterBy,MSPeakLists-method}
\title{Class containing MS Peak Lists}
\usage{
\S4method{peakLists}{MSPeakLists}(obj)

\S4method{analyses}{MSPeakLists}(obj)

\S4method{groupNames}{MSPeakLists}(obj)

\S4method{algorithm}{MSPeakLists}(obj)

\S4method{length}{MSPeakLists}(x)

\S4method{show}{MSPeakLists}(object)

\S4method{[}{MSPeakLists,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{filter}{MSPeakLists}(obj, absMSIntThr = NULL,
  absMSMSIntThr = NULL, relMSIntThr = NULL, relMSMSIntThr = NULL,
  topMSPeaks = NULL, topMSMSPeaks = NULL, deIsotopeMS = FALSE,
  deIsotopeMSMS = FALSE)

\S4method{filterBy}{MSPeakLists}(obj, fGroups, negate, onlyMSMS = FALSE)
}
\arguments{
\item{obj, x, object}{The \code{\link{MSPeakLists}} object to access.}

\item{i}{Either a numeric, character or logical \code{vector} that is used to select analyses by their index, name and logical selection, respectively (for the order/names see \code{analyses()}). If missing all analyses are selected.}

\item{j}{Either a numeric, character or logical \code{vector} that is used to select feature groups by their index, name and logical selection, respectively (for the order/names see \code{groupNames()}). If missing all feature groups are selected.}

\item{drop}{ignored.}

\item{absMSIntThr, absMSMSIntThr, relMSIntThr, relMSMSIntThr}{Absolute/relative
intensity threshold for MS or MS/MS peak lists. \code{NULL} for none.}

\item{topMSPeaks, topMSMSPeaks}{Only consider this amount of MS or MS/MS peaks
with highest intensity. \code{NULL} to consider all.}

\item{deIsotopeMS, deIsotopeMSMS}{Remove any isotopic peaks in MS or MS/MS
peak lists. This may improve data processing steps which do not assume the
presence of isotopic peaks (e.g. MetFrag for MS/MS). Note that
\code{getMzRPeakLists} does not (yet) support flagging of isotopes.}

\item{fGroups}{A \code{\link{featureGroups}} object which should be filtered.}

\item{negate}{If set to \code{TRUE} then the filter is applied in opposite
manner.}

\item{onlyMSMS}{If \code{TRUE} then only keep feature groups that have MS/MS
peaklists.}
}
\value{
\code{peakLists} returns a nested list containing MS (and MS/MS where
  available) peak lists per feature group and per analysis. The format is:
  \code{[[analysis]][[featureGroupName]][[MSType]][[PeakLists]]} where
  \code{MSType} is either \code{"MS"} or \code{"MSMS"} and \code{PeakLists} a
  \code{\link{data.table}} containing all \emph{m/z} values (\code{mz}
  column) and their intensities (\code{intensity} column). In addition, the
  peak list tables may contain a \code{cmp} column which contains an unique
  alphabatical identifier to which isotopic cluster (or "compound") a mass
  belongs (only supported by MS peak lists generated by Bruker tools at the
  moment).

\code{filterBy} returns a filtered \code{\link{featureGroups}}
  object.
}
\description{
Contains all MS (and MS/MS where available) peak lists for a
\code{\link{featureGroups}} object.
}
\details{
Objects for this class are returned by \link[=MSPeakLists-generation]{MS peak
lists generators}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{peakLists}: Accessor method to obtain the MS peak lists.

\item \code{analyses}: returns a \code{character} vector with the names of the
analyses from which data is present in this object.

\item \code{groupNames}: returns a \code{character} vector with the names of the
feature groups from which data is present in this object.

\item \code{algorithm}: Retrieve the algorithm (a character string) used to
generate the MS peak lists.

\item \code{length}: Obtain total number of \emph{m/z} values.

\item \code{show}: Shows summary information for this object.

\item \code{[}: Operator to subset on analyses and/or feature groups.

\item \code{filter}: provides post filtering of generated MS
peak lists, which may further enhance quality of subsequent workflow steps
(\emph{e.g.} formulae calculation and compounds identification) and/or
speed up these processes.

\item \code{filterBy}: Only keep feature groups in a given
\code{\link{featureGroups}} object that are also present in the specified
\code{MSPeakLists} object.
}}

\section{Slots}{

\describe{
\item{\code{peakLists}}{Contains a list of all MS (and MS/MS) peak lists. Use the
\code{peakLists} method for access.}

\item{\code{algorithm}}{The algorithm that was used to generate the MS peak lists.
Use the \code{algorithm} method for access.}
}}

