% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspeaklists.R
\docType{class}
\name{MSPeakLists-class}
\alias{MSPeakLists-class}
\alias{MSPeakLists}
\alias{peakLists,MSPeakLists-method}
\alias{peakLists}
\alias{averagedPeakLists,MSPeakLists-method}
\alias{analyses,MSPeakLists-method}
\alias{groupNames,MSPeakLists-method}
\alias{length,MSPeakLists-method}
\alias{show,MSPeakLists-method}
\alias{[,MSPeakLists,ANY,ANY,missing-method}
\alias{[[,MSPeakLists,ANY,ANY-method}
\alias{as.data.table,MSPeakLists-method}
\alias{filter,MSPeakLists-method}
\alias{plotSpec,MSPeakLists-method}
\title{Class containing MS Peak Lists}
\usage{
\S4method{peakLists}{MSPeakLists}(obj)

\S4method{averagedPeakLists}{MSPeakLists}(obj)

\S4method{analyses}{MSPeakLists}(obj)

\S4method{groupNames}{MSPeakLists}(obj)

\S4method{length}{MSPeakLists}(x)

\S4method{show}{MSPeakLists}(object)

\S4method{[}{MSPeakLists,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{MSPeakLists,ANY,ANY}(x, i, j)

\S4method{as.data.table}{MSPeakLists}(x, fGroups = NULL,
  averaged = TRUE)

\S4method{filter}{MSPeakLists}(obj, absMSIntThr = NULL,
  absMSMSIntThr = NULL, relMSIntThr = NULL, relMSMSIntThr = NULL,
  topMSPeaks = NULL, topMSMSPeaks = NULL, deIsotopeMS = FALSE,
  deIsotopeMSMS = FALSE)

\S4method{plotSpec}{MSPeakLists}(obj, groupName, analysis = NULL,
  MSLevel = 2, useGGPlot2 = FALSE)
}
\arguments{
\item{obj, x, object}{The \code{\link{MSPeakLists}} object to access.}

\item{i, j}{A numeric or character value which is used to select analyses/feature groups by
their index or name, respectively (for the order/names see \code{analyses()/groupNames()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all analyses/feature groups are selected.\cr\cr For \code{[[}: should be a skalar value. If \code{j} is not specified, \code{i} selects by feature groups instead.}

\item{drop}{ignored.}

\item{fGroups}{The \code{\link{featureGroups}} object that was used to
generate this object. If not \code{NULL} it is used to add feature group
information (retention and \emph{m/z} values).}

\item{averaged}{If \code{TRUE} then feature group averaged peak list data is
used.}

\item{absMSIntThr, absMSMSIntThr, relMSIntThr, relMSMSIntThr}{Absolute/relative
intensity threshold for MS or MS/MS peak lists. \code{NULL} for none.}

\item{topMSPeaks, topMSMSPeaks}{Only consider this amount of MS or MS/MS peaks
with highest intensity. \code{NULL} to consider all.}

\item{deIsotopeMS, deIsotopeMSMS}{Remove any isotopic peaks in MS or MS/MS
peak lists. This may improve data processing steps which do not assume the
presence of isotopic peaks (e.g. MetFrag for MS/MS). Note that
\code{getMzRPeakLists} does not (yet) support flagging of isotopes.}

\item{groupName}{The name of the feature group for which a plot should be
made.}

\item{analysis}{The name of the analysis for which a plot should be made. If
\code{NULL} then data from the feature group averaged peak list is used.}

\item{MSLevel}{The MS level: \samp{1} for regular MS, \samp{2} for MSMS.}

\item{useGGPlot2}{If \code{TRUE} then \pkg{\link{ggplot2}} is used for
plotting, otherwise base plot used. For \code{plotSpec}, \code{ggplot2}
allows nicely repelled text for annotation. However, base plot is generally
faster.}
}
\value{
\code{peakLists} returns a nested list containing MS (and MS/MS where
  available) peak lists per feature group and per analysis. The format is:
  \code{[[analysis]][[featureGroupName]][[MSType]][[PeakLists]]} where
  \code{MSType} is either \code{"MS"} or \code{"MSMS"} and \code{PeakLists} a
  \code{\link{data.table}} containing all \emph{m/z} values (\code{mz}
  column) and their intensities (\code{intensity} column). In addition, the
  peak list tables may contain a \code{cmp} column which contains an unique
  alphabatical identifier to which isotopic cluster (or "compound") a mass
  belongs (only supported by MS peak lists generated by Bruker tools at the
  moment).

\code{averagedPeakLists} returns a nested list of feature group
  averaged peak lists in a similar format as \code{peakLists}.

\code{plotSpec} will return a \code{\link[=ggplot2]{ggplot object}}
  if \code{useGGPlot2} is \code{TRUE}.
}
\description{
Contains all MS (and MS/MS where available) peak lists for a
\code{\link{featureGroups}} object.
}
\details{
Objects for this class are returned by \link[=MSPeakLists-generation]{MS peak
lists generators}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{peakLists}: Accessor method to obtain the MS peak lists.

\item \code{averagedPeakLists}: Accessor method to obtain the feature group averaged
MS peak lists.

\item \code{analyses}: returns a \code{character} vector with the names of the
analyses for which data is present in this object.

\item \code{groupNames}: returns a \code{character} vector with the names of the
feature groups for which data is present in this object.

\item \code{length}: Obtain total number of \emph{m/z} values.

\item \code{show}: Shows summary information for this object.

\item \code{[}: Subset on analyses/feature groups.

\item \code{[[}: Extract a list with MS and MS/MS (if available) peak
lists. If the second argument (\code{j}) is not specified the averaged peak
lists for the group specified by the first argument (\code{i}) will be
returned.

\item \code{as.data.table}: Returns all MS peak list data in a table.

\item \code{filter}: provides post filtering of generated MS
peak lists, which may further enhance quality of subsequent workflow steps
(\emph{e.g.} formulae calculation and compounds identification) and/or
speed up these processes.

\item \code{plotSpec}: Plots a spectrum using MS or MS/MS peak lists for a
given feature group.
}}

\section{Slots}{

\describe{
\item{\code{peakLists}}{Contains a list of all MS (and MS/MS) peak lists. Use the
\code{peakLists} method for access.}

\item{\code{averagedPeakLists}}{A \code{list} with averaged MS (and MS/MS) peak lists
for each feature group.}

\item{\code{avgPeakListArgs}}{A \code{list} with arguments used to generate feature
group averaged MS(/MS) peak lists.}

\item{\code{algorithm}}{The algorithm that was used to generate the MS peak lists.
Use the \code{algorithm} method for access.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{workflowStep}}}   \itemize{     \item{\strong{\code{\link{MSPeakLists}}}}   } }
}

