% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mspeaklists.R,
%   R/mspeaklists-set.R, R/utils-mspeaklists.R
\docType{class}
\name{peakLists}
\alias{peakLists}
\alias{averagedPeakLists}
\alias{spectrumSimilarity}
\alias{MSPeakLists-class}
\alias{MSPeakLists}
\alias{peakLists,MSPeakLists-method}
\alias{averagedPeakLists,MSPeakLists-method}
\alias{analyses,MSPeakLists-method}
\alias{groupNames,MSPeakLists-method}
\alias{length,MSPeakLists-method}
\alias{show,MSPeakLists-method}
\alias{[,MSPeakLists,ANY,ANY,missing-method}
\alias{[[,MSPeakLists,ANY,ANY-method}
\alias{$,MSPeakLists-method}
\alias{as.data.table,MSPeakLists-method}
\alias{delete,MSPeakLists-method}
\alias{filter,MSPeakLists-method}
\alias{plotSpectrum,MSPeakLists-method}
\alias{spectrumSimilarity,MSPeakLists-method}
\alias{MSPeakListsSet-class}
\alias{MSPeakListsSet}
\alias{analysisInfo,MSPeakListsSet-method}
\alias{show,MSPeakListsSet-method}
\alias{[,MSPeakListsSet,ANY,ANY,missing-method}
\alias{as.data.table,MSPeakListsSet-method}
\alias{delete,MSPeakListsSet-method}
\alias{filter,MSPeakListsSet-method}
\alias{plotSpectrum,MSPeakListsSet-method}
\alias{spectrumSimilarity,MSPeakListsSet-method}
\alias{MSPeakListsUnset-class}
\alias{MSPeakListsUnset}
\alias{unset,MSPeakListsSet-method}
\alias{getDefIsolatePrecParams}
\title{Class containing MS Peak Lists}
\usage{
peakLists(obj, ...)

averagedPeakLists(obj, ...)

spectrumSimilarity(obj, ...)

\S4method{peakLists}{MSPeakLists}(obj)

\S4method{averagedPeakLists}{MSPeakLists}(obj)

\S4method{analyses}{MSPeakLists}(obj)

\S4method{groupNames}{MSPeakLists}(obj)

\S4method{length}{MSPeakLists}(x)

\S4method{show}{MSPeakLists}(object)

\S4method{[}{MSPeakLists,ANY,ANY,missing}(x, i, j, ..., reAverage = FALSE, drop = TRUE)

\S4method{[[}{MSPeakLists,ANY,ANY}(x, i, j)

\S4method{$}{MSPeakLists}(x, name)

\S4method{as.data.table}{MSPeakLists}(x, fGroups = NULL, averaged = TRUE)

\S4method{delete}{MSPeakLists}(obj, i = NULL, j = NULL, k = NULL, reAverage = FALSE, ...)

\S4method{filter}{MSPeakLists}(
  obj,
  absMSIntThr = NULL,
  absMSMSIntThr = NULL,
  relMSIntThr = NULL,
  relMSMSIntThr = NULL,
  topMSPeaks = NULL,
  topMSMSPeaks = NULL,
  minMSMSPeaks = NULL,
  isolatePrec = NULL,
  deIsotopeMS = FALSE,
  deIsotopeMSMS = FALSE,
  withMSMS = FALSE,
  annotatedBy = NULL,
  retainPrecursorMSMS = TRUE,
  reAverage = FALSE,
  negate = FALSE
)

\S4method{plotSpectrum}{MSPeakLists}(
  obj,
  groupName,
  analysis = NULL,
  MSLevel = 1,
  title = NULL,
  specSimParams = getDefSpecSimParams(),
  xlim = NULL,
  ylim = NULL,
  ...
)

\S4method{spectrumSimilarity}{MSPeakLists}(
  obj,
  groupName1,
  groupName2 = NULL,
  analysis1 = NULL,
  analysis2 = NULL,
  MSLevel = 1,
  specSimParams = getDefSpecSimParams(),
  NAToZero = FALSE,
  drop = TRUE
)

\S4method{analysisInfo}{MSPeakListsSet}(obj)

\S4method{show}{MSPeakListsSet}(object)

\S4method{[}{MSPeakListsSet,ANY,ANY,missing}(x, i, j, ..., reAverage = FALSE, sets = NULL, drop = TRUE)

\S4method{as.data.table}{MSPeakListsSet}(x, fGroups = NULL, averaged = TRUE)

\S4method{delete}{MSPeakListsSet}(obj, i = NULL, j = NULL, k = NULL, reAverage = FALSE, ...)

\S4method{filter}{MSPeakListsSet}(
  obj,
  ...,
  annotatedBy = NULL,
  retainPrecursorMSMS = TRUE,
  reAverage = FALSE,
  negate = FALSE,
  sets = NULL
)

\S4method{plotSpectrum}{MSPeakListsSet}(
  obj,
  groupName,
  analysis = NULL,
  MSLevel = 1,
  title = NULL,
  specSimParams = getDefSpecSimParams(),
  xlim = NULL,
  ylim = NULL,
  perSet = TRUE,
  mirror = TRUE,
  ...
)

\S4method{spectrumSimilarity}{MSPeakListsSet}(
  obj,
  groupName1,
  groupName2 = NULL,
  analysis1 = NULL,
  analysis2 = NULL,
  MSLevel = 1,
  specSimParams = getDefSpecSimParams(),
  NAToZero = FALSE,
  drop = TRUE
)

\S4method{unset}{MSPeakListsSet}(obj, set)

getDefIsolatePrecParams(...)
}
\arguments{
\item{obj, x, object}{The \code{\link{MSPeakLists}} object to access.}

\item{\dots}{Further arguments passed to \code{\link[graphics]{plot}}.

  \setsPassedArgs1{MSPeakLists}}

\item{i, j}{For \code{[}/\code{[[}: A numeric or character value which is used to select analyses/feature groups by
their index or name, respectively (for the order/names see \code{analyses()/groupNames()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all analyses/feature groups are selected.\cr\cr For \code{[[}: should be a scalar value. If \code{j} is not specified, \code{i} selects by feature groups instead.\cr\cr For \code{delete}: The data to remove from. \code{i} are the
feature groups as numeric index, logical or character, \code{j} the MS peaks as numeric indices (rows). If either is
\code{NULL} then data for all is removed. \code{j} may also be a function: it will be called for each feature group,
with the peak list table (a \code{data.table}), feature group name, analysis (\code{NULL} if averaged peak list), type (\code{"MS"} or \code{"MSMS"}) and any other arguments passed as \code{\dots} to \code{delete}. The return value of this
function specifies the peak list indices (rows) to be removed (specified as an \code{integer} or \code{logical} vector).}

\item{reAverage}{Set to \code{TRUE} to regenerate group averaged MS peak lists. \strong{NOTE} it is very important
that any annotation data relying on MS peak lists (formulae/compounds) are regenerated afterwards! Otherwise it is
likely that \emph{e.g.} plotting methods will use wrong MS/MS data.}

\item{drop}{If set to \code{TRUE} and if the comparison is made between two spectra then \code{\link{drop}} is used
to reduce the \code{matrix} return value to a \code{numeric} vector.}

\item{name}{The feature group name (partially matched).}

\item{fGroups}{The \code{\link{featureGroups}} object that was used to
generate this object. If not \code{NULL} it is used to add feature group
information (retention and \emph{m/z} values).}

\item{averaged}{If \code{TRUE} then feature group averaged peak list data is
used.}

\item{k}{A vector with analyses (\code{character} with names or \code{integer} with indices) for which the data
should be deleted. If \code{k!=NULL} then deletions will \emph{not} occur on group averaged peak lists. Otherwise,
if \code{k=NULL} then deletion occurs on \emph{both} group averaged and analysis specific peak lists.}

\item{absMSIntThr, absMSMSIntThr, relMSIntThr, relMSMSIntThr}{Absolute/relative intensity threshold for MS or MS/MS peak
lists. \code{NULL} for none.}

\item{topMSPeaks, topMSMSPeaks}{Only consider this amount of MS or MS/MS peaks with highest intensity. \code{NULL} to
consider all.}

\item{minMSMSPeaks}{If the number of peaks in an MS/MS peak list (\strong{excluding} the precursor peak) is lower
than this it will be completely removed. Set to \code{NULL} to ignore.}

\item{isolatePrec}{If not \code{NULL} then value should be a \code{list} with parameters used for isolating the
precursor and its isotopes in MS peak lists (see \verb{Isolating precursor data}). Alternatively, \code{TRUE} to
apply the filter with default settings (as given with \code{getDefIsolatePrecParams}).}

\item{deIsotopeMS, deIsotopeMSMS}{Remove any isotopic peaks in MS or MS/MS peak lists. This may improve data
processing steps which do not assume the presence of isotopic peaks (e.g. MetFrag for MS/MS). Note that
\code{getMzRPeakLists} does not (yet) support flagging of isotopes.}

\item{withMSMS}{If set to \code{TRUE} then only results will be retained for which MS/MS data is available. if
\code{negate=TRUE} then only results \emph{without} MS/MS data will be retained.}

\item{annotatedBy}{Either a \code{\link{formulas}} or \code{\link{compounds}} object, or a \code{list} with both. Any
MS/MS peaks that are \emph{not} annotated by any of the candidates in the specified objects are removed.}

\item{retainPrecursorMSMS}{If \code{TRUE} then precursor peaks will never be filtered out from MS/MS peak lists (note
that precursors are never removed from MS peak lists). The \code{negate} argument does not affect this setting.}

\item{negate}{If \code{TRUE} then filters are applied in opposite manner.}

\item{groupName}{The name of the feature group for which a plot should be made. To compare spectra, two group names
can be specified.}

\item{analysis}{The name of the analysis for which a plot should be made. If \code{NULL} then data from the feature
group averaged peak list is used. When comparing spectra, either \code{NULL} or the analyses for both spectra
should be specified.}

\item{MSLevel}{The MS level: \samp{1} for regular MS, \samp{2} for MSMS.}

\item{title}{The title of the plot. If \code{NULL} a title will be automatically made.}

\item{specSimParams}{A named \code{list} with parameters that influence the calculation of MS spectra similarities.
See the \link[=specSimParams]{spectral similarity parameters} documentation for more details.}

\item{xlim, ylim}{Sets the plot size limits used by
\code{\link[graphics]{plot}}. Set to \code{NULL} for automatic plot sizing.}

\item{groupName1, groupName2}{The names of the feature groups for which the comparison should be made. If both
arguments are specified then a comparison is made with the spectra specified by \code{groupName1} \emph{vs} those
specified by \code{groupName2}. The length of either can be \samp{>1} to generate a comparison matrix.
Alternatively, if \code{groupName2} is \code{NULL} then all the spectra specified in \code{groupName1} will be
compared with eachother, \emph{i.e.} resulting in a square similarity matrix.}

\item{analysis1, analysis2}{The name of the analysis (analyses) for the comparison. If \code{NULL} then data from the
feature group averaged peak list is used. Otherwise, should be the same length as
\code{groupName1}/\code{groupName2}.}

\item{NAToZero}{Set to \code{TRUE} to convert \code{NA} similarities (\emph{i.e.} when no similarity could be
calculated) to zero values.}

\item{sets}{\setsWF A \code{character} with name(s) of the sets to keep (or remove if \code{negate=TRUE}).}

\item{perSet, mirror}{\setsWF If \code{perSet=TRUE} then the set specific mass peaks are annotated separately.
Furthermore, if \code{mirror=TRUE} (and there are two sets in the object) then a mirror plot is generated.}

\item{set}{\setsWF The name of the set.}
}
\value{
\code{peakLists} returns a nested list containing MS (and MS/MS where
  available) peak lists per feature group and per analysis. The format is:
  \code{[[analysis]][[featureGroupName]][[MSType]][[PeakLists]]} where
  \code{MSType} is either \code{"MS"} or \code{"MSMS"} and \code{PeakLists} a
  \code{\link{data.table}} containing all \emph{m/z} values (\code{mz}
  column) and their intensities (\code{intensity} column). In addition, the
  peak list tables may contain a \code{cmp} column which contains an unique
  alphabetical identifier to which isotopic cluster (or "compound") a mass
  belongs (only supported by MS peak lists generated by Bruker tools at the
  moment).

\code{averagedPeakLists} returns a nested list of feature group
  averaged peak lists in a similar format as \code{peakLists}.

\code{delete} returns the object for which the specified data was removed.
}
\description{
Contains all MS (and MS/MS where available) peak lists for a \code{\link{featureGroups}} object.
}
\details{
Objects for this class are returned by \code{\link{generateMSPeakLists}}.

The \code{getDefIsolatePrecParams} is used to create a parameter
  list for isolating the precursor and its isotopes (see \verb{Isolating precursor data}).
}
\section{Methods (by generic)}{
\itemize{
\item \code{peakLists(MSPeakLists)}: Accessor method to obtain the MS peak lists.

\item \code{averagedPeakLists(MSPeakLists)}: Accessor method to obtain the feature group averaged
MS peak lists.

\item \code{analyses(MSPeakLists)}: returns a \code{character} vector with the names of the
analyses for which data is present in this object.

\item \code{groupNames(MSPeakLists)}: returns a \code{character} vector with the names of the
feature groups for which data is present in this object.

\item \code{length(MSPeakLists)}: Obtain total number of \emph{m/z} values.

\item \code{show(MSPeakLists)}: Shows summary information for this object.

\item \code{x[i}: Subset on analyses/feature groups.

\item \code{x[[i}: Extract a list with MS and MS/MS (if available) peak
lists. If the second argument (\code{j}) is not specified the averaged peak
lists for the group specified by the first argument (\code{i}) will be
returned.

\item \code{$}: Extract group averaged MS peaklists for a feature group.

\item \code{as.data.table(MSPeakLists)}: Returns all MS peak list data in a table.

\item \code{delete(MSPeakLists)}: Completely deletes specified peaks from MS peak lists.

\item \code{filter(MSPeakLists)}: provides post filtering of generated MS peak lists, which may further enhance quality of
subsequent workflow steps (\emph{e.g.} formulae calculation and compounds identification) and/or speed up these
processes. The filters are applied to peak lists for each feature and feature group. The feature group peak lists
are \emph{not} re-averaged by default (see the \code{reAverage} argument). \emph{not} filtered afterwards.

\item \code{plotSpectrum(MSPeakLists)}: Plots a spectrum using MS or MS/MS peak lists for a given feature group. Two spectra can be
compared when two feature groups are specified.

\item \code{spectrumSimilarity(MSPeakLists)}: Calculates the spectral similarity between two or more spectra.

}}
\section{Slots}{

\describe{
\item{\code{peakLists}}{Contains a list of all MS (and MS/MS) peak lists. Use the \code{peakLists} method for access.}

\item{\code{metadata}}{Metadata for all spectra used to generate peak lists. Follows the format of the \code{peakLists} slot.}

\item{\code{averagedPeakLists}}{A \code{list} with averaged MS (and MS/MS) peak lists for each feature group.}

\item{\code{avgPeakListArgs}}{A \code{list} with arguments used to generate feature group averaged MS(/MS) peak lists.}

\item{\code{origFGNames}}{A \code{character} with the original input feature group names.}

\item{\code{analysisInfo}}{\setsWF  \link[=analysis-information]{Analysis information}. Use the \code{analysisInfo} method
for access.}
}}

\section{Isolating precursor data}{
 Formula calculation typically relies on evaluating the measured isotopic pattern
  from the precursor to score candidates. Some algorithms (currently only \command{GenForm}) penalize candidates if
  mass peaks are present in MS1 spectra that do not contribute to the isotopic pattern. Since these spectra are
  typically very 'noisy' due to background and co-eluting ions, an additional filtering step may be recommended prior
  to formula calculation. During this precursor isolation step all mass peaks are removed that are (1) not the
  precursor and (2) not likely to be an isotopologue of the precursor. To determine potential isotopic peaks the
  following parameters are used:

  \itemize{

  \item \code{maxIsotopes} The maximum number of isotopes to consider. For instance, a value of \samp{5} means that
  \code{M+0} (\emph{i.e.} the monoisotopic peak) till \code{M+5} is considered. All mass peaks outside this range are
  removed.

  \item \code{mzDefectRange} A two-sized \code{vector} specifying the minimum (can be negative) and maximum
  \emph{m/z} defect deviation compared to the precursor \emph{m/z} defect. When chlorinated, brominated or other
  compounds with strong \emph{m/z} defect in their isotopologues are to be considered a higher range may be desired.
  On the other hand, for natural compounds this range may be tightened. Note that the search range is propegated with
  increasing distance from the precursor, \emph{e.g.} the search range is doubled for \code{M+2}, tripled for
  \code{M+3} etc.

  \item \code{intRange} A two-sized \code{vector} specifying the minimum and maximum relative intensity range
  compared to the precursor. For instance, \code{c(0.001, 2)} removes all peaks that have an intensity below 0.1\% or
  above 200\% of that of the precursor.

  \item \code{z} The \code{z} value (\emph{i.e.} absolute charge) to be considerd. For instance, a value of \code{2}
  would look for \code{M+0.5}, \code{M+1} etc. Note that the \code{mzDefectRange} is adjusted accordingly
  (\emph{e.g.} halved if \code{z=2}).

  \item \code{maxGap} The maximum number of missing adjacent isotopic peaks ('gaps'). If the (rounded) \emph{m/z}
  difference to the previous peak exceeds this value then this and all next peaks will be removed. Similar to
  \code{z}, the maximum gap is automatically adjusted for \code{charge}.

  }

  These parameters should be in a \code{list} that is passed to the \code{isolatePrec} argument to \code{filter}. The
  default values can be obtained with the \code{getDefIsolatePrecParams} function:

\code{maxIsotopes=5}; \code{mzDefectRange=c(-0.01, 0.01)}; \code{intRange=c(0.001, 2)}; \code{z=1}; \code{maxGap=2}
}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{workflowStep}}}   \itemize{     \item{\strong{\code{\link{MSPeakLists}}}}     \itemize{       \item{\code{\link{MSPeakListsSet}}}       \item{\code{\link{MSPeakListsUnset}}}     }   } }
}

\section{Source}{
 \code{spectrumSimilarity}: The principles of spectral binning and cosine similarity calculations
  were loosely was based on the code from \code{SpectrumSimilarity()} function of \pkg{OrgMassSpecR}.
}

\section{Sets workflows}{
 \setsWFClass{MSPeakListsSet}{MSPeakLists}

  \setsWFNewMethodsSOExtra{MSPeakListsUnset}{Only the MS peaks that are present in the specified set are kept.}{

  \item \code{analysisInfo} Returns the  \link[=analysis-information]{analysis info} for this object.

  }

  \setsWFChangedMethods{

  \item \code{filter} and the subset operator (\code{[}) Can be used to select data that is only present for selected
  sets. The \code{filter} method is applied for each set individually, and afterwards the results are combined again
  (see \code{\link{generateMSPeakLists}}). Note that this has important implications for \emph{e.g.} relative
  intensity filters (\code{relMSIntThr}/\code{relMSMSIntThr}), \code{topMSPeaks}/\code{topMSMSPeaks} and
  \code{minMSMSPeaks}. Similarly, when the \code{annotatedBy} filter is applied, each set specific MS peak list is
  filtered by the annotation results from only that set.

  \item \code{plotSpectrum} Is able to highlight set specific mass peaks (\code{perSet} and \code{mirror} arguments).

  \item \code{spectrumSimilarity} First calculates similarities for each spectral pair per set (\emph{e.g.} all
  positive mode spectra are compared and then all negative mode spectra are compared). This data is then combined
  into an overall similarity value. How this combination is performed depends on the \code{setCombineMethod} field of
  the \code{\link{specSimParams}} argument.

  }
}

\author{
For \code{spectrumSimilarity}: major contributions by Bas van de Velde for spectral binning and similarity
  calculation.
}
