% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/mspeaklists.R,
%   R/mspeaklists-bruker.R, R/mspeaklists-mzr.R, R/utils-mspeaklists.R
\docType{methods}
\name{MSPeakLists-generation}
\alias{MSPeakLists-generation}
\alias{generateMSPeakLists,featureGroups-method}
\alias{generateMSPeakLists}
\alias{generateMSPeakListsDA}
\alias{generateMSPeakListsDAFMF}
\alias{generateMSPeakListsMzR}
\alias{getDefAvgPListParams}
\title{Generation of MS Peak Lists}
\usage{
\S4method{generateMSPeakLists}{featureGroups}(fGroups, algorithm, ...)

generateMSPeakListsDA(fGroups, bgsubtr = TRUE, maxRtMSWidth = 20,
  minMSIntensity = 500, minMSMSIntensity = 500, clear = TRUE,
  save = TRUE, MSMSType = "MSMS",
  avgFGroupParams = getDefAvgPListParams())

generateMSPeakListsDAFMF(fGroups, minMSIntensity = 500,
  minMSMSIntensity = 500, avgFGroupParams = getDefAvgPListParams())

generateMSPeakListsMzR(fGroups, maxRtMSWidth = 20,
  precursorMzWindow = 8, topMost = NULL,
  avgFeatParams = getDefAvgPListParams(),
  avgFGroupParams = getDefAvgPListParams())

getDefAvgPListParams(...)
}
\arguments{
\item{fGroups}{The \code{\link{featureGroups}} object from which MS peak
lists should be extracted.}

\item{algorithm}{A character string describing the algorithm that should be
used: \code{"bruker"}, \code{"brukerfmf"}, \code{"mzr"}}

\item{\dots}{For \code{generateMSPeakLists}: Any parameters to be passed to
  the selected MS peak lists generation algorithm.

  For \code{getDefAvgPListParams}: Optional named arguments that override
  defaults.}

\item{bgsubtr}{If \code{TRUE} background will be subtracted using the
'spectral' algorithm.}

\item{maxRtMSWidth}{Maximum total chromatographic peak width (seconds) used
for spectrum averaging. If \code{NULL} all spectra from a feature will be
taken into account. Lower to decrease processing time.}

\item{minMSIntensity, minMSMSIntensity}{Minimum intensity for peak lists
obtained with DataAnalysis. Highly recommended to set \samp{>0} as DA tends
to report many very low intensity peaks.}

\item{clear}{Remove any existing chromatogram traces/mass spectra prior to
making new ones.}

\item{save}{Save data files after completion.}

\item{MSMSType}{The type of MS/MS experiment performed: \code{"MSMS"} for
MRM/AutoMSMS or \code{"BBCID"} for broadband CID.}

\item{precursorMzWindow}{The \emph{m/z} window (in Da) to find MS/MS spectra
of a precursor. This is typically used for Data-Dependent like MS/MS data
and should correspond to the isolation \emph{m/z} width that was used to
collect the data. For Data-Independent MS/MS experiments, where precursor
ions are not isolated prior to fragmentation (\emph{e.g.} bbCID, MSe,
all-ion, ...) the value should be \code{NULL}.}

\item{topMost}{Only extract MS peak lists from a maximum of \code{topMost}
analyses with highest intensity. If \code{NULL} all analyses will be used.}

\item{avgFeatParams, avgFGroupParams}{A \code{list} with parameters used for
averaging of peak lists for individual features and feature groups,
respectively (see below).}
}
\value{
A \code{\link{MSPeakLists}} object that can be used for formulae
  calculation and compound identification.
}
\description{
Functionality to generate MS peak lists.
}
\details{
Formula calculation and identification tools rely on mass spectra that belong
to features of interest. For processing, MS (and MS/MS) spectra are typically
reduced to a table with a column containing measured \emph{m/z} values and a
column containing their intensities. These 'MS peak lists' can then be used
for \link[=formula-generation]{formula generation} and
\link[=compound-generation]{compound generation}.

MS and MS/MS peak lists are first generated for all features (or a subset, if
the \code{topMost} argument is set). During this step multiple spectra over
the feature elution profile are averaged. Subsequently, peak lists will be
generated for each feature group by averaging peak lists of the features
within the group. Functionality that uses peak lists will either use data
from individual features or from group averaged peak lists. For instance, the
former may be used by formulae calculation, while compound identification and
plotting functionality typically uses group averaged peak lists.

Several functions exist to automatically extract MS peak lists for feature
groups.

\code{generateMSPeakLists} is a generic function that will generate MS peak lists
  using one of the supported algorithms. The actual functionality is provided
  by algorithm specific functions such as \code{generateMSPeakListsMzR} and
  \code{generateMSPeakListsDA}. While these functions may be called directly,
  \code{generateMSPeakLists} provides a generic interface and is therefore usually
  preferred.

\code{generateMSPeakListsDA} uses Bruker DataAnalysis to generate MS
  peak lists. Naturally, this only works with analyses in the Bruker data
  format (\file{.d}). This function leverages DataAnalysis functionality to
  support averaging of spectra, background subtraction and identification of
  isotopes. In order to obtain mass spectra TICs will be added of the MS and
  relevant MS/MS signals.

\code{generateMSPeakListsDAFMF} is similar to \code{generateMSPeakListsDA},
  but uses compounds that were generated by the Find Molecular Features (FMF)
  algorithm to extract MS peak lists. This is generally much faster than
  \code{generateMSPeakListsDA}, however, it only works when features were obtained
  using the \code{\link{findFeaturesBruker}} function. Since all MS spectra
  are generated in advance by Bruker DataAnalysis, no further parameters
  exist to customize its operation.

\code{generateMSPeakListsMzR} uses the \pkg{\link{mzR}} package to
  extract MS peak lists. For this analyses should be either in \file{.mzXML}
  or \file{.mzML} format. This function averages multiple spectra over a
  chromatgraphic peak to improve accuracy.

The \code{getDefAvgPListParams} is used to create a parameter list
  for peak list averaging (discussed below).
}
\note{
\code{generateMSPeakListsDA} requires that the \option{Component}
  column is active (Method-->Parameters-->Layouts-->Mass List Layout).
}
\section{Peak list averaging parameters}{
 The parameters set used for
  averaging peak lists are set by the \code{avgFeatParams} and
  \code{avgFGroupParams} arguments. This should be a named \code{list} with
  the following values:
  \itemize{
  \item \code{clusterMzWindow} \emph{m/z} window (in Da) used for clustering
  \emph{m/z} values when spectra are averaged. Too small windows will prevent
  clustering \emph{m/z} values (thus erroneously creating 'extra' values),
  whereas too big windows may cluster unrelated \emph{m/z} values from
  different or even the same spectrum together.
  \item \code{topMost} Only retain this maximum number of MS peaks when
  generating averaged spectra. Lowering this number may exclude more
  irrelevant (noisy) MS peaks and decrease processing time, whereas higher
  values may avoid excluding lower intense MS peaks that may still be of
  interest.
  \item \code{minIntensityPre} MS peaks with intensities below this value
  will be removed (applied prior to selection by \code{topMost}) before
  averaging.
  \item \code{minIntensityPost} MS peaks with intensities below this value
  will be removed after averaging.
  \item \code{avgFun} Function that is used to calculate average \emph{m/z}
  values.
  \item \code{method} Method used for producing averaged MS spectra. Valid
  values are \code{"hclust"}, used for hierarchical clustering (using the
  \pkg{\link{fastcluster}} package), and \code{"distance"}, to use the
  between peak distance. The latter method may reduces processing time and
  memory requirements, at the potential cost of reduced accuracy.
  }
  The \code{getDefAvgPListParams} function can be used to generate a default
  parameter list.

  Note that when Bruker algorithms are used these parameters only control
  generation of feature groups averaged peak lists: how peak lists for
  features are generated is controlled by DataAnalysis.
}

\section{Source}{
 Averaging of mass spectra algorithms used by are based on
  the \href{https://github.com/zeehio/msProcess}{msProcess} R package (now
  archived on CRAN).
}

\references{
\cr\cr \addCitations{mzR}{1} \cr\cr \addCitations{mzR}{2} \cr\cr
  \addCitations{mzR}{3} \cr\cr \addCitations{mzR}{4} \cr\cr
  \addCitations{mzR}{5} \cr\cr \addCitations{fastcluster}{1}
}
\seealso{
\code{\link{MSPeakLists-class}}
}
