% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/defunct.R
\name{compoundViewer}
\alias{compoundViewer}
\alias{patRoon-defunct}
\alias{compoundViewer,featureGroups,MSPeakLists,compounds-method}
\alias{groupFeaturesScreening}
\alias{checkChromatograms}
\title{Defunct functions.}
\usage{
compoundViewer(fGroups, MSPeakLists, compounds)

\S4method{compoundViewer}{featureGroups,MSPeakLists,compounds}(fGroups, MSPeakLists, compounds)

groupFeaturesScreening(...)

checkChromatograms(...)
}
\arguments{
\item{MSPeakLists}{A \code{\link{MSPeakLists}} object.}

\item{compounds}{A \code{\link{compounds}} object.}

\item{...}{Ignore.}
}
\description{
These functions do not work anymore and may be updated in the future.
future.
}
\details{
The \code{compoundViewer} method is used to view compound
  identification results. It will display available candidate information
  such as scorings and identifiers, MS/MS spectra with explained peaks and
  chemical structures.

\code{groupFeaturesScreening} was replaced by
  \code{\link{screenSuspects}}. Please refer to this function and the
  handbook for the updated interface.

\code{checkChromatograms} was replaced by
  \code{\link{checkFeatures}}. Please refer to this function and the
  handbook for the updated interface.
}
\keyword{internal}
