% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{patRoon-package}
\alias{patRoon}
\alias{patRoon-package}
\title{Workflow solutions for mass-spectrometry based non-target analysis.}
\description{
\Sexpr[results=text,echo=FALSE]{packageDescription("patRoon", fields =
"Description")}
}
\section{Package options}{


  The following package options (see \code{\link{options}}) can be set:

  \itemize{

  \item \code{patRoon.cache.mode}: A \code{character} setting the current
  caching mode: \code{"save"} and \code{"load"} will only save/load results
  to/from the cache, \code{"both"} (default) will do both and \code{"none"}
  to completely disable caching. This option can be changed anytime, which
  might be useful, for instance, to temporarily disable cached results before
  running a function.

  \item \code{patRoon.cache.fileName}: a \code{character} specifying the name
  of the cache file (default is \file{cache.sqlite}).

  \item \code{patRoon.MP.maxProcs}: The maximum number of processes that
  should be initiated in parallel. A good starting point is the number of
  physical cores, which is the default as detected by
  \code{\link[parallel]{detectCores}}. This option is only used when
  \option{patRoon.MP.method="classic"}.
  
  \item \code{patRoon.MP.method}: Either \code{"classic"} or \code{"future"}.
  The former is the default and uses \code{\link{processx}} to execute
  multiple commands in parallel. When \code{"future"} the
  \code{\link{future.apply}} package is used for parallelization, which is
  especially useful for \emph{e.g.} cluster computing.
  
  \item \code{MP.futureSched}: Sets the \code{future.scheduling} function
  argument for \code{\link{future_lapply}}. Only used if
  \option{patRoon.MP.method="future"}.
  
  \item \code{patRoon.MP.logPath}: The path used for logging of output from
  commands executed by multiprocess.
  
  \item \code{patRoon.path.pwiz}: The path in which the
  \command{ProteoWizard} binaries are installed. If unset an attempt is made
  to find this directory from the Windows registry and \option{PATH}
  environment variable.

  \item \code{patRoon.path.GenForm}: The path to the \command{GenForm}
  executable. If not set (the default) the internal \code{GenForm} binary is
  used. Only set if you want to override the executable.

  \item \code{patRoon.path.MetFragCL}: The complete file path to the MetFrag
  CL \file{jar} file that \emph{must} be set when using
  \code{\link{generateCompoundsMetFrag}}. Example:
  \code{"C:/MetFrag2.4.2-CL.jar"}.

  \item \code{patRoon.path.MetFragCompTox}: The complete file path to the
  CompTox database \file{csv} file. See \code{\link{generateCompounds}} for
  more details.
  
  \item \code{patRoon.path.MetFragPubChemLite}: The complete file path to the
  PubChem database \file{csv} file. See \code{\link{generateCompounds}} for
  more details.

  \item \code{patRoon.path.SIRIUS}: The directory in which SIRIUS is
  installed. Unless the binaries can be located via the \option{PATH}
  environment variable, this \emph{must} be set when using
  \code{\link{generateFormulasSIRIUS}} or
  \code{\link{generateCompoundsSIRIUS}}. Example:
  \code{"C:/sirius-win64-3.5.1"}.

  \item \code{patRoon.path.OpenMS}: The path in which the \command{OpenMS}
  binaries are installed. Usually the location is added to the \option{PATH}
  environment variable when OpenMS is installed, in which case this option
  can be left empty.

  \item \code{patRoon.path.pngquant}: The path of the \command{pngquant}
  binary that is used when optimizing \file{.png} plots generated by
  \code{\link{reportHTML}} (with \code{optimizePng} set to \code{TRUE}). If the
  binary can be located through the \option{PATH} environment variable this
  option can remain empty. Note that some of the functionality of
  \code{reportHTML} only locates the binary through the \option{PATH}
  environment variable, hence, it is recommended to set up \option{PATH}
  instead.

  \item \code{patRoon.path.obabel}: The path in which the \command{OpenBabel}
  binaries are installed. Usually the location is added to the \option{PATH}
  environment variable when OpenBabel is installed, in which case this option
  can be left empty.
  
  }
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rickhelmus/patRoon}
  \item Report bugs at \url{https://github.com/rickhelmus/patRoon/issues}
}

}
\author{
\strong{Maintainer}: Rick Helmus \email{r.helmus@uva.nl} (\href{https://orcid.org/0000-0001-9401-3133}{ORCID})

Other contributors:
\itemize{
  \item Olaf Brock (\href{https://orcid.org/0000-0003-4727-8459}{ORCID}) [contributor]
  \item Vittorio Albergamo (\href{https://orcid.org/0000-0002-5347-1362}{ORCID}) [contributor]
  \item Andrea Brunner (\href{https://orcid.org/0000-0002-2801-1751}{ORCID}) [contributor]
  \item Emma Schymanski (\href{https://orcid.org/0000-0001-6868-8145}{ORCID}) [contributor]
}

}
