% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\docType{methods}
\name{reporting}
\alias{reporting}
\alias{reportCSV,featureGroups-method}
\alias{reportCSV}
\alias{reportPDF,featureGroups-method}
\alias{reportPDF}
\alias{reportMD,featureGroups-method}
\alias{reportMD}
\title{Report feature group data}
\usage{
\S4method{reportCSV}{featureGroups}(fGroups, path = "report",
  reportFGroupsAsRows = TRUE, reportFGroupsAnalysisInfo = TRUE,
  reportFGroupsRetMz = TRUE, reportFeatures = FALSE, formConsensus = NULL,
  compounds = NULL, compoundNormalizeScores = TRUE, compsCluster = NULL,
  components = NULL, cInfo = NULL, clusterK = NULL, retMin = TRUE,
  clearPath = FALSE)

\S4method{reportPDF}{featureGroups}(fGroups, path = "report",
  reportFGroups = TRUE, formConsensus = NULL, reportFormulaSpectra = TRUE,
  compounds = NULL, compoundNormalizeScores = TRUE, components = NULL,
  cInfo = NULL, clusterK = NULL, silInfo = NULL, clusterMaxLabels = 250,
  MSPeakLists = NULL, retMin = TRUE, EICGrid = c(2, 1),
  EICRtWindow = 20, EICMzWindow = 0.005, EICTopMost = NULL,
  EICOnlyPresent = TRUE, clearPath = FALSE)

\S4method{reportMD}{featureGroups}(fGroups, path = "report",
  reportChord = TRUE, reportFGroups = TRUE, formConsensus = NULL,
  reportFormulaSpectra = TRUE, compounds = NULL,
  compoundNormalizeScores = TRUE, components = NULL, cInfo = NULL,
  clusterK = NULL, silInfo = NULL, interactiveHeat = FALSE,
  clusterMaxLabels = 250, MSPeakLists = NULL, retMin = TRUE,
  EICRtWindow = 20, EICMzWindow = 0.005, EICTopMost = NULL,
  EICOnlyPresent = TRUE, selfContained = TRUE, optimizePng = FALSE,
  maxProcAmount = getOption("patRoon.maxProcAmount"), clearPath = FALSE)
}
\arguments{
\item{fGroups}{The \code{\link{featureGroups}} object that should be used for
reporting data.}

\item{path}{The destination file path for files generated during reporting.
Will be generated if needed.}

\item{reportFGroupsAsRows}{Report feature groups as rows (instead of columns)
within the resulting \file{.csv} file.}

\item{reportFGroupsAnalysisInfo}{Include analyses information (reference and
replicate groups) in the reported feature groups table \file{.csv} file.}

\item{reportFGroupsRetMz}{Include feature group information (retention time
and \emph{m/z}) within the reported feature groups table \file{.csv} file.}

\item{reportFeatures}{If set to \code{TRUE} then for each analysis a
\file{.csv} file will be generated with information about its detected
features.}

\item{formConsensus, compounds, components}{Further objects
(\code{\link{formulaConsensus}}, \code{\link{compounds}},
\code{\link{components}}) that should be reported. Specify \code{NULL} to
skip reporting a particular object.}

\item{compoundNormalizeScores}{If \code{TRUE} compound identification scores
will be normalized to maximum values.}

\item{cInfo}{If not \code{NULL} a \code{\link{clusterInfo}} object to be used
for reporting hierarchical clustering results.}

\item{clusterK}{Number of branches the hierarchical cluster should be cut
into. If specified (\emph{i.e.} not \code{NULL}) then individual clusters
will be reported.}

\item{retMin}{If \code{TRUE} then report retention times in minutes
(otehrwise seconds).}

\item{clearPath}{If \code{TRUE} then the destination path will be
(recursively) removed prior to reporting.}

\item{reportFGroups}{If \code{TRUE} then feature group data will be reported.}

\item{reportFormulaSpectra}{If \code{TRUE} then explained MS/MS spectra (if
available) for candidate formulae will be reported. Specifying
\code{formConsensus} and setting this argument to \code{FALSE} still allows
further annotation of compound MS/MS spectra.}

\item{silInfo}{A \code{\link{silhouetteInfo}} object to be used for reporting
a silhouette plot.}

\item{clusterMaxLabels}{Maximum amount of clusters when labels are still
drawn within a dendrogram.}

\item{MSPeakLists}{A \code{\link{MSPeakLists}} object that is
\emph{mandatory} when spectra for formulae and/or compounds will be
reported.}

\item{EICGrid}{An integer vector in the form \code{c(columns, rows)} that is
used to determine the plotting grid when reporting EICs in PDF files.}

\item{EICRtWindow, EICMzWindow, EICTopMost, EICOnlyPresent}{Plotting parameters
passed to \code{\link{plotEIC}} (\emph{i.e.} \code{rtWindow},
\code{mzWindow}, \code{topMost} and \code{onlyPresent} arguments).}

\item{reportChord}{If \code{TRUE} then a chord diagram for all feature groups
is plotted (data will be averaged among replicates).}

\item{interactiveHeat}{If \code{TRUE} an interactive heatmap HTML widget will
be generated to display hierarchical clustering results. Set to
\code{FALSE} for a 'regular' static plot.}

\item{selfContained}{If \code{TRUE} the output will be a standalone HTML file
which contains all graphics and script dependencies. When \code{FALSE}, the
latter will be placed in an additional directory (\file{report_files})
which should remain present when viewing the output file. Especially on
Windows, a non-self contained output might be desirable when reporting
large amounts of data to prevent \command{pandoc} from running out of
memory.}

\item{optimizePng}{If \code{TRUE} then \command{pngquant} is used to reduce
the size of generated graphics. A signficant reduction in disk space usage
may be seen, however, at the cost additional processing time.}

\item{maxProcAmount}{Maximum amount of \command{pngquant} commands to run in
parallel. Higher numbers will decrease processing time, with an optimum
usually close to the amount of CPU cores.}
}
\description{
Functionality to report data produced by most workflow steps such as
features, feature groups, calculated chemical formulae and tentatively
identified compounds.
}
\details{
These functions are usually called at the very end of the workflow. It is
used to report various data on features and feature groups. In addition,
these functions may be used for reporting formulae and/or compounds that were
generated for the specified feature groups. Data can be reported in tabular
form (\emph{i.e.} \file{.csv} files) by \code{reportCSV} or graphically by
\code{reportPDF} and \code{reportMD}. The latter functions will plot for
instance chromatograms and annotated mass spectra, which are useful to get a
graphical overview of results.

All functions have a wide variety of arguments that influence the reporting
process. Nevertheless, most parameters are optional and only required to be
given for fine tuning. In addition, only those objects (\emph{e.g.} formulae,
compounds, clustering) that are desired to be reported need to be specified.

\code{reportCSV} generates tabular data (\emph{i.e.} \file{.csv}
  files) for given data to be reported. This may also be useful to allow
  import by other tools for post processing.

\code{reportPDF} will report graphical data (\emph{e.g.}
  chromatograms and mass spectra) within PDF files. Compared
  to \code{reportMD} this function may be faster and yield smaller report
  files, however, its functionality is a bit more basic and generated data is
  more 'scattered' around.

\code{reportMD} will report graphical data (\emph{e.g.}
  chromatograms and mass spectra) and summary information in an easy
  browsable \code{HTML} file using \link{rmarkdown}, \link{flexdashboard} and
  \link{knitr}.
}
\note{
Any formulae and compounds for feature groups which are not present
  within \code{fGroups} (\emph{i.e.} because it has been subset afterwards)
  will not be reported.
}
\references{
\addCitations{knitr}{2} \cr\cr
  \addCitations{knitr}{3}
}
