% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{reporting}
\alias{reporting}
\alias{reportCSV,featureGroups-method}
\alias{reportCSV}
\alias{reportPDF,featureGroups-method}
\alias{reportPDF}
\alias{reportHTML,featureGroups-method}
\alias{reportHTML}
\title{Report feature group data}
\usage{
\S4method{reportCSV}{featureGroups}(
  fGroups,
  path = "report",
  reportFGroupsAsRows = TRUE,
  reportFGroupsAnalysisInfo = TRUE,
  reportFGroupsRetMz = TRUE,
  reportFeatures = FALSE,
  formulas = NULL,
  formulasNormalizeScores = "max",
  formulasExclNormScores = NULL,
  compounds = NULL,
  compoundsNormalizeScores = "max",
  compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount"),
  compsCluster = NULL,
  components = NULL,
  retMin = TRUE,
  clearPath = FALSE
)

\S4method{reportPDF}{featureGroups}(
  fGroups,
  path = "report",
  reportFGroups = TRUE,
  formulas = NULL,
  formulasTopMost = 5,
  formulasNormalizeScores = "max",
  formulasExclNormScores = NULL,
  reportFormulaSpectra = TRUE,
  compounds = NULL,
  compoundsNormalizeScores = "max",
  compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount"),
  compoundsOnlyUsedScorings = TRUE,
  compoundsTopMost = 5,
  compsCluster = NULL,
  components = NULL,
  MSPeakLists = NULL,
  retMin = TRUE,
  EICGrid = c(2, 1),
  EICRtWindow = 20,
  EICMzWindow = 0.005,
  EICTopMost = NULL,
  EICOnlyPresent = TRUE,
  clearPath = FALSE
)

\S4method{reportHTML}{featureGroups}(
  fGroups,
  path = "report",
  reportPlots = c("chord", "venn", "upset", "eics", "formulas"),
  formulas = NULL,
  formulasTopMost = 5,
  formulasNormalizeScores = "max",
  formulasExclNormScores = NULL,
  compounds = NULL,
  compoundsNormalizeScores = "max",
  compoundsExclNormScores = c("score", "individualMoNAScore", "annoTypeCount"),
  compoundsOnlyUsedScorings = TRUE,
  compoundsTopMost = 5,
  compsCluster = NULL,
  includeMFWebLinks = "compounds",
  components = NULL,
  interactiveHeat = FALSE,
  MSPeakLists = NULL,
  retMin = TRUE,
  EICRtWindow = 20,
  EICMzWindow = 0.005,
  EICTopMost = NULL,
  EICOnlyPresent = TRUE,
  selfContained = TRUE,
  optimizePng = FALSE,
  clearPath = FALSE,
  openReport = TRUE,
  noDate = FALSE
)
}
\arguments{
\item{fGroups}{The \code{\link{featureGroups}} object that should be used for
reporting data.}

\item{path}{The destination file path for files generated during reporting.
Will be generated if needed.}

\item{reportFGroupsAsRows}{Report feature groups as rows (instead of columns)
within the resulting \file{.csv} file.}

\item{reportFGroupsAnalysisInfo}{Include analyses information (reference and
replicate groups) in the reported feature groups table \file{.csv} file.}

\item{reportFGroupsRetMz}{Include feature group information (retention time
and \emph{m/z}) within the reported feature groups table \file{.csv} file.}

\item{reportFeatures}{If set to \code{TRUE} then for each analysis a
\file{.csv} file will be generated with information about its detected
features.}

\item{formulas, compounds, compsCluster, components}{Further objects
(\code{\link{formulas}}, \code{\link{compounds}},
\code{\link{compoundsCluster}}, \code{\link{components}}) that should be
reported. Specify \code{NULL} to skip reporting a particular object.}

\item{compoundsNormalizeScores, formulasNormalizeScores}{A \code{character} that specifies how normalization of
compound scorings occurs. Either \code{"none"} (no normalization),
\code{"max"} (normalize to max value) or \code{"minmax"} (perform min-max
normalization). Note that normalization of negative scores (e.g. output by
\command{SIRIUS}) is always performed as min-max. Furthermore, currently
normalization for \code{compounds} takes the original min/max scoring
values into account when candidates were generated. Thus, for
\code{compounds} scoring, normalization is not affected when candidate
results were removed after they were generated (\emph{e.g.} by use of
\code{filter}).}

\item{compoundsExclNormScores, formulasExclNormScores}{A \code{character} vector specifying any compound
  scoring names that should \emph{not} be normalized. Set to \code{NULL} to
  normalize all scorings. Note that whether any normalization occurs is set
  by the \code{compoundsExclNormScores,formulasExclNormScores} argument.

  For \code{compounds}: By default \code{score} and
  \code{individualMoNAScore} are set to mimic the behavior of the
  \command{MetFrag} web interface.}

\item{retMin}{If \code{TRUE} then report retention times in minutes
(otehrwise seconds).}

\item{clearPath}{If \code{TRUE} then the destination path will be
(recursively) removed prior to reporting.}

\item{reportFGroups}{If \code{TRUE} then feature group data will be reported.}

\item{formulasTopMost, compoundsTopMost}{Only this amount of top ranked
candidate formulae/compounds are reported. Lower values may significantly
speed up reporting. Set to \code{NULL} to ignore.}

\item{reportFormulaSpectra}{If \code{TRUE} then explained MS/MS spectra (if
available) for candidate formulae will be reported. Specifying
\code{formulas} and setting this argument to \code{FALSE} still allows
further annotation of compound MS/MS spectra.}

\item{compoundsOnlyUsedScorings}{If \code{TRUE} then only scorings are plotted
that actually have been used to rank data (see the \code{scoreTypes}
argument to \code{\link{generateCompoundsMetfrag}} for more details).}

\item{MSPeakLists}{A \code{\link{MSPeakLists}} object that is
\emph{mandatory} when spectra for formulae and/or compounds will be
reported.}

\item{EICGrid}{An integer vector in the form \code{c(columns, rows)} that is
used to determine the plotting grid when reporting EICs in PDF files.}

\item{EICRtWindow, EICMzWindow, EICTopMost, EICOnlyPresent}{Plotting parameters
passed to \code{\link{plotEIC}} (\emph{i.e.} \code{rtWindow},
\code{mzWindow}, \code{topMost} and \code{onlyPresent} arguments).}

\item{reportPlots}{A character vector specifying what should be plotted.
Valid options are: \code{"chord"}, \code{"venn"}, \code{"upset"} (plot a
chord, Venn and UpSet diagram, respectively), \code{"eics"} (plot EICs for
individual feature groups) and \code{"formulas"} (plot annotated formula
spectra). Set to \code{"none"} to plot none of these.}

\item{includeMFWebLinks}{A \code{character} specifying to which feature
groups a web link should be added in the annotation page to
\href{https://msbi.ipb-halle.de/MetFragBeta/index.xhtml}{MetFragWeb}.
Options are: \code{"compounds"} (only to those with compounds results),
\code{"MSMS"} (only to those with MSMS peak lists) or \code{"none"}.}

\item{interactiveHeat}{If \code{TRUE} an interactive heatmap HTML widget will
be generated to display hierarchical clustering results. Set to
\code{FALSE} for a 'regular' static plot.}

\item{selfContained}{If \code{TRUE} the output will be a standalone HTML file
which contains all graphics and script dependencies. When \code{FALSE}, the
latter will be placed in an additional directory (\file{report_files})
which should remain present when viewing the output file. Especially on
Windows, a non-self contained output might be desirable when reporting
large amounts of data to prevent \command{pandoc} from running out of
memory.}

\item{optimizePng}{If \code{TRUE} then \command{pngquant} is used to reduce
the size of generated graphics. A signficant reduction in disk space usage
may be seen, however, at the cost additional processing time.}

\item{openReport}{If set to \code{TRUE} then the output report file will be
opened with the system browser.}

\item{noDate}{If \code{TRUE} then the current date is not added to the
report. This is mainly used for testing and its main purpose is to
guarentees equal report files when `reportHTML()` is called multiple times
with equal arguments.}
}
\description{
Functionality to report data produced by most workflow steps such as
features, feature groups, calculated chemical formulae and tentatively
identified compounds.
}
\details{
These functions are usually called at the very end of the workflow. It is
used to report various data on features and feature groups. In addition,
these functions may be used for reporting formulae and/or compounds that were
generated for the specified feature groups. Data can be reported in tabular
form (\emph{i.e.} \file{.csv} files) by \code{reportCSV} or graphically by
\code{reportPDF} and \code{reportHTML}. The latter functions will plot for
instance chromatograms and annotated mass spectra, which are useful to get a
graphical overview of results.

All functions have a wide variety of arguments that influence the reporting
process. Nevertheless, most parameters are optional and only required to be
given for fine tuning. In addition, only those objects (\emph{e.g.} formulae,
compounds, clustering) that are desired to be reported need to be specified.

\code{reportCSV} generates tabular data (\emph{i.e.} \file{.csv}
  files) for given data to be reported. This may also be useful to allow
  import by other tools for post processing.

\code{reportPDF} will report graphical data (\emph{e.g.}
  chromatograms and mass spectra) within PDF files. Compared
  to \code{reportHTML} this function may be faster and yield smaller report
  files, however, its functionality is a bit more basic and generated data is
  more 'scattered' around.

\code{reportHTML} will report graphical data (\emph{e.g.}
  chromatograms and mass spectra) and summary information in an easy
  browsable \code{HTML} file using \link{rmarkdown}, \link{flexdashboard} and
  \link{knitr}.
}
\note{
Any formulae and compounds for feature groups which are not present
  within \code{fGroups} (\emph{i.e.} because it has been subset afterwards)
  will not be reported.
}
\references{
Creating MetFrag landing page URLs based on code from
  \href{https://github.com/Treutler/MetFamily}{MetFamily} R package. \cr\cr
  \addCitations{knitr}{2} \cr\cr \addCitations{knitr}{3}
}
