% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/feature_groups-screening.R,
%   R/feature_groups-tasq.R, R/features-tasq.R
\docType{class}
\name{suspect-screening}
\alias{suspect-screening}
\alias{screenSuspects,featureGroups-method}
\alias{screenSuspects}
\alias{numericIDLevel}
\alias{genIDLevelRulesFile}
\alias{featureGroupsBrukerTASQ-class}
\alias{featureGroupsBrukerTASQ}
\alias{importFeatureGroupsBrukerTASQ}
\alias{featuresBrukerTASQ-class}
\alias{featuresBrukerTASQ}
\title{Target and suspect screening}
\usage{
\S4method{screenSuspects}{featureGroups}(
  fGroups,
  suspects,
  rtWindow,
  mzWindow,
  adduct,
  skipInvalid,
  onlyHits
)

numericIDLevel(level)

genIDLevelRulesFile(out, inLevels = NULL, exLevels = NULL)

importFeatureGroupsBrukerTASQ(path, analysisInfo, clusterRTWindow = 12)
}
\arguments{
\item{fGroups}{The \code{\link{featureGroups}} object that should be
screened.}

\item{suspects}{A \code{data.frame} with suspect information. See the
\verb{Suspect list format} section below.}

\item{rtWindow, mzWindow}{The retention time window (in seconds) and
\emph{m/z} window that will be used for matching a suspect (+/- feature
data).}

\item{adduct}{An \code{\link{adduct}} object (or something that can be
converted to it with \code{\link{as.adduct}}). Examples: \code{"[M-H]-"},
\code{"[M+Na]+"}. May be \code{NULL}, see \verb{Suspect list format}
section below.}

\item{skipInvalid}{If set to \code{TRUE} then suspects with invalid data
(\emph{e.g.} missing names or other missing data) will be ignored with a
warning. Similarly, any suspects for which mass calculation failed (when no
\code{mz} column is present in the suspect list), for instance, due to
invalid \code{SMILES}, will be ignored with a warning.}

\item{onlyHits}{If \code{TRUE} then all feature groups not matched by any of
the suspects will be removed.}

\item{level}{The identification level to be converted.}

\item{out}{The file path to the target file.}

\item{inLevels, exLevels}{A \link[=regex]{regular expression} for the
identification levels to include or exclude, respectively. For instance,
\code{exLevels="4|5"} would exclude level 4 and 5 from the output file. Set
to \code{NULL} to ignore.}

\item{path}{The file path to an Excel export of the Global results table from
TASQ, converted to \file{.csv} format.}

\item{analysisInfo}{A table with \link[=analysis-information]{analysis
information}.}

\item{clusterRTWindow}{This retention time window (in seconds) is used to
group hits across analyses together. See also the details section.}
}
\value{
\code{screenSuspects} returns a \code{\link{featureGroupsScreening}}
  object, which is a copy of the input \code{fGroups} object amended with
  additional screening information.

\code{importFeatureGroupsBrukerTASQ} returns a new
  \code{featureGroups} object containing converted screening results from
  Bruker TASQ.
}
\description{
Utilities to screen for analytes with known or suspected identity.
}
\details{
Besides 'full non-target analysis', where compounds may be identified with
little to no prior knowledge, a common strategy is to screen for compounds
with known or suspected identity. This may be a generally favorable approach
if possible, as it can significantly reduce the load on data interpretation.

\code{screenSuspects} is used to perform suspect screening. The
  input \code{\link{featureGroups}} object will be screened for suspects by
  \emph{m/z} values and optionally retention times. Afterwards, any feature
  groups not matched may be kept or removed, depending whether a full
  non-target analysis is desired.

\code{numericIDLevel} Extracts the numeric part of a given
  identification level (\emph{e.g.} \code{"3a"} becomes \samp{3}).

\code{genIDLevelRulesFile} Generates a template YAML file that is
  used to configure the rules for automatic estimation of identification
  levels. This file can then be used as input for
  \code{\link{annotateSuspects}}.

\code{importFeatureGroupsBrukerTASQ} will convert screening results
  from Bruker TASQ to a \code{\link{featureGroups}} object. The feature
  groups across analyses are formed based on the name of suspects and their
  closeness in retention time. The latter is necessary because TASQ does not
  necessarily perform checks on retention times and may therefore assign a
  suspect to peaks with different retention times across analyses (or within
  a single analysis). Hence, suspects with equal names are hierarchically
  clustered on their retention times (using \pkg{\link{fastcluster}}) to form
  the feature groups. The cut-off value for this is specified by the
  \code{clusterRTWindow} argument. The input for this function is obtained by
  generating an Excel export of the 'global' results and subsequently
  converting the file to \file{.csv} format.
}
\note{
Both \code{screenSuspects} and \code{importFeatureGroupsBrukerTASQ} may
  use the suspect names to base file names used for reporting, logging etc.
  Therefore, it is important that these are file-compatible names.

For \code{screenSuspects} in some cases you may need to install
  \href{http://openbabel.org/wiki/Main_Page}{OpenBabel} (\emph{e.g.} when
  only InChI data is available for mass calculation).

\code{importFeatureGroupsBrukerTASQ} will use estimated min/max values
  for retention times and dummy min/max \emph{m/z} values for conversion to
  features, since this information is not (readily) available. Hence, when
  plotting, for instance, extracted ion chromatograms (with
  \code{\link{plotChroms}}) the integrated chromatographic peak range shown
  is incorrect.
}
\section{Suspect list format}{
 the \code{suspects} argument for
  \code{screenSuspects} should be a \code{data.frame} with the following
  mandatory and optional columns:

  \itemize{

  \item \code{name} The suspect name. Must be file-compatible.
  (\strong{mandatory})

  \item \code{rt} The retention time (in seconds) for the suspect. If
  specified the suspect will only be matched if its retention matches the
  experimental value (tolerance defined by the \code{rtWindow} argument).
  (\strong{optional})

  \item \code{neutralMass},\code{formula},\code{SMILES},\code{InChI} The
  neutral monoisotopic mass, chemical formula, SMILES or InChI for the
  suspect. (data from one of these columns are \strong{mandatory} in case no
  value from the \code{mz} column is available for a suspect)

  \item \code{mz} The ionized \emph{m/z} of the suspect. (\strong{mandatory}
  unless it can be calculated from one of the aforementioned columns)

  \item \code{adduct} A \code{character} that can be converted with
  \code{\link{as.adduct}}. (\strong{mandatory} unless data from the \code{mz}
  column is available or the \code{adduct} argument is set)

  \item \code{fragments_mz},\code{fragments_formula} One or more MS/MS
  fragments (specified as \emph{m/z} or formulae, respectively). Multiple
  values can be specified by separating them with a semicolon (\verb{;}).
  This data is used by \code{\link{annotateSuspects}} to report detected
  MS/MS fragments and calculate identification levels. (\strong{optional})

  }
}

\references{
\insertRef{OBoyle2011}{patRoon}

\addCitations{fastcluster}{1}
}
\seealso{
\code{featureGroupsScreening}
}
