% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/features.R, R/feature_groups.R,
%   R/feature_groups-screening.R, R/feature_groups-tasq.R, R/features-tasq.R
\docType{class}
\name{suspect-screening}
\alias{suspect-screening}
\alias{screenSuspects}
\alias{screenSuspects,features-method}
\alias{screenSuspects,featureGroups-method}
\alias{featureGroupsScreening-class}
\alias{featureGroupsScreening}
\alias{groupFeaturesScreening,featureGroups-method}
\alias{groupFeaturesScreening}
\alias{featureGroupsBrukerTASQ-class}
\alias{featureGroupsBrukerTASQ}
\alias{importFeatureGroupsBrukerTASQ}
\alias{featuresBrukerTASQ-class}
\alias{featuresBrukerTASQ}
\title{Target and suspect screening}
\usage{
screenSuspects(
  obj,
  suspects,
  rtWindow = 12,
  mzWindow = 0.005,
  adduct = NULL,
  skipInvalid = TRUE
)

\S4method{screenSuspects}{features}(
  obj,
  suspects,
  rtWindow = 12,
  mzWindow = 0.005,
  adduct = NULL,
  skipInvalid = TRUE
)

\S4method{screenSuspects}{featureGroups}(
  obj,
  suspects,
  rtWindow = 12,
  mzWindow = 0.005,
  adduct = NULL,
  skipInvalid = TRUE
)

\S4method{groupFeaturesScreening}{featureGroups}(fGroups, scr)

importFeatureGroupsBrukerTASQ(path, analysisInfo, clusterRTWindow = 12)
}
\arguments{
\item{obj}{The object that should be screened (\emph{i.e.}
\code{\link{features}} or \code{\link{featureGroups}}).}

\item{suspects}{A \code{data.frame} that must contain a \code{"name"} column
(the analyte name) and \emph{at least} a \code{"mz"}, \code{"neutralMass"},
\code{"formula"}, \code{"SMILES"}, \code{"InChI"} column (with the ionized
m/z value, the neutral monoisotopic mass or the chemical
formula/SMILES/InChI character string for the molecule, respectively). If
an ion mass needs to be calculated (\emph{i.e.} no valid data is available
in the \code{mz} column) then data is tried to be used from the columns in
the aforementioned order. Furthermore, if ion masses need to be calculated
then the adduct must be specified either with the \code{adduct} function
argument or by an \code{"adduct"} column containing a \code{character} that
can be converted with \code{as.adduct} (\emph{e.g.} \code{"[M+H]+"}). In
addition, a column \code{"rt"} can be included to specify the retention
time (if unspecified no retention times are checked).}

\item{rtWindow, mzWindow}{The retention time window (in seconds) and
\emph{m/z} window that will be used for matching a suspect (+/- feature
data).}

\item{adduct}{An \code{\link{adduct}} object (or something that can be
converted to it with \code{\link{as.adduct}}). Examples: \code{"[M-H]-"},
\code{"[M+Na]+"}. Can be \code{NULL} if either a \code{"mz"} or
\code{"adduct"} column is present in the \code{suspects} argument.}

\item{skipInvalid}{If set to \code{TRUE} then suspects with invalid data
(\emph{e.g.} missing names or other missing data) will be ignored with a
warning. Similarly, any suspects for which mass calculation failed (when no
\code{mz} column is present in the suspect list), for instance, due to
invalid \code{SMILES}, will be ignored with a warning.}

\item{fGroups}{The \code{\link{featureGroups}} object that should be
transformed (and was used to obtain the screening results).}

\item{scr}{The screening results table returned by \code{screenSuspects}.}

\item{path}{The file path to an Excel export of the Global results table from
TASQ, converted to \file{.csv} format.}

\item{analysisInfo}{A table with \link[=analysis-information]{analysis
information}.}

\item{clusterRTWindow}{This retention time window (in seconds) is used to
group hits across analyses together. See also the details section.}
}
\value{
\code{screenSuspects} will return a table (a
  \code{\link{data.table}}) with detected suspects and details such as
  retention and \emph{m/z} values. If a suspect is matched on multiple
  features/feature groups then each hit is reported as a separate row.

\code{groupFeaturesScreening} returns a modified \code{featureGroups}
  object in which those feature groups that were not matched by any suspects
  are removed and others renamed by the respective suspect name. In case of
  duplicate suspect results, feature group names are made unique with
  \code{\link{make.unique}}.

\code{importFeatureGroupsBrukerTASQ} returns a new
  \code{featureGroups} object containing converted screening results from
  Bruker TASQ.
}
\description{
Utilities to screen for analytes with known or suspected identity.
}
\details{
Besides 'full non-target analysis', where compounds may be identitified with
little to no prior knowledge, a common strategy is to screen for compounds
with known or suspected identity. This may be a generally favourable approach
if possible, as it can significantly reduce the load on data interpretation.

\code{screenSuspects} will screen a set of suspects (provided as a
  \code{data.frame}) within an \code{\link{features}} or
  \code{\link{featureGroups}} object.

\code{groupFeaturesScreening} uses results from \code{screenSuspects}
  to transform an existing \code{\link{featureGroups}} object by (1) renaming
  any matched feature groups by the respective name of the suspect and (2)
  filtering out any feature groups that were not matched. A common workflow
  is to first obtain and group features (with \emph{e.g.}
  \code{\link{findFeatures}} and \code{\link{groupFeatures}}), screen them
  with \code{screenSuspects}, convert the \code{featureGroups} object that was
  used for screening with this method function and continue any further
  workflow steps such as compound identification as with 'regular'
  \code{featureGroups}.

\code{importFeatureGroupsBrukerTASQ} will convert screening results
  from Bruker TASQ to a \code{\link{featureGroups}} object. The feature
  groups across analyses are formed based on the name of suspects and their
  closeness in retention time. The latter is necessary because TASQ does not
  necessarily perform checks on retention times and may therefore assign a
  suspect to peaks with different retention times across analyses (or within
  a single analysis). Hence, suspects with equal names are hierarchically
  clustered on their retention times (using \pkg{\link{fastcluster}}) to form
  the feature groups. The cut-off value for this is specified by the
  \code{clusterRTWindow} argument. The input for this function is obtained by
  generating an Excel export of the 'global' results and subsequently
  converting the file to \file{.csv} format. Similar to
  \code{groupFeaturesScreening}, this method will return an object that is
  suitable for any further workflow processes.
}
\note{
Both \code{groupFeaturesScreening} and
  \code{importFeatureGroupsBrukerTASQ} use names from targets/suspects as
  feature group names, therefore, it is important that these are
  file-compatible names when \link[=reporting]{reporting data} \file{csv} or
  \file{pdf} data.

For \code{screenSuspects} in some cases you may need to install
  \href{http://openbabel.org/wiki/Main_Page}{OpenBabel} (\emph{e.g.} when
  only InChI data is available for mass calculation).

Please note that \code{groupFeaturesScreening} method can only
  transform the \code{featureGroups} object that was used to obtain the given
  screening results.

\code{importFeatureGroupsBrukerTASQ} will use estimated min/max values
  for retention times and dummy min/max \emph{m/z} values for conversion to
  features, since this information is not (readily) available. Hence, when
  plotting, for instance, extracted ion chromatograms (with
  \code{\link{plotEIC}}) the integrated chromatographic peak range shown is
  incorrect.
}
\references{
\addCitations{fastcluster}{1}
}
