% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP.R, R/TP-logic.R
\docType{class}
\name{transformationProducts-class}
\alias{transformationProducts-class}
\alias{transformationProducts}
\alias{parents,transformationProducts-method}
\alias{parents}
\alias{products,transformationProducts-method}
\alias{products}
\alias{names,transformationProducts-method}
\alias{show,transformationProducts-method}
\alias{[,transformationProducts,ANY,missing,missing-method}
\alias{[[,transformationProducts,ANY,missing-method}
\alias{$,transformationProducts-method}
\alias{as.data.table,transformationProducts-method}
\alias{convertToSuspects,transformationProducts-method}
\alias{transformationProductsLogic-class}
\alias{transformationProductsLogic}
\title{Base transformation products (TP) class}
\usage{
\S4method{parents}{transformationProducts}(TPs)

\S4method{products}{transformationProducts}(TPs)

\S4method{names}{transformationProducts}(x)

\S4method{show}{transformationProducts}(object)

\S4method{[}{transformationProducts,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{transformationProducts,ANY,missing}(x, i, j)

\S4method{$}{transformationProducts}(x, name)

\S4method{as.data.table}{transformationProducts}(x)

\S4method{convertToSuspects}{transformationProducts}(TPs, includeParents = TRUE)
}
\arguments{
\item{TPs, x, object}{\code{transformationProducts} object to be accessed}

\item{i}{For \code{[}/\code{[[}: A numeric or character value which is used to select parents by
their index or name, respectively (for the order/names see \code{names()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all parents are selected.\cr\cr For \code{[[}: should be a scalar value.}

\item{\dots}{Unused.}

\item{drop, j}{ignored.}

\item{name}{The parent name (partially matched).}

\item{includeParents}{If \code{TRUE} then parents are also included in the returned suspect list.}
}
\description{
Holds information for all TPs for a set of parents.
}
\details{
This class holds all generated data for transformation products for a set of parents. The class is \code{virtual} and
derived objects are created by \link[=TP-generation]{TP generators}.

The TP data in objects from this class include a \code{retDir} column. These are \code{numeric} values that hint what
the the chromatographic retention order of a TP might be compared to its parent: a value of \samp{-1} means it will
elute earlier, \samp{1} it will elute later and \samp{0} that there is no significant difference or the direction is
unknown. These values are based on a typical reversed phase separation. When \code{\link{generateTPsBioTransformer}}
or \code{\link{generateTPsLibrary}} was used to generate the data, the \code{retDir} values are based on calculated
\code{log P} values of the parent and its TPs.
}
\section{Methods (by generic)}{
\itemize{
\item \code{parents}: Accessor method for the \code{parents} slot of a
\code{transformationProducts} class.

\item \code{products}: Accessor method for the \code{products} slot.

\item \code{names}: Obtain the names of all parents in this object.

\item \code{show}: Show summary information for this object.

\item \code{[}: Subset on parents.

\item \code{[[}: Extracts a table with TPs for a parent.

\item \code{$}: Extracts a table with TPs for a parent.

\item \code{as.data.table}: Returns all TP data in a table.

\item \code{convertToSuspects}: Converts this object to a suspect list that can be used as input for
\code{\link{screenSuspects}}.
}}

\section{Slots}{

\describe{
\item{\code{parents}}{A \code{\link{data.table}} with metadata for all parents that have TPs in this object. Use the
\code{parents} method for access.}

\item{\code{products}}{A \code{list} with \code{\link{data.table}} entries with TP information for each parent. Use the
\code{products} method for access.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{workflowStep}}}   \itemize{     \item{\strong{\code{\link{transformationProducts}}}}     \itemize{       \item{\code{\link{transformationProductsBT}}}       \item{\code{\link{transformationProductsComponents}}}       \item{\code{\link{transformationProductsLibrary}}}       \item{\code{\link{transformationProductsLogic}}}     }   } }
}

\seealso{
Derived classes \code{\link{transformationProductsBT}} and \code{\link{transformationProductsLibrary}} for
  specific algorithm methods and \code{\link{TP-generation}}
}
