% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP.R
\docType{class}
\name{transformationProducts-class}
\alias{transformationProducts-class}
\alias{transformationProducts}
\alias{parents,transformationProducts-method}
\alias{parents}
\alias{products,transformationProducts-method}
\alias{products}
\alias{names,transformationProducts-method}
\alias{show,transformationProducts-method}
\alias{[,transformationProducts,ANY,missing,missing-method}
\alias{[[,transformationProducts,ANY,missing-method}
\alias{$,transformationProducts-method}
\alias{as.data.table,transformationProducts-method}
\title{Base transformation product (TP) products class}
\usage{
\S4method{parents}{transformationProducts}(TPs)

\S4method{products}{transformationProducts}(TPs)

\S4method{names}{transformationProducts}(x)

\S4method{show}{transformationProducts}(object)

\S4method{[}{transformationProducts,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[[}{transformationProducts,ANY,missing}(x, i, j)

\S4method{$}{transformationProducts}(x, name)

\S4method{as.data.table}{transformationProducts}(x)
}
\arguments{
\item{i}{A numeric or character value which is used to select parents by
their index or name, respectively (for the order/names see \code{names()}).\cr\cr For \code{[}: Can also be logical to perform logical selection
(similar to regular vectors). If missing all parents are selected.\cr\cr For \code{[[}: should be a scalar value.}

\item{drop, j}{ignored.}

\item{name}{The parent name (partially matched).}

\item{obj, x, object}{\code{transformationProducts} object to be accessed}
}
\description{
Holds information for all TPs for a set of parents.
}
\details{
This class holds all generated data for transformation products for
a set of parents. The class is \code{virtual} and derived objects are
created by \link[=TP-generation]{TP predictors}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{parents}: Accessor method for the \code{parents} slot of a
\code{transformationProducts} class. This is a \code{data.table} with all parents
used for products.

\item \code{products}: Accessor method for the \code{products} slot of
a \code{transformationProducts} class. Each TP result is stored as a
\code{\link{data.table}}.

\item \code{names}: Obtain the names of all parents with TPs.

\item \code{show}: Show summary information for this object.

\item \code{[}: Subset on parents.

\item \code{[[}: Extracts a table with TPs for a parent.

\item \code{$}: Extracts a table with TPs for a parent.

\item \code{as.data.table}: Returns all TP data in a table.
}}

\section{Slots}{

\describe{
\item{\code{parents}}{Table with all parents with products. Use the
\code{parents} method for access.}

\item{\code{products}}{List of TPs for each parent. Use the
\code{products} method for access.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\code{\link{workflowStep}}}   \itemize{     \item{\strong{\code{\link{transformationProducts}}}}     \itemize{       \item{\code{\link{transformationProductsBT}}}       \item{\code{\link{transformationProductsComponents}}}       \item{\code{\link{transformationProductsLibrary}}}       \item{\code{\link{transformationProductsLogic}}}     }   } }
}

\seealso{
\code{\link{TP-generation}}
}
