% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-step.R
\docType{class}
\name{workflowStep-class}
\alias{workflowStep-class}
\alias{workflowStep}
\alias{algorithm,workflowStep-method}
\alias{as.data.table,workflowStep-method}
\alias{as.data.frame,workflowStep-method}
\alias{show,workflowStep-method}
\title{(Virtual) Base class for all workflow objects.}
\usage{
\S4method{algorithm}{workflowStep}(obj)

\S4method{as.data.table}{workflowStep}(x, keep.rownames = FALSE, ...)

\S4method{as.data.frame}{workflowStep}(x, row.names = NULL, optional = FALSE, ...)

\S4method{show}{workflowStep}(object)
}
\arguments{
\item{obj, x, object}{An object (derived from) this class.}

\item{keep.rownames}{Ignored.}

\item{\dots}{Method specific arguments. Please see the documentation of the
derived classes.}

\item{row.names, optional}{Ignored.}
}
\description{
All workflow objects (\emph{e.g.} \code{\link{featureGroups}},
\code{\link{compounds}}, etc) are derived from this class. Objects from this
class are never created directly.
}
\section{Methods (by generic)}{
\itemize{
\item \code{algorithm}: Returns the algorithm that was used to generate an
object.

\item \code{as.data.table}: Summarizes the data in this object and returns this
as a \code{\link{data.table}}.

\item \code{as.data.frame}: This method simply calls \code{as.data.table} and
converts the result to a classic a \code{data.frame}.

\item \code{show}: Shows summary information for this object.
}}

\section{Slots}{

\describe{
\item{\code{algorithm}}{The algorithm that was used to generate this object. Use the
\code{algorithm} method for access.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\strong{\code{\link{workflowStep}}}}   \itemize{     \item{\code{\link{transformationProducts}}}     \itemize{       \item{\code{\link{transformationProductsBT}}}       \item{\code{\link{transformationProductsComponents}}}       \item{\code{\link{transformationProductsLibrary}}}       \item{\code{\link{transformationProductsLogic}}}     }     \item{\code{\link{features}}}     \itemize{       \item{\code{\link{featuresSet}}}       \item{\code{\link{featuresUnset}}}       \item{\code{\link{featuresFromFeatGroups}}}       \item{\code{\link{featuresConsensus}}}       \item{\code{\link{featuresBruker}}}       \item{\code{\link{featuresEnviPick}}}       \item{\code{\link{featuresKPIC2}}}       \item{\code{\link{featuresOpenMS}}}       \item{\code{\link{featuresSAFD}}}       \item{\code{\link{featuresSIRIUS}}}       \item{\code{\link{featuresBrukerTASQ}}}       \item{\code{\link{featuresXCMS}}}       \item{\code{\link{featuresXCMS3}}}     }     \item{\code{\link{featureGroups}}}     \itemize{       \item{\code{\link{featureGroupsSet}}}       \itemize{         \item{\code{\link{featureGroupsScreeningSet}}}       }       \item{\code{\link{featureGroupsUnset}}}       \item{\code{\link{featureGroupsScreening}}}       \itemize{         \item{\code{\link{featureGroupsSetScreeningUnset}}}       }       \item{\code{\link{featureGroupsBruker}}}       \item{\code{\link{featureGroupsConsensus}}}       \item{\code{\link{featureGroupsEnviMass}}}       \item{\code{\link{featureGroupsKPIC2}}}       \item{\code{\link{featureGroupsOpenMS}}}       \item{\code{\link{featureGroupsSIRIUS}}}       \item{\code{\link{featureGroupsBrukerTASQ}}}       \item{\code{\link{featureGroupsXCMS}}}       \item{\code{\link{featureGroupsXCMS3}}}     }     \item{\code{\link{components}}}     \itemize{       \item{\code{\link{componentsSet}}}       \itemize{         \item{\code{\link{componentsNTSet}}}       }       \item{\code{\link{componentsUnset}}}       \item{\code{\link{componentsCamera}}}       \item{\code{\link{componentsFeatures}}}       \itemize{         \item{\code{\link{componentsCliqueMS}}}         \item{\code{\link{componentsOpenMS}}}       }       \item{\code{\link{componentsClust}}}       \itemize{         \item{\code{\link{componentsIntClust}}}         \item{\code{\link{componentsSpecClust}}}       }       \item{\code{\link{componentsNT}}}       \itemize{         \item{\code{\link{componentsNTUnset}}}       }       \item{\code{\link{componentsRC}}}       \item{\code{\link{componentsTPs}}}     }     \item{\code{\link{featureAnnotations}}}     \itemize{       \item{\code{\link{formulas}}}       \itemize{         \item{\code{\link{formulasConsensus}}}         \item{\code{\link{formulasSet}}}         \itemize{           \item{\code{\link{formulasConsensusSet}}}         }         \item{\code{\link{formulasUnset}}}       }       \item{\code{\link{compounds}}}       \itemize{         \item{\code{\link{compoundsConsensus}}}         \item{\code{\link{compoundsMF}}}         \item{\code{\link{compoundsSet}}}         \itemize{           \item{\code{\link{compoundsConsensusSet}}}         }         \item{\code{\link{compoundsUnset}}}       }     }     \item{\code{\link{MSPeakLists}}}     \itemize{       \item{\code{\link{MSPeakListsSet}}}       \item{\code{\link{MSPeakListsUnset}}}     }   } }
}

