% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-step.R
\docType{class}
\name{workflowStep-class}
\alias{workflowStep-class}
\alias{workflowStep}
\alias{algorithm,workflowStep-method}
\alias{as.data.table,workflowStep-method}
\alias{as.data.frame,workflowStep-method}
\alias{show,workflowStep-method}
\title{(Virtual) Base class for all workflow objects.}
\usage{
\S4method{algorithm}{workflowStep}(obj)

\S4method{as.data.table}{workflowStep}(x, keep.rownames = FALSE, ...)

\S4method{as.data.frame}{workflowStep}(x, row.names = NULL, optional = FALSE, ...)

\S4method{show}{workflowStep}(object)
}
\arguments{
\item{obj, x, object}{An object (derived from) this class.}

\item{keep.rownames}{Ignored.}

\item{\dots}{Method specific arguments. Please see the documentation of the
derived classes.}

\item{row.names, optional}{Ignored.}
}
\description{
All workflow objects (\emph{e.g.} \code{\link{featureGroups}},
\code{\link{compounds}}, etc) are derived from this class. Objects from this
class are never created directly.
}
\section{Methods (by generic)}{
\itemize{
\item \code{algorithm}: Returns the algorithm that was used to generate an
object.

\item \code{as.data.table}: Summarizes the data in this object and returns this
as a \code{\link{data.table}}.

\item \code{as.data.frame}: This method simply calls \code{as.data.table} and
converts the result to a classic a \code{data.frame}.

\item \code{show}: Shows summary information for this object.
}}

\section{Slots}{

\describe{
\item{\code{algorithm}}{The algorithm that was used to generate this object. Use the
\code{algorithm} method for access.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\strong{\code{\link{workflowStep}}}}   \itemize{     \item{\code{\link{features}}}     \itemize{       \item{\code{\link{featuresFromFeatGroups}}}       \item{\code{\link{featuresConsensus}}}       \item{\code{\link{featuresBruker}}}       \item{\code{\link{featuresEnviPick}}}       \item{\code{\link{featuresOpenMS}}}       \item{\code{\link{featuresSAFD}}}       \item{\code{\link{featuresBrukerTASQ}}}       \item{\code{\link{featuresXCMS}}}       \item{\code{\link{featuresXCMS3}}}     }     \item{\code{\link{featureGroups}}}     \itemize{       \item{\code{\link{featureGroupsBruker}}}       \item{\code{\link{featureGroupsConsensus}}}       \item{\code{\link{featureGroupsEnviMass}}}       \item{\code{\link{featureGroupsOpenMS}}}       \item{\code{\link{featureGroupsScreening}}}       \item{\code{\link{featureGroupsBrukerTASQ}}}       \item{\code{\link{featureGroupsXCMS}}}       \item{\code{\link{featureGroupsXCMS3}}}     }     \item{\code{\link{components}}}     \itemize{       \item{\code{\link{componentsReduced}}}       \item{\code{\link{componentsCamera}}}       \item{\code{\link{componentsIntClust}}}       \item{\code{\link{componentsNT}}}       \item{\code{\link{componentsRC}}}     }     \item{\code{\link{compounds}}}     \itemize{       \item{\code{\link{compoundsConsensus}}}       \item{\code{\link{compoundsMF}}}     }     \item{\code{\link{formulas}}}     \item{\code{\link{MSPeakLists}}}   } }
}

