% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow-step-set.R
\docType{class}
\name{workflowStepSet-class}
\alias{workflowStepSet-class}
\alias{workflowStepSet}
\alias{setObjects,workflowStepSet-method}
\alias{sets,workflowStepSet-method}
\alias{show,workflowStepSet-method}
\title{(Virtual) base class for sets related workflow objects}
\usage{
\S4method{setObjects}{workflowStepSet}(obj)

\S4method{sets}{workflowStepSet}(obj)

\S4method{show}{workflowStepSet}(object)
}
\arguments{
\item{obj, object}{An object that is derived from \code{workflowStepSet}.}
}
\description{
This class is the base for many \link[=sets-workflow]{sets workflows} related classes. This class is virtual, and
therefore never created directly.
}
\details{
The most important purpose of this class is to hold data that is specific for a set. These \emph{set objects} are
typically objects with classes from a regular non-sets workflow (\emph{e.g.} \code{\link{components}},
\code{\link{compounds}}), and are used by the sets workflow object to \emph{e.g.} form a consensus. Since the set
objects may contain additional data, such as algorithm specific slots, it may in some cases be of interest to access
them directly with the \code{setObjects} method (described below).
}
\section{Methods (by generic)}{
\itemize{
\item \code{setObjects(workflowStepSet)}: Accessor for the \code{setObjects} slot.

\item \code{sets(workflowStepSet)}: Returns the names for each set in this object.

\item \code{show(workflowStepSet)}: Shows summary information for this object.

}}
\section{Slots}{

\describe{
\item{\code{setObjects}}{A \code{list} with the \emph{set objects} (see the \verb{Details} section). The \code{list} is named
with the set names.}
}}

\section{S4 class hierarchy}{
 \itemize{   \item{\strong{\code{\link{workflowStepSet}}}}   \itemize{     \item{\code{\link{componentsSet}}}     \itemize{       \item{\code{\link{componentsNTSet}}}     }     \item{\code{\link{featureGroupsScreeningSet}}}     \item{\code{\link{compoundsSet}}}     \itemize{       \item{\code{\link{compoundsConsensusSet}}}     }     \item{\code{\link{formulasSet}}}     \itemize{       \item{\code{\link{formulasConsensusSet}}}     }     \item{\code{\link{MSPeakListsSet}}}   } }
}

